% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{aggregateFlows}
\alias{aggregateFlows}
\title{aggregateFlows}
\usage{
aggregateFlows(x, y)
}
\arguments{
\item{x}{= a df of 3 columns. These are: a) ReportDate, a column of yyyymmdd dates (first column) b) FundId, a column of integers (second column) c) a column of numbers representing Flows or AUM}

\item{y}{= a matrix/df of \% allocations. Row space is always <yyyymm>-<FundId> so as to match up with the first two columns of <x>}
}
\description{
a matrix indexed by time, the columns of which correspond to <y>
}
\keyword{aggregateFlows}
