% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{angle}
\alias{angle}
\title{angle}
\usage{
angle(x, y, n)
}
\arguments{
\item{x}{= number representing distance between points A & B}

\item{y}{= number representing distance between points B & C}

\item{n}{= number representing distance between points A & C}
}
\description{
angle ABC
}
\keyword{angle}
