% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{args.canonical}
\alias{args.canonical}
\title{args.canonical}
\usage{
args.canonical(x = T)
}
\arguments{
\item{x}{= a boolean (standard/reverse order)}
}
\description{
relabels arguments
}
\seealso{
Other args: \code{\link{args.rename.body}},
  \code{\link{args.rename}}
}
\keyword{args.canonical}
