% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{args.comment}
\alias{args.comment}
\title{args.comment}
\usage{
args.comment()
}
\description{
functions, arguments and comments
}
\seealso{
Other args: \code{\link{args.canonical}},
  \code{\link{args.missing}},
  \code{\link{args.rename.body}}, \code{\link{args.rename}}
}
\keyword{args.comment}
