% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{args.rename.body}
\alias{args.rename.body}
\title{args.rename.body}
\usage{
args.rename.body(x, y)
}
\arguments{
\item{x}{= an expression}

\item{y}{= a string vector (new args indexed by old)}
}
\description{
relabels arguments <y> by <n>
}
\seealso{
Other args: \code{\link{args.canonical}},
  \code{\link{args.comment}},
  \code{\link{args.missing.bad}},
  \code{\link{args.missing}}, \code{\link{args.rename}}
}
\keyword{args.rename.body}
