% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{args.rename}
\alias{args.rename}
\title{args.rename}
\usage{
args.rename(fcn, x = T)
}
\arguments{
\item{fcn}{= a function}

\item{x}{= a boolean (standard/reverse order)}
}
\description{
relabels arguments
}
\seealso{
Other args: \code{\link{args.canonical}},
  \code{\link{args.comment}},
  \code{\link{args.missing.bad}},
  \code{\link{args.missing}},
  \code{\link{args.rename.body}}
}
\keyword{args.rename}
