% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{array.bind}
\alias{array.bind}
\title{array.bind}
\usage{
array.bind(...)
}
\arguments{
\item{...}{= arrays}
}
\description{
binds together along the dimension they differ on
}
\seealso{
Other array: \code{\link{array.ex.list}},
  \code{\link{array.unlist}}
}
\keyword{array.bind}
