% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{array.ex.list}
\alias{array.ex.list}
\title{array.ex.list}
\usage{
array.ex.list(x, y, n)
}
\arguments{
\item{x}{= a numeric vector/matrix/data frame list}

\item{y}{= a boolean (row union/intersection)}

\item{n}{= a boolean (column union/intersection)}
}
\description{
array
}
\seealso{
Other array: \code{\link{array.bind}},
  \code{\link{array.unlist}}
}
\keyword{array.ex.list}
