% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{array.relist}
\alias{array.relist}
\title{array.relist}
\usage{
array.relist(x, y, n = F)
}
\arguments{
\item{x}{= data frame of dimnames (must be a list)}

\item{y}{= vector of values}

\item{n}{= T/F depending on whether zav is to be applied}
}
\description{
array indexed by all combinations of elements of <x>
}
\seealso{
Other array: \code{\link{array.unlist}}
}
\keyword{array.relist}
