% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{avg.model}
\alias{avg.model}
\title{avg.model}
\usage{
avg.model(x, y = 0)
}
\arguments{
\item{x}{= vector of results}

\item{y}{= lag length (to handle overlapping returns)     When using N-period returns, y is N - 1. So for     annual or qtrly periods of monthly data, y is 11 or 2.}
}
\description{
constant-only (zero-variable) regression model
}
\seealso{
Other avg: \code{\link{avg.winsorized}},
  \code{\link{avg.wtd}}
}
\keyword{avg.model}
