% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{avg.winsorized}
\alias{avg.winsorized}
\title{avg.winsorized}
\usage{
avg.winsorized(x, y = 100)
}
\arguments{
\item{x}{= a numeric vector}

\item{y}{= an integer (number of bins)}
}
\description{
mean is computed over the quantiles 2 through <y> - 1
}
\seealso{
Other avg: \code{\link{avg.model}}, \code{\link{avg.wtd}}
}
\keyword{avg.winsorized}
