% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{avg.wtd}
\alias{avg.wtd}
\title{avg.wtd}
\usage{
avg.wtd(x, y)
}
\arguments{
\item{x}{= a numeric vector}

\item{y}{= a numeric vector of weights}
}
\description{
returns the weighted mean of <x> given weights <n>
}
\keyword{avg.wtd}
