% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{base64encode.wrapper}
\alias{base64encode.wrapper}
\title{base64encode.wrapper}
\usage{
base64encode.wrapper(x)
}
\arguments{
\item{x}{= a file}
}
\description{
file encoded in base 64
}
\keyword{base64encode.wrapper}
