% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bbk.bin.rets.geom.summ}
\alias{bbk.bin.rets.geom.summ}
\title{bbk.bin.rets.geom.summ}
\usage{
bbk.bin.rets.geom.summ(x, y)
}
\arguments{
\item{x}{= a matrix/df with rows indexed by time and columns indexed by bins}

\item{y}{= number of rows of <x> needed to cover an entire calendar year}
}
\description{
Summarizes bin excess returns geometrically
}
\keyword{bbk.bin.rets.geom.summ}
