% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bbk.bin.rets.prd.summ}
\alias{bbk.bin.rets.prd.summ}
\title{bbk.bin.rets.prd.summ}
\usage{
bbk.bin.rets.prd.summ(fcn, x, y, n)
}
\arguments{
\item{fcn}{= function you use to summarize results}

\item{x}{= a matrix/df with rows indexed by time and columns indexed by bins}

\item{y}{= a vector corresponding to the rows of <x> that maps each row to a sub-period of interest (e.g. calendar year)}

\item{n}{= number of rows of <x> needed to cover an entire calendar year}
}
\description{
Summarizes bin excess returns by sub-periods of interest (as defined by <y>)
}
\keyword{bbk.bin.rets.prd.summ}
