% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bbk.bin.rets.summ}
\alias{bbk.bin.rets.summ}
\title{bbk.bin.rets.summ}
\usage{
bbk.bin.rets.summ(x, y, n = F)
}
\arguments{
\item{x}{= a matrix/df with rows indexed by time and columns indexed by bins}

\item{y}{= number of rows of <x> needed to cover an entire calendar year}

\item{n}{= T/F depending on if you want to count number of periods}
}
\description{
Summarizes bin excess returns arithmetically
}
\keyword{bbk.bin.rets.summ}
