% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bbk.bin.xRet}
\alias{bbk.bin.xRet}
\title{bbk.bin.xRet}
\usage{
bbk.bin.xRet(x, y, n = 5, w = F, h = F)
}
\arguments{
\item{x}{= a matrix/df of predictors, the rows of which are indexed by time}

\item{y}{= a matrix/df of the same dimensions as <x> containing associated forward returns}

\item{n}{= number of desired bins}

\item{w}{= T/F depending on whether universe return is desired}

\item{h}{= T/F depending on whether full detail or bin returns are needed}
}
\description{
Returns equal weight bin returns through time
}
\seealso{
Other bbk: \code{\link{bbk.bin.rets.prd.summ}},
  \code{\link{bbk.bin.rets.summ}}, \code{\link{bbk.data}},
  \code{\link{bbk.drawdown}}, \code{\link{bbk.fwdRet}},
  \code{\link{bbk.holidays}}, \code{\link{bbk.matrix}},
  \code{\link{bbk.summ}}, \code{\link{bbk.turnover}},
  \code{\link{bbk}}
}
\keyword{bbk.bin.xRet}
