% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bbk.data}
\alias{bbk.data}
\title{bbk.data}
\usage{
bbk.data(x, y, n, w, h, u, v, g, r, s)
}
\arguments{
\item{x}{= a matrix/data frame (predictors)}

\item{y}{= a matrix/data frame (total return indices)}

\item{n}{= an integer (flow window in days/months)}

\item{w}{= a boolean (sum/compound)}

\item{h}{= an integer (lag in days/months)}

\item{u}{= an integer (delay in days/months)}

\item{v}{= an integer (0 = Sun, 1 = Mon, etc., the day you trade)}

\item{g}{= an integer (return window in days/months)}

\item{r}{= a string (index within which you trade)}

\item{s}{= a boolean (spread changes/returns)}
}
\description{
fetches data required to compute standard model output
}
\seealso{
Other bbk: \code{\link{bbk.bin.rets.prd.summ}},
  \code{\link{bbk.bin.rets.summ}},
  \code{\link{bbk.bin.xRet}}, \code{\link{bbk.drawdown}},
  \code{\link{bbk.fwdRet}}, \code{\link{bbk.holidays}},
  \code{\link{bbk.matrix}}, \code{\link{bbk.summ}},
  \code{\link{bbk.turnover}}, \code{\link{bbk}}
}
\keyword{bbk.data}
