% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bbk.doW.bulk}
\alias{bbk.doW.bulk}
\title{bbk.doW.bulk}
\usage{
bbk.doW.bulk(x, y, n)
}
\arguments{
\item{x}{= a matrix/data-frame indexed by <yyyymmdd> dates}

\item{y}{= a day of the week from 0:6 (Sun:Sat)}

\item{n}{= an essential column that cannot be NA}
}
\description{
Adds rows to <x> so that day <y> of the week is never missing
}
\keyword{bbk.doW.bulk}
