% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bbk.drawdown}
\alias{bbk.drawdown}
\title{bbk.drawdown}
\usage{
bbk.drawdown(x)
}
\arguments{
\item{x}{= a numeric vector}
}
\description{
returns a logical vector identifying the contiguous periods #		:	corresponding to max drawdown
}
\keyword{bbk.drawdown}
