% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bbk.drawdown.underlying}
\alias{bbk.drawdown.underlying}
\title{bbk.drawdown.underlying}
\usage{
bbk.drawdown.underlying(x)
}
\arguments{
\item{x}{= an integer vector}
}
\description{
position of trough & peak
}
\seealso{
Other bbk: \code{\link{bbk.bin.rets.prd.summ}},
  \code{\link{bbk.bin.rets.summ}},
  \code{\link{bbk.bin.xRet}}, \code{\link{bbk.data}},
  \code{\link{bbk.drawdown}}, \code{\link{bbk.fwdRet}},
  \code{\link{bbk.holidays}}, \code{\link{bbk.matrix}},
  \code{\link{bbk.summ}}, \code{\link{bbk.turnover}},
  \code{\link{bbk}}
}
\keyword{bbk.drawdown.underlying}
