% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bbk.fanChart}
\alias{bbk.fanChart}
\title{bbk.fanChart}
\usage{
bbk.fanChart(x)
}
\arguments{
\item{x}{= "rets" part of the output of function bbk}
}
\description{
quintile fan charts
}
\seealso{
Other bbk: \code{\link{bbk.bin.rets.geom.summ}},
  \code{\link{bbk.bin.rets.prd.summ}},
  \code{\link{bbk.bin.rets.summ}},
  \code{\link{bbk.bin.xRet}}, \code{\link{bbk.data}},
  \code{\link{bbk.doW.bulk}}, \code{\link{bbk.drawdown}},
  \code{\link{bbk.fwdRet}}, \code{\link{bbk.histogram}},
  \code{\link{bbk.holidays}}, \code{\link{bbk.summ}},
  \code{\link{bbk.turnover}}, \code{\link{bbk}}
}
\keyword{bbk.fanChart}
