% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bbk.holidays}
\alias{bbk.holidays}
\title{bbk.holidays}
\usage{
bbk.holidays(x, y)
}
\arguments{
\item{x}{= a matrix/data frame (predictors)}

\item{y}{= a matrix/data frame (forward returns)}
}
\description{
Sets <x> to NA whenever <y> is NA
}
\seealso{
Other bbk: \code{\link{bbk.bin.rets.prd.summ}},
  \code{\link{bbk.bin.rets.summ}},
  \code{\link{bbk.bin.xRet}}, \code{\link{bbk.data}},
  \code{\link{bbk.drawdown}}, \code{\link{bbk.fwdRet}},
  \code{\link{bbk.matrix}}, \code{\link{bbk.summ}},
  \code{\link{bbk.turnover}}, \code{\link{bbk.underlying}},
  \code{\link{bbk}}
}
\keyword{bbk.holidays}
