% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bbk}
\alias{bbk}
\title{bbk}
\usage{
bbk(x, y, floW = 1, retW = 5, nBin = 5, doW = NULL, sum.flows = F,
  lag = 0, delay = 2, idx = NULL, sprds = F)
}
\arguments{
\item{x}{= predictor indexed by yyyymmdd or yyyymm}

\item{y}{= total return index indexed by the same date format as <x>}

\item{floW}{= number of <prd.size>'s over which the predictor should be compounded/summed}

\item{retW}{= return window in days or months depending on whether <x> is YYYYMMDD or YYYYMM}

\item{nBin}{= number of bins to divide the variable into}

\item{doW}{= day of the week you will trade on (5 = Fri)}

\item{sum.flows}{= T/F depending on whether <x> should be summed or compounded}

\item{lag}{= predictor lag in days or months depending on whether <x> is YYYYMMDD or YYYYMM}

\item{delay}{= delay in knowing data in days or months depending on whether <x> is YYYYMMDD or YYYYMM}

\item{idx}{= the index within which you are trading}

\item{sprds}{= T/F depending on whether spread changes, rather than returns, are needed}
}
\description{
standard model output
}
\seealso{
Other bbk: \code{\link{bbk.bin.rets.prd.summ}},
  \code{\link{bbk.bin.rets.summ}},
  \code{\link{bbk.bin.xRet}}, \code{\link{bbk.data}},
  \code{\link{bbk.drawdown}}, \code{\link{bbk.fanChart}},
  \code{\link{bbk.fwdRet}}, \code{\link{bbk.histogram}},
  \code{\link{bbk.holidays}}, \code{\link{bbk.matrix}},
  \code{\link{bbk.summ}}, \code{\link{bbk.turnover}}
}
\keyword{bbk}
