% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bbk.summ}
\alias{bbk.summ}
\title{bbk.summ}
\usage{
bbk.summ(x, y, n)
}
\arguments{
\item{x}{= bin returns}

\item{y}{= bin memberships}

\item{n}{= return window in days or months depending on whether <x> is YYYYMMDD or YYYYMM}
}
\description{
summarizes by year and overall. Assumes periods are non-overlapping.
}
\keyword{bbk.summ}
