% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bbk.turnover}
\alias{bbk.turnover}
\title{bbk.turnover}
\usage{
bbk.turnover(x)
}
\arguments{
\item{x}{= a matrix/df of positive integers}
}
\description{
returns average name turnover per bin
}
\keyword{bbk.turnover}
