% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{best.linear.strategy.blend}
\alias{best.linear.strategy.blend}
\title{best.linear.strategy.blend}
\usage{
best.linear.strategy.blend(x, y)
}
\arguments{
\item{x}{= a return stream from a strategy}

\item{y}{= an isomekic return stream from a strategy}
}
\description{
Returns optimal weights to put on <x> and <y>
}
\keyword{best.linear.strategy.blend}
