% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{binomial.trial}
\alias{binomial.trial}
\title{binomial.trial}
\usage{
binomial.trial(x, y, n, w)
}
\arguments{
\item{x}{= probability of success in a 1/0 Bernoulli trial}

\item{y}{= number of coin flips}

\item{n}{= number of heads}

\item{w}{= a boolean (upper/lower tail)}
}
\description{
returns the likelihood of getting <n> or more/fewer heads depending on whether <w> is T/F
}
\keyword{binomial.trial}
