% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bond.curve.expand}
\alias{bond.curve.expand}
\title{bond.curve.expand}
\usage{
bond.curve.expand(x)
}
\arguments{
\item{x}{= a named vector of interest rates}
}
\description{
full yield curve
}
\keyword{bond.curve.expand}
