% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{brinson}
\alias{brinson}
\title{brinson}
\usage{
brinson(x, y, n, w)
}
\arguments{
\item{x}{= a numeric vector (benchmark weights)}

\item{y}{= a numeric vector (active weights)}

\item{n}{= a numeric vector (returns)}

\item{w}{= a numeric vector (groups)}
}
\description{
performs a Brinson attribution
}
\keyword{brinson}
