% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{britten.jones.data}
\alias{britten.jones.data}
\title{britten.jones.data}
\usage{
britten.jones.data(x, y, n, w = NULL)
}
\arguments{
\item{x}{= a matrix/data frame (predictors)}

\item{y}{= a matrix/data frame (total return indices)}

\item{n}{= an integer (return window)}

\item{w}{= a string (index within which you trade)}
}
\description{
returns data needed for a Britten-Jones analysis
}
\seealso{
Other britten: \code{\link{britten.jones.data.stack}},
  \code{\link{britten.jones}}
}
\keyword{britten.jones.data}
