% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{britten.jones.data.stack}
\alias{britten.jones.data.stack}
\title{britten.jones.data.stack}
\usage{
britten.jones.data.stack(x, y, n, w, h)
}
\arguments{
\item{x}{=}

\item{y}{=}

\item{n}{=}

\item{w}{=}

\item{h}{=}
}
\description{
applies the Britten-Jones transformation to a subset and then stacks
}
\seealso{
Other britten: \code{\link{britten.jones.data}},
  \code{\link{britten.jones}}
}
\keyword{britten.jones.data.stack}
