% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{char.lag}
\alias{char.lag}
\title{char.lag}
\usage{
char.lag(x, y)
}
\arguments{
\item{x}{= a vector of characters}

\item{y}{= a number}
}
\description{
lags <x> by <y>
}
\seealso{
Other char: \code{\link{char.ex.int}},
  \code{\link{char.seq}}, \code{\link{char.to.int}},
  \code{\link{char.to.num}}
}
\keyword{char.lag}
