% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{char.seq}
\alias{char.seq}
\title{char.seq}
\usage{
char.seq(x, y, n = 1)
}
\arguments{
\item{x}{= a SINGLE character}

\item{y}{= a SINGLE character}

\item{n}{= quantum size}
}
\description{
returns a sequence of ASCII characters between (and including) x and y
}
\keyword{char.seq}
