% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{char.to.int}
\alias{char.to.int}
\title{char.to.int}
\usage{
char.to.int(x)
}
\arguments{
\item{x}{= a string of single characters}
}
\description{
ascii values
}
\seealso{
Other char: \code{\link{char.ex.int}},
  \code{\link{char.lag}}, \code{\link{char.seq}},
  \code{\link{char.to.num}}
}
\keyword{char.to.int}
