% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{char.to.num}
\alias{char.to.num}
\title{char.to.num}
\usage{
char.to.num(x)
}
\arguments{
\item{x}{= a vector of strings}
}
\description{
coerces to numeric without generating warnings
}
\seealso{
Other char: \code{\link{char.ex.int}},
  \code{\link{char.lag}}, \code{\link{char.seq}},
  \code{\link{char.to.int}}
}
\keyword{char.to.num}
