% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{col.lag}
\alias{col.lag}
\title{col.lag}
\usage{
col.lag(x, y)
}
\arguments{
\item{x}{= a string (excel column)}

\item{y}{= an integer vector}
}
\description{
Lags <x> by <y> columns
}
\seealso{
Other col: \code{\link{col.ex.int}},
  \code{\link{col.to.int}}
}
\keyword{col.lag}
