% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{col.offset}
\alias{col.offset}
\title{col.offset}
\usage{
col.offset(x, y)
}
\arguments{
\item{x}{= string representation of an excel column}

\item{y}{= an integer representing the desired column offset}
}
\description{
Offsets <x> by <y> columns
}
\seealso{
Other col: \code{\link{col.ex.int}},
  \code{\link{col.smallest}}, \code{\link{col.to.int}}
}
\keyword{col.offset}
