% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{col.to.int}
\alias{col.to.int}
\title{col.to.int}
\usage{
col.to.int(x)
}
\arguments{
\item{x}{= a SINGLE string representation of an excel column}
}
\description{
Returns the relevant associated integer (1 = "A", 2 = "B", etc.)
}
\seealso{
Other col: \code{\link{col.ex.int}},
  \code{\link{col.offset}}, \code{\link{col.smallest}}
}
\keyword{col.to.int}
