% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{combinations.to.int}
\alias{combinations.to.int}
\title{combinations.to.int}
\usage{
combinations.to.int(x)
}
\arguments{
\item{x}{= a logical vector}
}
\description{
maps each particular way to choose <sum(x)> things amongst <length(x)> things to the number line
}
\keyword{combinations.to.int}
