% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{common.fund.flow.shock}
\alias{common.fund.flow.shock}
\title{common.fund.flow.shock}
\usage{
common.fund.flow.shock(x, y, n)
}
\arguments{
\item{x}{= a YYYYMM}

\item{y}{= a connection string}

\item{n}{= number of months}
}
\description{
common fund flow shock of Dou, Kogan & Wu (2022)
}
\keyword{common.fund.flow.shock}
