% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{compound.flows.fast}
\alias{compound.flows.fast}
\title{compound.flows.fast}
\usage{
compound.flows.fast(x, y, n = F)
}
\arguments{
\item{x}{= a matrix/data-frame of percentage flows}

\item{y}{= number of trailing rows to compound/sum}

\item{n}{= if T, flows get summed. Otherwise they get compounded.}
}
\description{
compounded flows over <n> trailing periods indexed by last day in the flow window
}
\seealso{
Other compound: \code{\link{compound.flows.initial}},
  \code{\link{compound.flows.underlying}},
  \code{\link{compound.flows}}, \code{\link{compound.sf}},
  \code{\link{compound}}
}
\keyword{compound.flows.fast}
