% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{compound.flows.initial}
\alias{compound.flows.initial}
\title{compound.flows.initial}
\usage{
compound.flows.initial(x, y)
}
\arguments{
\item{x}{= a matrix/data-frame of percentage flows}

\item{y}{= an integer representing the size of the window needed}
}
\description{
T/F depending on whether output for a row is to be set to NA
}
\seealso{
Other compound: \code{\link{compound.flows.underlying}},
  \code{\link{compound.flows}}, \code{\link{compound.sf}},
  \code{\link{compound}}
}
\keyword{compound.flows.initial}
