% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{compound.flows}
\alias{compound.flows}
\title{compound.flows}
\usage{
compound.flows(x, y, n = F)
}
\arguments{
\item{x}{= a matrix/data-frame of percentage flows}

\item{y}{= number of trailing rows to compound/sum}

\item{n}{= if T, flows get summed. Otherwise they get compounded.}
}
\description{
compounded flows over <n> trailing periods indexed by last day in the flow window
}
\keyword{compound.flows}
