% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{compound.flows.underlying}
\alias{compound.flows.underlying}
\title{compound.flows.underlying}
\usage{
compound.flows.underlying(fcn, x, y, n, w)
}
\arguments{
\item{fcn}{= function used to compound flows}

\item{x}{= a matrix/data-frame of percentage flows}

\item{y}{= flow window in terms of the number of trailing periods to compound}

\item{n}{= if T simple positional lagging is used. If F, yyyymm.lag is invoked}

\item{w}{= size of each period in terms of days if the rows of <x> are yyyymmdd or months otherwise}
}
\description{
compounded flows over <y> trailing periods indexed by last day in the flow window
}
\keyword{compound.flows.underlying}
