% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{compound}
\alias{compound}
\title{compound}
\usage{
compound(x)
}
\arguments{
\item{x}{= a numeric vector (percentage returns)}
}
\description{
Outputs the compounded return
}
\seealso{
Other compound: \code{\link{compound.flows}},
  \code{\link{compound.sf}}
}
\keyword{compound}
