% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{compound.sf}
\alias{compound.sf}
\title{compound.sf}
\usage{
compound.sf(x, y)
}
\arguments{
\item{x}{= a matrix/data-frame of percentage flows}

\item{y}{= if T, flows get summed. Otherwise they get compounded.}
}
\description{
compounds flows
}
\keyword{compound.sf}
