% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{compound.stock.flows}
\alias{compound.stock.flows}
\title{compound.stock.flows}
\usage{
compound.stock.flows(x, y)
}
\arguments{
\item{x}{= a matrix/data-frame of percentage flows}

\item{y}{= if T, flows get summed. Otherwise they get compounded.}
}
\description{
compounds flows
}
\seealso{
Other compound: \code{\link{compound.flows.initial}},
  \code{\link{compound.flows.underlying}},
  \code{\link{compound.flows}}, \code{\link{compound}}
}
\keyword{compound.stock.flows}
