% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{contains}
\alias{contains}
\title{contains}
\usage{
contains(x, y)
}
\arguments{
\item{x}{= a vector of strings}

\item{y}{= a single string}
}
\description{
returns the elements of <x> that contain <y>
}
\keyword{contains}
