% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{correl.PrcMo}
\alias{correl.PrcMo}
\title{correl.PrcMo}
\usage{
correl.PrcMo(x, y, n, w)
}
\arguments{
\item{x}{= one-day flow percentage}

\item{y}{= total return index}

\item{n}{= flow window}

\item{w}{= the number of days needed for the flow data to be known}
}
\description{
returns correlation of <n> day flows with price momentum (175d lag 10)
}
\keyword{correl.PrcMo}
