% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{correl}
\alias{correl}
\title{correl}
\usage{
correl(x, y, n = T)
}
\arguments{
\item{x}{= a numeric vector/matrix/data frame}

\item{y}{= either missing or a numeric isomekic vector}

\item{n}{= T/F depending on whether rank correlations are desired}
}
\description{
the estimated correlation between <x> and <y> or the columns of <x>
}
\keyword{correl}
