% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{covar}
\alias{covar}
\title{covar}
\usage{
covar(x, y, n = F)
}
\arguments{
\item{x}{= a numeric vector}

\item{y}{= a numeric isomekic vector}

\item{n}{= T/F depending on whether rank correlations are desired}
}
\description{
the estimated covariance between <x> and <y> or the columns of <x>
}
\keyword{covar}
