% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{cpt.FloAlphaLt.Ctry}
\alias{cpt.FloAlphaLt.Ctry}
\title{cpt.FloAlphaLt.Ctry}
\usage{
cpt.FloAlphaLt.Ctry(x)
}
\arguments{
\item{x}{= factor folder}
}
\description{
Computes flow alpha for countries. indexed by data date. Does not account for delay.
}
\keyword{cpt.FloAlphaLt.Ctry}
