% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{cpt.RgnSecJP}
\alias{cpt.RgnSecJP}
\title{cpt.RgnSecJP}
\usage{
cpt.RgnSecJP(x, y)
}
\arguments{
\item{x}{= a vector of Sectors}

\item{y}{= a vector of country codes}
}
\description{
makes Region-Sector groupings
}
\keyword{cpt.RgnSecJP}
