% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{cptRollingAverageWeights}
\alias{cptRollingAverageWeights}
\title{cptRollingAverageWeights}
\usage{
cptRollingAverageWeights(x = 4, y = 100, n = 0)
}
\arguments{
\item{x}{= number of trailing weeks to use}

\item{y}{= weight on the earliest as a percentage of weight on latest week}

\item{n}{= number of additional weeks to lag data}
}
\description{
Returns weights on individual weeks with the most recent week being to the RIGHT
}
\keyword{cptRollingAverageWeights}
