% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{Ctry.info}
\alias{Ctry.info}
\title{Ctry.info}
\usage{
Ctry.info(x, y)
}
\arguments{
\item{x}{= a country code vector}

\item{y}{= a string (column in classif-Ctry)}
}
\description{
looks up information from classif-Ctry
}
\examples{
Ctry.info("PK", "CtryNm")
}
\seealso{
Other Ctry: \code{\link{Ctry.msci.index.changes}},
  \code{\link{Ctry.msci.members.rng}},
  \code{\link{Ctry.msci.members}}, \code{\link{Ctry.msci}},
  \code{\link{Ctry.to.CtryGrp}}
}
\keyword{Ctry.info}
