% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{Ctry.msci.index.changes}
\alias{Ctry.msci.index.changes}
\title{Ctry.msci.index.changes}
\usage{
Ctry.msci.index.changes(x, y)
}
\arguments{
\item{x}{= a matrix/data frame (total return indices indexed by trade date)}

\item{y}{= a string (MSCI index like ACWI/EAFE/EM)}
}
\description{
handles the addition and removal of countries from an index
}
\seealso{
Other Ctry: \code{\link{Ctry.info}},
  \code{\link{Ctry.msci.members.rng}},
  \code{\link{Ctry.msci.members}}, \code{\link{Ctry.msci}},
  \code{\link{Ctry.to.CtryGrp}}
}
\keyword{Ctry.msci.index.changes}
