% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{Ctry.msci.members}
\alias{Ctry.msci.members}
\title{Ctry.msci.members}
\usage{
Ctry.msci.members(x, y)
}
\arguments{
\item{x}{= an index name such as ACWI/EAFE/EM}

\item{y}{= one of the following: (a) a YYYYMM date (b) a YYYYMMDD date (c) "" for a static series}
}
\description{
lists countries in an index at <y>
}
\seealso{
Other Ctry: \code{\link{Ctry.info}},
  \code{\link{Ctry.msci.index.changes}},
  \code{\link{Ctry.msci.members.rng}},
  \code{\link{Ctry.msci.sql}}, \code{\link{Ctry.msci}},
  \code{\link{Ctry.to.CtryGrp}}
}
\keyword{Ctry.msci.members}
