% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{Ctry.msci.members.rng}
\alias{Ctry.msci.members.rng}
\title{Ctry.msci.members.rng}
\usage{
Ctry.msci.members.rng(x, y, n)
}
\arguments{
\item{x}{= an index name such as ACWI/EAFE/EM}

\item{y}{= a YYYYMM/YYYYMMDD date}

\item{n}{= after <y> and of the same date type}
}
\description{
lists countries that were ever in an index between <y> and <n>
}
\keyword{Ctry.msci.members.rng}
