% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{Ctry.msci.sql}
\alias{Ctry.msci.sql}
\title{Ctry.msci.sql}
\usage{
Ctry.msci.sql(fcn, x, y, n)
}
\arguments{
\item{fcn}{= function to convert from yyyymm to yyyymmdd}

\item{x}{= output of Ctry.msci}

\item{y}{= single two-character country code}

\item{n}{= date field such as DayEnding or WeightDate}
}
\description{
SQL query to get date restriction
}
\seealso{
Other Ctry: \code{\link{Ctry.info}},
  \code{\link{Ctry.msci.index.changes}},
  \code{\link{Ctry.msci.members.rng}},
  \code{\link{Ctry.msci.members}}, \code{\link{Ctry.msci}},
  \code{\link{Ctry.to.CtryGrp}}
}
\keyword{Ctry.msci.sql}
