% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{Ctry.to.CtryGrp}
\alias{Ctry.to.CtryGrp}
\title{Ctry.to.CtryGrp}
\usage{
Ctry.to.CtryGrp(x)
}
\arguments{
\item{x}{= a vector of country codes}
}
\description{
makes Country groups
}
\keyword{Ctry.to.CtryGrp}
