% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{dataset.subset}
\alias{dataset.subset}
\title{dataset.subset}
\usage{
dataset.subset(x, y, n)
}
\arguments{
\item{x}{= a local folder (e.g. "C:\\\\temp\\\\crap")}

\item{y}{= column on which to subset}

\item{n}{= a vector of identifiers}
}
\description{
subsets all files in <x> so that column <y> is made up of elements of <n>. Original files are overwritten.
}
\keyword{dataset.subset}
