% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{day.ex.date}
\alias{day.ex.date}
\title{day.ex.date}
\usage{
day.ex.date(x)
}
\arguments{
\item{x}{= a vector of R dates}
}
\description{
calendar dates
}
\seealso{
Other day: \code{\link{day.ex.int}}, \code{\link{day.lag}},
  \code{\link{day.seq}}, \code{\link{day.to.int}},
  \code{\link{day.to.weekday}}, \code{\link{day.to.week}}
}
\keyword{day.ex.date}
