% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{day.ex.int}
\alias{day.ex.int}
\title{day.ex.int}
\usage{
day.ex.int(x)
}
\arguments{
\item{x}{= an integer or vector of integers}
}
\description{
calendar dates
}
\seealso{
Other day: \code{\link{day.ex.date}},
  \code{\link{day.lag}}, \code{\link{day.seq}},
  \code{\link{day.to.int}}, \code{\link{day.to.weekday}}
}
\keyword{day.ex.int}
