% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{day.lag}
\alias{day.lag}
\title{day.lag}
\usage{
day.lag(x, y)
}
\arguments{
\item{x}{= a vector of calendar dates}

\item{y}{= an integer or vector of integers (if <x> and <y> are vectors then <y> isomekic)}
}
\description{
lags <x> by <y> days.
}
\seealso{
Other day: \code{\link{day.ex.date}},
  \code{\link{day.ex.int}}, \code{\link{day.seq}},
  \code{\link{day.to.date}}, \code{\link{day.to.int}},
  \code{\link{day.to.ui}}, \code{\link{day.to.weekday}},
  \code{\link{day.to.week}}
}
\keyword{day.lag}
