% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{day.to.ui}
\alias{day.to.ui}
\title{day.to.ui}
\usage{
day.to.ui(x)
}
\arguments{
\item{x}{= a YYYYMMDD date}
}
\description{
<x> rendered in Engineering team's format
}
\seealso{
Other day: \code{\link{day.ex.date}},
  \code{\link{day.ex.int}}, \code{\link{day.lag}},
  \code{\link{day.seq}}, \code{\link{day.to.date}},
  \code{\link{day.to.int}}, \code{\link{day.to.weekday}},
  \code{\link{day.to.week}}
}
\keyword{day.to.ui}
