% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{day.to.weekday}
\alias{day.to.weekday}
\title{day.to.weekday}
\usage{
day.to.weekday(x)
}
\arguments{
\item{x}{= a vector of yyyymmdd}
}
\description{
Converts to 0 = Sun, 1 = Mon, ..., 6 = Sat
}
\seealso{
Other day: \code{\link{day.ex.date}},
  \code{\link{day.ex.int}}, \code{\link{day.lag}},
  \code{\link{day.seq}}, \code{\link{day.to.date}},
  \code{\link{day.to.int}}, \code{\link{day.to.week}}
}
\keyword{day.to.weekday}
