% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{decimal.format}
\alias{decimal.format}
\title{decimal.format}
\usage{
decimal.format(x, y)
}
\arguments{
\item{x}{= a numeric vector}

\item{y}{= a positive integer}
}
\description{
rounds <x> to <y> decimals and renders as nice character vector
}
\keyword{decimal.format}
