% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{dir.all.files}
\alias{dir.all.files}
\title{dir.all.files}
\usage{
dir.all.files(x, y)
}
\arguments{
\item{x}{= a file}

\item{y}{= a regular expression}
}
\description{
Returns all files in the folder including sub-directories
}
\keyword{dir.all.files}
