% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{dir.clear}
\alias{dir.clear}
\title{dir.clear}
\usage{
dir.clear(x, y)
}
\arguments{
\item{x}{= a file}

\item{y}{= a regular expression}
}
\description{
rids <x> of files of type <y>
}
\seealso{
Other dir: \code{\link{dir.all.files}},
  \code{\link{dir.ensure}}, \code{\link{dir.kill}},
  \code{\link{dir.make}}, \code{\link{dir.parameters}},
  \code{\link{dir.parent}}, \code{\link{dir.publications}},
  \code{\link{dir.size}}
}
\keyword{dir.clear}
