% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{dir.ensure}
\alias{dir.ensure}
\title{dir.ensure}
\usage{
dir.ensure(x)
}
\arguments{
\item{x}{= a string of full paths}
}
\description{
Creates necessary folders so files can be copied to <x>
}
\keyword{dir.ensure}
