% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{dir.kill}
\alias{dir.kill}
\title{dir.kill}
\usage{
dir.kill(x)
}
\arguments{
\item{x}{= a folder vector}
}
\description{
removes <x>
}
\seealso{
Other dir: \code{\link{dir.all.files}},
  \code{\link{dir.clear}}, \code{\link{dir.ensure}},
  \code{\link{dir.make}}, \code{\link{dir.parameters}},
  \code{\link{dir.parent}}, \code{\link{dir.publications}},
  \code{\link{dir.size}}
}
\keyword{dir.kill}
