% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{dir.make}
\alias{dir.make}
\title{dir.make}
\usage{
dir.make(x)
}
\arguments{
\item{x}{= a vector of full paths to folder}
}
\description{
Creates a folder in the location(s) specified by <x>
}
\seealso{
Other dir: \code{\link{dir.all.files}},
  \code{\link{dir.ensure}}, \code{\link{dir.parameters}},
  \code{\link{dir.parent}}
}
\keyword{dir.make}
