% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{dir.parameters}
\alias{dir.parameters}
\title{dir.parameters}
\usage{
dir.parameters(x)
}
\arguments{
\item{x}{= desired sub-folder}
}
\description{
returns full path to relevant parameters sub-folder
}
\keyword{dir.parameters}
