% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{dir.parent}
\alias{dir.parent}
\title{dir.parent}
\usage{
dir.parent(x)
}
\arguments{
\item{x}{= a string of full paths}
}
\description{
returns paths to the parent directory
}
\seealso{
Other dir: \code{\link{dir.all.files}},
  \code{\link{dir.ensure}}, \code{\link{dir.kill}},
  \code{\link{dir.make}}, \code{\link{dir.parameters}}
}
\keyword{dir.parent}
