% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{dir.publications}
\alias{dir.publications}
\title{dir.publications}
\usage{
dir.publications(x)
}
\arguments{
\item{x}{= desired sub-folder}
}
\description{
desired output directory for relevant publication
}
\seealso{
Other dir: \code{\link{dir.all.files}},
  \code{\link{dir.clear}}, \code{\link{dir.ensure}},
  \code{\link{dir.kill}}, \code{\link{dir.make}},
  \code{\link{dir.parameters}}, \code{\link{dir.parent}},
  \code{\link{dir.size}}
}
\keyword{dir.publications}
