% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{dir.size}
\alias{dir.size}
\title{dir.size}
\usage{
dir.size(x)
}
\arguments{
\item{x}{= a string of full paths}
}
\description{
size of directory <x> in KB
}
\seealso{
Other dir: \code{\link{dir.all.files}},
  \code{\link{dir.ensure}}, \code{\link{dir.kill}},
  \code{\link{dir.make}}, \code{\link{dir.parameters}},
  \code{\link{dir.parent}}
}
\keyword{dir.size}
