% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{EHD}
\alias{EHD}
\title{EHD}
\usage{
EHD(x, y, n, w, h, u = NULL)
}
\arguments{
\item{x}{= a connection string/connection}

\item{y}{= a string (Flow/AssetsStart/AssetsEnd/PortfolioChange)}

\item{n}{= a frequency (T/F for daily/weekly or D/W/M)}

\item{w}{= a YYYYMMDD (begin date)}

\item{h}{= a YYYYMMDD (end date)}

\item{u}{= a filter vector}
}
\description{
named vector of item between <w> and <h> sorted ascending
}
\keyword{EHD}
