% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{email.kill}
\alias{email.kill}
\title{email.kill}
\usage{
email.kill(x)
}
\arguments{
\item{x}{= a string (report name)}
}
\description{
deletes entry <x> in the email record. Returns nothing.
}
\seealso{
Other email: \code{\link{email.exists}},
  \code{\link{email.list}}, \code{\link{email.record}},
  \code{\link{email}}
}
\keyword{email.kill}
