% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{email.list}
\alias{email.list}
\title{email.list}
\usage{
email.list()
}
\description{
named vector of emails and sent dates
}
\seealso{
Other email: \code{\link{email.exists}},
  \code{\link{email.kill}}, \code{\link{email.record}}
}
\keyword{email.list}
