% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{email.record}
\alias{email.record}
\title{email.record}
\usage{
email.record(x, y)
}
\arguments{
\item{x}{= report name}

\item{y}{= date for which you sent the report}
}
\description{
updates the email record. Returns nothing.
}
\seealso{
Other email: \code{\link{email.exists}},
  \code{\link{email.kill}}, \code{\link{email.list}},
  \code{\link{email}}
}
\keyword{email.record}
