% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{err.raise}
\alias{err.raise}
\title{err.raise}
\usage{
err.raise(x, y, n)
}
\arguments{
\item{x}{= a string vector}

\item{y}{= a boolean (output many lines/one line)}

\item{n}{= a string (main line of error message)}
}
\description{
error message
}
\seealso{
Other err: \code{\link{err.raise.txt}}
}
\keyword{err.raise}
