% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{event.read}
\alias{event.read}
\title{event.read}
\usage{
event.read(x)
}
\arguments{
\item{x}{= a file of dates in dd/mm/yyyy format}
}
\description{
data frame with events sorted and numbered
}
\keyword{event.read}
