% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{extract.AnnMn.sf.wrapper}
\alias{extract.AnnMn.sf.wrapper}
\title{extract.AnnMn.sf.wrapper}
\usage{
extract.AnnMn.sf.wrapper(x, y = "AnnMn")
}
\arguments{
\item{x}{= a list (each element a 3D array, items/bins/parameters)}

\item{y}{= an item (AnnMn/AnnSd/Sharp/HitRate)}
}
\description{
Subsets to "AnnMn" and re-labels columns
}
\seealso{
Other extract: \code{\link{extract.AnnMn.sf}}
}
\keyword{extract.AnnMn.sf.wrapper}
