% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{factordump.rds}
\alias{factordump.rds}
\title{factordump.rds}
\usage{
factordump.rds(x, y, n, w, h, u)
}
\arguments{
\item{x}{= variable name (e.g. "HerdingLSV")}

\item{y}{= local folder (e.g. "C:\\\\temp\\\\mystuff")}

\item{n}{= starting QTR}

\item{w}{= ending QTR}

\item{h}{= list object containing the following items: a) classif - classif file b) conn - a connection, the output of odbcDriverConnect c) fldr - stock-flows folder}

\item{u}{= output variable name}
}
\description{
Dumps variable <x> to folder <y> in standard text format
}
\seealso{
Other factordump: \code{\link{factordump.sql}},
  \code{\link{factordump.write}}
}
\keyword{factordump.rds}
