% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{factordump.sql}
\alias{factordump.sql}
\title{factordump.sql}
\usage{
factordump.sql(x, y, n, w, h)
}
\arguments{
\item{x}{= variable name (e.g. "Herfindahl")}

\item{y}{= local folder (e.g. "C:\\\\temp\\\\mystuff")}

\item{n}{= starting QTR}

\item{w}{= ending QTR}

\item{h}{= one of StockFlows/Regular/Quant}
}
\description{
Dumps variable <x> to folder <y> in standard text format
}
\seealso{
Other factordump: \code{\link{factordump.rds}},
  \code{\link{factordump.write}}
}
\keyword{factordump.sql}
