% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{factordump.write}
\alias{factordump.write}
\title{factordump.write}
\usage{
factordump.write(x, y)
}
\arguments{
\item{x}{= a matrix/data-frame}

\item{y}{= output path}
}
\description{
Dumps variable <x> to path <y> in standard text format
}
\seealso{
Other factordump: \code{\link{factordump.rds}},
  \code{\link{factordump.sql}}
}
\keyword{factordump.write}
