% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{farben}
\alias{farben}
\title{farben}
\usage{
farben(x, y)
}
\arguments{
\item{x}{= number of colours needed}

\item{y}{= T/F depending on whether fill or border is wanted}
}
\description{
vector of R colours
}
\keyword{farben}
