% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.all.roxygenize}
\alias{fcn.all.roxygenize}
\title{fcn.all.roxygenize}
\usage{
fcn.all.roxygenize(x)
}
\arguments{
\item{x}{= the output file}
}
\description{
roxygenizes all functions
}
\keyword{fcn.all.roxygenize}
