% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.all.sub}
\alias{fcn.all.sub}
\title{fcn.all.sub}
\usage{
fcn.all.sub(x)
}
\arguments{
\item{x}{= a vector of function names}
}
\description{
a string vector of names of all sub-functions
}
\keyword{fcn.all.sub}
