% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.all.super}
\alias{fcn.all.super}
\title{fcn.all.super}
\usage{
fcn.all.super(x)
}
\arguments{
\item{x}{= a vector of function names}
}
\description{
names of all functions that depend on <x>
}
\keyword{fcn.all.super}
