% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.args.comment}
\alias{fcn.args.comment}
\title{fcn.args.comment}
\usage{
fcn.args.comment(x)
}
\arguments{
\item{x}{= a function name}
}
\description{
named vector of commented arguments
}
\keyword{fcn.args.comment}
