% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.canonical}
\alias{fcn.canonical}
\title{fcn.canonical}
\usage{
fcn.canonical(x)
}
\arguments{
\item{x}{= a function name}
}
\description{
T/F depending on whether <x> is in standard form
}
\keyword{fcn.canonical}
