% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.comments.parse}
\alias{fcn.comments.parse}
\title{fcn.comments.parse}
\usage{
fcn.comments.parse(x)
}
\arguments{
\item{x}{= comments section of a function}
}
\description{
extracts information from the comments
}
\keyword{fcn.comments.parse}
