% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.dates.parse}
\alias{fcn.dates.parse}
\title{fcn.dates.parse}
\usage{
fcn.dates.parse(x)
}
\arguments{
\item{x}{= date item from fcn.comments.parse}
}
\description{
dates a function was modified
}
\keyword{fcn.dates.parse}
