% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.direct.sub}
\alias{fcn.direct.sub}
\title{fcn.direct.sub}
\usage{
fcn.direct.sub(x)
}
\arguments{
\item{x}{= a SINGLE function name}
}
\description{
a string vector of names of all direct sub-functions
}
\keyword{fcn.direct.sub}
