% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.direct.super}
\alias{fcn.direct.super}
\title{fcn.direct.super}
\usage{
fcn.direct.super(x)
}
\arguments{
\item{x}{= a SINGLE function name}
}
\description{
names of all functions that directly depend on <x>
}
\keyword{fcn.direct.super}
