% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.expressions.count}
\alias{fcn.expressions.count}
\title{fcn.expressions.count}
\usage{
fcn.expressions.count(x)
}
\arguments{
\item{x}{= a function name}
}
\description{
number of expressions
}
\keyword{fcn.expressions.count}
