% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.extract.args}
\alias{fcn.extract.args}
\title{fcn.extract.args}
\usage{
fcn.extract.args(x)
}
\arguments{
\item{x}{= string vector representing argument section of comments}
}
\description{
vector of arguments with explanations
}
\keyword{fcn.extract.args}
