% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.extract.out}
\alias{fcn.extract.out}
\title{fcn.extract.out}
\usage{
fcn.extract.out(x)
}
\arguments{
\item{x}{= string vector representing output section of comments}
}
\description{
extracts output
}
\keyword{fcn.extract.out}
