% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.indent.proper}
\alias{fcn.indent.proper}
\title{fcn.indent.proper}
\usage{
fcn.indent.proper(x)
}
\arguments{
\item{x}{= a SINGLE function name}
}
\description{
T/F depending on whether the function is indented properly
}
\keyword{fcn.indent.proper}
