% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.indirect}
\alias{fcn.indirect}
\title{fcn.indirect}
\usage{
fcn.indirect(fcn, x)
}
\arguments{
\item{fcn}{= a function to apply}

\item{x}{= vector of function names}
}
\description{
applies <fcn> recursively
}
\keyword{fcn.indirect}
