% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.lines.code}
\alias{fcn.lines.code}
\title{fcn.lines.code}
\usage{
fcn.lines.code(x, y)
}
\arguments{
\item{x}{= a function name}

\item{y}{= T/F depending on whether internal comments count}
}
\description{
lines of actual code
}
\keyword{fcn.lines.code}
