% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.lines.count}
\alias{fcn.lines.count}
\title{fcn.lines.count}
\usage{
fcn.lines.count(x, y = T)
}
\arguments{
\item{x}{= a function name}

\item{y}{= T/F depending on whether internal comments count}
}
\description{
number of lines of code
}
\keyword{fcn.lines.count}
