% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.list}
\alias{fcn.list}
\title{fcn.list}
\usage{
fcn.list(x = "*")
}
\arguments{
\item{x}{= pattern you want to see in returned objects}
}
\description{
Returns the names of objects that are or are not functions
}
\keyword{fcn.list}
