% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.mat.col}
\alias{fcn.mat.col}
\title{fcn.mat.col}
\usage{
fcn.mat.col(fcn, x, y, n)
}
\arguments{
\item{fcn}{= function mapping two vectors to a single value}

\item{x}{= a vector/matrix/dataframe}

\item{y}{= either missing or a numeric isomekic vector}

\item{n}{= T/F depending on whether inputs should be ranked}
}
\description{
applies <fcn> to the columns of <x> pairwise
}
\keyword{fcn.mat.col}
