% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.mat.vec}
\alias{fcn.mat.vec}
\title{fcn.mat.vec}
\usage{
fcn.mat.vec(fcn, x, y, n)
}
\arguments{
\item{fcn}{= function mapping vector(s) to an isomekic vector}

\item{x}{= a vector/matrix/dataframe}

\item{y}{= a number/vector or matrix/dataframe with the same dimensions as <x>}

\item{n}{= T/F depending on whether you want <fcn> applied to columns or rows}
}
\description{
applies <fcn> to <x> if a vector or the columns/rows of <x> otherwise
}
\keyword{fcn.mat.vec}
