% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.matrix}
\alias{fcn.matrix}
\title{fcn.matrix}
\usage{
fcn.matrix(fcn, x, y)
}
\arguments{
\item{fcn}{= a function mapping values to values}

\item{x}{= a matrix/df}

\item{y}{= missing, isomekic vector, or isomekic isoplatic matrix/df}
}
\description{
applies <fcn> to the elements of <x> and <y>
}
\keyword{fcn.matrix}
