% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.nonNA}
\alias{fcn.nonNA}
\title{fcn.nonNA}
\usage{
fcn.nonNA(fcn, x)
}
\arguments{
\item{fcn}{= a function that maps a vector to a vector}

\item{x}{= a vector}
}
\description{
applies <fcn> to the non-NA values of <x>
}
\keyword{fcn.nonNA}
