% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.num.nonNA}
\alias{fcn.num.nonNA}
\title{fcn.num.nonNA}
\usage{
fcn.num.nonNA(fcn, x, y, n)
}
\arguments{
\item{fcn}{= a function that maps a vector to a number}

\item{x}{= a vector}

\item{y}{= either missing or an isomekic vector}

\item{n}{= T/F depending on whether inputs should be ranked}
}
\description{
applies <fcn> to the non-NA values of <x> and <y>
}
\keyword{fcn.num.nonNA}
