% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.pair.comment}
\alias{fcn.pair.comment}
\title{fcn.pair.comment}
\usage{
fcn.pair.comment(x, y, n, w)
}
\arguments{
\item{x}{= a string (function name, higher)}

\item{y}{= a string (function name, lower)}

\item{n}{= a data frame (function name/arg/comment)}

\item{w}{= a data frame (sub classes)}
}
\description{
comment consitency check
}
\seealso{
Other fcn: \code{\link{fcn.all.canonical}},
  \code{\link{fcn.all.roxygenize}},
  \code{\link{fcn.all.sub}}, \code{\link{fcn.all.super}},
  \code{\link{fcn.args.actual}},
  \code{\link{fcn.args.comment}},
  \code{\link{fcn.canonical}},
  \code{\link{fcn.comments.parse}},
  \code{\link{fcn.dates.parse}}, \code{\link{fcn.date}},
  \code{\link{fcn.direct.map}},
  \code{\link{fcn.direct.sub}},
  \code{\link{fcn.direct.super}}, \code{\link{fcn.dir}},
  \code{\link{fcn.expressions.count}},
  \code{\link{fcn.expressions.to.txt}},
  \code{\link{fcn.extract.args}},
  \code{\link{fcn.extract.out}}, \code{\link{fcn.has}},
  \code{\link{fcn.indent.proper}},
  \code{\link{fcn.lines.code}},
  \code{\link{fcn.lines.count}}, \code{\link{fcn.list}},
  \code{\link{fcn.lite}}, \code{\link{fcn.mat.col}},
  \code{\link{fcn.mat.num}}, \code{\link{fcn.mat.vec}},
  \code{\link{fcn.nonNA}}, \code{\link{fcn.num.nonNA}},
  \code{\link{fcn.order}},
  \code{\link{fcn.pair.comment.wrapper}},
  \code{\link{fcn.path}}, \code{\link{fcn.roxygenize}},
  \code{\link{fcn.sho}}, \code{\link{fcn.simple}},
  \code{\link{fcn.to.comments}}, \code{\link{fcn.to.txt}},
  \code{\link{fcn.vec.grp}}, \code{\link{fcn.vec.num}}
}
\keyword{fcn.pair.comment}
