% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.roxygenize}
\alias{fcn.roxygenize}
\title{fcn.roxygenize}
\usage{
fcn.roxygenize(x, y)
}
\arguments{
\item{x}{= function name}

\item{y}{= library to import}
}
\description{
roxygenized function format
}
\keyword{fcn.roxygenize}
