% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.simple}
\alias{fcn.simple}
\title{fcn.simple}
\usage{
fcn.simple(x)
}
\arguments{
\item{x}{= a function name}
}
\description{
T/F depending on whether <x> has multi-line expressions
}
\keyword{fcn.simple}
