% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.to.comments}
\alias{fcn.to.comments}
\title{fcn.to.comments}
\usage{
fcn.to.comments(x)
}
\arguments{
\item{x}{= a SINGLE function name}
}
\description{
returns the comment section
}
\keyword{fcn.to.comments}
