% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.vec.grp}
\alias{fcn.vec.grp}
\title{fcn.vec.grp}
\usage{
fcn.vec.grp(fcn, x, y)
}
\arguments{
\item{fcn}{= function to be applied within groups}

\item{x}{= a vector/matrix/dataframe}

\item{y}{= a vector of groups (e.g. GSec)}
}
\description{
applies <fcn> to <x> within groups <y>
}
\keyword{fcn.vec.grp}
