% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.vec.num}
\alias{fcn.vec.num}
\title{fcn.vec.num}
\usage{
fcn.vec.num(fcn, x, y)
}
\arguments{
\item{fcn}{= function mapping vector(s) to a single value}

\item{x}{= an element or vector}

\item{y}{= an element or isomekic vector}
}
\description{
applies <fcn> to <x>
}
\keyword{fcn.vec.num}
