% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fetch}
\alias{fetch}
\title{fetch}
\usage{
fetch(x, y, n, w, h)
}
\arguments{
\item{x}{= either a single variable or a vector of variable names}

\item{y}{= the YYYYMM or YYYYMMDD for which you want data}

\item{n}{= number of daily/monthly trailing periods}

\item{w}{= R-object folder}

\item{h}{= classif file}
}
\description{
fetches <x> for the trailing <n> periods ending at <y>
}
\keyword{fetch}
