% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ff.frequency}
\alias{ff.frequency}
\title{ff.frequency}
\usage{
ff.frequency(x)
}
\arguments{
\item{x}{= a string vector (row names of indicator file)}
}
\description{
D/W/M depending on frequency of the data
}
\seealso{
Other ff: \code{\link{ff.publish.last}},
  \code{\link{ff.refresh.predictors.sql}},
  \code{\link{ff.refresh.predictors}}
}
\keyword{ff.frequency}
