% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ff.FundHistory}
\alias{ff.FundHistory}
\title{ff.FundHistory}
\usage{
ff.FundHistory(x)
}
\arguments{
\item{x}{= a string (Flow/AssetsStart/AssetsEnd/PortfolioChange)}
}
\description{
additional columns needed from FundHistory
}
\seealso{
Other ff: \code{\link{ff.frequency}},
  \code{\link{ff.item}}, \code{\link{ff.publish.last}},
  \code{\link{ff.refresh.predictors.sql}},
  \code{\link{ff.refresh.predictors}}
}
\keyword{ff.FundHistory}
