% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ff.item}
\alias{ff.item}
\title{ff.item}
\usage{
ff.item(x)
}
\arguments{
\item{x}{= a string (Flow/AssetsStart/AssetsEnd/PortfolioChange)}
}
\description{
select statement item
}
\seealso{
Other ff: \code{\link{ff.FundHistory}},
  \code{\link{ff.frequency}},
  \code{\link{ff.publish.last}},
  \code{\link{ff.refresh.predictors.sql}},
  \code{\link{ff.refresh.predictors}}
}
\keyword{ff.item}
