% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ff.publish.last}
\alias{ff.publish.last}
\title{ff.publish.last}
\usage{
ff.publish.last(x)
}
\arguments{
\item{x}{= a frequency (D/W/M)}
}
\description{
D/W/M depending on frequency of the data
}
\seealso{
Other ff: \code{\link{ff.frequency}},
  \code{\link{ff.refresh.predictors.sql}},
  \code{\link{ff.refresh.predictors}}
}
\keyword{ff.publish.last}
