% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ff.refresh.predictors}
\alias{ff.refresh.predictors}
\title{ff.refresh.predictors}
\usage{
ff.refresh.predictors(x, y, n = NULL)
}
\arguments{
\item{x}{= a file}

\item{y}{= a string (Flow/AssetsStart/AssetsEnd/PortfolioChange)}

\item{n}{= a filter vector}
}
\description{
refreshes <x> all the way up to the present
}
\seealso{
Other ff: \code{\link{ff.frequency}},
  \code{\link{ff.publish.last}},
  \code{\link{ff.refresh.predictors.sql}}
}
\keyword{ff.refresh.predictors}
