% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ff.refresh.predictors.sql}
\alias{ff.refresh.predictors.sql}
\title{ff.refresh.predictors.sql}
\usage{
ff.refresh.predictors.sql(x, y, n)
}
\arguments{
\item{x}{= a matrix/data frame}

\item{y}{= a string (Flow/AssetsStart/AssetsEnd/PortfolioChange)}

\item{n}{= a frequency (D/W/M)}
}
\description{
SQL query to refresh <x>
}
\seealso{
Other ff: \code{\link{ff.frequency}},
  \code{\link{ff.publish.last}},
  \code{\link{ff.refresh.predictors}}
}
\keyword{ff.refresh.predictors.sql}
