% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{file.bkp}
\alias{file.bkp}
\title{file.bkp}
\usage{
file.bkp(x, y)
}
\arguments{
\item{x}{= a string of full paths}

\item{y}{= an isomekic string of full paths}
}
\description{
Copies <x> to <y>
}
\seealso{
Other file: \code{\link{file.break}},
  \code{\link{file.date}}, \code{\link{file.kill}},
  \code{\link{file.mtime.to.time}},
  \code{\link{file.time}}, \code{\link{file.to.last}}
}
\keyword{file.bkp}
