% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{file.break}
\alias{file.break}
\title{file.break}
\usage{
file.break(x)
}
\arguments{
\item{x}{= path to a file}
}
\description{
breaks up the file into 1GB chunks and rewrites to same directory with a "-001", "-002", etc extension
}
\seealso{
Other file: \code{\link{file.bkp}},
  \code{\link{file.date}}, \code{\link{file.kill}},
  \code{\link{file.mtime.to.time}},
  \code{\link{file.time}}, \code{\link{file.to.last}}
}
\keyword{file.break}
