% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{file.date}
\alias{file.date}
\title{file.date}
\usage{
file.date(x)
}
\arguments{
\item{x}{= a vector of full file paths}
}
\description{
Returns the last modified date in yyyymmdd format
}
\keyword{file.date}
