% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{file.kill}
\alias{file.kill}
\title{file.kill}
\usage{
file.kill(x)
}
\arguments{
\item{x}{= a file vector}
}
\description{
Deletes designated files
}
\seealso{
Other file: \code{\link{file.bkp}},
  \code{\link{file.date}},
  \code{\link{file.mtime.to.time}},
  \code{\link{file.time}}, \code{\link{file.to.last}}
}
\keyword{file.kill}
