% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{file.mtime.to.time}
\alias{file.mtime.to.time}
\title{file.mtime.to.time}
\usage{
file.mtime.to.time(x)
}
\arguments{
\item{x}{= a vector of dates}
}
\description{
Converts to HHMMSS times
}
\keyword{file.mtime.to.time}
