% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{file.to.last}
\alias{file.to.last}
\title{file.to.last}
\usage{
file.to.last(x)
}
\arguments{
\item{x}{= csv file containing the predictors}
}
\description{
the last YYYYMMDD or the last day of the YYYYMM for which we have data
}
\keyword{file.to.last}
