% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{find.data}
\alias{find.data}
\title{find.data}
\usage{
find.data(x, y = T)
}
\arguments{
\item{x}{= a logical vector}

\item{y}{= a boolean (position of first/last true value)}
}
\description{
returns the position of the first/last true value of x
}
\seealso{
Other find: \code{\link{find.gaps}}
}
\keyword{find.data}
