% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{find.gaps}
\alias{find.gaps}
\title{find.gaps}
\usage{
find.gaps(x)
}
\arguments{
\item{x}{= a logical vector}
}
\description{
returns the position of the first and last true value of x #		:	together with the first positions of all gaps
}
\seealso{
Other find: \code{\link{find.data}}
}
\keyword{find.gaps}
