% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{find.whitespace.trail}
\alias{find.whitespace.trail}
\title{find.whitespace.trail}
\usage{
find.whitespace.trail(x)
}
\arguments{
\item{x}{= the name of a function}
}
\description{
cats 2 lines above and below lines with trailing white space
}
\seealso{
Other find: \code{\link{find.data}},
  \code{\link{find.gaps}}
}
\keyword{find.whitespace.trail}
