% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{flowdate.diff}
\alias{flowdate.diff}
\title{flowdate.diff}
\usage{
flowdate.diff(x, y)
}
\arguments{
\item{x}{= a flowdate vector}

\item{y}{= a flowdate vector}
}
\description{
returns <x - y> in terms of flowdates
}
\keyword{flowdate.diff}
