% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{flowdate.ex.AllocMo}
\alias{flowdate.ex.AllocMo}
\title{flowdate.ex.AllocMo}
\usage{
flowdate.ex.AllocMo(x, y = 23)
}
\arguments{
\item{x}{= a single yyyymm}

\item{y}{= calendar day in the next month when allocations are known (usually the 23rd)}
}
\description{
Returns the flowdates corresponding to <x> (inverse of yyyymmdd.to.AllocMo)
}
\keyword{flowdate.ex.AllocMo}
