% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{flowdate.ex.yyyymm}
\alias{flowdate.ex.yyyymm}
\title{flowdate.ex.yyyymm}
\usage{
flowdate.ex.yyyymm(x, y = T)
}
\arguments{
\item{x}{= a YYYYMM vector}

\item{y}{= a boolean (return last/all flowdates)}
}
\description{
last/all trading days daily flow-publication dates in <x>
}
\seealso{
Other flowdate: \code{\link{flowdate.diff}},
  \code{\link{flowdate.ex.AllocMo}},
  \code{\link{flowdate.ex.int}},
  \code{\link{flowdate.exists}},
  \code{\link{flowdate.lag}}, \code{\link{flowdate.seq}},
  \code{\link{flowdate.to.int}}
}
\keyword{flowdate.ex.yyyymm}
