% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{flowdate.lag}
\alias{flowdate.lag}
\title{flowdate.lag}
\usage{
flowdate.lag(x, y)
}
\arguments{
\item{x}{= a vector of daily flow-publication dates}

\item{y}{= an integer}
}
\description{
lags <x> by <y> daily flow-publication dates
}
\seealso{
Other flowdate: \code{\link{flowdate.ex.int}},
  \code{\link{flowdate.ex.yyyymm}},
  \code{\link{flowdate.exists}},
  \code{\link{flowdate.seq}}, \code{\link{flowdate.to.int}}
}
\keyword{flowdate.lag}
