% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{flowdate.to.int}
\alias{flowdate.to.int}
\title{flowdate.to.int}
\usage{
flowdate.to.int(x)
}
\arguments{
\item{x}{= a vector of flow dates in YYYYMMDD format}
}
\description{
number of daily flow-publication dates after Thursday, January 1, 1970
}
\keyword{flowdate.to.int}
