% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fop.Bin}
\alias{fop.Bin}
\title{fop.Bin}
\usage{
fop.Bin(x, y, n, w, h)
}
\arguments{
\item{x}{= a matrix/df with rows indexed by time and columns indexed by bins}

\item{y}{= a matrix/data frame of returns of the same dimension as <x>}

\item{n}{= a vector corresponding to the rows of <x> that maps each row to a sub-period of interest (e.g. calendat year)}

\item{w}{= return horizon in weekdays or months}

\item{h}{= number of bins into which you are going to divide your predictors}
}
\description{
Summarizes bin excess returns by sub-periods of interest (as defined by <vec>)
}
\keyword{fop.Bin}
