% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fop.correl}
\alias{fop.correl}
\title{fop.correl}
\usage{
fop.correl(x, y, n)
}
\arguments{
\item{x}{= a matrix/df with rows indexed by time and columns indexed by bins}

\item{y}{= a matrix/data frame of returns of the same dimension as <x>}

\item{n}{= an argument which is not used}
}
\description{
computes IC
}
\seealso{
Other fop: \code{\link{fop.Bin.underlying}},
  \code{\link{fop.Bin}}, \code{\link{fop.IC.summ}},
  \code{\link{fop.IC}}, \code{\link{fop.grp.map}},
  \code{\link{fop.rank.xRet}}, \code{\link{fop.wrapper}},
  \code{\link{fop}}
}
\keyword{fop.correl}
