% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fop.IC}
\alias{fop.IC}
\title{fop.IC}
\usage{
fop.IC(x, y, n, w, h)
}
\arguments{
\item{x}{= a matrix/df with rows indexed by time and columns indexed by bins}

\item{y}{= a matrix/data frame of returns of the same dimension as <x>}

\item{n}{= a vector corresponding to the rows of <x> that maps each row to a sub-period of interest (e.g. calendat year)}

\item{w}{= return horizon in weekdays}

\item{h}{= an argument which is not used}
}
\description{
Summarizes bin excess returns by sub-periods of interest (as defined by <vec>)
}
\seealso{
Other fop: \code{\link{fop.Bin}},
  \code{\link{fop.IC.summ}}, \code{\link{fop.grp.map}},
  \code{\link{fop.rank.xRet}}, \code{\link{fop.stats}},
  \code{\link{fop.subset}}, \code{\link{fop.underlying}},
  \code{\link{fop}}
}
\keyword{fop.IC}
