% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fop.IC.summ}
\alias{fop.IC.summ}
\title{fop.IC.summ}
\usage{
fop.IC.summ(x, y)
}
\arguments{
\item{x}{= a vector of IC's}

\item{y}{= an argument which is not used}
}
\description{
Summarizes IC's
}
\seealso{
Other fop: \code{\link{fop.Bin}}, \code{\link{fop.IC}},
  \code{\link{fop.grp.map}}, \code{\link{fop.rank.xRet}},
  \code{\link{fop.stats}}, \code{\link{fop.subset}},
  \code{\link{fop.wrapper}}, \code{\link{fop}}
}
\keyword{fop.IC.summ}
