% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fop.stats}
\alias{fop.stats}
\title{fop.stats}
\usage{
fop.stats(x, y, n)
}
\arguments{
\item{x}{= output of <fop>}

\item{y}{= a column in <x>}

\item{n}{= another column in <x> containing values of interest}
}
\description{
puts all the entries corresponding to <y> on one row
}
\keyword{fop.stats}
