% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fop.underlying}
\alias{fop.underlying}
\title{fop.underlying}
\usage{
fop.underlying(x, y, n, lag, delay, nBins, grp.map, findOptimalParametersFcn,
  resample)
}
\arguments{
\item{x}{= a matrix/data frame of predictors}

\item{y}{= a matrix/data frame of returns of the same dimension as <x>}

\item{n}{= the number of days of forward returns that need to be predicted}

\item{lag}{= the number of days the predictors are lagged}

\item{delay}{= the number of days needed for the predictors to be known}

\item{nBins}{= number of bins into which you are going to divide your predictors}

\item{grp.map}{= a vector that maps yyyymmdd dates to groups of interest (e.g. day of the week)}

\item{findOptimalParametersFcn}{= the function you are using to summarize your results}

\item{resample}{= F/T depending on whether you want one simulation or the median across simulations run by removing one universe element}
}
\description{
returns a table with rows indexed by flow window length
}
\seealso{
Other fop: \code{\link{fop.Bin}},
  \code{\link{fop.IC.summ}}, \code{\link{fop.IC}},
  \code{\link{fop.grp.map}}, \code{\link{fop.rank.xRet}},
  \code{\link{fop.stats}}, \code{\link{fop.subset}},
  \code{\link{fop}}
}
\keyword{fop.underlying}
