% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fop.wrapper}
\alias{fop.wrapper}
\title{fop.wrapper}
\usage{
fop.wrapper(x, y, retW, prd.size = 5, sum.flows = F, lag = 0, delay = 2,
  floW = 1:20, nBin = 5, reverse.vbl = F, sprds = F)
}
\arguments{
\item{x}{= a matrix/data frame of predictors, the rows of which are YYYYMM or YYYYMMDD}

\item{y}{= a matrix/data frame of total return indices, the rows of which are YYYYMM or YYYYMMDD}

\item{retW}{= a numeric vector of forward return windows}

\item{prd.size}{= size of each compounding period in terms of days (days = 1, wks = 5, etc.) if <x> is indexed by YYYYMMDD or months if <x> is indexed by YYYYMM}

\item{sum.flows}{= if T, flows get summed. Otherwise they get compounded.}

\item{lag}{= an integer of predictor lags}

\item{delay}{= the number of days needed for the predictors to be known}

\item{floW}{= a numeric vector of trailing flow windows}

\item{nBin}{= a non-negative integer}

\item{reverse.vbl}{= T/F depending on whether you want the variable reversed}

\item{sprds}{= T/F depending on whether spread changes, rather than returns, are needed}
}
\description{
a table of Sharpes, IC's and annualized mean excess returns for: Q1 - a strategy that goes long the top fifth and short the equal-weight universe TxB - a strategy that goes long and short the top and bottom fifth respectively
}
\seealso{
Other fop: \code{\link{fop.Bin.underlying}},
  \code{\link{fop.Bin}}, \code{\link{fop.IC.summ}},
  \code{\link{fop.IC}}, \code{\link{fop.correl}},
  \code{\link{fop.grp.map}}, \code{\link{fop.rank.xRet}},
  \code{\link{fop}}
}
\keyword{fop.wrapper}
