% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.all.dir}
\alias{ftp.all.dir}
\title{ftp.all.dir}
\usage{
ftp.all.dir(x, y, n, w, h, u)
}
\arguments{
\item{x}{= a remote folder}

\item{y}{= ftp site (defaults to standard)}

\item{n}{= user id (defaults to standard)}

\item{w}{= password (defaults to standard)}

\item{h}{= protocol (either "ftp" or "sftp")}

\item{u}{= T/F flag for ftp.use.epsv argument of getCurlHandle}
}
\description{
remote-site directory listing of all sub-folders
}
\keyword{ftp.all.dir}
