% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.all.files.in.dir}
\alias{ftp.all.files.in.dir}
\title{ftp.all.files.in.dir}
\usage{
ftp.all.files.in.dir(x, y, n, w)
}
\arguments{
\item{x}{= a folder on our ftp site}

\item{y}{= ftp site}

\item{n}{= user id}

\item{w}{= password}
}
\description{
remote site directory listing of files (incl. sub-folders)
}
\keyword{ftp.all.files.in.dir}
