% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.all.files.underlying}
\alias{ftp.all.files.underlying}
\title{ftp.all.files.underlying}
\usage{
ftp.all.files.underlying(x, y, n, w, h = "ftp", u, v)
}
\arguments{
\item{x}{= remote folder on an ftp site (e.g. "/ftpdata/mystuff")}

\item{y}{= ftp site (defaults to standard)}

\item{n}{= user id (defaults to standard)}

\item{w}{= password (defaults to standard)}

\item{h}{= protocol (either "ftp" or "sftp")}

\item{u}{= T/F flag for ftp.use.epsv argument of getCurlHandle}

\item{v}{= T/F depending on whether you want files or folders}
}
\description{
remote-site directory listing of files or folders
}
\keyword{ftp.all.files.underlying}
