% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.del}
\alias{ftp.del}
\title{ftp.del}
\usage{
ftp.del(x, y, n, w, h, u = "ftp")
}
\arguments{
\item{x}{= a remote folder/file}

\item{y}{= a SINGLE file (e.g. "foo.txt") or missing if <x> is a file}

\item{n}{= ftp site (can be missing)}

\item{w}{= user id (can be missing)}

\item{h}{= password (can be missing)}

\item{u}{= a string (ftp/sftp)}
}
\description{
deletes file <x> or file <y> on remote folder <x>
}
\seealso{
Other ftp: \code{\link{ftp.all.dir}},
  \code{\link{ftp.all.files.underlying}},
  \code{\link{ftp.all.files}},
  \code{\link{ftp.credential}},
  \code{\link{ftp.dir.parse.ftp}},
  \code{\link{ftp.dir.parse.sftp}}, \code{\link{ftp.dir}},
  \code{\link{ftp.download}}, \code{\link{ftp.exists}},
  \code{\link{ftp.file}}, \code{\link{ftp.get}},
  \code{\link{ftp.info}}, \code{\link{ftp.kill}},
  \code{\link{ftp.list}},
  \code{\link{ftp.missing.underlying}},
  \code{\link{ftp.missing}}, \code{\link{ftp.parent}},
  \code{\link{ftp.put}}, \code{\link{ftp.record}},
  \code{\link{ftp.rmdir}}, \code{\link{ftp.sql.factor}},
  \code{\link{ftp.upload}}
}
\keyword{ftp.del}
