% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.dir.excise.crap}
\alias{ftp.dir.excise.crap}
\title{ftp.dir.excise.crap}
\usage{
ftp.dir.excise.crap(x)
}
\arguments{
\item{x}{= output from ftp directory listing}
}
\description{
cleans up output
}
\keyword{ftp.dir.excise.crap}
