% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.dir.parse.sftp}
\alias{ftp.dir.parse.sftp}
\title{ftp.dir.parse.sftp}
\usage{
ftp.dir.parse.sftp(x)
}
\arguments{
\item{x}{= string vector (raw output of ftp)}
}
\description{
data frame with ftp information
}
\keyword{ftp.dir.parse.sftp}
