% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.dir}
\alias{ftp.dir}
\title{ftp.dir}
\usage{
ftp.dir(x, y, n, w, h = F, u = "ftp", v)
}
\arguments{
\item{x}{= a remote folder}

\item{y}{= ftp site (defaults to standard)}

\item{n}{= user id (defaults to standard)}

\item{w}{= password (defaults to standard)}

\item{h}{= T/F depending on whether you want time stamps}

\item{u}{= protocol (either "ftp" or "sftp")}

\item{v}{= T/F flag for ftp.use.epsv argument of getURL}
}
\description{
logical or YYYYMMDD vector indexed by remote file names
}
\keyword{ftp.dir}
