% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.download}
\alias{ftp.download}
\title{ftp.download}
\usage{
ftp.download(x, y, n, w, h, u = "ftp", v)
}
\arguments{
\item{x}{= remote folder on an ftp site (e.g. "/ftpdata/mystuff")}

\item{y}{= local folder (e.g. "C:\\\\temp\\\\mystuff")}

\item{n}{= ftp site (defaults to standard)}

\item{w}{= user id (defaults to standard)}

\item{h}{= password (defaults to standard)}

\item{u}{= protocol (either "ftp" or "sftp")}

\item{v}{= T/F flag for ftp.use.epsv argument of getCurlHandle}
}
\description{
replicates <x> in folder <y>
}
\keyword{ftp.download}
