% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.download.script}
\alias{ftp.download.script}
\title{ftp.download.script}
\usage{
ftp.download.script(x, y, n, w, h)
}
\arguments{
\item{x}{= remote folder on our ftp site}

\item{y}{= local folder}

\item{n}{= ftp site}

\item{w}{= user id}

\item{h}{= password}
}
\description{
creates bat/ftp files to get all files from an ftp folder
}
\keyword{ftp.download.script}
