% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.exists}
\alias{ftp.exists}
\title{ftp.exists}
\usage{
ftp.exists(x, y)
}
\arguments{
\item{x}{= report name}

\item{y}{= date for which you want to send the report}
}
\description{
T/F depending on whether upload already happened
}
\keyword{ftp.exists}
