% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.get}
\alias{ftp.get}
\title{ftp.get}
\usage{
ftp.get(x, y, n, w, h, u = "ftp", v)
}
\arguments{
\item{x}{= a remote file}

\item{y}{= local folder (e.g. "C:\\\\temp")}

\item{n}{= ftp site (defaults to standard)}

\item{w}{= user id (defaults to standard)}

\item{h}{= password (defaults to standard)}

\item{u}{= protocol (either "ftp" or "sftp")}

\item{v}{= T/F flag for ftp.use.epsv argument of getCurlHandle}
}
\description{
file <x> from remote site
}
\keyword{ftp.get}
