% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.info}
\alias{ftp.info}
\title{ftp.info}
\usage{
ftp.info(x, y, n, w)
}
\arguments{
\item{x}{= M/W/D depending on whether flows are monthly/weekly/daily}

\item{y}{= T/F depending on whether you want to check Fund or Share-Class level data}

\item{n}{= one of sql.table/date.field/ftp.path}

\item{w}{= filter (e.g. Aggregate/Active/Passive/ETF/Mutual)}
}
\description{
parameter <n> associated with <x> flows at the <y> level with the <w> filter
}
\keyword{ftp.info}
