% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.is.dir}
\alias{ftp.is.dir}
\title{ftp.is.dir}
\usage{
ftp.is.dir(x, y)
}
\arguments{
\item{x}{= string vector}

\item{y}{= T/F flag for ftp.use.epsv argument of getURL}
}
\description{
T/F depending on whether each line of <x> corresponds to a folder or file
}
\seealso{
Other ftp: \code{\link{ftp.all.dir}},
  \code{\link{ftp.all.files.underlying}},
  \code{\link{ftp.all.files}}, \code{\link{ftp.break}},
  \code{\link{ftp.credential}}, \code{\link{ftp.del}},
  \code{\link{ftp.dir}}, \code{\link{ftp.download}},
  \code{\link{ftp.exists}}, \code{\link{ftp.file.size}},
  \code{\link{ftp.file}}, \code{\link{ftp.get}},
  \code{\link{ftp.info}}, \code{\link{ftp.kill}},
  \code{\link{ftp.list}}, \code{\link{ftp.parent}},
  \code{\link{ftp.put}}, \code{\link{ftp.record}},
  \code{\link{ftp.rmdir}}, \code{\link{ftp.sql.factor}},
  \code{\link{ftp.sql.other}}, \code{\link{ftp.upload}}
}
\keyword{ftp.is.dir}
