% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.is.file.underlying}
\alias{ftp.is.file.underlying}
\title{ftp.is.file.underlying}
\usage{
ftp.is.file.underlying(x, y, n, w)
}
\arguments{
\item{x}{= path to remote file or folder (e.g. "/ftpdata/mystuff")}

\item{y}{= ftp site}

\item{n}{= user id}

\item{w}{= password}
}
\description{
T/F depending on whether <x> represents a file or folder
}
\seealso{
Other ftp: \code{\link{ftp.all.dir}},
  \code{\link{ftp.all.files.underlying}},
  \code{\link{ftp.all.files}},
  \code{\link{ftp.delete.script.underlying}},
  \code{\link{ftp.delete.script}},
  \code{\link{ftp.dir.excise.crap}},
  \code{\link{ftp.dir.ftp.code}}, \code{\link{ftp.dir}},
  \code{\link{ftp.download.script}},
  \code{\link{ftp.file.size}}, \code{\link{ftp.get}},
  \code{\link{ftp.info}}, \code{\link{ftp.is.file}},
  \code{\link{ftp.put}}, \code{\link{ftp.sql}},
  \code{\link{ftp.txt}},
  \code{\link{ftp.upload.script.underlying}},
  \code{\link{ftp.upload.script}}
}
\keyword{ftp.is.file.underlying}
