% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.kill}
\alias{ftp.kill}
\title{ftp.kill}
\usage{
ftp.kill(x)
}
\arguments{
\item{x}{= report name}
}
\description{
deletes entry <x> in the ftp record. Returns nothing.
}
\seealso{
Other ftp: \code{\link{ftp.all.dir}},
  \code{\link{ftp.all.files.underlying}},
  \code{\link{ftp.all.files}},
  \code{\link{ftp.credential}}, \code{\link{ftp.del}},
  \code{\link{ftp.dir.parse.ftp}},
  \code{\link{ftp.dir.parse.sftp}}, \code{\link{ftp.dir}},
  \code{\link{ftp.download}}, \code{\link{ftp.exists}},
  \code{\link{ftp.file}}, \code{\link{ftp.get}},
  \code{\link{ftp.info}}, \code{\link{ftp.list}},
  \code{\link{ftp.missing.underlying}},
  \code{\link{ftp.missing}}, \code{\link{ftp.parent}},
  \code{\link{ftp.put}}, \code{\link{ftp.record}},
  \code{\link{ftp.rmdir}}, \code{\link{ftp.sql.factor}},
  \code{\link{ftp.sql.other}}, \code{\link{ftp.upload}}
}
\keyword{ftp.kill}
