% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.missing}
\alias{ftp.missing}
\title{ftp.missing}
\usage{
ftp.missing(x, y)
}
\arguments{
\item{x}{= arguments from higher function}

\item{y}{= argument names}
}
\description{
supplies missing arguments
}
\keyword{ftp.missing}
