% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.missing.underlying}
\alias{ftp.missing.underlying}
\title{ftp.missing.underlying}
\usage{
ftp.missing.underlying(x, y, n, w, h)
}
\arguments{
\item{x}{= ftp site (defaults to standard)}

\item{y}{= user id (defaults to standard)}

\item{n}{= password (defaults to standard)}

\item{w}{= protocol (either "ftp" or "sftp")}

\item{h}{= T/F flag for ftp.use.epsv argument of getURL}
}
\description{
logical or YYYYMMDD vector indexed by remote file names
}
\keyword{ftp.missing.underlying}
