% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.put}
\alias{ftp.put}
\title{ftp.put}
\usage{
ftp.put(x, y, n, w, h, u = "ftp", v)
}
\arguments{
\item{x}{= remote folder on an ftp site (e.g. "/ftpdata/mystuff")}

\item{y}{= a vector of local file path(s) (e.g. "C:\\\\temp\\\\foo.txt")}

\item{n}{= ftp site (defaults to standard)}

\item{w}{= user id (defaults to standard)}

\item{h}{= password (defaults to standard)}

\item{u}{= protocol (either "ftp" or "sftp")}

\item{v}{= T/F flag for ftp.use.epsv argument of getCurlHandle}
}
\description{
puts file <y> to remote site <x>, creating folders as needed
}
\keyword{ftp.put}
