% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.record}
\alias{ftp.record}
\title{ftp.record}
\usage{
ftp.record(x, y)
}
\arguments{
\item{x}{= report name}

\item{y}{= date for which you sent the report}
}
\description{
updates the email record. Returns nothing.
}
\keyword{ftp.record}
