% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.rmdir}
\alias{ftp.rmdir}
\title{ftp.rmdir}
\usage{
ftp.rmdir(x, y, n, w, h = "ftp")
}
\arguments{
\item{x}{= a remote folder}

\item{y}{= ftp site (defaults to standard)}

\item{n}{= user id (defaults to standard)}

\item{w}{= password (defaults to standard)}

\item{h}{= protocol (either "ftp" or "sftp")}
}
\description{
removes directory <x> (e.g. "mystuff")
}
\keyword{ftp.rmdir}
