% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.sql.factor}
\alias{ftp.sql.factor}
\title{ftp.sql.factor}
\usage{
ftp.sql.factor(x, y, n, w, h)
}
\arguments{
\item{x}{= factor}

\item{y}{= a flowdate}

\item{n}{= fund filter (e.g. Aggregate/Active/Passive/ETF/Mutual)}

\item{w}{= stock filter (e.g. All/China/Japan)}

\item{h}{= breakdown filter (e.g. All/GeoId/DomicileId)}
}
\description{
SQL code to validate <x> flows at the <y> level
}
\keyword{ftp.sql.factor}
