% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.sql.other}
\alias{ftp.sql.other}
\title{ftp.sql.other}
\usage{
ftp.sql.other(x, y, n)
}
\arguments{
\item{x}{= M/W/D/C/I/S depending on flows or allocations}

\item{y}{= flow date in YYYYMMDD format}

\item{n}{= filter (e.g. Aggregate/Active/Passive/ETF/Mutual)}
}
\description{
SQL code to validate <x> flows at the <y> level
}
\seealso{
Other ftp: \code{\link{ftp.action}},
  \code{\link{ftp.all.dir}},
  \code{\link{ftp.all.files.underlying}},
  \code{\link{ftp.all.files}},
  \code{\link{ftp.credential}}, \code{\link{ftp.del}},
  \code{\link{ftp.dir.excise.crap}},
  \code{\link{ftp.dir.ftp.code}}, \code{\link{ftp.dir}},
  \code{\link{ftp.download}}, \code{\link{ftp.file.size}},
  \code{\link{ftp.file}}, \code{\link{ftp.get}},
  \code{\link{ftp.info}}, \code{\link{ftp.mkdir}},
  \code{\link{ftp.parent}}, \code{\link{ftp.put}},
  \code{\link{ftp.remove}}, \code{\link{ftp.rmdir}},
  \code{\link{ftp.sql.factor}}, \code{\link{ftp.txt}},
  \code{\link{ftp.upload}}
}
\keyword{ftp.sql.other}
