% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.upload}
\alias{ftp.upload}
\title{ftp.upload}
\usage{
ftp.upload(x, y, n, w, h, u = "ftp", v)
}
\arguments{
\item{x}{= empty a remote folder}

\item{y}{= local folder containing the data (e.g. "C:\\\\temp\\\\mystuff")}

\item{n}{= ftp site (defaults to standard)}

\item{w}{= user id (defaults to standard)}

\item{h}{= password (defaults to standard)}

\item{u}{= protocol (either "ftp" or "sftp")}

\item{v}{= T/F flag for ftp.use.epsv argument of getCurlHandle}
}
\description{
Copies up files from the local machine
}
\keyword{ftp.upload}
