% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.upload.script.underlying}
\alias{ftp.upload.script.underlying}
\title{ftp.upload.script.underlying}
\usage{
ftp.upload.script.underlying(x)
}
\arguments{
\item{x}{= local folder containing the data (e.g. "C:\\\\temp\\\\mystuff")}
}
\description{
returns ftp script to copy up files from the local machine
}
\seealso{
Other ftp: \code{\link{ftp.all.dir}},
  \code{\link{ftp.all.files.underlying}},
  \code{\link{ftp.all.files}},
  \code{\link{ftp.delete.script.underlying}},
  \code{\link{ftp.delete.script}},
  \code{\link{ftp.dir.excise.crap}},
  \code{\link{ftp.dir.ftp.code}}, \code{\link{ftp.dir}},
  \code{\link{ftp.download.script}},
  \code{\link{ftp.file.size}}, \code{\link{ftp.get}},
  \code{\link{ftp.info}},
  \code{\link{ftp.is.file.underlying}},
  \code{\link{ftp.is.file}}, \code{\link{ftp.put}},
  \code{\link{ftp.sql.factor}},
  \code{\link{ftp.sql.other}}, \code{\link{ftp.txt}},
  \code{\link{ftp.upload.script}}
}
\keyword{ftp.upload.script.underlying}
