% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftpQuant.write}
\alias{ftpQuant.write}
\title{ftpQuant.write}
\usage{
ftpQuant.write(x)
}
\arguments{
\item{x}{= a string (report name, can be missing)}
}
\description{
writes to ftpQuant table
}
\keyword{ftpQuant.write}
