% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{glome.ex.R3.prime}
\alias{glome.ex.R3.prime}
\title{glome.ex.R3.prime}
\usage{
glome.ex.R3.prime(x, y, n)
}
\arguments{
\item{x}{= a number between 0 and 1}

\item{y}{= a number between 0 and 1}

\item{n}{= a number between 0 and 1}
}
\description{
derivative of glome.ex.R3, a 4 X 3 matrix
}
\seealso{
Other glome: \code{\link{glome.ex.R3}},
  \code{\link{glome.near}}
}
\keyword{glome.ex.R3.prime}
