% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{glome.ex.R3}
\alias{glome.ex.R3}
\title{glome.ex.R3}
\usage{
glome.ex.R3(x, y, n)
}
\arguments{
\item{x}{= a number or numeric vector between 0 and 1}

\item{y}{= a number or numeric vector between 0 and 1}

\item{n}{= a number or numeric vector between 0 and 1}
}
\description{
maps unit cube to the glome (sphere in 4D)
}
\seealso{
Other glome: \code{\link{glome.ex.R3.prime}},
  \code{\link{glome.near}}
}
\keyword{glome.ex.R3}
