% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{glome.near}
\alias{glome.near}
\title{glome.near}
\usage{
glome.near(x, y, n, w)
}
\arguments{
\item{x}{= a number between 0 and 1}

\item{y}{= a number between 0 and 1}

\item{n}{= a number between 0 and 1}

\item{w}{= distance metric}
}
\description{
returns six points nearest glome.ex.R3(x, y, z)
}
\seealso{
Other glome: \code{\link{glome.ex.R3.prime}},
  \code{\link{glome.ex.R3}}
}
\keyword{glome.near}
