% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{grp.unique}
\alias{grp.unique}
\title{grp.unique}
\usage{
grp.unique(x)
}
\arguments{
\item{x}{= a string vector}
}
\description{
list of unique groups. disentangles memberships separated by "-"
}
\keyword{grp.unique}
