% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{html.and}
\alias{html.and}
\title{html.and}
\usage{
html.and(x)
}
\arguments{
\item{x}{= a string vector}
}
\description{
<x> stated in a grammatical phrase
}
\seealso{
Other html: \code{\link{html.email.done}},
  \code{\link{html.email}}, \code{\link{html.ex.utf8}},
  \code{\link{html.flow.CountInOut}},
  \code{\link{html.flow.breakdown}},
  \code{\link{html.flow.english}},
  \code{\link{html.flow.underlying}},
  \code{\link{html.image}}, \code{\link{html.list}},
  \code{\link{html.positioning}},
  \code{\link{html.problem.underlying}},
  \code{\link{html.problem}}, \code{\link{html.signature}},
  \code{\link{html.tbl}}, \code{\link{html.tenure}}
}
\keyword{html.and}
