% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{html.ex.utf8}
\alias{html.ex.utf8}
\title{html.ex.utf8}
\usage{
html.ex.utf8(x)
}
\arguments{
\item{x}{= string vector}
}
\description{
code to represent <x> in html
}
\seealso{
Other html: \code{\link{html.and}},
  \code{\link{html.email}},
  \code{\link{html.flow.breakdown}},
  \code{\link{html.flow.english}},
  \code{\link{html.flow.underlying}},
  \code{\link{html.image}}, \code{\link{html.list}},
  \code{\link{html.positioning}},
  \code{\link{html.problem.underlying}},
  \code{\link{html.problem}}, \code{\link{html.signature}},
  \code{\link{html.tbl}}, \code{\link{html.tenure}}
}
\keyword{html.ex.utf8}
