% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{html.image}
\alias{html.image}
\title{html.image}
\usage{
html.image(x, y)
}
\arguments{
\item{x}{= path to the image}

\item{y}{= percentage magnification}
}
\description{
html to attach an image
}
\keyword{html.image}
