% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{html.positioning}
\alias{html.positioning}
\title{html.positioning}
\usage{
html.positioning(x, y)
}
\arguments{
\item{x}{= matrix of indicator values}

\item{y}{= security names (corresponding to columns of <x>)}
}
\description{
writes a positioning report
}
\keyword{html.positioning}
