% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{html.signature}
\alias{html.signature}
\title{html.signature}
\usage{
html.signature()
}
\description{
signature at the end of an email
}
\seealso{
Other html: \code{\link{html.and}},
  \code{\link{html.flow.breakdown}},
  \code{\link{html.flow.english}},
  \code{\link{html.flow.underlying}},
  \code{\link{html.image}}, \code{\link{html.list}},
  \code{\link{html.positioning}}, \code{\link{html.tbl}},
  \code{\link{html.tenure}}
}
\keyword{html.signature}
