% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{html.tbl}
\alias{html.tbl}
\title{html.tbl}
\usage{
html.tbl(x, y)
}
\arguments{
\item{x}{= a matrix/data frame}

\item{y}{= a boolean (do/don't format integers)}
}
\description{
renders <x> in html
}
\seealso{
Other html: \code{\link{html.and}},
  \code{\link{html.email}}, \code{\link{html.ex.utf8}},
  \code{\link{html.flow.breakdown}},
  \code{\link{html.flow.english}},
  \code{\link{html.flow.underlying}},
  \code{\link{html.image}}, \code{\link{html.list}},
  \code{\link{html.positioning}},
  \code{\link{html.problem.underlying}},
  \code{\link{html.problem}}, \code{\link{html.signature}},
  \code{\link{html.tenure}}
}
\keyword{html.tbl}
