% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{html.tenure}
\alias{html.tenure}
\title{html.tenure}
\usage{
html.tenure(x, y, n)
}
\arguments{
\item{x}{= a named vector of integers}

\item{y}{= vector of length two for positive descriptions}

\item{n}{= vector of length two for negative descriptions}
}
\description{
describes how long securities/factors have belonged to a group
}
\keyword{html.tenure}
