% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{indicator.track}
\alias{indicator.track}
\title{indicator.track}
\usage{
indicator.track(x, y)
}
\arguments{
\item{x}{= a flowdate}

\item{y}{= a boolean (regular/Asia process)}
}
\description{
writes report for date <x>
}
\keyword{indicator.track}
