% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{int.exists}
\alias{int.exists}
\title{int.exists}
\usage{
int.exists(x)
}
\arguments{
\item{x}{= string vector}
}
\description{
T/F depending on whether <x> contains only integers 0-9
}
\seealso{
Other int: \code{\link{int.format}},
  \code{\link{int.random}}, \code{\link{int.to.prime}}
}
\keyword{int.exists}
