% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{int.format}
\alias{int.format}
\title{int.format}
\usage{
int.format(x)
}
\arguments{
\item{x}{= a vector of integers}
}
\description{
adds commas "1,234,567"
}
\seealso{
Other int: \code{\link{int.random}},
  \code{\link{int.to.prime}}
}
\keyword{int.format}
