% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{int.random}
\alias{int.random}
\title{int.random}
\usage{
int.random(x = 5)
}
\arguments{
\item{x}{= a natural number}
}
\description{
random integer between 1 and <x>
}
\seealso{
Other int: \code{\link{int.exists}},
  \code{\link{int.format}}, \code{\link{int.to.prime}}
}
\keyword{int.random}
