% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{int.to.prime}
\alias{int.to.prime}
\title{int.to.prime}
\usage{
int.to.prime(x)
}
\arguments{
\item{x}{= an integer}
}
\description{
prime factors of <x>
}
\seealso{
Other int: \code{\link{int.format}}
}
\keyword{int.to.prime}
