% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{isin.check.digit}
\alias{isin.check.digit}
\title{isin.check.digit}
\usage{
isin.check.digit(x, y)
}
\arguments{
\item{x}{= a string of 9-digit nsin's (national security id's)}

\item{y}{= the code of the country of origin}
}
\description{
The check digit, derived using the 'Modulus 10 Double Add Double' technique.
}
\seealso{
Other isin: \code{\link{isin.ex.cusip}},
  \code{\link{isin.ex.sedol}}
}
\keyword{isin.check.digit}
