% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{isin.ex.cusip}
\alias{isin.ex.cusip}
\title{isin.ex.cusip}
\usage{
isin.ex.cusip(x, y)
}
\arguments{
\item{x}{= a string of cusips}

\item{y}{= the country of origin (either "CA" or "US")}
}
\description{
a string of isin's
}
\seealso{
Other isin: \code{\link{isin.check.digit}},
  \code{\link{isin.ex.sedol}}
}
\keyword{isin.ex.cusip}
