% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{isin.ex.sedol}
\alias{isin.ex.sedol}
\title{isin.ex.sedol}
\usage{
isin.ex.sedol(x, y)
}
\arguments{
\item{x}{= a string of 7-digit sedols}

\item{y}{= the country of origin}
}
\description{
a string of isin's
}
\keyword{isin.ex.sedol}
