% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{knapsack.process}
\alias{knapsack.process}
\title{knapsack.process}
\usage{
knapsack.process(x, y, n, w, h, u, v)
}
\arguments{
\item{x}{= a matrix}

\item{y}{= a string vector (binning group)}

\item{n}{= an integer (beg comb)}

\item{w}{= an integer (end comb)}

\item{h}{= a list (result object)}

\item{u}{= a string (log file, can be missing)}

\item{v}{= an integer (sim number, can be missing)}
}
\description{
best AnnMn/Sharpe
}
\seealso{
Other knapsack: \code{\link{knapsack.count}},
  \code{\link{knapsack.ex.int}},
  \code{\link{knapsack.next}}, \code{\link{knapsack.prev}},
  \code{\link{knapsack.to.int}}
}
\keyword{knapsack.process}
