% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{knapsack}
\alias{knapsack}
\title{knapsack}
\usage{
knapsack(x, y)
}
\arguments{
\item{x}{= a non-negative integer}

\item{y}{= a positive integer exceeding unity}
}
\description{
all the ways to subdivide <x> things amongst <y> people
}
\keyword{knapsack}
