% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{knapsack.to.int}
\alias{knapsack.to.int}
\title{knapsack.to.int}
\usage{
knapsack.to.int(x)
}
\arguments{
\item{x}{= a vector of non-negative integers}
}
\description{
maps each particular way to subdivide <sum(x)> things amongst <length(x)> people to the number line
}
\keyword{knapsack.to.int}
