% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{latin.ex.arabic}
\alias{latin.ex.arabic}
\title{latin.ex.arabic}
\usage{
latin.ex.arabic(x)
}
\arguments{
\item{x}{= a numeric vector}
}
\description{
lower-case latin representation of <x>
}
\seealso{
Other latin: \code{\link{latin.to.arabic}}
}
\keyword{latin.ex.arabic}
