% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{latin.to.arabic.underlying}
\alias{latin.to.arabic.underlying}
\title{latin.to.arabic.underlying}
\usage{
latin.to.arabic.underlying()
}
\description{
basic map of latin to arabic numerals
}
\keyword{latin.to.arabic.underlying}
