% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{lead.lag.effects}
\alias{lead.lag.effects}
\title{lead.lag.effects}
\usage{
lead.lag.effects(fcn, x, y, n = seq(-10, 10))
}
\arguments{
\item{fcn}{= a function that maps yyyymmdd dates to groups of interest (e.g. day.to.weekday)}

\item{x}{= a matrix/data frame of predictors}

\item{y}{= a vector of lags corresponding to each column}

\item{n}{= a vector of integers}
}
\description{
Correlates predictor with return columns with various leads/lags. Zero is contemporaneous. +ve numbers associated with future returns
}
\keyword{lead.lag.effects}
