% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{list.bind}
\alias{list.bind}
\title{list.bind}
\usage{
list.bind(x, y = "yyyymmdd")
}
\arguments{
\item{x}{= a list of dataframes}

\item{y}{= column name to store names of <x>}
}
\description{
rbinds elements of <x> with added column <y>
}
\seealso{
Other list: \code{\link{list.rename}}
}
\keyword{list.bind}
