% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{list.common.row.space}
\alias{list.common.row.space}
\title{list.common.row.space}
\usage{
list.common.row.space(fcn, x, y)
}
\arguments{
\item{fcn}{= function used to combine row spaces}

\item{x}{= a list of mat objects}

\item{y}{= column containing row names}
}
\description{
list object with the elements mapped to the common row space
}
\keyword{list.common.row.space}
