% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{list.rename}
\alias{list.rename}
\title{list.rename}
\usage{
list.rename(x, y, n)
}
\arguments{
\item{x}{= list}

\item{y}{= a string vector (old names)}

\item{n}{= a string vector (new names)}
}
\description{
renamed list
}
\keyword{list.rename}
