% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{load.dy.vbl.1obj}
\alias{load.dy.vbl.1obj}
\title{load.dy.vbl.1obj}
\usage{
load.dy.vbl.1obj(beg, end, mk.fcn, optional.args, vbl.name, mo, env)
}
\arguments{
\item{beg}{= a single YYYYMMDD}

\item{end}{= a single YYYYMMDD}

\item{mk.fcn}{= a function}

\item{optional.args}{= passed down to <mk.fcn>}

\item{vbl.name}{= name under which the variable is to be stored}

\item{mo}{= the YYYYMM for which the object is to be made}

\item{env}{= stock-flows environment}
}
\description{
Loads a daily variable
}
\seealso{
Other load: \code{\link{load.dy.vbl}},
  \code{\link{load.mo.vbl.1obj}}, \code{\link{load.mo.vbl}}
}
\keyword{load.dy.vbl.1obj}
