% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{load.dy.vbl}
\alias{load.dy.vbl}
\title{load.dy.vbl}
\usage{
load.dy.vbl(beg, end, mk.fcn, optional.args, vbl.name, out.fldr, env)
}
\arguments{
\item{beg}{= a single YYYYMMDD}

\item{end}{= a single YYYYMMDD}

\item{mk.fcn}{= a function}

\item{optional.args}{= passed down to <mk.fcn>}

\item{vbl.name}{= name under which the variable is to be stored}

\item{out.fldr}{= R-object folder}

\item{env}{= stock-flows environment}
}
\description{
Loads a daily variable
}
\keyword{load.dy.vbl}
