% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{load.dy.vbl.underlying}
\alias{load.dy.vbl.underlying}
\title{load.dy.vbl.underlying}
\usage{
load.dy.vbl.underlying(beg, end, mk.fcn, optional.args, vbl.name, out.fldr, env,
  fcn.conv, fcn.load)
}
\arguments{
\item{beg}{= a single YYYYMMDD}

\item{end}{= a single YYYYMMDD}

\item{mk.fcn}{= a function}

\item{optional.args}{= passed down to <mk.fcn>}

\item{vbl.name}{= name under which the variable is to be stored}

\item{out.fldr}{= R-object folder}

\item{env}{= stock-flows environment}

\item{fcn.conv}{= conversion from period of columns to period of objects}

\item{fcn.load}{= function to load one object}
}
\description{
Loads a variable
}
\seealso{
Other load: \code{\link{load.dy.vbl.1obj}},
  \code{\link{load.dy.vbl}},
  \code{\link{load.mo.vbl.1obj}}, \code{\link{load.mo.vbl}}
}
\keyword{load.dy.vbl.underlying}
