% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{load.mo.vbl.1obj}
\alias{load.mo.vbl.1obj}
\title{load.mo.vbl.1obj}
\usage{
load.mo.vbl.1obj(fcn, x, y, n, w, h, u)
}
\arguments{
\item{fcn}{= a function}

\item{x}{= a YYYYMM}

\item{y}{= a YYYYMM}

\item{n}{= passed down to <fcn>}

\item{w}{= a string (name to store variable under)}

\item{h}{= the period for which the object is to be made}

\item{u}{= a StockFlows environment}
}
\description{
Loads a monthly variable
}
\seealso{
Other load: \code{\link{load.dy.vbl.1obj}},
  \code{\link{load.dy.vbl}}, \code{\link{load.mo.vbl}},
  \code{\link{load.vbl.underlying}}
}
\keyword{load.mo.vbl.1obj}
