% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{load.mo.vbl}
\alias{load.mo.vbl}
\title{load.mo.vbl}
\usage{
load.mo.vbl(fcn, x, y, n, w, h, u)
}
\arguments{
\item{fcn}{= a function}

\item{x}{= a YYYYMM}

\item{y}{= a YYYYMM}

\item{n}{= a string vector (passed down to <fcn>)}

\item{w}{= a string (name to store variable under)}

\item{h}{= R-object folder}

\item{u}{= a StockFlows environment}
}
\description{
Loads a monthly variable
}
\seealso{
Other load: \code{\link{load.dy.vbl.1obj}},
  \code{\link{load.dy.vbl}},
  \code{\link{load.mo.vbl.1obj}},
  \code{\link{load.vbl.underlying}}
}
\keyword{load.mo.vbl}
