% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{load.vbl.underlying}
\alias{load.vbl.underlying}
\title{load.vbl.underlying}
\usage{
load.vbl.underlying(fcn, x, y, n, w, h, u, v)
}
\arguments{
\item{fcn}{= a function}

\item{x}{= a YYYYMMDD}

\item{y}{= a YYYYMMDD}

\item{n}{= passed down to <fcn>}

\item{w}{= name under which the variable is to be stored}

\item{h}{= R-object folder}

\item{u}{= stock-flows environment}

\item{v}{= T/F depending on daily/monthly}
}
\description{
Loads a variable
}
\keyword{load.vbl.underlying}
