% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{machine.info}
\alias{machine.info}
\title{machine.info}
\usage{
machine.info(x)
}
\arguments{
\item{x}{= a column in the classif-Machines file}
}
\description{
folder of function source file
}
\keyword{machine.info}
