% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{machine}
\alias{machine}
\title{machine}
\usage{
machine(x, y)
}
\arguments{
\item{x}{= argument that applies to Vik's machine}

\item{y}{= argument that applies to prod dev}
}
\description{
folder of function source file
}
\keyword{machine}
