% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{map.classif}
\alias{map.classif}
\title{map.classif}
\usage{
map.classif(x, y, n)
}
\arguments{
\item{x}{= a named vector}

\item{y}{= <classif>}

\item{n}{= something like "isin" or "HSId"}
}
\description{
Maps data to the row space of <y>
}
\keyword{map.classif}
