% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{map.rname}
\alias{map.rname}
\title{map.rname}
\usage{
map.rname(x, y)
}
\arguments{
\item{x}{= a string vector/matrix/data frame}

\item{y}{= a string vector}
}
\description{
returns a matrix/data frame, the row names of which match up with <y>
}
\seealso{
Other map: \code{\link{map.classif}}
}
\keyword{map.rname}
