% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.combine}
\alias{mat.combine}
\title{mat.combine}
\usage{
mat.combine(fcn, x, y)
}
\arguments{
\item{fcn}{= the function you want applied to row space of <x> and <y>}

\item{x}{= a matrix/df}

\item{y}{= a matrix/df}
}
\description{
Combines <x> and <y>
}
\keyword{mat.combine}
