% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.correl}
\alias{mat.correl}
\title{mat.correl}
\usage{
mat.correl(x, y)
}
\arguments{
\item{x}{= a vector/matrix/data-frame}

\item{y}{= an isomekic vector or isomekic isoplatic matrix/data-frame}
}
\description{
Returns the correlation of <x> & <y> if <x> is a vector or those between the rows of <x> and <y> otherwise
}
\keyword{mat.correl}
