% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.daily.to.monthly}
\alias{mat.daily.to.monthly}
\title{mat.daily.to.monthly}
\usage{
mat.daily.to.monthly(x, y = F)
}
\arguments{
\item{x}{= a matrix/df of daily data}

\item{y}{= T/F depending on whether data points must be from month ends}
}
\description{
returns latest data in each month indexed by <yyyymm> ascending
}
\keyword{mat.daily.to.monthly}
