% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.daily.to.weekly}
\alias{mat.daily.to.weekly}
\title{mat.daily.to.weekly}
\usage{
mat.daily.to.weekly(x, y)
}
\arguments{
\item{x}{= a matrix/df of daily data}

\item{y}{= an integer representing the day the week ends on 0 is Sun, 1 is Mon, ..., 6 is Sat}
}
\description{
returns latest data in each week in ascending order
}
\seealso{
Other mat: \code{\link{mat.compound}},
  \code{\link{mat.correl}}, \code{\link{mat.count}},
  \code{\link{mat.daily.to.monthly}},
  \code{\link{mat.ex.array3d}}, \code{\link{mat.ex.array}},
  \code{\link{mat.ex.matrix}}, \code{\link{mat.ex.vec}},
  \code{\link{mat.fake}}, \code{\link{mat.index}},
  \code{\link{mat.lag}}, \code{\link{mat.last.to.first}},
  \code{\link{mat.rank}}, \code{\link{mat.read}},
  \code{\link{mat.reverse}}, \code{\link{mat.same}},
  \code{\link{mat.sort}}, \code{\link{mat.subset}},
  \code{\link{mat.to.first.data.row}},
  \code{\link{mat.to.lags}}, \code{\link{mat.to.last.Idx}},
  \code{\link{mat.to.matrix}}, \code{\link{mat.to.obs}},
  \code{\link{mat.to.xlModel}}, \code{\link{mat.write}},
  \code{\link{mat.zScore}}
}
\keyword{mat.daily.to.weekly}
