% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.diff}
\alias{mat.diff}
\title{mat.diff}
\usage{
mat.diff(x, y)
}
\arguments{
\item{x}{= a matrix/df}

\item{y}{= a non-negative integer}
}
\description{
difference between <x> and itself lagged <y>
}
\keyword{mat.diff}
