% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.ex.array}
\alias{mat.ex.array}
\title{mat.ex.array}
\usage{
mat.ex.array(x, y)
}
\arguments{
\item{x}{= any numerical array}

\item{y}{= a vector of names for the columns of the output corresponding to the dimensions of <x>}
}
\description{
unlists the contents of an array
}
\keyword{mat.ex.array}
