% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.ex.array3d}
\alias{mat.ex.array3d}
\title{mat.ex.array3d}
\usage{
mat.ex.array3d(x, y = "C", n = "A")
}
\arguments{
\item{x}{= a three-dimensional numerical array}

\item{y}{= dimension which becomes the panel header}

\item{n}{= dimension which forms the rows}
}
\description{
unlists the contents of an array to a data frame
}
\keyword{mat.ex.array3d}
