% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.ex.list}
\alias{mat.ex.list}
\title{mat.ex.list}
\usage{
mat.ex.list(x, y = "yyyymmdd")
}
\arguments{
\item{x}{= a list of dataframes}

\item{y}{= column name to store names of <x>}
}
\description{
rbinds elements of <x> with added column <y>
}
\keyword{mat.ex.list}
