% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.ex.matrix}
\alias{mat.ex.matrix}
\title{mat.ex.matrix}
\usage{
mat.ex.matrix(x, y = NULL)
}
\arguments{
\item{x}{= a matrix}

\item{y}{= desired row names (defaults to NULL)}
}
\description{
converts into a data frame
}
\keyword{mat.ex.matrix}
