% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.ex.qtl}
\alias{mat.ex.qtl}
\title{mat.ex.qtl}
\usage{
mat.ex.qtl(x, y)
}
\arguments{
\item{x}{= a vector}

\item{y}{= an isomekic vector}
}
\description{
transforms into a 1/0 matrix of bin memberships if <y> is missing or the values of <y> otherwise
}
\keyword{mat.ex.qtl}
