% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.ex.vec}
\alias{mat.ex.vec}
\title{mat.ex.vec}
\usage{
mat.ex.vec(x, y, n = T)
}
\arguments{
\item{x}{= a numeric or character vector}

\item{y}{= an isomekic vector of associated values}

\item{n}{= T/F depending on whether "Q" is to be appended to column headers}
}
\description{
transforms into a 1/0 matrix of bin memberships if <y> is missing or the values of <y> otherwise
}
\keyword{mat.ex.vec}
