% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.index}
\alias{mat.index}
\title{mat.index}
\usage{
mat.index(x, y = 1, n = T)
}
\arguments{
\item{x}{= a matrix/df}

\item{y}{= columns}

\item{n}{= T/F depending on whether you remove columns <y>}
}
\description{
indexes <x> by, and, if <n>, removes, columns <y>
}
\keyword{mat.index}
