% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.lag.cols}
\alias{mat.lag.cols}
\title{mat.lag.cols}
\usage{
mat.lag.cols(x, y)
}
\arguments{
\item{x}{= a matrix/data frame of predictors}

\item{y}{= a vector of lags corresponding to each column}
}
\description{
lags columns of <x> individually based on <y>
}
\keyword{mat.lag.cols}
