% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.lag}
\alias{mat.lag}
\title{mat.lag}
\usage{
mat.lag(x, y)
}
\arguments{
\item{x}{= a matrix/df indexed by time running FORWARDS}

\item{y}{= number of periods over which to lag}
}
\description{
Returns data lagged <y> periods with the same row space as <x>
}
\keyword{mat.lag}
