% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.last.to.first}
\alias{mat.last.to.first}
\title{mat.last.to.first}
\usage{
mat.last.to.first(x, y = 1)
}
\arguments{
\item{x}{= a matrix/df}

\item{y}{= a non-negative integer}
}
\description{
Re-orders so the last <y> columns come first
}
\keyword{mat.last.to.first}
