% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.rank}
\alias{mat.rank}
\title{mat.rank}
\usage{
mat.rank(x)
}
\arguments{
\item{x}{= a vector/matrix/data-frame}
}
\description{
ranks <x> if <x> is a vector or the rows of <x> otherwise
}
\keyword{mat.rank}
