% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.rollsum}
\alias{mat.rollsum}
\title{mat.rollsum}
\usage{
mat.rollsum(x, y)
}
\arguments{
\item{x}{= a matrix/df}

\item{y}{= a non-negative integer}
}
\description{
rolling sum of <n> rows
}
\keyword{mat.rollsum}
