% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.same}
\alias{mat.same}
\title{mat.same}
\usage{
mat.same(x, y)
}
\arguments{
\item{x}{= a matrix/df}

\item{y}{= an isomekic isoplatic matrix/df}
}
\description{
T/F depending on whether <x> and <y> are identical
}
\keyword{mat.same}
