% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.sort}
\alias{mat.sort}
\title{mat.sort}
\usage{
mat.sort(x, y, n)
}
\arguments{
\item{x}{= a matrix/df}

\item{y}{= string vector of column names of <x>}

\item{n}{= logical vector of the same length as <y>}
}
\description{
sorts <x> by <y> in decreasing order if <n> is T
}
\keyword{mat.sort}
