% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.to.first.data.row}
\alias{mat.to.first.data.row}
\title{mat.to.first.data.row}
\usage{
mat.to.first.data.row(x)
}
\arguments{
\item{x}{= a matrix/data-frame}
}
\description{
the row number of the first row containing data
}
\keyword{mat.to.first.data.row}
