% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.to.lags}
\alias{mat.to.lags}
\title{mat.to.lags}
\usage{
mat.to.lags(x, y, n = T, w = 1)
}
\arguments{
\item{x}{= a matrix/df indexed by time running FORWARDS}

\item{y}{= number of lagged values desired plus one}

\item{n}{= if T simple positional lagging is used. If F, yyyymm.lag is invoked}

\item{w}{= size of each period in terms of YYYYMMDD or YYYYMM depending on the rows of <x>}
}
\description{
a 3D array of <x> together with itself lagged 1, ..., <y> - 1 times
}
\keyword{mat.to.lags}
