% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.to.last.Idx}
\alias{mat.to.last.Idx}
\title{mat.to.last.Idx}
\usage{
mat.to.last.Idx(x)
}
\arguments{
\item{x}{= a matrix/df}
}
\description{
the last row index for which we have data
}
\keyword{mat.to.last.Idx}
