% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.to.matrix}
\alias{mat.to.matrix}
\title{mat.to.matrix}
\usage{
mat.to.matrix(x)
}
\arguments{
\item{x}{= a matrix/data-frame with 3 columns corresponding respectively with the rows, columns and entries of the resulting matrix}
}
\description{
converts <x> to a matrix
}
\keyword{mat.to.matrix}
