% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.to.xlModel}
\alias{mat.to.xlModel}
\title{mat.to.xlModel}
\usage{
mat.to.xlModel(x, y = 2, n = 5, w = F)
}
\arguments{
\item{x}{= a data frame indexed by data dates or trade open dates}

\item{y}{= number of days needed for flow data to be known}

\item{n}{= return horizon in weekdays}

\item{w}{= T/F depending on whether the index is data or trade-open date}
}
\description{
prepends the trade open and close dates and re-indexes by data date (as needed)
}
\keyword{mat.to.xlModel}
