% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.weekly.to.daily}
\alias{mat.weekly.to.daily}
\title{mat.weekly.to.daily}
\usage{
mat.weekly.to.daily(x)
}
\arguments{
\item{x}{= a matrix/df of weekly data}
}
\description{
daily file having latest weekly data known by each flow date
}
\keyword{mat.weekly.to.daily}
