% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.write}
\alias{mat.write}
\title{mat.write}
\usage{
mat.write(x, y = "C:\\\\temp\\\\write.csv", n = ",")
}
\arguments{
\item{x}{= any matrix/df}

\item{y}{= file intended to receive the output}

\item{n}{= the separator}
}
\description{
Writes <x> as a <n>-separated file to <y>
}
\keyword{mat.write}
