% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.zScore}
\alias{mat.zScore}
\title{mat.zScore}
\usage{
mat.zScore(x, y, n)
}
\arguments{
\item{x}{= a vector/matrix/data-frame}

\item{y}{= a 1/0 membership vector}

\item{n}{= a vector of groups (e.g. GSec)}
}
\description{
zScores <x> within groups <n> using weights <y>
}
\keyword{mat.zScore}
