% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{maturity.bucket}
\alias{maturity.bucket}
\title{maturity.bucket}
\usage{
maturity.bucket(x)
}
\arguments{
\item{x}{= named numeric vector}
}
\description{
where clauses for SQL case statement
}
\keyword{maturity.bucket}
