% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.1dFloMo.CtrySG}
\alias{mk.1dFloMo.CtrySG}
\title{mk.1dFloMo.CtrySG}
\usage{
mk.1dFloMo.CtrySG(x, y, n, w, h, u = "E", v = F)
}
\arguments{
\item{x}{= starting flowdate/YYYYMMDD depending on whether daily/weekly}

\item{y}{= item (Flow/AssetsStart/AssetsEnd/PortfolioChange)}

\item{n}{= country list (one of Ctry/FX)}

\item{w}{= input to or output of sql.connect}

\item{h}{= T/F depending on whether daily/weekly}

\item{u}{= vector of filters}

\item{v}{= T/F to use institutional or all share classes}
}
\description{
SQL query for daily/weekly regional flow momentum
}
\keyword{mk.1dFloMo.CtrySG}
