% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.1dFloMo.FI}
\alias{mk.1dFloMo.FI}
\title{mk.1dFloMo.FI}
\usage{
mk.1dFloMo.FI(x, y, n, w, h = "All", u = F)
}
\arguments{
\item{x}{= starting flowdate/YYYYMMDD depending on whether daily/weekly}

\item{y}{= item (Flow/AssetsStart/AssetsEnd/PortfolioChange)}

\item{n}{= input to or output of sql.connect}

\item{w}{= T/F depending on whether daily/weekly}

\item{h}{= vector of filters}

\item{u}{= T/F to use institutional or all share classes}
}
\description{
SQL query for daily/weekly regional flow momentum
}
\keyword{mk.1dFloMo.FI}
