% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.1dFloMo}
\alias{mk.1dFloMo}
\title{mk.1dFloMo}
\usage{
mk.1dFloMo(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMMDD}

\item{y}{= a string vector of variables to build with the last element specifying the type of funds to use (All/Act/Num, defaults to "All")}

\item{n}{= list object containing the following items: a) classif - classif file b) conn - a connection, the output of odbcDriverConnect c) DB - any of StockFlows/Japan/CSI300/Energy}
}
\description{
Returns a flow variable with the same row space as <n>
}
\seealso{
Other mk: \code{\link{mk.1mAllocMo}},
  \code{\link{mk.1mPerfTrend}}, \code{\link{mk.ActWt}},
  \code{\link{mk.Alpha.daily}}, \code{\link{mk.Alpha}},
  \code{\link{mk.FloAlphaLt.Ctry}},
  \code{\link{mk.FundsMem}},
  \code{\link{mk.JensensAlpha.fund}}, \code{\link{mk.Mem}},
  \code{\link{mk.Wt}}, \code{\link{mk.avail}},
  \code{\link{mk.beta}}, \code{\link{mk.isin}},
  \code{\link{mk.sqlDump}},
  \code{\link{mk.trail.sum.dly.vbl}},
  \code{\link{mk.trail.sum.mo.vbl}},
  \code{\link{mk.vbl.chg}}, \code{\link{mk.vbl.diff}},
  \code{\link{mk.vbl.lag}}, \code{\link{mk.vbl.max}},
  \code{\link{mk.vbl.ratio}}
}
\keyword{mk.1dFloMo}
