% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.1mFloMo.Ctry}
\alias{mk.1mFloMo.Ctry}
\title{mk.1mFloMo.Ctry}
\usage{
mk.1mFloMo.Ctry(x, y, n, w, h = "E")
}
\arguments{
\item{x}{= a YYYYMM month}

\item{y}{= item (one of Flow/AssetsStart/AssetsEnd)}

\item{n}{= country list (one of Ctry/FX/Sector)}

\item{w}{= input to or output of sql.connect}

\item{h}{= vector of filters}
}
\description{
SQL query for monthly CBE flow momentum
}
\keyword{mk.1mFloMo.Ctry}
