% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.1mPerfTrend}
\alias{mk.1mPerfTrend}
\title{mk.1mPerfTrend}
\usage{
mk.1mPerfTrend(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMM}

\item{y}{= variable to build}

\item{n}{= list object containing the following items: a) classif - classif file b) conn - a connection, the output of odbcDriverConnect c) uiconn - a connection to EPFRUI, the output of odbcDriverConnect d) DB - any of StockFlows/Japan/CSI300/Energy}
}
\description{
Returns a variable with the same row space as <n>
}
\keyword{mk.1mPerfTrend}
