% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.FloAlphaLt.Ctry}
\alias{mk.FloAlphaLt.Ctry}
\title{mk.FloAlphaLt.Ctry}
\usage{
mk.FloAlphaLt.Ctry(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMM}

\item{y}{= an object name (preceded by #) or the path to a ".csv" file}

\item{n}{= list object containing the following items: a) classif - classif file}
}
\description{
Monthly Country Flow Alpha
}
\keyword{mk.FloAlphaLt.Ctry}
