% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.FundsMem}
\alias{mk.FundsMem}
\title{mk.FundsMem}
\usage{
mk.FundsMem(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMM}

\item{y}{= a string vector, the elements of which are: 1) column to match in classif (e.g. "FundType") 2) column value (e.g. "E" or "B")}

\item{n}{= list object containing the following items: a) classif - classif file b) fldr - stock-flows folder}
}
\description{
Returns a 1/0 vector with the same row space as <n> that is 1 whenever it has the right fund type as well as one-month forward return.
}
\keyword{mk.FundsMem}
