% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.HoldValTot}
\alias{mk.HoldValTot}
\title{mk.HoldValTot}
\usage{
mk.HoldValTot(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMM}

\item{y}{= one of All/Act/Pas/Etf/Mutual/JP/xJP/CBE}

\item{n}{= list object containing the following items: a) classif - classif file b) conn - a connection, the output of odbcDriverConnect}
}
\description{
Total Holding Value ($MM)
}
\seealso{
Other mk: \code{\link{mk.1dFloMo}},
  \code{\link{mk.1mAllocMo}}, \code{\link{mk.1mPerfTrend}},
  \code{\link{mk.ActWt}}, \code{\link{mk.Alpha.daily}},
  \code{\link{mk.Alpha}}, \code{\link{mk.FloAlphaLt.Ctry}},
  \code{\link{mk.Fragility}}, \code{\link{mk.FundsMem}},
  \code{\link{mk.HerdingLSV}},
  \code{\link{mk.JensensAlpha.fund}}, \code{\link{mk.Mem}},
  \code{\link{mk.SRIMem}}, \code{\link{mk.SatoMem}},
  \code{\link{mk.Wt}}, \code{\link{mk.avail}},
  \code{\link{mk.beta}}, \code{\link{mk.isin}},
  \code{\link{mk.sqlDump}}, \code{\link{mk.vbl.chg}},
  \code{\link{mk.vbl.diff}}, \code{\link{mk.vbl.lag}},
  \code{\link{mk.vbl.max}}, \code{\link{mk.vbl.ratio}},
  \code{\link{mk.vbl.scale}}, \code{\link{mk.vbl.sum}},
  \code{\link{mk.vbl.trail.fetch}},
  \code{\link{mk.vbl.trail.sum}}, \code{\link{mk.vbl.vol}}
}
\keyword{mk.HoldValTot}
