% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.JensensAlpha.fund}
\alias{mk.JensensAlpha.fund}
\title{mk.JensensAlpha.fund}
\usage{
mk.JensensAlpha.fund(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMM}

\item{y}{= number of months of trailing returns to use}

\item{n}{= list object containing the following items: a) classif - classif file b) fldr - stock-flows folder c) CATRETS - category returns}
}
\description{
Returns variable with the same row space as <n>
}
\keyword{mk.JensensAlpha.fund}
