% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.PrcMo}
\alias{mk.PrcMo}
\title{mk.PrcMo}
\usage{
mk.PrcMo(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMM}

\item{y}{= 1m variable to fetch (e.g. "Ret")}

\item{n}{= list object containing the following items: a) classif - classif file b) fldr - stock-flows folder}
}
\description{
makes 5 month lagged one month PrcMo
}
\keyword{mk.PrcMo}
