% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.trail.sum.dly.vbl}
\alias{mk.trail.sum.dly.vbl}
\title{mk.trail.sum.dly.vbl}
\usage{
mk.trail.sum.dly.vbl(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMM}

\item{y}{= a vector of string, the elements of which are: 1) 1d variable to fetch (e.g. "1dFloMo") 2)  T to sum or F to compound (e.g. "F") 3) number of weekdays to compound over (e.g. "65") 4) number of days to lag (defaults to "0") 5) sub-folder to fetch basic variable from (defaults to "derived")}

\item{n}{= list object containing the following items: a) classif - classif file b) fldr - stock-flows folder}
}
\description{
Compounds/sums over some trailing period
}
\keyword{mk.trail.sum.dly.vbl}
