% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.vbl.trail.sum.dly}
\alias{mk.vbl.trail.sum.dly}
\title{mk.vbl.trail.sum.dly}
\usage{
mk.vbl.trail.sum.dly(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMM}

\item{y}{= a vector of string, the elements of which are: 1) 1d variable to fetch (e.g. "1dFloMo") 2)  T to sum or F to compound (e.g. "F") 3) number of weekdays to compound over (e.g. "65") 4) number of days to lag (defaults to "0") 5) sub-folder to fetch basic variable from (defaults to "derived")}

\item{n}{= list object containing the following items: a) classif - classif file b) fldr - stock-flows folder}
}
\description{
Compounds/sums over some trailing period
}
\seealso{
Other mk: \code{\link{mk.1dFloMo}},
  \code{\link{mk.1mAllocMo}}, \code{\link{mk.1mPerfTrend}},
  \code{\link{mk.ActWt}}, \code{\link{mk.Alpha.daily}},
  \code{\link{mk.Alpha}}, \code{\link{mk.FloAlphaLt.Ctry}},
  \code{\link{mk.FundsMem}},
  \code{\link{mk.JensensAlpha.fund}}, \code{\link{mk.Mem}},
  \code{\link{mk.Wt}}, \code{\link{mk.avail}},
  \code{\link{mk.beta}}, \code{\link{mk.isin}},
  \code{\link{mk.sqlDump}}, \code{\link{mk.vbl.chg}},
  \code{\link{mk.vbl.diff}}, \code{\link{mk.vbl.lag}},
  \code{\link{mk.vbl.max}}, \code{\link{mk.vbl.ratio}},
  \code{\link{mk.vbl.trail.fetch}},
  \code{\link{mk.vbl.trail.sum.mo}},
  \code{\link{mk.vbl.vol}}
}
\keyword{mk.vbl.trail.sum.dly}
