% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{multi.asset}
\alias{multi.asset}
\title{multi.asset}
\usage{
multi.asset(x)
}
\arguments{
\item{x}{= a vector of paths to files}
}
\description{
Reads in data relevant to the multi-asset strategy
}
\keyword{multi.asset}
