% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{nameTo}
\alias{nameTo}
\title{nameTo}
\usage{
nameTo(x, y)
}
\arguments{
\item{x}{= a logical vector/matrix/dataframe without NA's}

\item{y}{= a logical value, isomekic vector or isomekic isoplatic matrix/df without NA's}
}
\description{
pct name turnover between <x> and <y> if <x> is a vector or their rows otherwise
}
\keyword{nameTo}
