% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{nameTo.underlying}
\alias{nameTo.underlying}
\title{nameTo.underlying}
\usage{
nameTo.underlying(x, y, n)
}
\arguments{
\item{x}{= an integer vector (counts over the current period)}

\item{y}{= an integer vector (counts over the prior period)}

\item{n}{= an integer vector (count of common names)}
}
\description{
percent name turnover
}
\keyword{nameTo.underlying}
