% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{num.exists}
\alias{num.exists}
\title{num.exists}
\usage{
num.exists(x, y)
}
\arguments{
\item{x}{= a string vector}

\item{y}{= number type}
}
\description{
T/F depending on whether <x> is a number of type <y>
}
\keyword{num.exists}
