% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{nyse.holidays}
\alias{nyse.holidays}
\title{nyse.holidays}
\usage{
nyse.holidays(x = "yyyymmdd")
}
\arguments{
\item{x}{= either "yyyymmdd" or "reason"}
}
\description{
returns full day NYSE holidays from the year 2000 and after
}
\seealso{
Other nyse: \code{\link{nyse.lag}}
}
\keyword{nyse.holidays}
