% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{nyse.lag}
\alias{nyse.lag}
\title{nyse.lag}
\usage{
nyse.lag(x)
}
\arguments{
\item{x}{= a flowdate}
}
\description{
falls back one non-NYSE holidays
}
\seealso{
Other nyse: \code{\link{nyse.holidays}}
}
\keyword{nyse.lag}
