% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{obj.diff}
\alias{obj.diff}
\title{obj.diff}
\usage{
obj.diff(fcn, x, y)
}
\arguments{
\item{fcn}{= a function (vector to integer vector)}

\item{x}{= an object vector}

\item{y}{= an object vector}
}
\description{
returns <x - y>
}
\seealso{
Other obj: \code{\link{obj.lag}}, \code{\link{obj.seq}}
}
\keyword{obj.diff}
