% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{obj.list}
\alias{obj.list}
\title{obj.list}
\usage{
obj.list(x = F, y = "*")
}
\arguments{
\item{x}{= T/F variable depending on whether you want functions or objects}

\item{y}{= pattern you want to see in returned objects}
}
\description{
Returns the names of objects that are or are not functions
}
\keyword{obj.list}
