% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{parameters.ex.file}
\alias{parameters.ex.file}
\title{parameters.ex.file}
\usage{
parameters.ex.file(x, y)
}
\arguments{
\item{x}{= a folder vector}

\item{y}{= a string (file names)}
}
\description{
path to function source file
}
\keyword{parameters.ex.file}
