% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{permutations.buckets.many}
\alias{permutations.buckets.many}
\title{permutations.buckets.many}
\usage{
permutations.buckets.many(x)
}
\arguments{
\item{x}{= a list object of string vectors without NA's}
}
\description{
all possible choices of one element from each vector
}
\seealso{
Other permutations: \code{\link{permutations.next}},
  \code{\link{permutations}}
}
\keyword{permutations.buckets.many}
