% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{permutations.next}
\alias{permutations.next}
\title{permutations.next}
\usage{
permutations.next(x)
}
\arguments{
\item{x}{= an integer vector (seq_along(<x>) in some order)}
}
\description{
returns the next permutation in dictionary order
}
\keyword{permutations.next}
