% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{pivot.1d}
\alias{pivot.1d}
\title{pivot.1d}
\usage{
pivot.1d(fcn, x, y)
}
\arguments{
\item{fcn}{= a summary function}

\item{x}{= a grouping vector}

\item{y}{= a numeric vector/matrix/data-frame}
}
\description{
returns a table, having the same column space of <x>, the rows of which are unique members of <grp> The cells of the table are the summ.fcn of <x> whenever <grp> takes on its respective value
}
\keyword{pivot.1d}
