% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{pivot}
\alias{pivot}
\title{pivot}
\usage{
pivot(fcn, x, y, n)
}
\arguments{
\item{fcn}{= a summary function}

\item{x}{= a numeric vector}

\item{y}{= a grouping vector}

\item{n}{= a grouping vector}
}
\description{
returns a table, the rows and columns of which are unique members of rowIdx and colIdx The cells of the table are the <fcn> of <x> whenever <y> and <n> take on their respective values
}
\keyword{pivot}
