% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{plurality.map}
\alias{plurality.map}
\title{plurality.map}
\usage{
plurality.map(x, y)
}
\arguments{
\item{x}{= a vector}

\item{y}{= an isomekic vector}
}
\description{
returns a map from <x> to <y>
}
\keyword{plurality.map}
