% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{portfolio.beta}
\alias{portfolio.beta}
\title{portfolio.beta}
\usage{
portfolio.beta(x, y, n)
}
\arguments{
\item{x}{= a numeric vector/matrix/data-frame}

\item{y}{= an isomekic numeric vector}

\item{n}{= T/F depending on whether all observations required}
}
\description{
beta of <x> with respect to <y>
}
\seealso{
Other portfolio: \code{\link{portfolio.beta.wrapper}},
  \code{\link{portfolio.residual}}
}
\keyword{portfolio.beta}
