% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{portfolio.beta.wrapper}
\alias{portfolio.beta.wrapper}
\title{portfolio.beta.wrapper}
\usage{
portfolio.beta.wrapper(x, y, n)
}
\arguments{
\item{x}{= a matrix/data frame (total return indices)}

\item{y}{= a string (benchmark like "ACWorld")}

\item{n}{= an integer (lookback in days)}
}
\description{
<n> day beta of columns of <x> with respect to benchmark <y>
}
\seealso{
Other portfolio: \code{\link{portfolio.residual}}
}
\keyword{portfolio.beta.wrapper}
