% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{portfolio.residual}
\alias{portfolio.residual}
\title{portfolio.residual}
\usage{
portfolio.residual(x, y)
}
\arguments{
\item{x}{= a numeric vector}

\item{y}{= an isomekic numeric vector}
}
\description{
residual of <x> after factoring out <y>
}
\seealso{
Other portfolio: \code{\link{portfolio.beta.wrapper}},
  \code{\link{portfolio.beta}}
}
\keyword{portfolio.residual}
