% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{position.floPct}
\alias{position.floPct}
\title{position.floPct}
\usage{
position.floPct(x, y, n)
}
\arguments{
\item{x}{= a file (strategy)}

\item{y}{= subset}

\item{n}{= last publication date}
}
\description{
Latest four-week flow percentage
}
\keyword{position.floPct}
