% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{principal.components}
\alias{principal.components}
\title{principal.components}
\usage{
principal.components(x, y = 2)
}
\arguments{
\item{x}{= a matrix/df}

\item{y}{= number of principal components desired}
}
\description{
first <y> principal components
}
\seealso{
Other principal: \code{\link{principal.components.covar}}
}
\keyword{principal.components}
