% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{principal.components.underlying}
\alias{principal.components.underlying}
\title{principal.components.underlying}
\usage{
principal.components.underlying(x, y)
}
\arguments{
\item{x}{= a matrix/df}

\item{y}{= number of principal components desired}
}
\description{
first <y> principal components
}
\seealso{
Other principal: \code{\link{principal.components.covar}},
  \code{\link{principal.components}}
}
\keyword{principal.components.underlying}
