% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{proc.count}
\alias{proc.count}
\title{proc.count}
\usage{
proc.count(x = 10)
}
\arguments{
\item{x}{= number of records to return (0 = everything)}
}
\description{
returns top <x> processes by number running
}
\keyword{proc.count}
