% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{production.upload}
\alias{production.upload}
\title{production.upload}
\usage{
production.upload(x, y, n)
}
\arguments{
\item{x}{= a string (report name)}

\item{y}{= a file}

\item{n}{= date for which you sent the report}
}
\description{
uploads <y> to where <x> needs to be delivered
}
\seealso{
Other production: \code{\link{production.write}}
}
\keyword{production.upload}
