% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{production.upload.underlying}
\alias{production.upload.underlying}
\title{production.upload.underlying}
\usage{
production.upload.underlying(x, y, n, w)
}
\arguments{
\item{x}{= a file}

\item{y}{= a remote folder vector}

\item{n}{= a string (ftp/sftp)}

\item{w}{= a boolean (don't/do overwrite)}
}
\description{
uploads <y> to where <x> needs to be delivered
}
\seealso{
Other production: \code{\link{production.upload}},
  \code{\link{production.write}}
}
\keyword{production.upload.underlying}
