% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{production.write}
\alias{production.write}
\title{production.write}
\usage{
production.write(x, y)
}
\arguments{
\item{x}{= a matrix/data frame}

\item{y}{= a file (output)}
}
\description{
Writes production output if warranted
}
\seealso{
Other production: \code{\link{production.upload}}
}
\keyword{production.write}
