% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ptile}
\alias{ptile}
\title{ptile}
\usage{
ptile(x)
}
\arguments{
\item{x}{= a numeric vector/matrix/data frame}
}
\description{
Converts <x>, if a numeric vector, or the rows of <x> otherwise, to a ptile
}
\keyword{ptile}
