% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{publications.data.kill}
\alias{publications.data.kill}
\title{publications.data.kill}
\usage{
publications.data.kill(x, y, n)
}
\arguments{
\item{x}{= a flowdate vector}

\item{y}{= a folder}

\item{n}{= a string (regular expression)}
}
\description{
removes stale data
}
\seealso{
Other publications: \code{\link{publications.data}}
}
\keyword{publications.data.kill}
