% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{publications.data}
\alias{publications.data}
\title{publications.data}
\usage{
publications.data(x, y, n, w)
}
\arguments{
\item{x}{= a flowdate vector}

\item{y}{= a SQL query/function}

\item{n}{= a folder}

\item{w}{= a connection string}
}
\description{
additional data is got and stale data removed
}
\seealso{
Other publications: \code{\link{publications.data.kill}}
}
\keyword{publications.data}
