% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{publish.date}
\alias{publish.date}
\title{publish.date}
\usage{
publish.date(x)
}
\arguments{
\item{x}{= a vector of yyyymm months}
}
\description{
the date on which country/sector allocations are published
}
\keyword{publish.date}
