% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{publish.monthly.last}
\alias{publish.monthly.last}
\title{publish.monthly.last}
\usage{
publish.monthly.last(x, y = 23, n = 0)
}
\arguments{
\item{x}{= a YYYYMMDD}

\item{y}{= calendar day allocations are known the next month}

\item{n}{= additional monthly lag (defaults to zero)}
}
\description{
date of last monthly publication
}
\keyword{publish.monthly.last}
