% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{publish.weekly.last}
\alias{publish.weekly.last}
\title{publish.weekly.last}
\usage{
publish.weekly.last(x)
}
\arguments{
\item{x}{= a YYYYMMDD (can be missing)}
}
\description{
date of last weekly publication
}
\seealso{
Other publish: \code{\link{publish.daily.last}},
  \code{\link{publish.monthly.last}}
}
\keyword{publish.weekly.last}
