% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{qa.columns}
\alias{qa.columns}
\title{qa.columns}
\usage{
qa.columns(x)
}
\arguments{
\item{x}{= M/W/D depending on whether flows are monthly/weekly/daily}
}
\description{
columns expected in ftp file
}
\seealso{
Other qa: \code{\link{qa.filter.map}},
  \code{\link{qa.index}}, \code{\link{qa.mat.read}}
}
\keyword{qa.columns}
