% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{qa.filter.map}
\alias{qa.filter.map}
\title{qa.filter.map}
\usage{
qa.filter.map(x)
}
\arguments{
\item{x}{= a filter (e.g. Aggregate/Active/Passive/ETF/Mutual)}
}
\description{
maps to appropriate code on the R side
}
\seealso{
Other qa: \code{\link{qa.mat.read}}
}
\keyword{qa.filter.map}
