% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{qa.index}
\alias{qa.index}
\title{qa.index}
\usage{
qa.index(x, y, n)
}
\arguments{
\item{x}{= data frame}

\item{y}{= T/F depending on whether <x> pertains to a macro strategy}

\item{n}{= T/F depending on whether <x> pertains to a factor}
}
\description{
unique index for <x>
}
\seealso{
Other qa: \code{\link{qa.columns}},
  \code{\link{qa.filter.map}}, \code{\link{qa.mat.read}}
}
\keyword{qa.index}
