% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{qa.mat.read}
\alias{qa.mat.read}
\title{qa.mat.read}
\usage{
qa.mat.read(x, y, n, w, h, u, v)
}
\arguments{
\item{x}{= a remote file}

\item{y}{= local folder (e.g. "C:\\\\temp")}

\item{n}{= ftp site (can be missing)}

\item{w}{= user id (can be missing)}

\item{h}{= password (can be missing)}

\item{u}{= a string (ftp/sftp)}

\item{v}{= a boolean (ftp.use.epsv argument of getCurlHandle)}
}
\description{
contents of <x> as a data frame
}
\seealso{
Other qa: \code{\link{qa.filter.map}}
}
\keyword{qa.mat.read}
