% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{qa.secMenu}
\alias{qa.secMenu}
\title{qa.secMenu}
\usage{
qa.secMenu(x, y, n, w)
}
\arguments{
\item{x}{= a YYYYMM month}

\item{y}{= SecMenuM/SecMenuD}

\item{n}{= a connection, the output of odbcDriverConnect}

\item{w}{= stock filter (e.g. All/China/Japan)}
}
\description{
compares HSecurityId/ReportDate pairs in Security Menu versus Flow Dollar
}
\seealso{
Other qa: \code{\link{qa.columns}},
  \code{\link{qa.filter.map}}, \code{\link{qa.flow}},
  \code{\link{qa.index}}, \code{\link{qa.mat.read}}
}
\keyword{qa.secMenu}
