% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{qtl.single.grp}
\alias{qtl.single.grp}
\title{qtl.single.grp}
\usage{
qtl.single.grp(x, y)
}
\arguments{
\item{x}{= a data frame (2 columns, numeric, no NA's)}

\item{y}{= an integer (number of bins)}
}
\description{
an equal-weight binning so that the first column of <x> is divided into <y> equal bins. Weights determined by the 2nd column
}
\seealso{
Other qtl: \code{\link{qtl.eq}}, \code{\link{qtl.fast}},
  \code{\link{qtl}}
}
\keyword{qtl.single.grp}
