% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{qtl.underlying}
\alias{qtl.underlying}
\title{qtl.underlying}
\usage{
qtl.underlying(x, y, n)
}
\arguments{
\item{x}{= a vector with no NA's}

\item{y}{= an isomekic vector lacking NA's or zeroes}

\item{n}{= a positive integer}
}
\description{
divided <x> into <n> equal bins of roughly equal weight (as defined by <y>)
}
\keyword{qtl.underlying}
