% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{qtr.ex.int}
\alias{qtr.ex.int}
\title{qtr.ex.int}
\usage{
qtr.ex.int(x)
}
\arguments{
\item{x}{= a vector of integers}
}
\description{
returns a vector of <yyyymm> months
}
\seealso{
Other qtr: \code{\link{qtr.seq}}, \code{\link{qtr.to.int}}
}
\keyword{qtr.ex.int}
