% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{qtr.lag}
\alias{qtr.lag}
\title{qtr.lag}
\usage{
qtr.lag(x, y)
}
\arguments{
\item{x}{= a qtr vector}

\item{y}{= an integer vector}
}
\description{
lags <x> by <y> quarters
}
\seealso{
Other qtr: \code{\link{qtr.ex.int}}, \code{\link{qtr.seq}},
  \code{\link{qtr.to.int}}
}
\keyword{qtr.lag}
