% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{qtr.seq}
\alias{qtr.seq}
\title{qtr.seq}
\usage{
qtr.seq(x, y, n = 1)
}
\arguments{
\item{x}{= a QTR}

\item{y}{= a QTR}

\item{n}{= quantum size in QTR}
}
\description{
returns a sequence of QTR between (and including) x and y
}
\seealso{
Other qtr: \code{\link{qtr.ex.int}},
  \code{\link{qtr.to.int}}
}
\keyword{qtr.seq}
