% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{quant.info}
\alias{quant.info}
\title{quant.info}
\usage{
quant.info(x, y)
}
\arguments{
\item{x}{= unique identifier of the quant}

\item{y}{= a column in the classif-Quants file}
}
\description{
folder of function source file
}
\keyword{quant.info}
