% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{read.prcRet}
\alias{read.prcRet}
\title{read.prcRet}
\usage{
read.prcRet(x)
}
\arguments{
\item{x}{= a string (object name preceded by # or a file)}
}
\description{
returns the contents of the file
}
\seealso{
Other read: \code{\link{read.EPFR}}
}
\keyword{read.prcRet}
