% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{read.split.adj.prices}
\alias{read.split.adj.prices}
\title{read.split.adj.prices}
\usage{
read.split.adj.prices(x, y)
}
\arguments{
\item{x}{= full path to a file that has the following columns: a) PRC containing raw prices b) CFACPR containing split factor that you divide PRC by c) CUSIP containing eight-digit cusip d) date containing date in yyyymmdd format}

\item{y}{= classif file}
}
\description{
reads the split-adjusted prices that Matt provides
}
\seealso{
Other read: \code{\link{read.EPFR}},
  \code{\link{read.prcRet}}
}
\keyword{read.split.adj.prices}
