% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{recipient.exists}
\alias{recipient.exists}
\title{recipient.exists}
\usage{
recipient.exists(x)
}
\arguments{
\item{x}{= report name}
}
\description{
T/F depending on whether recipient list exists
}
\keyword{recipient.exists}
