% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{recipient.read}
\alias{recipient.read}
\title{recipient.read}
\usage{
recipient.read(x, y = F)
}
\arguments{
\item{x}{= a string (report name)}

\item{y}{= a boolean (SendGrid/regular)}
}
\description{
vector of recipient tranches
}
\seealso{
Other recipient: \code{\link{recipient.exists}}
}
\keyword{recipient.read}
