% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{record.exists}
\alias{record.exists}
\title{record.exists}
\usage{
record.exists(x, y, n)
}
\arguments{
\item{x}{= a string (report name)}

\item{y}{= date for which you want to send the report}

\item{n}{= a file}
}
\description{
T/F depending on whether action already taken
}
\seealso{
Other record: \code{\link{record.kill}},
  \code{\link{record.read}}, \code{\link{record.track}},
  \code{\link{record.write}}
}
\keyword{record.exists}
