% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{record.read}
\alias{record.read}
\title{record.read}
\usage{
record.read(x)
}
\arguments{
\item{x}{= file name}
}
\description{
named vector of records and sent dates
}
\seealso{
Other record: \code{\link{record.exists}},
  \code{\link{record.kill}}, \code{\link{record.write}}
}
\keyword{record.read}
