% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{record.track.target}
\alias{record.track.target}
\title{record.track.target}
\usage{
record.track.target(x, y, n)
}
\arguments{
\item{x}{= a flowdate}

\item{y}{= a data frame}

\item{n}{= a boolean (regular/Asia process)}
}
\description{
target/today for date <x> and data frame <y>
}
\seealso{
Other record: \code{\link{record.exists}},
  \code{\link{record.kill}}, \code{\link{record.read}},
  \code{\link{record.track}}, \code{\link{record.write}}
}
\keyword{record.track.target}
