% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{refresh.predictors.append}
\alias{refresh.predictors.append}
\title{refresh.predictors.append}
\usage{
refresh.predictors.append(x, y)
}
\arguments{
\item{x}{= a matrix/data frame}

\item{y}{= a data frame}
}
\description{
Appends new to old data after performing checks
}
\seealso{
Other refresh: \code{\link{refresh.predictors.daily}},
  \code{\link{refresh.predictors.script}},
  \code{\link{refresh.predictors.troika}},
  \code{\link{refresh.predictors}}
}
\keyword{refresh.predictors.append}
