% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{refresh.predictors.monthly}
\alias{refresh.predictors.monthly}
\title{refresh.predictors.monthly}
\usage{
refresh.predictors.monthly(x, y, n, w, h)
}
\arguments{
\item{x}{= csv file containing the predictors}

\item{y}{= query needed to get full history}

\item{n}{= last part of the query that goes after the date restriction}

\item{w}{= input to sql.connect}

\item{h}{= when T, ignores the fact that data for the last row has changed}
}
\description{
refreshes the text file contains flows data from SQL
}
\keyword{refresh.predictors.monthly}
