% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{refresh.predictors}
\alias{refresh.predictors}
\title{refresh.predictors}
\usage{
refresh.predictors(fcn, x, y, n)
}
\arguments{
\item{fcn}{= a function to generate new data}

\item{x}{= a file (predictors)}

\item{y}{= a boolean (ignore/note data changes)}

\item{n}{= last complete publication period}
}
\description{
refreshes the text file contains flows data from SQL
}
\seealso{
Other refresh: \code{\link{refresh.predictors.append}},
  \code{\link{refresh.predictors.daily}},
  \code{\link{refresh.predictors.monthly}},
  \code{\link{refresh.predictors.script}},
  \code{\link{refresh.predictors.weekly}}
}
\keyword{refresh.predictors}
