% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{refresh.predictors.script}
\alias{refresh.predictors.script}
\title{refresh.predictors.script}
\usage{
refresh.predictors.script(x, y, n, w)
}
\arguments{
\item{x}{= query needed to get full history}

\item{y}{= last part of the query that goes after the date restriction}

\item{n}{= column corresponding to date in relevant sql table}

\item{w}{= last date for which you already have data}
}
\description{
generates the SQL script to refresh predictors
}
\keyword{refresh.predictors.script}
