% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{refresh.predictors.troika}
\alias{refresh.predictors.troika}
\title{refresh.predictors.troika}
\usage{
refresh.predictors.troika(fcn, x, y, n = list())
}
\arguments{
\item{fcn}{= a function to generate new data}

\item{x}{= a file vector (Flow/AssetsStart/Result)}

\item{y}{= last complete publication period}

\item{n}{= auxiliary arguments to <fcn>}
}
\description{
refreshes the text file contains flows data from SQL
}
\seealso{
Other refresh: \code{\link{refresh.predictors.append}},
  \code{\link{refresh.predictors.daily}},
  \code{\link{refresh.predictors.monthly}},
  \code{\link{refresh.predictors.script}},
  \code{\link{refresh.predictors.weekly}},
  \code{\link{refresh.predictors}}
}
\keyword{refresh.predictors.troika}
