% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{renorm}
\alias{renorm}
\title{renorm}
\usage{
renorm(x)
}
\arguments{
\item{x}{= a numeric vector}
}
\description{
renormalizes, so the absolute weights sum to 100, <x>, if a vector, or the rows of <x> otherwise
}
\keyword{renorm}
