% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{report.flow.annual.cumulative}
\alias{report.flow.annual.cumulative}
\title{report.flow.annual.cumulative}
\usage{
report.flow.annual.cumulative(x)
}
\arguments{
\item{x}{= a numeric vector indexed by YYYYMMDD}
}
\description{
summarizes flows for one year
}
\seealso{
Other report: \code{\link{report.flow.annual}},
  \code{\link{report.flow}}
}
\keyword{report.flow.annual.cumulative}
