% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{report.flow.english}
\alias{report.flow.english}
\title{report.flow.english}
\usage{
report.flow.english(x, y, n, w)
}
\arguments{
\item{x}{= a named vector of integers (numbers need to be rounded)}

\item{y}{= a named text vector}

\item{n}{= line number at which to insert a statement}

\item{w}{= statement to be inserted}
}
\description{
writes a flow report in English
}
\seealso{
Other report: \code{\link{report.flow.underlying}}
}
\keyword{report.flow.english}
