% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{report.flow}
\alias{report.flow}
\title{report.flow}
\usage{
report.flow(x, y, n, w, h, u = T)
}
\arguments{
\item{x}{= a numeric vector indexed by YYYYMMDD}

\item{y}{= character string representing name of the asset class}

\item{n}{= numeric value representing AUM}

\item{w}{= line number at which to insert a statement}

\item{h}{= statement to be inserted}

\item{u}{= T/F depending on whether you want to use html}
}
\description{
writes a flow report
}
\seealso{
Other report: \code{\link{report.flow.annual.cumulative}},
  \code{\link{report.flow.annual}}
}
\keyword{report.flow}
