% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{report.flow.underlying}
\alias{report.flow.underlying}
\title{report.flow.underlying}
\usage{
report.flow.underlying(x)
}
\arguments{
\item{x}{= a numeric vector indexed by YYYYMMDD}
}
\description{
list object containing the following items: a) text - dates and text information about flows b) numbers - numeric summary of the flows
}
\seealso{
Other report: \code{\link{report.flow.english}}
}
\keyword{report.flow.underlying}
