% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{reshape.long}
\alias{reshape.long}
\title{reshape.long}
\usage{
reshape.long(x, y, n = "id", w = "val")
}
\arguments{
\item{x}{= a data frame}

\item{y}{= a string vector (columns to be compressed)}

\item{n}{= a string (name for identifier column)}

\item{w}{= a string (name for values column)}
}
\description{
<x> in long format
}
\seealso{
Other reshape: \code{\link{reshape.wide}}
}
\keyword{reshape.long}
