% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{reshape.wide}
\alias{reshape.wide}
\title{reshape.wide}
\usage{
reshape.wide(x)
}
\arguments{
\item{x}{= a matrix/data frame (last columns correspond to array entries)}
}
\description{
converts <x> to an array
}
\seealso{
Other reshape: \code{\link{reshape.long}}
}
\keyword{reshape.wide}
