% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ret.ex.idx}
\alias{ret.ex.idx}
\title{ret.ex.idx}
\usage{
ret.ex.idx(x, y, n, w)
}
\arguments{
\item{x}{= a file of total return indices indexed so that time runs forward}

\item{y}{= number of rows over which the return is computed}

\item{n}{= if T the result is labelled by the beginning of the period, else by the end.}

\item{w}{= T/F depending on whether returns or spread changes are needed}
}
\description{
computes return
}
\keyword{ret.ex.idx}
