% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ret.idx.gaps.fix}
\alias{ret.idx.gaps.fix}
\title{ret.idx.gaps.fix}
\usage{
ret.idx.gaps.fix(x)
}
\arguments{
\item{x}{= a matrix/data frame (indexed by YYYYMMDD)}
}
\description{
replaces NA's by latest available total return index (i.e. zero return over that period)
}
\seealso{
Other ret: \code{\link{ret.ex.idx}},
  \code{\link{ret.outliers}}, \code{\link{ret.to.idx}},
  \code{\link{ret.to.log}}
}
\keyword{ret.idx.gaps.fix}
