% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ret.to.idx}
\alias{ret.to.idx}
\title{ret.to.idx}
\usage{
ret.to.idx(x)
}
\arguments{
\item{x}{= a file of total returns indexed so that time runs forward}
}
\description{
computes a total-return index
}
\keyword{ret.to.idx}
