% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ret.to.log}
\alias{ret.to.log}
\title{ret.to.log}
\usage{
ret.to.log(x)
}
\arguments{
\item{x}{= a numeric vector (returns)}
}
\description{
converts to logarithmic return
}
\seealso{
Other ret: \code{\link{ret.ex.idx}},
  \code{\link{ret.idx.gaps.fix}},
  \code{\link{ret.outliers}}, \code{\link{ret.to.idx}}
}
\keyword{ret.to.log}
