% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{rgb.diff}
\alias{rgb.diff}
\title{rgb.diff}
\usage{
rgb.diff(x, y)
}
\arguments{
\item{x}{= an integer vector (length three, elements between 0 and 256)}

\item{y}{= an integer vector (length three, elements between 0 and 256)}
}
\description{
distance between RGB colours <x> and <y>
}
\keyword{rgb.diff}
