% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{rpt.email}
\alias{rpt.email}
\title{rpt.email}
\usage{
rpt.email(x, y, n, w)
}
\arguments{
\item{x}{= report name}

\item{y}{= output type (".csv", ".pdf", etc.)}

\item{n}{= T/F depending on whether you're checking latest}

\item{w}{= T/F depending on whether you're sending the email or testing}
}
\description{
emails report
}
\keyword{rpt.email}
