% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{rpt.email.underlying}
\alias{rpt.email.underlying}
\title{rpt.email.underlying}
\usage{
rpt.email.underlying(x, y, n, w, h, u)
}
\arguments{
\item{x}{= report name}

\item{y}{= output type (".csv", ".pdf", etc.)}

\item{n}{= T/F depending on whether you're checking latest}

\item{w}{= T/F depending on whether you're sending the email or testing}

\item{h}{= the email address(es) of the recipient(s)}

\item{u}{= path to log file}
}
\description{
emails report
}
\seealso{
Other rpt: \code{\link{rpt.email}}
}
\keyword{rpt.email.underlying}
