% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{rquaternion}
\alias{rquaternion}
\title{rquaternion}
\usage{
rquaternion(x)
}
\arguments{
\item{x}{= number of quaternions desired}
}
\description{
n x 4 matrix of randomly generated number of unit size
}
\keyword{rquaternion}
