% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{rrw.factors}
\alias{rrw.factors}
\title{rrw.factors}
\usage{
rrw.factors(x)
}
\arguments{
\item{x}{= a data frame (returns in the first column)}
}
\description{
Returns the t-values of factors that best predict return
}
\seealso{
Other rrw: \code{\link{rrw.underlying}}, \code{\link{rrw}}
}
\keyword{rrw.factors}
