% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{rrw}
\alias{rrw}
\title{rrw}
\usage{
rrw(x, y, n, w, h, u, v, g = NULL, r)
}
\arguments{
\item{x}{= a YYYYMM (first-return date, backtest start)}

\item{y}{= a YYYYMM (first-return date, backtest end)}

\item{n}{= a variable vector (to regress return against)}

\item{w}{= a universe (e.g. "R1Mem", or c("EafeMem", 1, "CountryCode", "JP"))}

\item{h}{= a variable (neutrality group)}

\item{u}{= a variable (return)}

\item{v}{= a folder (must have sub-folders data/derived)}

\item{g}{= a variable (to orthogonalize against)}

\item{r}{= classif file}
}
\description{
regression results
}
\seealso{
Other rrw: \code{\link{rrw.factors}},
  \code{\link{rrw.underlying}}
}
\keyword{rrw}
