% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{run.cs.reg}
\alias{run.cs.reg}
\title{run.cs.reg}
\usage{
run.cs.reg(x, y)
}
\arguments{
\item{x}{= a matrix of n columns (usually stocks go down and returns go across)}

\item{y}{= a matrix of n rows (whatever vectors you're regressing on)}
}
\description{
regresses each row of <x> on design matrix <y>
}
\keyword{run.cs.reg}
