% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{separating.hyperplane}
\alias{separating.hyperplane}
\title{separating.hyperplane}
\usage{
separating.hyperplane(x, y)
}
\arguments{
\item{x}{= a matrix, the first column being a 1/0 vector}

\item{y}{= a unit vector of length dim(x)[2] - 1}
}
\description{
number of errors and distance from origin for best separating hyperlane
}
\keyword{separating.hyperplane}
