% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf.daily}
\alias{sf.daily}
\title{sf.daily}
\usage{
sf.daily(prdBeg, prdEnd, vbl.nm, univ, grp.nm, ret.nm, trails, sum.flows, fldr,
  vbl.lag, dly.vbl = T, classif)
}
\arguments{
\item{prdBeg}{= first-return date in YYYYMM}

\item{prdEnd}{= first-return date in YYYYMM after <prdBeg>}

\item{vbl.nm}{= variable}

\item{univ}{= membership (e.g. "EafeMem" or c("GemMem", 1))}

\item{grp.nm}{= group within which binning is to be performed}

\item{ret.nm}{= return variable}

\item{trails}{= number of trailing periods to compound/sum over}

\item{sum.flows}{= T/F depending on whether you want flows summed or compounded.}

\item{fldr}{= data folder}

\item{vbl.lag}{= lags by <vbl.lag> weekdays or months depending on whether <dly.vbl> is true.}

\item{dly.vbl}{= if T then a daily predictor is assumed else a monthly one}

\item{classif}{= classif file}
}
\description{
runs stock-flows simulation
}
\keyword{sf.daily}
