% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf.daily.summ}
\alias{sf.daily.summ}
\title{sf.daily.summ}
\usage{
sf.daily.summ(x, y, n, w, h)
}
\arguments{
\item{x}{= bin returns by period}

\item{y}{= the number of days in the return window}

\item{n}{= the number of days the predictors are lagged}

\item{w}{= the number of days needed for the predictors to be known}

\item{h}{= if F grp.fcn is applied to formation dates. Otherwise it is applied to the first day in forward the return window.}
}
\description{
Summarizes bin excess returns by sub-periods of interest (as defined by <vec>)
}
\seealso{
Other sf: \code{\link{sf.bin.nms}}, \code{\link{sf.daily}},
  \code{\link{sf.single.bsim}}, \code{\link{sf.subset}},
  \code{\link{sf.summ}}, \code{\link{sf.underlying.data}},
  \code{\link{sf.underlying.summ}},
  \code{\link{sf.underlying}}, \code{\link{sf}}
}
\keyword{sf.daily.summ}
