% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf}
\alias{sf}
\title{sf}
\usage{
sf(fcn, prdBeg, prdEnd, univ, grp.nm, ret.nm, trails, fldr, nBins = 5,
  geom.comp = F, retHz = 1, classif)
}
\arguments{
\item{fcn}{= function that fetches data for the appropriate period and parameter}

\item{prdBeg}{= first-return date in YYYYMM}

\item{prdEnd}{= first-return date in YYYYMM after <prdBeg>}

\item{univ}{= membership (e.g. "EafeMem" or c("GemMem", 1))}

\item{grp.nm}{= group within which binning is to be performed}

\item{ret.nm}{= return variable}

\item{trails}{= variable parameter}

\item{fldr}{= data folder}

\item{nBins}{= number of bins}

\item{geom.comp}{= T/F depending on whether you want bin excess returns summarized geometrically or arithmetically}

\item{retHz}{= forward return horizon in months}

\item{classif}{= classif file}
}
\description{
runs a stock-flows simulation
}
\seealso{
Other sf: \code{\link{sf.bin.nms}},
  \code{\link{sf.detail}}, \code{\link{sf.single.bsim}},
  \code{\link{sf.subset}},
  \code{\link{sf.underlying.data}},
  \code{\link{sf.underlying.summ}}
}
\keyword{sf}
