% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf.single.bsim}
\alias{sf.single.bsim}
\title{sf.single.bsim}
\usage{
sf.single.bsim(fcn, prdBeg, prdEnd, univ, grp.nm, ret.nm, fldr, trail,
  uRet = F, nBins = 5, retHz = 1, classif, weighting.factor = NULL)
}
\arguments{
\item{fcn}{= function that fetches data for the appropriate period and parameter}

\item{prdBeg}{= first-return date in YYYYMM}

\item{prdEnd}{= first-return date in YYYYMM after <prdBeg>}

\item{univ}{= membership (e.g. "EafeMem" or c("GemMem", 1))}

\item{grp.nm}{= group within which binning is to be performed}

\item{ret.nm}{= return variable}

\item{fldr}{= data folder}

\item{trail}{= variable parameter}

\item{uRet}{= T/F depending on whether the equal-weight universe return is desired}

\item{nBins}{= number of bins or numeric vector with last element T/F for dependent/independent}

\item{retHz}{= forward return horizon in months}

\item{classif}{= classif file}

\item{weighting.factor}{= factor you want to use for Cap-weighted back-tests (defaults to NULL)}
}
\description{
runs a single quintile simulation
}
\seealso{
Other sf: \code{\link{sf.bin.nms}},
  \code{\link{sf.detail}}, \code{\link{sf.subset}},
  \code{\link{sf.underlying.data.bin}},
  \code{\link{sf.underlying.data}},
  \code{\link{sf.underlying.summ}},
  \code{\link{sf.vec.to.array}}, \code{\link{sf}}
}
\keyword{sf.single.bsim}
