% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf.summ}
\alias{sf.summ}
\title{sf.summ}
\usage{
sf.summ(fcn, x, y)
}
\arguments{
\item{fcn}{= a function that summarizes the data}

\item{x}{= a df of bin returns indexed by time}

\item{y}{= forward return horizon in months}
}
\description{
summarizes the back test
}
\keyword{sf.summ}
