% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf.underlying.data}
\alias{sf.underlying.data}
\title{sf.underlying.data}
\usage{
sf.underlying.data(fcn, x, y, n, w, h, u, v, g, r, s)
}
\arguments{
\item{fcn}{= a fetch function}

\item{x}{= a YYYYMM/flowdate (period for which you want returns)}

\item{y}{= a variable (factor for cap weighting)}

\item{n}{= a universe (e.g. "EafeMem" or c("GemMem", 1))}

\item{w}{= a variable (neutrality group)}

\item{h}{= a variable (return)}

\item{u}{= a string (parameter)}

\item{v}{= a folder (data)}

\item{g}{= an integer vector (if vector, last element T/F for dep/indep binning))}

\item{r}{= classif file}

\item{s}{= a positive integer (return window)}
}
\description{
Gets data needed to back-test a single period
}
\seealso{
Other sf: \code{\link{sf.bin.nms}}, \code{\link{sf.daily}},
  \code{\link{sf.detail}}, \code{\link{sf.single.bsim}},
  \code{\link{sf.subset}},
  \code{\link{sf.underlying.data.bin}},
  \code{\link{sf.underlying.summ}}, \code{\link{sf}}
}
\keyword{sf.underlying.data}
