% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf.vec.to.array}
\alias{sf.vec.to.array}
\title{sf.vec.to.array}
\usage{
sf.vec.to.array(x, y, n)
}
\arguments{
\item{x}{= named vector of characteristics}

\item{y}{= variable names}

\item{n}{= number of bins}
}
\description{
expresses <x> as an array
}
\seealso{
Other sf: \code{\link{sf.bin.nms}},
  \code{\link{sf.detail}}, \code{\link{sf.single.bsim}},
  \code{\link{sf.subset}},
  \code{\link{sf.underlying.data.bin}},
  \code{\link{sf.underlying.data}},
  \code{\link{sf.underlying.summ}}, \code{\link{sf}}
}
\keyword{sf.vec.to.array}
