% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sfpd.ActWtTrend}
\alias{sfpd.ActWtTrend}
\title{sfpd.ActWtTrend}
\usage{
sfpd.ActWtTrend(x, y, n, w, h)
}
\arguments{
\item{x}{= flows file}

\item{y}{= holdings file}

\item{n}{= fund history file}

\item{w}{= desired flow date to be reported}

\item{h}{= one of ActWtTrend/ActWtDiff/ActWtDiff2}
}
\description{
<h> factor
}
\seealso{
Other sfpd: \code{\link{sfpd.FloDollar}},
  \code{\link{sfpd.FloMo.underlying}},
  \code{\link{sfpd.FloMo}}, \code{\link{sfpd.FloTrend}},
  \code{\link{sfpd.Holdings}}, \code{\link{sfpd.ION}},
  \code{\link{sfpd.Wt}}, \code{\link{sfpd.filter}}
}
\keyword{sfpd.ActWtTrend}
