% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sfpd.AllocMo}
\alias{sfpd.AllocMo}
\title{sfpd.AllocMo}
\usage{
sfpd.AllocMo(x, y, n, w)
}
\arguments{
\item{x}{= a matrix/data frame (holdings)}

\item{y}{= a matrix/data frame (prior holdings)}

\item{n}{= a matrix/data frame (security history)}

\item{w}{= a variable (AllocMo/AllocTrend/AllocDiff)}
}
\description{
<w> factor
}
\seealso{
Other sfpd: \code{\link{sfpd.ActWtTrend}},
  \code{\link{sfpd.AllocSkew}},
  \code{\link{sfpd.FloDollar}},
  \code{\link{sfpd.FloMo.underlying}},
  \code{\link{sfpd.FloMo}},
  \code{\link{sfpd.FloTrend.holdings}},
  \code{\link{sfpd.FloTrend.parameterize}},
  \code{\link{sfpd.FloTrend.underlying}},
  \code{\link{sfpd.FloTrend}}, \code{\link{sfpd.Flow}},
  \code{\link{sfpd.Holdings}}, \code{\link{sfpd.ION}},
  \code{\link{sfpd.Wt}}, \code{\link{sfpd.filter}}
}
\keyword{sfpd.AllocMo}
