% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sfpd.filter}
\alias{sfpd.filter}
\title{sfpd.filter}
\usage{
sfpd.filter(x, y)
}
\arguments{
\item{x}{= fund history file}

\item{y}{= filter (one of All/Act/Pas/Etf/Mutual)}
}
\description{
T/F depending on whether records of <x> pass filter <y>
}
\keyword{sfpd.filter}
