% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sfpd.FloDollar}
\alias{sfpd.FloDollar}
\title{sfpd.FloDollar}
\usage{
sfpd.FloDollar(x, y, n, w)
}
\arguments{
\item{x}{= flows file}

\item{y}{= holdings file}

\item{n}{= fund history file}

\item{w}{= a flowdate}
}
\description{
FloMo factor
}
\keyword{sfpd.FloDollar}
