% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sfpd.FloMo}
\alias{sfpd.FloMo}
\title{sfpd.FloMo}
\usage{
sfpd.FloMo(x, y, n, w)
}
\arguments{
\item{x}{= flows file}

\item{y}{= holdings file}

\item{n}{= fund history file}

\item{w}{= desired flow date to be reported}
}
\description{
FloMo factor
}
\seealso{
Other sfpd: \code{\link{sfpd.ActWtTrend}},
  \code{\link{sfpd.FloDollar}},
  \code{\link{sfpd.FloMo.underlying}},
  \code{\link{sfpd.FloTrend}}, \code{\link{sfpd.Holdings}},
  \code{\link{sfpd.ION}}, \code{\link{sfpd.Wt}},
  \code{\link{sfpd.filter}}
}
\keyword{sfpd.FloMo}
