% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sfpd.FloMo.underlying}
\alias{sfpd.FloMo.underlying}
\title{sfpd.FloMo.underlying}
\usage{
sfpd.FloMo.underlying(x, y, n, w, h = NULL)
}
\arguments{
\item{x}{= flows file}

\item{y}{= holdings file}

\item{n}{= fund history file}

\item{w}{= items (e.g. c("Flow", "AssetsStart"))}

\item{h}{= optional items from fund history}
}
\description{
FloMo factor
}
\keyword{sfpd.FloMo.underlying}
