% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sfpd.FloTrend.holdings}
\alias{sfpd.FloTrend.holdings}
\title{sfpd.FloTrend.holdings}
\usage{
sfpd.FloTrend.holdings(x, y, n)
}
\arguments{
\item{x}{= a matrix/data frame (holdings)}

\item{y}{= a matrix/data frame (prior holdings)}

\item{n}{= a matrix/data frame (security history)}
}
\description{
merges old and new holdings
}
\seealso{
Other sfpd: \code{\link{sfpd.ActWtTrend}},
  \code{\link{sfpd.AllocMo}}, \code{\link{sfpd.AllocSkew}},
  \code{\link{sfpd.FloDollar}},
  \code{\link{sfpd.FloMo.underlying}},
  \code{\link{sfpd.FloMo}},
  \code{\link{sfpd.FloTrend.parameterize}},
  \code{\link{sfpd.FloTrend.underlying}},
  \code{\link{sfpd.FloTrend}}, \code{\link{sfpd.Flow}},
  \code{\link{sfpd.Holdings}}, \code{\link{sfpd.ION}},
  \code{\link{sfpd.Wt}}, \code{\link{sfpd.filter}}
}
\keyword{sfpd.FloTrend.holdings}
