% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sfpd.FloTrend.parameterize}
\alias{sfpd.FloTrend.parameterize}
\title{sfpd.FloTrend.parameterize}
\usage{
sfpd.FloTrend.parameterize(x, y)
}
\arguments{
\item{x}{= a matrix/data frame}

\item{y}{= a variable}
}
\description{
numerator and denominator to create variable
}
\seealso{
Other sfpd: \code{\link{sfpd.ActWtTrend}},
  \code{\link{sfpd.AllocMo}}, \code{\link{sfpd.AllocSkew}},
  \code{\link{sfpd.FloDollar}},
  \code{\link{sfpd.FloMo.underlying}},
  \code{\link{sfpd.FloMo}},
  \code{\link{sfpd.FloTrend.holdings}},
  \code{\link{sfpd.FloTrend.underlying}},
  \code{\link{sfpd.FloTrend}}, \code{\link{sfpd.Flow}},
  \code{\link{sfpd.Holdings}}, \code{\link{sfpd.ION}},
  \code{\link{sfpd.Wt}}, \code{\link{sfpd.filter}}
}
\keyword{sfpd.FloTrend.parameterize}
