% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sfpd.FloTrend}
\alias{sfpd.FloTrend}
\title{sfpd.FloTrend}
\usage{
sfpd.FloTrend(x, y, n, w, h, u, v)
}
\arguments{
\item{x}{= flows file}

\item{y}{= holdings file}

\item{n}{= fund history file}

\item{w}{= a flowdate}

\item{h}{= one of FloTrend/FloDiff/FloDiff2}

\item{u}{= holdings file from prior month}

\item{v}{= security history file}
}
\description{
<h> factor
}
\keyword{sfpd.FloTrend}
