% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sfpd.Flow}
\alias{sfpd.Flow}
\title{sfpd.Flow}
\usage{
sfpd.Flow(x, y, n, w = 1)
}
\arguments{
\item{x}{= premium daily file}

\item{y}{= a flowdate}

\item{n}{= cutoff time (New York) (e.g. "07:00:00")}

\item{w}{= 0 for same day, 1 for next day}
}
\description{
subsets <x> to latest information known before time <n>
}
\seealso{
Other sfpd: \code{\link{sfpd.ActWtTrend}},
  \code{\link{sfpd.FloDollar}},
  \code{\link{sfpd.FloMo.underlying}},
  \code{\link{sfpd.FloMo}},
  \code{\link{sfpd.FloTrend.underlying}},
  \code{\link{sfpd.FloTrend}}, \code{\link{sfpd.Holdings}},
  \code{\link{sfpd.ION}}, \code{\link{sfpd.Wt}},
  \code{\link{sfpd.filter}}
}
\keyword{sfpd.Flow}
