% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sfpd.Holdings}
\alias{sfpd.Holdings}
\title{sfpd.Holdings}
\usage{
sfpd.Holdings(x, y)
}
\arguments{
\item{x}{= a YYYYMM}

\item{y}{= input to or output of sql.connect}
}
\description{
Generates the SQL query to get weights for individual stocks
}
\keyword{sfpd.Holdings}
