% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sfpd.ION}
\alias{sfpd.ION}
\title{sfpd.ION}
\usage{
sfpd.ION(x, y, n, w)
}
\arguments{
\item{x}{= flows file}

\item{y}{= holdings file}

\item{n}{= fund history file}

\item{w}{= a flowdate}
}
\description{
Inflow/Outflow factors
}
\keyword{sfpd.ION}
