% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sfpd.Wt}
\alias{sfpd.Wt}
\title{sfpd.Wt}
\usage{
sfpd.Wt(x)
}
\arguments{
\item{x}{= holdings file}
}
\description{
computes weight
}
\seealso{
Other sfpd: \code{\link{sfpd.ActWtTrend}},
  \code{\link{sfpd.FloDollar}},
  \code{\link{sfpd.FloMo.underlying}},
  \code{\link{sfpd.FloMo}}, \code{\link{sfpd.FloTrend}},
  \code{\link{sfpd.Flow}}, \code{\link{sfpd.Holdings}},
  \code{\link{sfpd.ION}}, \code{\link{sfpd.filter}}
}
\keyword{sfpd.Wt}
