% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{shell.wrapper}
\alias{shell.wrapper}
\title{shell.wrapper}
\usage{
shell.wrapper(x, y)
}
\arguments{
\item{x}{= a string to issue as command}

\item{y}{= timeout in seconds}
}
\description{
result of command <x>
}
\keyword{shell.wrapper}
