% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sim.limits}
\alias{sim.limits}
\title{sim.limits}
\usage{
sim.limits(x, y)
}
\arguments{
\item{x}{= a data frame}

\item{y}{= an integer vector (group limits, names correspond to columns in <x>)}
}
\description{
returns group active-weight limits applying to each stock
}
\seealso{
Other sim: \code{\link{sim.direction.buy}},
  \code{\link{sim.direction.sell}},
  \code{\link{sim.direction}}, \code{\link{sim.fetch}},
  \code{\link{sim.optimal}}, \code{\link{sim.overall}},
  \code{\link{sim.seed}}, \code{\link{sim.summ}},
  \code{\link{sim.trade.grp}}, \code{\link{sim.trade.stk}}
}
\keyword{sim.limits}
