% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sim.optimal}
\alias{sim.optimal}
\title{sim.optimal}
\usage{
sim.optimal(x, y, n, w)
}
\arguments{
\item{x}{= a data frame with columns Alp/Bmk/Sec/Ctry}

\item{y}{= initial portfolio weight}

\item{n}{= percentage single-stock active-weight name limit}

\item{w}{= vector of group limits (names correspond to columns in <x>)}
}
\description{
returns named vector of optimal portfolio weights
}
\seealso{
Other sim: \code{\link{sim.direction}},
  \code{\link{sim.fetch}}, \code{\link{sim.limits}},
  \code{\link{sim.overall}}, \code{\link{sim.summ}},
  \code{\link{sim.trade.grp}}, \code{\link{sim.trade.stk}}
}
\keyword{sim.optimal}
