% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sim.overall}
\alias{sim.overall}
\title{sim.overall}
\usage{
sim.overall(x, y, n)
}
\arguments{
\item{x}{= list object, elements are the output of <sim.optimal>}

\item{y}{= named vector of turnover}

\item{n}{= vector of group limits (names correspond to columns in elements of <x>)}
}
\description{
summarizes simulation
}
\keyword{sim.overall}
