% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sim.summ}
\alias{sim.summ}
\title{sim.summ}
\usage{
sim.summ(x, y)
}
\arguments{
\item{x}{= data frame, the output of <sim.optimal>}

\item{y}{= vector of group limits (names correspond to columns of <x>)}
}
\description{
summarizes simulation
}
\keyword{sim.summ}
