% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sim.trade.grp}
\alias{sim.trade.grp}
\title{sim.trade.grp}
\usage{
sim.trade.grp(x, y, n)
}
\arguments{
\item{x}{= a data frame, the output of <sim.fetch>}

\item{y}{= T/F depending on whether you buy/sell}

\item{n}{= vector of group limits (names correspond to columns in <x>)}
}
\description{
max you can trade without breaching group limits
}
\keyword{sim.trade.grp}
