% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sim.trade.stk}
\alias{sim.trade.stk}
\title{sim.trade.stk}
\usage{
sim.trade.stk(x, y, n, w)
}
\arguments{
\item{x}{= a data frame}

\item{y}{= a boolean (buy/sell)}

\item{n}{= an integer (name active limit)}

\item{w}{= a boolean (do/don't fully invest)}
}
\description{
max you can trade without breaching name limits
}
\seealso{
Other sim: \code{\link{sim.direction.buy}},
  \code{\link{sim.direction.sell}},
  \code{\link{sim.direction}}, \code{\link{sim.fetch}},
  \code{\link{sim.limits}}, \code{\link{sim.optimal}},
  \code{\link{sim.overall}}, \code{\link{sim.seed}},
  \code{\link{sim.summ}}, \code{\link{sim.trade.grp}}
}
\keyword{sim.trade.stk}
