% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{smear.Q1}
\alias{smear.Q1}
\title{smear.Q1}
\usage{
smear.Q1(x)
}
\arguments{
\item{x}{= any real number}
}
\description{
Returns weights associated with ranks 1:x so that #		:	a) every position in the top quintile has an equal positive weight #		:	b) every position in the bottom 3 quintiles has an equal negative weight #		:	c) second quintile positions get a linear interpolation #		:	d) the weights sum to zero #		:	e) the positive weights sum to 100
}
\keyword{smear.Q1}
