% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dActWtTrend.Final}
\alias{sql.1dActWtTrend.Final}
\title{sql.1dActWtTrend.Final}
\usage{
sql.1dActWtTrend.Final(x, y, n)
}
\arguments{
\item{x}{= temp table name (#CTRY/#SEC)}

\item{y}{= factor (one of ActWtTrend/ActWtDiff/ActWtDiff2)}

\item{n}{= identifier column (SectorId/CountryId)}
}
\description{
SQL query for daily ActWtTrend
}
\keyword{sql.1dActWtTrend.Final}
