% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dActWtTrend.Flow}
\alias{sql.1dActWtTrend.Flow}
\title{sql.1dActWtTrend.Flow}
\usage{
sql.1dActWtTrend.Flow(x, y)
}
\arguments{
\item{x}{= flowdate}

\item{y}{= vector of filters}
}
\description{
SQL query for flows to compute ActWtTrend for Ctry/Sec
}
\keyword{sql.1dActWtTrend.Flow}
