% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dActWtTrend}
\alias{sql.1dActWtTrend}
\title{sql.1dActWtTrend}
\usage{
sql.1dActWtTrend(x, y, n)
}
\arguments{
\item{x}{= the YYYYMMDD for which you want flows (known one day later)}

\item{y}{= a string vector of factors to be computed, the last element of which is the type of fund used.}

\item{n}{= any of StockFlows/Japan/CSI300/Energy}
}
\description{
Generates the SQL query to get the data for 1dActWtTrend
}
\keyword{sql.1dActWtTrend}
