% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dActWtTrend.select}
\alias{sql.1dActWtTrend.select}
\title{sql.1dActWtTrend.select}
\usage{
sql.1dActWtTrend.select(x)
}
\arguments{
\item{x}{= desired factor}
}
\description{
select statement to compute <x>
}
\keyword{sql.1dActWtTrend.select}
