% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloMo.aum}
\alias{sql.1dFloMo.aum}
\title{sql.1dFloMo.aum}
\usage{
sql.1dFloMo.aum(x, y)
}
\arguments{
\item{x}{= month end in YYYYMMDD format}

\item{y}{= name of AssetEnd column (e.g. "PortVal")}
}
\description{
Underlying part of SQL query to get 1dFloMo for individual stocks
}
\keyword{sql.1dFloMo.aum}
