% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloMo.CountryId.List}
\alias{sql.1dFloMo.CountryId.List}
\title{sql.1dFloMo.CountryId.List}
\usage{
sql.1dFloMo.CountryId.List(x, y = "")
}
\arguments{
\item{x}{= one of Ctry/FX/Sector/EMDM/Aux}

\item{y}{= one of the following: (a) a YYYYMM date (b) a YYYYMMDD date}
}
\description{
map of security to CountryId
}
\keyword{sql.1dFloMo.CountryId.List}
