% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloMo.Ctry.Allocations}
\alias{sql.1dFloMo.Ctry.Allocations}
\title{sql.1dFloMo.Ctry.Allocations}
\usage{
sql.1dFloMo.Ctry.Allocations(x, y)
}
\arguments{
\item{x}{= a string vector indexed by allocation-table names}

\item{y}{= one of Ctry/FX/Sector}
}
\description{
Generates the SQL query to get daily 1dFloMo for countries
}
\keyword{sql.1dFloMo.Ctry.Allocations}
