% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloMo.Ctry.Allocations.term}
\alias{sql.1dFloMo.Ctry.Allocations.term}
\title{sql.1dFloMo.Ctry.Allocations.term}
\usage{
sql.1dFloMo.Ctry.Allocations.term(x, y)
}
\arguments{
\item{x}{= a string vector of allocation-table names}

\item{y}{= NA or one of EM/EAFE/ACWI}
}
\description{
total weight allocated to countries <x> in index <y>
}
\seealso{
Other sql: \code{\link{sql.1dActWtTrend.Ctry.underlying}},
  \code{\link{sql.1dActWtTrend.select}},
  \code{\link{sql.1dActWtTrend.topline.from}},
  \code{\link{sql.1dActWtTrend.topline}},
  \code{\link{sql.1dActWtTrend.underlying}},
  \code{\link{sql.1dActWtTrend}},
  \code{\link{sql.1dFloMo.CountryId.List}},
  \code{\link{sql.1dFloMo.Ctry.Allocations.GF.avg}},
  \code{\link{sql.1dFloMo.Ctry.Allocations}},
  \code{\link{sql.1dFloMo.Ctry.List}},
  \code{\link{sql.1dFloMo.CtryFlow}},
  \code{\link{sql.1dFloMo.Ctry}},
  \code{\link{sql.1dFloMo.FI.underlying}},
  \code{\link{sql.1dFloMo.FI}},
  \code{\link{sql.1dFloMo.Rgn}},
  \code{\link{sql.1dFloMo.SecFlow}},
  \code{\link{sql.1dFloMo.filter}},
  \code{\link{sql.1dFloMo.grp}},
  \code{\link{sql.1dFloMo.select.wrapper}},
  \code{\link{sql.1dFloMo.select}},
  \code{\link{sql.1dFloMo.underlying}},
  \code{\link{sql.1dFloMoAggr}}, \code{\link{sql.1dFloMo}},
  \code{\link{sql.1dFloTrend.Ctry.topline}},
  \code{\link{sql.1dFloTrend.Ctry.underlying}},
  \code{\link{sql.1dFloTrend.Ctry}},
  \code{\link{sql.1dFloTrend.select}},
  \code{\link{sql.1dFloTrend.underlying}},
  \code{\link{sql.1dFloTrend}}, \code{\link{sql.1dFundCt}},
  \code{\link{sql.1dFundRet}}, \code{\link{sql.1dION}},
  \code{\link{sql.1mActPas.Ctry}},
  \code{\link{sql.1mActWt.underlying}},
  \code{\link{sql.1mActWtTrend.underlying}},
  \code{\link{sql.1mActWtTrend}},
  \code{\link{sql.1mActWt}},
  \code{\link{sql.1mAllocD.select}},
  \code{\link{sql.1mAllocD}},
  \code{\link{sql.1mAllocMo.select}},
  \code{\link{sql.1mAllocMo.underlying.from}},
  \code{\link{sql.1mAllocMo.underlying.pre}},
  \code{\link{sql.1mAllocMo}},
  \code{\link{sql.1mAllocSkew.topline.from}},
  \code{\link{sql.1mAllocSkew}},
  \code{\link{sql.1mChActWt}}, \code{\link{sql.1mFloMo}},
  \code{\link{sql.1mFloTrend.underlying}},
  \code{\link{sql.1mFloTrend}}, \code{\link{sql.1mFundCt}},
  \code{\link{sql.1wFlow.Corp}},
  \code{\link{sql.ActWtDiff2}}, \code{\link{sql.AllocTbl}},
  \code{\link{sql.Bullish.Ctry}},
  \code{\link{sql.Bullish}}, \code{\link{sql.DailyFlo}},
  \code{\link{sql.Diff}}, \code{\link{sql.Dispersion}},
  \code{\link{sql.FloMo.Funds}},
  \code{\link{sql.FundHistory.macro}},
  \code{\link{sql.FundHistory.sf}},
  \code{\link{sql.FundHistory}}, \code{\link{sql.HSIdmap}},
  \code{\link{sql.HerdingLSV}},
  \code{\link{sql.Herfindahl}},
  \code{\link{sql.Holdings.bulk}},
  \code{\link{sql.Holdings}}, \code{\link{sql.ION}},
  \code{\link{sql.MonthlyAlloc}},
  \code{\link{sql.MonthlyAssetsEnd}}, \code{\link{sql.Mo}},
  \code{\link{sql.Overweight}}, \code{\link{sql.RDSuniv}},
  \code{\link{sql.ReportDate}}, \code{\link{sql.SRI}},
  \code{\link{sql.TopDownAllocs.items}},
  \code{\link{sql.TopDownAllocs.underlying}},
  \code{\link{sql.TopDownAllocs}}, \code{\link{sql.Trend}},
  \code{\link{sql.and}}, \code{\link{sql.arguments}},
  \code{\link{sql.bcp}}, \code{\link{sql.close}},
  \code{\link{sql.connect.wrapper}},
  \code{\link{sql.connect}},
  \code{\link{sql.cross.border}},
  \code{\link{sql.datediff}}, \code{\link{sql.declare}},
  \code{\link{sql.drop}}, \code{\link{sql.exists}},
  \code{\link{sql.floTbl.to.Col}}, \code{\link{sql.into}},
  \code{\link{sql.in}}, \code{\link{sql.isin.old.to.new}},
  \code{\link{sql.label}}, \code{\link{sql.map.classif}},
  \code{\link{sql.mat.cofactor}},
  \code{\link{sql.mat.crossprod.vector}},
  \code{\link{sql.mat.crossprod}},
  \code{\link{sql.mat.determinant}},
  \code{\link{sql.mat.flip}},
  \code{\link{sql.mat.multiply}}, \code{\link{sql.median}},
  \code{\link{sql.nonneg}},
  \code{\link{sql.query.underlying}},
  \code{\link{sql.query}}, \code{\link{sql.regr}},
  \code{\link{sql.tbl}}, \code{\link{sql.ui}},
  \code{\link{sql.unbracket}}, \code{\link{sql.yyyymmdd}},
  \code{\link{sql.yyyymm}}
}
\keyword{sql.1dFloMo.Ctry.Allocations.term}
