% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloMo.Ctry.List}
\alias{sql.1dFloMo.Ctry.List}
\title{sql.1dFloMo.Ctry.List}
\usage{
sql.1dFloMo.Ctry.List(x)
}
\arguments{
\item{x}{= One of Ctry/FX/Sector/UBS/EMDM}
}
\description{
Generates the SQL query to get daily 1dFloMo for countries
}
\keyword{sql.1dFloMo.Ctry.List}
