% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloMo.Ctry}
\alias{sql.1dFloMo.Ctry}
\title{sql.1dFloMo.Ctry}
\usage{
sql.1dFloMo.Ctry(x)
}
\arguments{
\item{x}{= Ctry/FX/Sector}
}
\description{
Generates the SQL query to get daily 1dFloMo for countries
}
\keyword{sql.1dFloMo.Ctry}
