% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloMo.CtrySG}
\alias{sql.1dFloMo.CtrySG}
\title{sql.1dFloMo.CtrySG}
\usage{
sql.1dFloMo.CtrySG(x, y, n, w, h, u)
}
\arguments{
\item{x}{= starting flowdate/YYYYMMDD depending on whether daily/weekly}

\item{y}{= item (Flow/AssetsStart/AssetsEnd/PortfolioChange)}

\item{n}{= named vector of group definitions}

\item{w}{= frequency (T/F for daily/weekly or D/W/M)}

\item{h}{= vector of filters}

\item{u}{= T/F to use institutional or all share classes}
}
\description{
SQL query for daily/weekly flow momentum by group
}
\keyword{sql.1dFloMo.CtrySG}
