% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloMo.FI}
\alias{sql.1dFloMo.FI}
\title{sql.1dFloMo.FI}
\usage{
sql.1dFloMo.FI(x = "Flow", y)
}
\arguments{
\item{x}{= flow-table column (defaults to "Flow")}

\item{y}{= date from which you want data (can be missing)}
}
\description{
Generates the SQL query to get daily 1dFloMo for fixed income
}
\keyword{sql.1dFloMo.FI}
