% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloMo.FI.underlying}
\alias{sql.1dFloMo.FI.underlying}
\title{sql.1dFloMo.FI.underlying}
\usage{
sql.1dFloMo.FI.underlying()
}
\description{
Generates the SQL query to get daily 1dFloMo for fixed income
}
\keyword{sql.1dFloMo.FI.underlying}
