% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloMo.filter}
\alias{sql.1dFloMo.filter}
\title{sql.1dFloMo.filter}
\usage{
sql.1dFloMo.filter(x, y)
}
\arguments{
\item{x}{= a string vector of factors to be computed, the last elements of which are the type of fund used}

\item{y}{= breakdown filter (e.g. All/GeoId/DomicileId)}
}
\description{
implements filters for 1dFloMo
}
\keyword{sql.1dFloMo.filter}
