% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloMo.grp}
\alias{sql.1dFloMo.grp}
\title{sql.1dFloMo.grp}
\usage{
sql.1dFloMo.grp(x, y)
}
\arguments{
\item{x}{= T/F depending on whether you are checking ftp}

\item{y}{= one or more breakdown filters (e.g. All/GeoId/DomicileId)}
}
\description{
group by clause for 1dFloMo
}
\keyword{sql.1dFloMo.grp}
