% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloMo.hld}
\alias{sql.1dFloMo.hld}
\title{sql.1dFloMo.hld}
\usage{
sql.1dFloMo.hld(x, y, n = F)
}
\arguments{
\item{x}{= month end in YYYYMMDD format}

\item{y}{= "" or the SQL query to subset to securities desired}

\item{n}{= T/F depending on whether you want to introduce rounding error}
}
\description{
Query to insert <x> into flow table
}
\keyword{sql.1dFloMo.hld}
