% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloMo}
\alias{sql.1dFloMo}
\title{sql.1dFloMo}
\usage{
sql.1dFloMo(x, y, n)
}
\arguments{
\item{x}{= the date for which you want flows (known one day later)}

\item{y}{= either "All" or "Act" or "CBE"}

\item{n}{= any of StockFlows/Japan/CSI300/Energy}
}
\description{
Generates the SQL query to get the data for 1dFloMo for individual stocks
}
\keyword{sql.1dFloMo}
