% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloMo.Sec.topline}
\alias{sql.1dFloMo.Sec.topline}
\title{sql.1dFloMo.Sec.topline}
\usage{
sql.1dFloMo.Sec.topline(x, y, n, w)
}
\arguments{
\item{x}{= SectorId/IndustryId}

\item{y}{= item (one of Flow/AssetsStart/AssetsEnd)}

\item{n}{= name of SQL temp table (#SEC/#INDY)}

\item{w}{= frequency (T/F for daily/weekly or D/W/M)}
}
\description{
top line SQL statement for daily/weekly CBE flow momentum
}
\keyword{sql.1dFloMo.Sec.topline}
