% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloMoAggr}
\alias{sql.1dFloMoAggr}
\title{sql.1dFloMoAggr}
\usage{
sql.1dFloMoAggr(x, y, n)
}
\arguments{
\item{x}{= the YYYYMMDD for which you want flows (known two days later)}

\item{y}{= one or more of FwtdIn0/FwtdEx0/SwtdIn0/SwtdEx0}

\item{n}{= any of StockFlows/Japan/CSI300/Energy}
}
\description{
Generates the SQL query to get the data for aggregate 1dFloMo
}
\keyword{sql.1dFloMoAggr}
