% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloTrend.Alloc.data}
\alias{sql.1dFloTrend.Alloc.data}
\title{sql.1dFloTrend.Alloc.data}
\usage{
sql.1dFloTrend.Alloc.data(x, y, n)
}
\arguments{
\item{x}{= SQL statement}

\item{y}{= named vector of country/sector codes indexed by Id}

\item{n}{= input to or output of sql.connect}
}
\description{
gets data for FloTrend
}
\keyword{sql.1dFloTrend.Alloc.data}
