% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloTrend.Alloc.fetch}
\alias{sql.1dFloTrend.Alloc.fetch}
\title{sql.1dFloTrend.Alloc.fetch}
\usage{
sql.1dFloTrend.Alloc.fetch(x, y, n, w, h)
}
\arguments{
\item{x}{= YYYYMM month}

\item{y}{= identifier column (SectorId/CountryId)}

\item{n}{= vector of acceptable identifiers}

\item{w}{= T/F depending on whether sign needs to be reversed}

\item{h}{= T/F depending on whether to unbracket}
}
\description{
SQL query for allocation table for FloTrend
}
\keyword{sql.1dFloTrend.Alloc.fetch}
