% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloTrend.Alloc.final}
\alias{sql.1dFloTrend.Alloc.final}
\title{sql.1dFloTrend.Alloc.final}
\usage{
sql.1dFloTrend.Alloc.final(x, y, n, w)
}
\arguments{
\item{x}{= from statement}

\item{y}{= factor (one of FloTrend/FloDiff/FloDiff2)}

\item{n}{= identifier column (SectorId/CountryId)}

\item{w}{= frequency (T/F for daily/weekly or D/W/M)}
}
\description{
SQL query for daily/weekly FloTrend
}
\keyword{sql.1dFloTrend.Alloc.final}
