% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloTrend.Alloc.purge}
\alias{sql.1dFloTrend.Alloc.purge}
\title{sql.1dFloTrend.Alloc.purge}
\usage{
sql.1dFloTrend.Alloc.purge(x, y)
}
\arguments{
\item{x}{= a temp table}

\item{y}{= identifier column (SectorId/CountryId)}
}
\description{
Ensures two sets of entries
}
\keyword{sql.1dFloTrend.Alloc.purge}
