% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloTrend.Alloc}
\alias{sql.1dFloTrend.Alloc}
\title{sql.1dFloTrend.Alloc}
\usage{
sql.1dFloTrend.Alloc(x, y, n, w = NULL)
}
\arguments{
\item{x}{= YYYYMM month}

\item{y}{= temp table name (e.g. "#CTRY")}

\item{n}{= identifier column (SectorId/CountryId)}

\item{w}{= vector of acceptable identifiers}
}
\description{
SQL query for allocation table for FloTrend
}
\keyword{sql.1dFloTrend.Alloc}
