% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloTrend.Ctry}
\alias{sql.1dFloTrend.Ctry}
\title{sql.1dFloTrend.Ctry}
\usage{
sql.1dFloTrend.Ctry(x, y, n)
}
\arguments{
\item{x}{= one of Ctry/FX/Sector}

\item{y}{= one of Flo/ActWt}

\item{n}{= one of Diff/Diff2/Trend/AllocMo/AllocDiff/AllocTrend}
}
\description{
For Ctry/FX generates the SQL query to get daily 1d a) FloDiff		= sql.1dFloTrend.Ctry("?", "Flo", "Diff") b) FloTrend		= sql.1dFloTrend.Ctry("?", "Flo", "Trend") c) ActWtDiff		= sql.1dFloTrend.Ctry("?", "ActWt", "Diff") d) ActWtTrend		= sql.1dFloTrend.Ctry("?", "ActWt", "Trend") e) FloDiff2		= sql.1dFloTrend.Ctry("?", "Flo", "Diff2") f) ActWtDiff2		= sql.1dFloTrend.Ctry("?", "ActWt", "Diff2") g) AllocMo		= sql.1dFloTrend.Ctry("?", "Flo", "AllocMo") h) AllocDiff		= sql.1dFloTrend.Ctry("?", "Flo", "AllocDiff") i) AllocTrend		= sql.1dFloTrend.Ctry("?", "Flo", "AllocTrend") j) AllocSkew		= sql.1dFloTrend.Ctry("?", "ActWt", "AllocSkew")
}
\keyword{sql.1dFloTrend.Ctry}
