% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloTrend.Ctry.topline}
\alias{sql.1dFloTrend.Ctry.topline}
\title{sql.1dFloTrend.Ctry.topline}
\usage{
sql.1dFloTrend.Ctry.topline(x, y, n)
}
\arguments{
\item{x}{= one of Trend/Diff/Diff2/AllocMo/AllocDiff/AllocTrend/AllocSkew}

\item{y}{= country list}

\item{n}{= one of DailyData/WeeklyData/MonthlyData}
}
\description{
Generates the SQL query to get daily 1d Flo/ActWt Diff/Trend for Ctry/FX
}
\keyword{sql.1dFloTrend.Ctry.topline}
