% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloTrend.Ctry.underlying}
\alias{sql.1dFloTrend.Ctry.underlying}
\title{sql.1dFloTrend.Ctry.underlying}
\usage{
sql.1dFloTrend.Ctry.underlying(x, y, n)
}
\arguments{
\item{x}{= a string vector indexed by allocation-table names}

\item{y}{= the SQL table from which you get flows (DailyData/MonthlyData)}

\item{n}{= one of Ctry/FX/Sector}
}
\description{
Generates the SQL query to get daily 1dFloMo for countries
}
\keyword{sql.1dFloTrend.Ctry.underlying}
