% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloTrend}
\alias{sql.1dFloTrend}
\title{sql.1dFloTrend}
\usage{
sql.1dFloTrend(x, y, n, w)
}
\arguments{
\item{x}{= data date in YYYYMMDD (known two days later)}

\item{y}{= a string vector of factors to be computed,       the last element of which is the type of fund used.}

\item{n}{= the delay in knowing allocations}

\item{w}{= any of StockFlows/Japan/CSI300/Energy}
}
\description{
Generates the SQL query to get the data for 1dFloTrend
}
\keyword{sql.1dFloTrend}
