% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloTrend.underlying}
\alias{sql.1dFloTrend.underlying}
\title{sql.1dFloTrend.underlying}
\usage{
sql.1dFloTrend.underlying(x, y)
}
\arguments{
\item{x}{= either "All" or "Act" or "CBE"}

\item{y}{= any of StockFlows/Japan/CSI300/Energy}
}
\description{
Generates the SQL query to get the data for 1dFloTrend
}
\keyword{sql.1dFloTrend.underlying}
