% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFundCt}
\alias{sql.1dFundCt}
\title{sql.1dFundCt}
\usage{
sql.1dFundCt(x, y, n, w, h)
}
\arguments{
\item{x}{= vector of flow dates in YYYYMMDD (known two days later)}

\item{y}{= a string vector of factors to be computed, the last element of which is the type of fund used.}

\item{n}{= any of StockFlows/China/Japan/CSI300/Energy}

\item{w}{= T/F depending on whether you are checking ftp}

\item{h}{= one or more breakdown filters (e.g. All/GeoId/DomicileId)}
}
\description{
Generates FundCt, the ownership breadth measure set forth in Chen, Hong & Stein (2001)"Breadth of ownership and stock returns"
}
\keyword{sql.1dFundCt}
