% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFundRet}
\alias{sql.1dFundRet}
\title{sql.1dFundRet}
\usage{
sql.1dFundRet(x)
}
\arguments{
\item{x}{= a list of fund identifiers}
}
\description{
Generates the SQL query to get monthly AIS for countries
}
\keyword{sql.1dFundRet}
