% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dION}
\alias{sql.1dION}
\title{sql.1dION}
\usage{
sql.1dION(x, y, n, w)
}
\arguments{
\item{x}{= data date (known two days later)}

\item{y}{= a vector of variables, the last element of which is ignored}

\item{n}{= the delay in knowing allocations}

\item{w}{= any of StockFlows/Japan/CSI300/Energy}
}
\description{
Generates the SQL query to get the data for 1dION$ & 1dION\%
}
\keyword{sql.1dION}
