% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1mActWt.underlying}
\alias{sql.1mActWt.underlying}
\title{sql.1mActWt.underlying}
\usage{
sql.1mActWt.underlying(x, y)
}
\arguments{
\item{x}{= the month for which you want data (0 = latest, 1 = lagged one month, etc.)}

\item{y}{= characters you want put in front of the query}
}
\description{
Generates tail end of an SQL query
}
\keyword{sql.1mActWt.underlying}
