% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1mAllocD.data}
\alias{sql.1mAllocD.data}
\title{sql.1mAllocD.data}
\usage{
sql.1mAllocD.data(x, y, n, w, h, u = NULL)
}
\arguments{
\item{x}{= the YYYYMM for which you want data (known 26 days later)}

\item{y}{= filter (one of All/Act/Pas/Etf/Mutual)}

\item{n}{= T/F depending on whether to account for price action}

\item{w}{= T/F depending on whether to excise leveraged funds}

\item{h}{= T/F depending on whether to excise fake cash security -999}

\item{u}{= columns in addition to AssetsEnd}
}
\description{
SQL query to get the data for AllocD
}
\keyword{sql.1mAllocD.data}
