% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1mAllocD.data.underlying}
\alias{sql.1mAllocD.data.underlying}
\title{sql.1mAllocD.data.underlying}
\usage{
sql.1mAllocD.data.underlying(x, y)
}
\arguments{
\item{x}{= parameter vector}

\item{y}{= columns in addition to AssetsEnd}
}
\description{
SQL query to get raw data for AllocD
}
\keyword{sql.1mAllocD.data.underlying}
