% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1mAllocD}
\alias{sql.1mAllocD}
\title{sql.1mAllocD}
\usage{
sql.1mAllocD(x, y, n, w, h, u = NULL, v = T, g = "All")
}
\arguments{
\item{x}{= a YYYYMM or vector of flow dates (if u = "Flow")}

\item{y}{= a string vector of factors to be computed, #        :    the last element of which is the type of fund used.}

\item{n}{= DB - any of StockFlows/China/Japan/CSI300/Energy}

\item{w}{= T/F depending on whether you are checking ftp}

\item{h}{= T/F depending on whether to account for price action}

\item{u}{= one of AssetsStart/Flow/NULL}

\item{v}{= T/F depending on whether to chuck securities held by just one fund}

\item{g}{= share-class filter (one of All/Inst/Retail) (x not monthly!)}
}
\description{
SQL query for 1mAllocD
}
\keyword{sql.1mAllocD}
