% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1mAllocD.select}
\alias{sql.1mAllocD.select}
\title{sql.1mAllocD.select}
\usage{
sql.1mAllocD.select(x)
}
\arguments{
\item{x}{= the factor to be computed}
}
\description{
select term to compute <x>
}
\keyword{sql.1mAllocD.select}
