% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1mAllocMo}
\alias{sql.1mAllocMo}
\title{sql.1mAllocMo}
\usage{
sql.1mAllocMo(x, y, n)
}
\arguments{
\item{x}{= the YYYYMM for which you want data (known 26 days later)}

\item{y}{= a string vector of factors to be computed, the last element of which is the type of fund used.}

\item{n}{= any of StockFlows/Japan/CSI300/Energy}
}
\description{
Generates the SQL query to get the data for 1mAllocMo
}
\keyword{sql.1mAllocMo}
