% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1mAllocMo.underlying}
\alias{sql.1mAllocMo.underlying}
\title{sql.1mAllocMo.underlying}
\usage{
sql.1mAllocMo.underlying(x, y)
}
\arguments{
\item{x}{= either "All" or "Act"}

\item{y}{= any of StockFlows/Japan/CSI300/Energy}
}
\description{
FROM and WHERE for 1mAllocMo
}
\keyword{sql.1mAllocMo.underlying}
