% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1mAllocSkew}
\alias{sql.1mAllocSkew}
\title{sql.1mAllocSkew}
\usage{
sql.1mAllocSkew(x, y, n, w, h = "All")
}
\arguments{
\item{x}{= the YYYYMM for which you want data (known 26 days later)}

\item{y}{= a string vector of factors to be computed, the last element of which is the type of fund used.}

\item{n}{= any of StockFlows/China/Japan/CSI300/Energy}

\item{w}{= T/F depending on whether you are checking ftp}

\item{h}{= share-class filter (one of All/Inst/Retail)}
}
\description{
Generates the SQL query to get the data for 1mAllocTrend
}
\keyword{sql.1mAllocSkew}
