% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1mAllocSkew.topline.from}
\alias{sql.1mAllocSkew.topline.from}
\title{sql.1mAllocSkew.topline.from}
\usage{
sql.1mAllocSkew.topline.from(x, y)
}
\arguments{
\item{x}{= table with GeoIds}

\item{y}{= T/F depending on whether you are checking ftp}
}
\description{
from part of the final select statement in 1mAllocTrend
}
\keyword{sql.1mAllocSkew.topline.from}
