% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1mAllocSkew.topline}
\alias{sql.1mAllocSkew.topline}
\title{sql.1mAllocSkew.topline}
\usage{
sql.1mAllocSkew.topline(x, y, n)
}
\arguments{
\item{x}{= a string vector of factors to be computed}

\item{y}{= T/F depending on whether you are checking ftp}

\item{n}{= T to force ReportDate to be a column}
}
\description{
Generates the SQL query to get the data for 1mAllocTrend
}
\keyword{sql.1mAllocSkew.topline}
