% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1mAllocSkew.underlying.basic}
\alias{sql.1mAllocSkew.underlying.basic}
\title{sql.1mAllocSkew.underlying.basic}
\usage{
sql.1mAllocSkew.underlying.basic(x, y, n)
}
\arguments{
\item{x}{= date restriction}

\item{y}{= a vector of filters}

\item{n}{= T/F for daily or monthly data}
}
\description{
Query to insert <x> into flow table
}
\keyword{sql.1mAllocSkew.underlying.basic}
