% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1mAllocSkew.underlying}
\alias{sql.1mAllocSkew.underlying}
\title{sql.1mAllocSkew.underlying}
\usage{
sql.1mAllocSkew.underlying(x, y, n, w)
}
\arguments{
\item{x}{= single YYYYMM or vector of flow dates in YYYYMMDD (known two days later)}

\item{y}{= the type of fund used in the computation}

\item{n}{= "" or the SQL query to subset to securities desired}

\item{w}{= share-class filter (one of All/Inst/Retail)}
}
\description{
the SQL query to get the data for 1dActWtTrend
}
\keyword{sql.1mAllocSkew.underlying}
