% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1mBullish.Alloc}
\alias{sql.1mBullish.Alloc}
\title{sql.1mBullish.Alloc}
\usage{
sql.1mBullish.Alloc(x, y, n)
}
\arguments{
\item{x}{= SQL statement}

\item{y}{= SectorId/CountryId}

\item{n}{= a temp table}
}
\description{
SQL query for monthly Bullish sector indicator
}
\keyword{sql.1mBullish.Alloc}
