% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1mBullish.Final}
\alias{sql.1mBullish.Final}
\title{sql.1mBullish.Final}
\usage{
sql.1mBullish.Final(x, y)
}
\arguments{
\item{x}{= SectorId/CountryId}

\item{y}{= a temp table}
}
\description{
SQL query for monthly Bullish sector indicator
}
\keyword{sql.1mBullish.Final}
