% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1mFundCt}
\alias{sql.1mFundCt}
\title{sql.1mFundCt}
\usage{
sql.1mFundCt(x, y, n, w, h, u = 0, v = "All")
}
\arguments{
\item{x}{= the YYYYMM for which you want data (known 26 days later)}

\item{y}{= a string vector of factors to be computed, #        :    the last elements of which are the types of fund used.}

\item{n}{= any of StockFlows/China/Japan/CSI300/Energy}

\item{w}{= T/F depending on whether you are checking ftp}

\item{h}{= breakdown filter (e.g. All/GeoId/DomicileId)}

\item{u}{= when non-zero only the biggest <u> funds for each security matter}

\item{v}{= share-class filter (one of All/Inst/Retail)}
}
\description{
Generates FundCt, the ownership breadth measure set forth in #		:	Chen, Hong & Stein (2001)"Breadth of ownership and stock returns"
}
\keyword{sql.1mFundCt}
