% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1mHoldAum}
\alias{sql.1mHoldAum}
\title{sql.1mHoldAum}
\usage{
sql.1mHoldAum(x, y, n, w, h)
}
\arguments{
\item{x}{= the YYYYMM for which you want data (known 26 days later)}

\item{y}{= a string vector of factors to be computed, #        :    the last elements of which are the types of fund used.}

\item{n}{= any of StockFlows/China/Japan/CSI300/Energy}

\item{w}{= T/F depending on whether you are checking ftp}

\item{h}{= breakdown filter (e.g. All/GeoId/DomicileId)}
}
\description{
Total AUM of all funds owning a particular security
}
\keyword{sql.1mHoldAum}
