% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1mSRIAdvisorPct}
\alias{sql.1mSRIAdvisorPct}
\title{sql.1mSRIAdvisorPct}
\usage{
sql.1mSRIAdvisorPct(x, y, n, w)
}
\arguments{
\item{x}{= the YYYYMM for which you want data (known 26 days later)}

\item{y}{= a string vector of factors to be computed, #        :    the last element of which is the type of fund used.}

\item{n}{= DB - any of StockFlows/China/Japan/CSI300/Energy}

\item{w}{= T/F depending on whether you are checking ftp}
}
\description{
Generates the SQL query to get the data for 1mSRIAdvisorPct
}
\keyword{sql.1mSRIAdvisorPct}
