% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1wFlow.Corp}
\alias{sql.1wFlow.Corp}
\title{sql.1wFlow.Corp}
\usage{
sql.1wFlow.Corp(x)
}
\arguments{
\item{x}{= a YYYYMMDD from which flows are to be computed}
}
\description{
Generates the SQL query to get weekly corporate flow ($MM)
}
\keyword{sql.1wFlow.Corp}
