% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.ActWtDiff2}
\alias{sql.ActWtDiff2}
\title{sql.ActWtDiff2}
\usage{
sql.ActWtDiff2(x)
}
\arguments{
\item{x}{= a flowdate}
}
\description{
ActWtDiff2 on R1 Materials for positioning
}
\keyword{sql.ActWtDiff2}
