% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.AggrAllocations}
\alias{sql.AggrAllocations}
\title{sql.AggrAllocations}
\usage{
sql.AggrAllocations(x, y, n, w, h)
}
\arguments{
\item{x}{= one of FwtdIn0/FwtdEx0/SwtdIn0/SwtdEx0}

\item{y}{= the name of the table containing Holdings (e.g. "#HLDGS")}

\item{n}{= a date of the form "@allocDt" or "'20151231'"}

\item{w}{= the grouping column (e.g. "GeographicFocusId")}

\item{h}{= the temp table for output}
}
\description{
Generates the SQL query to get aggregate allocations for StockFlows
}
\keyword{sql.AggrAllocations}
