% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.Allocation}
\alias{sql.Allocation}
\title{sql.Allocation}
\usage{
sql.Allocation(x, y, n = NULL, w = "All", h, u, v)
}
\arguments{
\item{x}{= needed columns}

\item{y}{= one of Country/Sector/Industry}

\item{n}{= columns needed from FundHistory besides HFundId/FundId}

\item{w}{= a vector of FundHistory filters}

\item{h}{= where clause (can be missing)}

\item{u}{= group by clause (can be missing)}

\item{v}{= having clause (can be missing)}
}
\description{
SQL query to fetch Country/Sector/Industry allocations
}
\keyword{sql.Allocation}
