% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.Allocation.Sec.FinsExREst}
\alias{sql.Allocation.Sec.FinsExREst}
\title{sql.Allocation.Sec.FinsExREst}
\usage{
sql.Allocation.Sec.FinsExREst(x)
}
\arguments{
\item{x}{= column names of table #SEC}
}
\description{
SQL query to add FinsExREst sector allocations
}
\keyword{sql.Allocation.Sec.FinsExREst}
