% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.Allocation.Sec}
\alias{sql.Allocation.Sec}
\title{sql.Allocation.Sec}
\usage{
sql.Allocation.Sec(x, y = NULL, n = "All")
}
\arguments{
\item{x}{= list object of individual where clauses}

\item{y}{= columns needed from FundHistory besides HFundId/FundId}

\item{n}{= a vector of FundHistory filters}
}
\description{
SQL query for sector allocations for month ending <x>
}
\keyword{sql.Allocation.Sec}
