% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.Allocations.bulk.EqWtAvg}
\alias{sql.Allocations.bulk.EqWtAvg}
\title{sql.Allocations.bulk.EqWtAvg}
\usage{
sql.Allocations.bulk.EqWtAvg(x, y, n, w)
}
\arguments{
\item{x}{= name of column being bulked up}

\item{y}{= vector of columns in addition to <w> within which averages are computed}

\item{n}{= allocation table name}

\item{w}{= primary grouping within which averages are computed}
}
\description{
Bulks up allocations with equal-weight averages
}
\keyword{sql.Allocations.bulk.EqWtAvg}
