% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.Allocations.bulk.Single}
\alias{sql.Allocations.bulk.Single}
\title{sql.Allocations.bulk.Single}
\usage{
sql.Allocations.bulk.Single(x, y, n, w, h)
}
\arguments{
\item{x}{= name of column being bulked up}

\item{y}{= vector of columns in addition to <w> with which funds are tagged}

\item{n}{= allocation table name}

\item{w}{= allocation bulking group (e.g. GeographicFocus/BenchIndex)}

\item{h}{= single-group column and value}
}
\description{
Bulks up allocations with single-group funds
}
\keyword{sql.Allocations.bulk.Single}
