% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.and}
\alias{sql.and}
\title{sql.and}
\usage{
sql.and(x, y = "and")
}
\arguments{
\item{x}{= list object of individual where clauses}

\item{y}{= logical operator to use}
}
\description{
and segment of an SQL statement
}
\keyword{sql.and}
