% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.arguments}
\alias{sql.arguments}
\title{sql.arguments}
\usage{
sql.arguments(x)
}
\arguments{
\item{x}{= a string vector of variables to build with the last elements specifying the type of funds to use}
}
\description{
splits <x> into factor and filters
}
\keyword{sql.arguments}
