% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.bcp}
\alias{sql.bcp}
\title{sql.bcp}
\usage{
sql.bcp(x, y, n = "Quant", w = "EPFRUI", h = "dbo")
}
\arguments{
\item{x}{= SQL table to perform the bulk copy from}

\item{y}{= the location of the output file}

\item{n}{= One of "StockFlows", "Quant", "QuantSF" or "Regular"}

\item{w}{= the database on which <x> resides}

\item{h}{= the owner of <x>}
}
\description{
code to bcp data out of server
}
\keyword{sql.bcp}
