% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.BenchIndex.duplication}
\alias{sql.BenchIndex.duplication}
\title{sql.BenchIndex.duplication}
\usage{
sql.BenchIndex.duplication(x)
}
\arguments{
\item{x}{= name of table being updated}
}
\description{
updates BenchIndex field in table <x> to remove duplicates
}
\keyword{sql.BenchIndex.duplication}
