% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.breakdown}
\alias{sql.breakdown}
\title{sql.breakdown}
\usage{
sql.breakdown(x)
}
\arguments{
\item{x}{= one or more breakdown filters (e.g. All/GeoId/DomicileId)}
}
\description{
Returns
}
\keyword{sql.breakdown}
