% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.Bullish}
\alias{sql.Bullish}
\title{sql.Bullish}
\usage{
sql.Bullish(x, y, n, w)
}
\arguments{
\item{x}{= the YYYYMM for which you want data (known 26 days later)}

\item{y}{= a string vector of factors to be computed, the last element of which is the type of fund used.}

\item{n}{= any of StockFlows/China/Japan/CSI300/Energy}

\item{w}{= T/F depending on whether you are checking ftp}
}
\description{
SQL query for Bullish-sentiment factor
}
\keyword{sql.Bullish}
