% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.case}
\alias{sql.case}
\title{sql.case}
\usage{
sql.case(x, y, n, w = T)
}
\arguments{
\item{x}{= final label}

\item{y}{= a string vector of conditions}

\item{n}{= a string vector of length one more than <y> of labels}

\item{w}{= T/F depending on whether labels are numeric}
}
\description{
case statement assigning labels <n> based on conditions <y>
}
\keyword{sql.case}
