% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.common}
\alias{sql.common}
\title{sql.common}
\usage{
sql.common(x, y)
}
\arguments{
\item{x}{= pair of table names}

\item{y}{= column name}
}
\description{
ensures common records in <x> based on <y>
}
\keyword{sql.common}
