% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.connect.wrapper}
\alias{sql.connect.wrapper}
\title{sql.connect.wrapper}
\usage{
sql.connect.wrapper(x)
}
\arguments{
\item{x}{= input to or output of sql.connect}
}
\description{
Opens an SQL connection (if needed)
}
\keyword{sql.connect.wrapper}
