% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.cross.border}
\alias{sql.cross.border}
\title{sql.cross.border}
\usage{
sql.cross.border(x)
}
\arguments{
\item{x}{= T/F depending on whether StockFlows data are being used}
}
\description{
Returns a list object of cross-border Geo. Foci and their names
}
\keyword{sql.cross.border}
