% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.CtryFlow.Alloc}
\alias{sql.CtryFlow.Alloc}
\title{sql.CtryFlow.Alloc}
\usage{
sql.CtryFlow.Alloc(x, y, n, w, h)
}
\arguments{
\item{x}{= NULL or a vector of identifiers}

\item{y}{= FundType}

\item{n}{= allocation month, in YYYYMMMDD format}

\item{w}{= one of Country/Sector/Industry}

\item{h}{= vector of filters}
}
\description{
SQL query for allocations needed in country flows
}
\keyword{sql.CtryFlow.Alloc}
