% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.CtryFlow.Flow}
\alias{sql.CtryFlow.Flow}
\title{sql.CtryFlow.Flow}
\usage{
sql.CtryFlow.Flow(x, y, n, w, h)
}
\arguments{
\item{x}{= a YYYYMMDD}

\item{y}{= item(s) (any of Flow/AssetsStart/AssetsEnd)}

\item{n}{= characteristics}

\item{w}{= frequency (T/F for daily/weekly or D/W/M)}

\item{h}{= vector of filters}
}
\description{
SQL query for single-group flows
}
\keyword{sql.CtryFlow.Flow}
