% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.currprior}
\alias{sql.currprior}
\title{sql.currprior}
\usage{
sql.currprior(fcn, x, y, n, w)
}
\arguments{
\item{fcn}{= SQL-script generator}

\item{x}{= a YYYYMM}

\item{y}{= table names}

\item{n}{= primary parameter (could be missing)}

\item{w}{= secondary parameter (could be missing)}
}
\description{
SQL query for current & prior allocations
}
\keyword{sql.currprior}
