% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.DailyFlo}
\alias{sql.DailyFlo}
\title{sql.DailyFlo}
\usage{
sql.DailyFlo(x, y)
}
\arguments{
\item{x}{= the date for which you want flows (known one day later)}

\item{y}{= the temp table to hold output}
}
\description{
Generates the SQL query to get the data for daily Flow
}
\keyword{sql.DailyFlo}
