% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.datediff}
\alias{sql.datediff}
\title{sql.datediff}
\usage{
sql.datediff(x, y, n)
}
\arguments{
\item{x}{= column in the monthly table}

\item{y}{= column in the daily table}

\item{n}{= calendar day on which previous month's data available}
}
\description{
Before <n>, falls back two else one month
}
\keyword{sql.datediff}
