% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.declare}
\alias{sql.declare}
\title{sql.declare}
\usage{
sql.declare(x, y, n)
}
\arguments{
\item{x}{= a variable vector}

\item{y}{= variable types}

\item{n}{= values}
}
\description{
declare statement
}
\keyword{sql.declare}
