% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.Diff}
\alias{sql.Diff}
\title{sql.Diff}
\usage{
sql.Diff(x, y, n = "")
}
\arguments{
\item{x}{= bit of SQL string}

\item{y}{= bit of SQL string}

\item{n}{= one of ""/"Num"/"Den"}
}
\description{
SQL statement for diffusion
}
\keyword{sql.Diff}
