% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.exists}
\alias{sql.exists}
\title{sql.exists}
\usage{
sql.exists(x, y = T)
}
\arguments{
\item{x}{= SQL statement}

\item{y}{= T/F depending on whether exists/not exists}
}
\description{
<x> in <y> if <n> or <x> not in <y> otherwise
}
\keyword{sql.exists}
