% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.extra.domicile}
\alias{sql.extra.domicile}
\title{sql.extra.domicile}
\usage{
sql.extra.domicile(x, y, n)
}
\arguments{
\item{x}{= a flowdate/YYYYMMDD depending on whether daily/weekly}

\item{y}{= column in classif-Ctry corresponding to names of <x>}

\item{n}{= column in FundHistory corresponding to names of <x>}
}
\description{
where clauses to ensure foreign flow
}
\keyword{sql.extra.domicile}
