% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.FloMo.Funds}
\alias{sql.FloMo.Funds}
\title{sql.FloMo.Funds}
\usage{
sql.FloMo.Funds(x)
}
\arguments{
\item{x}{= the month/day for which you want \% flow, \% portfolio change, & assets end}
}
\description{
Generates the SQL query to get monthly/daily data for Funds
}
\keyword{sql.FloMo.Funds}
