% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.floTbl.to.Col}
\alias{sql.floTbl.to.Col}
\title{sql.floTbl.to.Col}
\usage{
sql.floTbl.to.Col(x, y)
}
\arguments{
\item{x}{= one of DailyData/WeeklyData/MonthlyData}

\item{y}{= T/F depending on whether you want the date formatted.}
}
\description{
derived the appropriate date column from the flow table name
}
\keyword{sql.floTbl.to.Col}
