% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.Flow}
\alias{sql.Flow}
\title{sql.Flow}
\usage{
sql.Flow(x, y, n = "All", w = NULL, h = T, u, v)
}
\arguments{
\item{x}{= needed columns}

\item{y}{= a list of where clauses, the first being a flowdate}

\item{n}{= a vector of FundHistory filters}

\item{w}{= columns needed from FundHistory besides HFundId/FundId}

\item{h}{= frequency (T/F for daily/weekly or D/W/M)}

\item{u}{= group by clause (can be missing)}

\item{v}{= having clause (can be missing)}
}
\description{
SQL query to fetch daily/weekly/monthly flows
}
\keyword{sql.Flow}
