% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.Flow.tbl}
\alias{sql.Flow.tbl}
\title{sql.Flow.tbl}
\usage{
sql.Flow.tbl(x, y)
}
\arguments{
\item{x}{= frequency (T/F for daily/weekly or D/W/M)}

\item{y}{= T/F for table/date field (e.g. DailyData/DayEnding)}
}
\description{
table/date field name
}
\keyword{sql.Flow.tbl}
