% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.FundHistory}
\alias{sql.FundHistory}
\title{sql.FundHistory}
\usage{
sql.FundHistory(x, y, n)
}
\arguments{
\item{x}{= a vector of filters}

\item{y}{= T/F depending on whether StockFlows data are being used}

\item{n}{= columns needed in addition to HFundId}
}
\description{
SQL query to restrict to Global and Regional equity funds
}
\keyword{sql.FundHistory}
