% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.FundHistory.sf}
\alias{sql.FundHistory.sf}
\title{sql.FundHistory.sf}
\usage{
sql.FundHistory.sf(x)
}
\arguments{
\item{x}{= a vector of filters}
}
\description{
SQL query where clause
}
\keyword{sql.FundHistory.sf}
