% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.FundWtdExcl0}
\alias{sql.FundWtdExcl0}
\title{sql.FundWtdExcl0}
\usage{
sql.FundWtdExcl0(x)
}
\arguments{
\item{x}{= YYYYMM for which you want allocations (known 24 days after month end)}
}
\description{
Generates the SQL query to get Geo Focus average weights
}
\keyword{sql.FundWtdExcl0}
