% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.get}
\alias{sql.get}
\title{sql.get}
\usage{
sql.get(fcn, x, y, n, w = NULL)
}
\arguments{
\item{fcn}{= get function}

\item{x}{= list of vectors of flowdates}

\item{y}{= name of an SQL connection (e.g. "StockFlows", "NEWUI" etc.)}

\item{n}{= maximum number of queries using the same connection}

\item{w}{= argument passed down to <fcn>}
}
\description{
gets data using <fcn>
}
\keyword{sql.get}
