% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.Holdings}
\alias{sql.Holdings}
\title{sql.Holdings}
\usage{
sql.Holdings(x, y, n)
}
\arguments{
\item{x}{= where clause}

\item{y}{= columns you want fetched}

\item{n}{= a temp table}
}
\description{
query to access stock-holdings data
}
\keyword{sql.Holdings}
