% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.in}
\alias{sql.in}
\title{sql.in}
\usage{
sql.in(x, y, n = T)
}
\arguments{
\item{x}{= column}

\item{y}{= SQL statement}

\item{n}{= T/F depending on whether <x> is in <y>}
}
\description{
<x> in <y> if <n> or <x> not in <y> otherwise
}
\keyword{sql.in}
