% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.index}
\alias{sql.index}
\title{sql.index}
\usage{
sql.index(x, y)
}
\arguments{
\item{x}{= table name}

\item{y}{= string of column labels to index by (e.g. "DayEnding, FundId")}
}
\description{
SQL for primary key on <x> by columns <y>
}
\keyword{sql.index}
