% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.into}
\alias{sql.into}
\title{sql.into}
\usage{
sql.into(x, y)
}
\arguments{
\item{x}{= SQL statement}

\item{y}{= the temp table for the output}
}
\description{
unbrackets and selects into <y>
}
\keyword{sql.into}
