% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.ION}
\alias{sql.ION}
\title{sql.ION}
\usage{
sql.ION(x, y)
}
\arguments{
\item{x}{= bit of SQL string}

\item{y}{= bit of SQL string}
}
\description{
sum(case when <x> > 0 then <y> else 0 end)/case when sum(abs(<y>)) > 0 then sum(abs(<y>)) else NULL end
}
\keyword{sql.ION}
