% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.isin.old.to.new}
\alias{sql.isin.old.to.new}
\title{sql.isin.old.to.new}
\usage{
sql.isin.old.to.new(x)
}
\arguments{
\item{x}{= Historical Isin}
}
\description{
Returns the latest isin
}
\keyword{sql.isin.old.to.new}
