% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.mat.cofactor}
\alias{sql.mat.cofactor}
\title{sql.mat.cofactor}
\usage{
sql.mat.cofactor(x)
}
\arguments{
\item{x}{= square character matrix}
}
\description{
SQL for the cofactor matrix
}
\keyword{sql.mat.cofactor}
