% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.mat.crossprod}
\alias{sql.mat.crossprod}
\title{sql.mat.crossprod}
\usage{
sql.mat.crossprod(x, y)
}
\arguments{
\item{x}{= vector of names}

\item{y}{= T/F depending on whether there's an intercept term}
}
\description{
SQL for entries of X'X
}
\keyword{sql.mat.crossprod}
