% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.mat.crossprod.vector}
\alias{sql.mat.crossprod.vector}
\title{sql.mat.crossprod.vector}
\usage{
sql.mat.crossprod.vector(x, y, n)
}
\arguments{
\item{x}{= vector of names}

\item{y}{= a string}

\item{n}{= T/F depending on whether there's an intercept term}
}
\description{
SQL for entries of X'Y
}
\keyword{sql.mat.crossprod.vector}
