% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.mat.determinant}
\alias{sql.mat.determinant}
\title{sql.mat.determinant}
\usage{
sql.mat.determinant(x)
}
\arguments{
\item{x}{= square character matrix}
}
\description{
SQL for the determinant
}
\keyword{sql.mat.determinant}
