% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.mat.flip}
\alias{sql.mat.flip}
\title{sql.mat.flip}
\usage{
sql.mat.flip(x)
}
\arguments{
\item{x}{= square character matrix}
}
\description{
flips the sign for a term in a matrix
}
\keyword{sql.mat.flip}
