% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.mat.multiply}
\alias{sql.mat.multiply}
\title{sql.mat.multiply}
\usage{
sql.mat.multiply(x, y)
}
\arguments{
\item{x}{= string}

\item{y}{= string}
}
\description{
SQL for the determinant
}
\keyword{sql.mat.multiply}
