% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.median}
\alias{sql.median}
\title{sql.median}
\usage{
sql.median(x, y, n, w = 0.5)
}
\arguments{
\item{x}{= column on which computation is run}

\item{y}{= column on which partitioning is performed}

\item{n}{= SQL statement}

\item{w}{= desired ptile break point}
}
\description{
median (or alternate ptile point) of <x> within <y>
}
\keyword{sql.median}
