% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.Mo}
\alias{sql.Mo}
\title{sql.Mo}
\usage{
sql.Mo(x, y, n, w)
}
\arguments{
\item{x}{= vector of "flow"}

\item{y}{= isomekic vector of "assets"}

\item{n}{= isomekic vector of "weights"}

\item{w}{= T/F depending on whether to handle division by zero}
}
\description{
SQL statement for momentum
}
\keyword{sql.Mo}
