% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.MonthlyAlloc}
\alias{sql.MonthlyAlloc}
\title{sql.MonthlyAlloc}
\usage{
sql.MonthlyAlloc(x, y = "All", n = F, w = F)
}
\arguments{
\item{x}{= a YYYYMMDD for which you want allocations}

\item{y}{= DB - any of StockFlows/China/Japan/CSI300/Energy}

\item{n}{= T/F depending on SharesHeld/HoldingValue}

\item{w}{= T/F depending on if SecurityId is wanted}
}
\description{
Generates the SQL query to get the data for monthly allocations for StockFlows
}
\keyword{sql.MonthlyAlloc}
