% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.MonthlyAssetsEnd}
\alias{sql.MonthlyAssetsEnd}
\title{sql.MonthlyAssetsEnd}
\usage{
sql.MonthlyAssetsEnd(x, y = "", n = F)
}
\arguments{
\item{x}{= YYYYMMDD for which you want flows (known one day later)}

\item{y}{= characters that get pasted in front of every line (usually tabs for indentation)}

\item{n}{= T/F variable depending on whether you want AssetsStart/AssetsEnd or just AssetsEnd}
}
\description{
Generates the SQL query to get the data for monthly Assets End
}
\keyword{sql.MonthlyAssetsEnd}
