% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.nonneg}
\alias{sql.nonneg}
\title{sql.nonneg}
\usage{
sql.nonneg(x)
}
\arguments{
\item{x}{= bit of sql string}
}
\description{
case when <x> > 0 then <x> else NULL end
}
\seealso{
Other sql: \code{\link{sql.1dActWtTrend.Ctry.underlying}},
  \code{\link{sql.1dActWtTrend.topline}},
  \code{\link{sql.1dActWtTrend.underlying}},
  \code{\link{sql.1dActWtTrend}},
  \code{\link{sql.1dFloMo.Ctry.Allocations.GF.avg}},
  \code{\link{sql.1dFloMo.Ctry.Allocations}},
  \code{\link{sql.1dFloMo.Ctry.List}},
  \code{\link{sql.1dFloMo.Ctry}},
  \code{\link{sql.1dFloMo.FI.underlying}},
  \code{\link{sql.1dFloMo.FI}},
  \code{\link{sql.1dFloMo.Rgn}},
  \code{\link{sql.1dFloMoAggr}}, \code{\link{sql.1dFloMo}},
  \code{\link{sql.1dFloTrend.Ctry.topline}},
  \code{\link{sql.1dFloTrend.Ctry.underlying}},
  \code{\link{sql.1dFloTrend.Ctry}},
  \code{\link{sql.1dFloTrend.underlying}},
  \code{\link{sql.1dFloTrend}},
  \code{\link{sql.1dFundRet}}, \code{\link{sql.1dION}},
  \code{\link{sql.1mActPas.Ctry}},
  \code{\link{sql.1mActWt.underlying}},
  \code{\link{sql.1mActWt}},
  \code{\link{sql.1mAllocMo.underlying}},
  \code{\link{sql.1mAllocMo}},
  \code{\link{sql.1mAllocSkew}},
  \code{\link{sql.1mChActWt}},
  \code{\link{sql.AggrAllocations}},
  \code{\link{sql.AllocTbl}}, \code{\link{sql.DailyFlo}},
  \code{\link{sql.Diff}}, \code{\link{sql.FloMo.Funds}},
  \code{\link{sql.FundHistory}},
  \code{\link{sql.FundWtdExcl0}},
  \code{\link{sql.Holdings.bulk}},
  \code{\link{sql.Holdings}}, \code{\link{sql.ION}},
  \code{\link{sql.MonthlyAlloc}},
  \code{\link{sql.MonthlyAssetsEnd}}, \code{\link{sql.Mo}},
  \code{\link{sql.RDSuniv}},
  \code{\link{sql.TopDownAllocs}}, \code{\link{sql.Trend}},
  \code{\link{sql.and}}, \code{\link{sql.bcp}},
  \code{\link{sql.connect}},
  \code{\link{sql.cross.border}},
  \code{\link{sql.datediff}}, \code{\link{sql.declare}},
  \code{\link{sql.drop}}, \code{\link{sql.floTbl.to.Col}},
  \code{\link{sql.into}}, \code{\link{sql.in}},
  \code{\link{sql.isin.old.to.new}},
  \code{\link{sql.label}}, \code{\link{sql.map.classif}},
  \code{\link{sql.query}},
  \code{\link{sql.stock.flows.wtd.avg}},
  \code{\link{sql.tbl}}, \code{\link{sql.unbracket}}
}
\keyword{sql.nonneg}
