% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.query}
\alias{sql.query}
\title{sql.query}
\usage{
sql.query(x, y)
}
\arguments{
\item{x}{= query needed for the update}

\item{y}{= one of StockFlows/Regular/Quant}
}
\description{
opens a connection, executes sql query, then closes the connection
}
\keyword{sql.query}
