% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.query.underlying}
\alias{sql.query.underlying}
\title{sql.query.underlying}
\usage{
sql.query.underlying(x, y, n = T)
}
\arguments{
\item{x}{= query needed for the update}

\item{y}{= a connection, the output of odbcDriverConnect}

\item{n}{= T/F depending on whether you wish to output number of rows of data got}
}
\description{
opens a connection, executes sql query, then closes the connection
}
\keyword{sql.query.underlying}
