% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.RDSuniv}
\alias{sql.RDSuniv}
\title{sql.RDSuniv}
\usage{
sql.RDSuniv(x)
}
\arguments{
\item{x}{= any of StockFlows/China/Japan/CSI300/Energy}
}
\description{
Generates the SQL query to get the row space for a #		:	stock flows research data set
}
\keyword{sql.RDSuniv}
