% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.regr}
\alias{sql.regr}
\title{sql.regr}
\usage{
sql.regr(x, y, n)
}
\arguments{
\item{x}{= a string vector (independent variable(s))}

\item{y}{= a string (dependent variable)}

\item{n}{= T/F depending on whether there's an intercept term}
}
\description{
SQL for regression coefficients
}
\keyword{sql.regr}
