% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.ReportDate}
\alias{sql.ReportDate}
\title{sql.ReportDate}
\usage{
sql.ReportDate(x)
}
\arguments{
\item{x}{= a YYYYMMDD}
}
\description{
SQL select statement for constant date <x>
}
\keyword{sql.ReportDate}
