% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.ShareClass}
\alias{sql.ShareClass}
\title{sql.ShareClass}
\usage{
sql.ShareClass(x, y)
}
\arguments{
\item{x}{= date restriction}

\item{y}{= share-class filter (one of All/Inst/Retail)}
}
\description{
Generates where clause for share-class filter
}
\keyword{sql.ShareClass}
