% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.SRI}
\alias{sql.SRI}
\title{sql.SRI}
\usage{
sql.SRI(x, y)
}
\arguments{
\item{x}{= the YYYYMM for which you want data (known 26 days later)}

\item{y}{= any of StockFlows/Japan/CSI300/Energy}
}
\description{
number of SRI funds holding the stock at time <x>
}
\keyword{sql.SRI}
