% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.stock.flows.wtd.avg}
\alias{sql.stock.flows.wtd.avg}
\title{sql.stock.flows.wtd.avg}
\usage{
sql.stock.flows.wtd.avg(x, y)
}
\arguments{
\item{x}{= YYYYMM at the end of which allocations are desired}

\item{y}{= a string. Must be one of All/Etf/MF.}
}
\description{
Computes Fund/Smpl weighted Incl/Excl zero for all names in the S&P
}
\keyword{sql.stock.flows.wtd.avg}
