% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.tbl}
\alias{sql.tbl}
\title{sql.tbl}
\usage{
sql.tbl(x, y, n, w, h, u)
}
\arguments{
\item{x}{= needed columns}

\item{y}{= from clause}

\item{n}{= where clause}

\item{w}{= "group by" clause}

\item{h}{= having clause}

\item{u}{= order by clause}
}
\description{
Full SQL statement
}
\keyword{sql.tbl}
