% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.TopDownAllocs.items}
\alias{sql.TopDownAllocs.items}
\title{sql.TopDownAllocs.items}
\usage{
sql.TopDownAllocs.items(x, y = T)
}
\arguments{
\item{x}{= a string vector specifying types of allocation wanted}

\item{y}{= T/F depending on whether select item or having entry is desired}
}
\description{
allocations to select in Top-Down Allocations SQL query
}
\keyword{sql.TopDownAllocs.items}
