% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.TopDownAllocs}
\alias{sql.TopDownAllocs}
\title{sql.TopDownAllocs}
\usage{
sql.TopDownAllocs(x, y, n)
}
\arguments{
\item{x}{= the YYYYMM for which you want data (known 26 days later)}

\item{y}{= a string vector, first the element of which is 0/1 depending on whether you want Active/Passive, while the subsequent elements are the types of top-down allocation wanted}

\item{n}{= any of StockFlows/Japan/CSI300/Energy}
}
\description{
Generates the SQL query to get Active/Passive Top-Down Allocations
}
\keyword{sql.TopDownAllocs}
