% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.Trend}
\alias{sql.Trend}
\title{sql.Trend}
\usage{
sql.Trend(x, y = "All")
}
\arguments{
\item{x}{= bit of SQL string}

\item{y}{= type of fund to use}
}
\description{
= sum(<x>)/case when sum(<x>) = 0 then NULL else sum(<x>) end
}
\keyword{sql.Trend}
