% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.update}
\alias{sql.update}
\title{sql.update}
\usage{
sql.update(x, y, n, w)
}
\arguments{
\item{x}{= table name}

\item{y}{= set argument}

\item{n}{= from argument (can be missing)}

\item{w}{= where clause}
}
\description{
update <x> set <y> from <n> where <w>
}
\keyword{sql.update}
