% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.yield.curve.1dFloMo}
\alias{sql.yield.curve.1dFloMo}
\title{sql.yield.curve.1dFloMo}
\usage{
sql.yield.curve.1dFloMo(x, y, n, w)
}
\arguments{
\item{x}{= value to match CompanyName field}

\item{y}{= value to match SecurityType field}

\item{n}{= value to match BondCurrency field}

\item{w}{= a YYYYMMDD vector}
}
\description{
daily FloMo by yield-curve bucket
}
\keyword{sql.yield.curve.1dFloMo}
