% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.yield.curve}
\alias{sql.yield.curve}
\title{sql.yield.curve}
\usage{
sql.yield.curve(x, y, n, w = "General", h = "FundId")
}
\arguments{
\item{x}{= value to match CompanyName field}

\item{y}{= value to match SecurityType field}

\item{n}{= value to match BondCurrency field}

\item{w}{= type of maturity buckets}

\item{h}{= fund identifier}
}
\description{
buckets holdings by maturities
}
\keyword{sql.yield.curve}
