% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.yyyymm}
\alias{sql.yyyymm}
\title{sql.yyyymm}
\usage{
sql.yyyymm(x, y)
}
\arguments{
\item{x}{= name of a datetime column in an SQL table}

\item{y}{= label after conversion (defaults to <x> if missing)}
}
\description{
SQL code to convert to YYYYMM
}
\keyword{sql.yyyymm}
