% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.yyyymmdd}
\alias{sql.yyyymmdd}
\title{sql.yyyymmdd}
\usage{
sql.yyyymmdd(x, y, n = F)
}
\arguments{
\item{x}{= name of a datetime column in an SQL table}

\item{y}{= label after conversion (defaults to <x> if missing)}

\item{n}{= T/F depending on whether you are checking ftp}
}
\description{
SQL code to convert to YYYYMMDD
}
\keyword{sql.yyyymmdd}
