% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sqlts.FloDollar.monthly}
\alias{sqlts.FloDollar.monthly}
\title{sqlts.FloDollar.monthly}
\usage{
sqlts.FloDollar.monthly(x)
}
\arguments{
\item{x}{= the security id for which you want data}
}
\description{
SQL query for monthly dollar flow
}
\seealso{
Other sqlts: \code{\link{sqlts.FloDollar.daily}},
  \code{\link{sqlts.TopDownAllocs}},
  \code{\link{sqlts.wrapper}}
}
\keyword{sqlts.FloDollar.monthly}
