% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sqlts.TopDownAllocs}
\alias{sqlts.TopDownAllocs}
\title{sqlts.TopDownAllocs}
\usage{
sqlts.TopDownAllocs(x, y)
}
\arguments{
\item{x}{= the security id for which you want data}

\item{y}{= a string vector specifying types of allocation wanted}
}
\description{
SQL query for Top-Down Allocations
}
\seealso{
Other sqlts: \code{\link{sqlts.FloDollar.daily}},
  \code{\link{sqlts.FloDollar.monthly}},
  \code{\link{sqlts.wrapper}}
}
\keyword{sqlts.TopDownAllocs}
