% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sqlts.wrapper}
\alias{sqlts.wrapper}
\title{sqlts.wrapper}
\usage{
sqlts.wrapper(x, y)
}
\arguments{
\item{x}{= a vector of security id's}

\item{y}{= data item wanted (Daily/Monthly/Allocation)}
}
\description{
SQL query for monthly dollar flow
}
\seealso{
Other sqlts: \code{\link{sqlts.FloDollar.daily}},
  \code{\link{sqlts.FloDollar.monthly}},
  \code{\link{sqlts.TopDownAllocs}}
}
\keyword{sqlts.wrapper}
