% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{strat.email}
\alias{strat.email}
\title{strat.email}
\usage{
strat.email(x, y, n, w = "All")
}
\arguments{
\item{x}{= vector of strategy names (e.g. "FX" or "FloPctCtry")}

\item{y}{= frequency (e.g. "daily", "weekly" or "monthly")}

\item{n}{= the email address(es) of the recipient(s)}

\item{w}{= the salutation}
}
\description{
emails strategies <x> of frequency <y>
}
\seealso{
Other strat: \code{\link{strat.dir}},
  \code{\link{strat.file}}, \code{\link{strat.path}}
}
\keyword{strat.email}
