% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{strat.file}
\alias{strat.file}
\title{strat.file}
\usage{
strat.file(x, y)
}
\arguments{
\item{x}{= vector of strategy names (e.g. "FX" or "FloPctCtry")}

\item{y}{= frequency (e.g. "daily", "weekly" or "monthly")}
}
\description{
the path to the factor file
}
\keyword{strat.file}
