% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{strat.list}
\alias{strat.list}
\title{strat.list}
\usage{
strat.list(x)
}
\arguments{
\item{x}{= frequency (e.g. "daily", "weekly" or "monthly")}
}
\description{
lists strategies of that frequency
}
\seealso{
Other strat: \code{\link{strat.dir}},
  \code{\link{strat.email}}, \code{\link{strat.file}},
  \code{\link{strat.path}}
}
\keyword{strat.list}
