% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{strategy.dir}
\alias{strategy.dir}
\title{strategy.dir}
\usage{
strategy.dir(x)
}
\arguments{
\item{x}{= "daily" or "weekly"}
}
\description{
factor folder
}
\seealso{
Other strategy: \code{\link{strategy.file}},
  \code{\link{strategy.path}}
}
\keyword{strategy.dir}
