% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{strategy.path}
\alias{strategy.path}
\title{strategy.path}
\usage{
strategy.path(x, y)
}
\arguments{
\item{x}{= name of the strategy (e.g. "FX" or "PremSec-JP")}

\item{y}{= "daily" or "weekly"}
}
\description{
Returns the full path to the factor file
}
\seealso{
Other strategy: \code{\link{strategy.dir}},
  \code{\link{strategy.file}}
}
\keyword{strategy.path}
