% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{stratrets.indicator}
\alias{stratrets.indicator}
\title{stratrets.indicator}
\usage{
stratrets.indicator(x, y, n, w, h)
}
\arguments{
\item{x}{= paths to strategy files}

\item{y}{= lookback}

\item{n}{= if T, flows get summed. Otherwise they get compounded}

\item{w}{= if T, sector-adjustment is performed}

\item{h}{= delay in knowing data (only used when <w>)}
}
\description{
data frame compounded across <y>
}
\keyword{stratrets.indicator}
