% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{stratrets}
\alias{stratrets}
\title{stratrets}
\usage{
stratrets(x, y)
}
\arguments{
\item{x}{= a variable}

\item{y}{= dataframe of back-test parameters (can be missing)}
}
\description{
data frame of TxB return spreads
}
\seealso{
Other stratrets: \code{\link{stratrets.bbk}},
  \code{\link{stratrets.beta}},
  \code{\link{stratrets.data}},
  \code{\link{stratrets.indicator}},
  \code{\link{stratrets.path}},
  \code{\link{stratrets.returns}},
  \code{\link{stratrets.subset.Ctry}},
  \code{\link{stratrets.subset}}
}
\keyword{stratrets}
