% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{stratrets.subset.Ctry}
\alias{stratrets.subset.Ctry}
\title{stratrets.subset.Ctry}
\usage{
stratrets.subset.Ctry(x, y)
}
\arguments{
\item{x}{= indicators}

\item{y}{= index to subset to}
}
\description{
determine which countries to subset to
}
\keyword{stratrets.subset.Ctry}
