% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{stratrets.subset}
\alias{stratrets.subset}
\title{stratrets.subset}
\usage{
stratrets.subset(x, y)
}
\arguments{
\item{x}{= indicators}

\item{y}{= index to subset to}
}
\description{
subsets to columns used in the back-test
}
\keyword{stratrets.subset}
