% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{stunden}
\alias{stunden}
\title{stunden}
\usage{
stunden(x = 5, y = 8)
}
\arguments{
\item{x}{= integer}

\item{y}{= integer}
}
\description{
vector of <x> random numbers within +/-1 of, and averaging to, <y>
}
\keyword{stunden}
