% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{summ.multi}
\alias{summ.multi}
\title{summ.multi}
\usage{
summ.multi(fcn, x, y)
}
\arguments{
\item{fcn}{= a function that summarizes the data}

\item{x}{= a df of bin returns indexed by time}

\item{y}{= offset associated with the rows of <x>}
}
\description{
summarizes the multi-period back test
}
\keyword{summ.multi}
