% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{tstat}
\alias{tstat}
\title{tstat}
\usage{
tstat(x, y)
}
\arguments{
\item{x}{= a matrix/data frame}

\item{y}{= a numeric vector (corresponds to columns of <x>)}
}
\description{
t-statistic associated with the regression of each row of <x> on <y>
}
\keyword{tstat}
