% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.anagram}
\alias{txt.anagram}
\title{txt.anagram}
\usage{
txt.anagram(x, y, n = 0)
}
\arguments{
\item{x}{= a SINGLE string}

\item{y}{= a file of potentially-usable capitalized words}

\item{n}{= vector of minimum number of characters for first few words}
}
\description{
all possible anagrams
}
\keyword{txt.anagram}
