% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.anagram.underlying}
\alias{txt.anagram.underlying}
\title{txt.anagram.underlying}
\usage{
txt.anagram.underlying(x, y, n)
}
\arguments{
\item{x}{= a SINGLE string}

\item{y}{= potentially-usable capitalized words}

\item{n}{= vector of minimum number of characters for first few words}
}
\description{
all possible anagrams
}
\keyword{txt.anagram.underlying}
