% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.core}
\alias{txt.core}
\title{txt.core}
\usage{
txt.core(x)
}
\arguments{
\item{x}{= a vector}
}
\description{
renders with upper-case letters, spaces and numbers only
}
\seealso{
Other txt: \code{\link{txt.anagram.underlying}},
  \code{\link{txt.anagram}}, \code{\link{txt.count}},
  \code{\link{txt.ex.file}}, \code{\link{txt.ex.int}},
  \code{\link{txt.excise}}, \code{\link{txt.expand}},
  \code{\link{txt.gunning}}, \code{\link{txt.has}},
  \code{\link{txt.hdr}}, \code{\link{txt.itrim}},
  \code{\link{txt.left}}, \code{\link{txt.levenshtein}},
  \code{\link{txt.name.format}}, \code{\link{txt.na}},
  \code{\link{txt.parse}}, \code{\link{txt.prepend}},
  \code{\link{txt.regr}}, \code{\link{txt.replace}},
  \code{\link{txt.reverse}}, \code{\link{txt.right}},
  \code{\link{txt.space}}, \code{\link{txt.to.char}},
  \code{\link{txt.trim.end}}, \code{\link{txt.trim.left}},
  \code{\link{txt.trim.right}}, \code{\link{txt.trim}},
  \code{\link{txt.words}}
}
\keyword{txt.core}
