% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.count}
\alias{txt.count}
\title{txt.count}
\usage{
txt.count(x, y)
}
\arguments{
\item{x}{= a vector of strings}

\item{y}{= a substring}
}
\description{
counts the number of occurences of <y> in each element of <x>
}
\keyword{txt.count}
