% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.ex.int.cardinal}
\alias{txt.ex.int.cardinal}
\title{txt.ex.int.cardinal}
\usage{
txt.ex.int.cardinal(x)
}
\arguments{
\item{x}{= an integer vector}
}
\description{
a string vector describing <x> in words (cardinal numbers)
}
\keyword{txt.ex.int.cardinal}
