% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.ex.int.ordinal}
\alias{txt.ex.int.ordinal}
\title{txt.ex.int.ordinal}
\usage{
txt.ex.int.ordinal(x)
}
\arguments{
\item{x}{= a vector of integers}
}
\description{
a string vector describing <x> in words (cardinal numbers)
}
\keyword{txt.ex.int.ordinal}
