% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.ex.int}
\alias{txt.ex.int}
\title{txt.ex.int}
\usage{
txt.ex.int(x, y = F)
}
\arguments{
\item{x}{= a vector of integers}

\item{y}{= T/F depending on whether ordinal numbers are wanted}
}
\description{
a string vector describing <x> in words
}
\keyword{txt.ex.int}
