% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.ex.int.underlying}
\alias{txt.ex.int.underlying}
\title{txt.ex.int.underlying}
\usage{
txt.ex.int.underlying(x, y, n, w, h)
}
\arguments{
\item{x}{= a vector of integers}

\item{y}{= odds & ends}

\item{n}{= units}

\item{w}{= tens}

\item{h}{= tens ordinal}
}
\description{
string vector describing <x> in words
}
\keyword{txt.ex.int.underlying}
