% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.expand}
\alias{txt.expand}
\title{txt.expand}
\usage{
txt.expand(x, y, n = "-", w = F)
}
\arguments{
\item{x}{= a string vector}

\item{y}{= a string vector}

\item{n}{= the separator}

\item{w}{= T/F variable controlling paste order}
}
\description{
Returns all combinations OF <x> and <y> pasted together
}
\keyword{txt.expand}
