% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.first}
\alias{txt.first}
\title{txt.first}
\usage{
txt.first(x, y)
}
\arguments{
\item{x}{= a string vector}

\item{y}{= a regular expression}
}
\description{
first occurrence of pattern <y> in <x> (NA if none)
}
\keyword{txt.first}
