% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.gunning}
\alias{txt.gunning}
\title{txt.gunning}
\usage{
txt.gunning(x, y, n)
}
\arguments{
\item{x}{= a string representing a text passage}

\item{y}{= a file of potentially-usable capitalized words}

\item{n}{= a file of potentially-usable capitalized words considered "simple"}
}
\description{
the Gunning fog index measuring the number of years of  schooling beyond kindergarten needed to comprehend <x>
}
\keyword{txt.gunning}
