% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.has}
\alias{txt.has}
\title{txt.has}
\usage{
txt.has(x, y, n = F)
}
\arguments{
\item{x}{= a vector of strings}

\item{y}{= a single string}

\item{n}{= T/F depending on whether a logical vector is desired}
}
\description{
the elements of <x> that contain <y> if <n> is F or a #		:	logical vector otherwise
}
\keyword{txt.has}
