% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.left}
\alias{txt.left}
\title{txt.left}
\usage{
txt.left(x, y)
}
\arguments{
\item{x}{= a string vector}

\item{y}{= a positive integer}
}
\description{
Returns the left <y> characters
}
\keyword{txt.left}
