% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.levenshtein}
\alias{txt.levenshtein}
\title{txt.levenshtein}
\usage{
txt.levenshtein(x, y)
}
\arguments{
\item{x}{= a string}

\item{y}{= a string}
}
\description{
Levenshtein distance between <x> and <y>. Similar to dtw
}
\keyword{txt.levenshtein}
