% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.name.format}
\alias{txt.name.format}
\title{txt.name.format}
\usage{
txt.name.format(x)
}
\arguments{
\item{x}{= a string vector}
}
\description{
capitalizes first letter of each word, rendering remaining letters in lower case
}
\keyword{txt.name.format}
