% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.palindrome}
\alias{txt.palindrome}
\title{txt.palindrome}
\usage{
txt.palindrome(x, y)
}
\arguments{
\item{x}{= a SINGLE string}

\item{y}{= potentially-usable capitalized words}
}
\description{
short palindromes that reflect just before the first letter of <x>, or just after the last letter of <x>, or somewhere in-between.
}
\seealso{
Other txt: \code{\link{txt.anagram.underlying}},
  \code{\link{txt.anagram}}, \code{\link{txt.core}},
  \code{\link{txt.count}}, \code{\link{txt.ex.file}},
  \code{\link{txt.excise}}, \code{\link{txt.expand}},
  \code{\link{txt.gunning}}, \code{\link{txt.has}},
  \code{\link{txt.hdr}}, \code{\link{txt.itrim}},
  \code{\link{txt.left}}, \code{\link{txt.levenshtein}},
  \code{\link{txt.name.format}}, \code{\link{txt.na}},
  \code{\link{txt.palindrome.entire}},
  \code{\link{txt.palindrome.partial}},
  \code{\link{txt.palindrome.tail}},
  \code{\link{txt.palindrome.underlying}},
  \code{\link{txt.parse}}, \code{\link{txt.prepend}},
  \code{\link{txt.regr}}, \code{\link{txt.replace}},
  \code{\link{txt.reverse}}, \code{\link{txt.right}},
  \code{\link{txt.space}}, \code{\link{txt.to.char}},
  \code{\link{txt.trim.end}}, \code{\link{txt.trim.left}},
  \code{\link{txt.trim.right}}, \code{\link{txt.trim}},
  \code{\link{txt.words}}
}
\keyword{txt.palindrome}
