% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.parse}
\alias{txt.parse}
\title{txt.parse}
\usage{
txt.parse(x, y)
}
\arguments{
\item{x}{= a string vector}

\item{y}{= a string that serves as a delimiter}
}
\description{
breaks up string <x> by <y>
}
\keyword{txt.parse}
