% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.prepend}
\alias{txt.prepend}
\title{txt.prepend}
\usage{
txt.prepend(x, y, n)
}
\arguments{
\item{x}{= a vector of strings}

\item{y}{= number of characters to add}

\item{n}{= the characters to add at the beginning}
}
\description{
bulks up each string to have at least <y> characters by adding <n> to the beginning of each string
}
\seealso{
Other txt: \code{\link{txt.count}},
  \code{\link{txt.ex.file}}, \code{\link{txt.expand}},
  \code{\link{txt.has}}, \code{\link{txt.hdr}},
  \code{\link{txt.itrim}}, \code{\link{txt.left}},
  \code{\link{txt.name.format}}, \code{\link{txt.na}},
  \code{\link{txt.parse}}, \code{\link{txt.regr}},
  \code{\link{txt.replace}}, \code{\link{txt.reverse}},
  \code{\link{txt.right}}, \code{\link{txt.space}},
  \code{\link{txt.to.char}}, \code{\link{txt.trim.end}},
  \code{\link{txt.trim.left}},
  \code{\link{txt.trim.right}}, \code{\link{txt.trim}}
}
\keyword{txt.prepend}
