% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.regex}
\alias{txt.regex}
\title{txt.regex}
\usage{
txt.regex(x)
}
\arguments{
\item{x}{= string}
}
\description{
converts <x> to a regular expression by padding certain characters with \\\\
}
\keyword{txt.regex}
