% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.regr}
\alias{txt.regr}
\title{txt.regr}
\usage{
txt.regr(x, y = T)
}
\arguments{
\item{x}{= a vector of column names}

\item{y}{= T/F depending on whether regression has an intercept}
}
\description{
returns the string you need to regress the first column on the others
}
\keyword{txt.regr}
