% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.right}
\alias{txt.right}
\title{txt.right}
\usage{
txt.right(x, y)
}
\arguments{
\item{x}{= a vector of string}

\item{y}{= a positive integer}
}
\description{
Returns the right <y> characters
}
\keyword{txt.right}
