% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.space}
\alias{txt.space}
\title{txt.space}
\usage{
txt.space(x, y = " ")
}
\arguments{
\item{x}{= an integer vector}

\item{y}{= a SINGLE string}
}
\description{
returns <x> iterations of <y> pasted together
}
\keyword{txt.space}
