% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.subclass.underlying}
\alias{txt.subclass.underlying}
\title{txt.subclass.underlying}
\usage{
txt.subclass.underlying(x, y, n)
}
\arguments{
\item{x}{= a string}

\item{y}{= a string}

\item{n}{= a data frame}
}
\description{
F/T depending on whether <x> is a subclass of <y>
}
\seealso{
Other txt: \code{\link{txt.anagram.underlying}},
  \code{\link{txt.anagram}}, \code{\link{txt.bracket}},
  \code{\link{txt.core}}, \code{\link{txt.count}},
  \code{\link{txt.ex.file}},
  \code{\link{txt.ex.int.cardinal.wrapper}},
  \code{\link{txt.ex.int.cardinal}},
  \code{\link{txt.ex.int.ordinal.wrapper}},
  \code{\link{txt.ex.int.ordinal}},
  \code{\link{txt.ex.int.underlying}},
  \code{\link{txt.ex.int}}, \code{\link{txt.excise}},
  \code{\link{txt.expand}}, \code{\link{txt.first}},
  \code{\link{txt.gunning}}, \code{\link{txt.hdr}},
  \code{\link{txt.itrim}}, \code{\link{txt.left}},
  \code{\link{txt.levenshtein}},
  \code{\link{txt.name.format}}, \code{\link{txt.na}},
  \code{\link{txt.parse.delimit}}, \code{\link{txt.parse}},
  \code{\link{txt.prepend}}, \code{\link{txt.regexp}},
  \code{\link{txt.regr}}, \code{\link{txt.reverse}},
  \code{\link{txt.right}}, \code{\link{txt.subclass.bulk}},
  \code{\link{txt.subclass.wrapper}},
  \code{\link{txt.subclass}}, \code{\link{txt.to.char}},
  \code{\link{txt.trim.left}},
  \code{\link{txt.trim.right}}, \code{\link{txt.trim}},
  \code{\link{txt.words}}
}
\keyword{txt.subclass.underlying}
