% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.trim.end}
\alias{txt.trim.end}
\title{txt.trim.end}
\usage{
txt.trim.end(fcn, x, y, n)
}
\arguments{
\item{fcn}{= a function that returns characters from the bad end}

\item{x}{= a vector of string}

\item{y}{= a vector of verboten strings, each of the same length}

\item{n}{= a functon that returns characters from the opposite end}
}
\description{
trims off leading or trailing elements of <y>
}
\seealso{
Other txt: \code{\link{txt.anagram.underlying}},
  \code{\link{txt.anagram}}, \code{\link{txt.core}},
  \code{\link{txt.count}}, \code{\link{txt.ex.file}},
  \code{\link{txt.excise}}, \code{\link{txt.expand}},
  \code{\link{txt.gunning}}, \code{\link{txt.has}},
  \code{\link{txt.hdr}}, \code{\link{txt.itrim}},
  \code{\link{txt.left}}, \code{\link{txt.levenshtein}},
  \code{\link{txt.name.format}}, \code{\link{txt.na}},
  \code{\link{txt.parse}}, \code{\link{txt.prepend}},
  \code{\link{txt.regr}}, \code{\link{txt.replace}},
  \code{\link{txt.reverse}}, \code{\link{txt.right}},
  \code{\link{txt.space}}, \code{\link{txt.to.char}},
  \code{\link{txt.trim.left}},
  \code{\link{txt.trim.right}}, \code{\link{txt.trim}},
  \code{\link{txt.words}}
}
\keyword{txt.trim.end}
