% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.trim.left}
\alias{txt.trim.left}
\title{txt.trim.left}
\usage{
txt.trim.left(x, y)
}
\arguments{
\item{x}{= a string vector}

\item{y}{= a single string or expression like "[\\t ]"}
}
\description{
trims off leading occurences of <y>
}
\keyword{txt.trim.left}
