% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.trim}
\alias{txt.trim}
\title{txt.trim}
\usage{
txt.trim(x, y = " ")
}
\arguments{
\item{x}{= a vector of string}

\item{y}{= a single string or expression like "[\\t ]"}
}
\description{
trims off leading/trailing occurences of <y>
}
\keyword{txt.trim}
