% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{urn.comb}
\alias{urn.comb}
\title{urn.comb}
\usage{
urn.comb(x)
}
\arguments{
\item{x}{= a matrix of integers}
}
\description{
various combinations
}
\seealso{
Other urn: \code{\link{urn.exact}}, \code{\link{urn.least}}
}
\keyword{urn.comb}
