% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{urn.exact}
\alias{urn.exact}
\title{urn.exact}
\usage{
urn.exact(x, y)
}
\arguments{
\item{x}{= a vector of integers}

\item{y}{= an isomekic vector of integers that is pointwise greater than or equal to <x>}
}
\description{
probability of drawing precisely <x> balls from an urn containing <y> balls
}
\keyword{urn.exact}
