% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{urn.least}
\alias{urn.least}
\title{urn.least}
\usage{
urn.least(x, y)
}
\arguments{
\item{x}{= a vector of integers}

\item{y}{= an isomekic vector of integers that is pointwise greater than or equal to <x>}
}
\description{
probability of drawing at least <x> balls from an urn containing <y> balls
}
\seealso{
Other urn: \code{\link{urn.comb}}, \code{\link{urn.exact}}
}
\keyword{urn.least}
