% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{utf8.to.quoted.printable}
\alias{utf8.to.quoted.printable}
\title{utf8.to.quoted.printable}
\usage{
utf8.to.quoted.printable(x)
}
\arguments{
\item{x}{= a single character}
}
\description{
quoted-printable representation of <x>
}
\keyword{utf8.to.quoted.printable}
