% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{variance.ratio.test}
\alias{variance.ratio.test}
\title{variance.ratio.test}
\usage{
variance.ratio.test(x, y)
}
\arguments{
\item{x}{= vector}

\item{y}{= an integer (greater than 1)}
}
\description{
tests whether <x> follows a random walk (i.e. <x> independent of prior values)
}
\keyword{variance.ratio.test}
