% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.count}
\alias{vec.count}
\title{vec.count}
\usage{
vec.count(x)
}
\arguments{
\item{x}{= a string vector}
}
\description{
Counts unique instances of <x>
}
\seealso{
Other vec: \code{\link{vec.cat}}, \code{\link{vec.cum}},
  \code{\link{vec.diff}}, \code{\link{vec.ex.filters}},
  \code{\link{vec.first}}, \code{\link{vec.lag}},
  \code{\link{vec.last}}, \code{\link{vec.max}},
  \code{\link{vec.min}}, \code{\link{vec.named}},
  \code{\link{vec.read}}, \code{\link{vec.same}},
  \code{\link{vec.swap}}, \code{\link{vec.to.list}},
  \code{\link{vec.unique}}
}
\keyword{vec.count}
