% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.diff}
\alias{vec.diff}
\title{vec.diff}
\usage{
vec.diff(x, y)
}
\arguments{
\item{x}{= a numeric vector}

\item{y}{= an integer}
}
\description{
difference between <x> and itself lagged <y>
}
\keyword{vec.diff}
