% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.ex.filters}
\alias{vec.ex.filters}
\title{vec.ex.filters}
\usage{
vec.ex.filters(x)
}
\arguments{
\item{x}{= either "sf" or "macro"}
}
\description{
SQL query where clauses associated with filters
}
\keyword{vec.ex.filters}
