% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.lag}
\alias{vec.lag}
\title{vec.lag}
\usage{
vec.lag(x, y)
}
\arguments{
\item{x}{= a vector indexed by time running FORWARDS}

\item{y}{= number of periods over which to lag}
}
\description{
<x> lagged <y> periods (simple positional lag)
}
\keyword{vec.lag}
