% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.last.element.increment}
\alias{vec.last.element.increment}
\title{vec.last.element.increment}
\usage{
vec.last.element.increment(x, y = 1)
}
\arguments{
\item{x}{= a numeric vector}

\item{y}{= increment (defaults to unity)}
}
\description{
increments last element of <x> by <y>
}
\seealso{
Other vec: \code{\link{vec.cat}}, \code{\link{vec.count}},
  \code{\link{vec.max}}, \code{\link{vec.min}},
  \code{\link{vec.named}}, \code{\link{vec.read}},
  \code{\link{vec.same}}, \code{\link{vec.swap}},
  \code{\link{vec.to.lags}}, \code{\link{vec.to.list}},
  \code{\link{vec.unique}}
}
\keyword{vec.last.element.increment}
