% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.last}
\alias{vec.last}
\title{vec.last}
\usage{
vec.last(x)
}
\arguments{
\item{x}{= a string vector}
}
\description{
latest element of a vector
}
\seealso{
Other vec: \code{\link{vec.cat}}, \code{\link{vec.count}},
  \code{\link{vec.cum}}, \code{\link{vec.diff}},
  \code{\link{vec.ex.filters}}, \code{\link{vec.first}},
  \code{\link{vec.lag}}, \code{\link{vec.max}},
  \code{\link{vec.min}}, \code{\link{vec.named}},
  \code{\link{vec.read}}, \code{\link{vec.swap}},
  \code{\link{vec.to.list}}, \code{\link{vec.unique}}
}
\keyword{vec.last}
