% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.max}
\alias{vec.max}
\title{vec.max}
\usage{
vec.max(x, y)
}
\arguments{
\item{x}{= a vector/matrix/dataframe}

\item{y}{= a number/vector or matrix/dataframe with the same dimensions as <x>}
}
\description{
Returns the piecewise maximum of <x> and <y>
}
\keyword{vec.max}
