% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.min}
\alias{vec.min}
\title{vec.min}
\usage{
vec.min(x, y)
}
\arguments{
\item{x}{= a vector/matrix/dataframe}

\item{y}{= a number/vector or matrix/dataframe with the same dimensions as <x>}
}
\description{
Returns the piecewise minimum of <x> and <y>
}
\keyword{vec.min}
