% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.named}
\alias{vec.named}
\title{vec.named}
\usage{
vec.named(x, y)
}
\arguments{
\item{x}{= a vector}

\item{y}{= an isomekic vector}
}
\description{
Returns a vector with values <x> and names <y>
}
\seealso{
Other vec: \code{\link{vec.cat}}, \code{\link{vec.count}},
  \code{\link{vec.last.element.increment}},
  \code{\link{vec.max}}, \code{\link{vec.min}},
  \code{\link{vec.read}}, \code{\link{vec.same}},
  \code{\link{vec.swap}}, \code{\link{vec.to.lags}},
  \code{\link{vec.to.list}}, \code{\link{vec.unique}}
}
\keyword{vec.named}
