% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.read}
\alias{vec.read}
\title{vec.read}
\usage{
vec.read(x)
}
\arguments{
\item{x}{= path to a named vector (no column headers)}
}
\description{
reads into a named vector
}
\seealso{
Other vec: \code{\link{vec.cat}}, \code{\link{vec.count}},
  \code{\link{vec.max}}, \code{\link{vec.min}},
  \code{\link{vec.named}}, \code{\link{vec.same}},
  \code{\link{vec.swap}}, \code{\link{vec.to.lags}},
  \code{\link{vec.to.list}}, \code{\link{vec.unique}},
  \code{\link{vec.zScore}}
}
\keyword{vec.read}
