% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.same}
\alias{vec.same}
\title{vec.same}
\usage{
vec.same(x, y)
}
\arguments{
\item{x}{= a vector}

\item{y}{= an isomekic vector}
}
\description{
T/F depending on whether <x> and <y> are identical
}
\keyword{vec.same}
