% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.swap}
\alias{vec.swap}
\title{vec.swap}
\usage{
vec.swap(x, y, n)
}
\arguments{
\item{x}{= a vector}

\item{y}{= an integer between 1 and length(<x>)}

\item{n}{= an integer between 1 and length(<x>)}
}
\description{
swaps elements <y> and <n> of vector <x>
}
\keyword{vec.swap}
