% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.to.lags}
\alias{vec.to.lags}
\title{vec.to.lags}
\usage{
vec.to.lags(x, y)
}
\arguments{
\item{x}{= a numeric vector (time flows forward)}

\item{y}{= number of lagged values desired plus one}
}
\description{
a data frame of <x> together with itself lagged 1, ..., <y> - 1 times
}
\keyword{vec.to.lags}
