% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.to.list}
\alias{vec.to.list}
\title{vec.to.list}
\usage{
vec.to.list(x, y = F)
}
\arguments{
\item{x}{= string vector}

\item{y}{= T/F depending on whether to use <x> as group vector}
}
\description{
list object
}
\keyword{vec.to.list}
