% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.write}
\alias{vec.write}
\title{vec.write}
\usage{
vec.write(x, y)
}
\arguments{
\item{x}{= a named vector}

\item{y}{= path to a vector}
}
\description{
writes <x> to <y>
}
\seealso{
Other vec: \code{\link{vec.cat}}, \code{\link{vec.count}},
  \code{\link{vec.cum}}, \code{\link{vec.diff}},
  \code{\link{vec.ex.filters}}, \code{\link{vec.lag}},
  \code{\link{vec.max}}, \code{\link{vec.min}},
  \code{\link{vec.named}}, \code{\link{vec.read}},
  \code{\link{vec.same}}, \code{\link{vec.swap}},
  \code{\link{vec.to.list}}, \code{\link{vec.unique}}
}
\keyword{vec.write}
