% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.zScore}
\alias{vec.zScore}
\title{vec.zScore}
\usage{
vec.zScore(x, y, n)
}
\arguments{
\item{x}{= a vector}

\item{y}{= a 1/0 membership vector}

\item{n}{= a vector of groups (e.g. GSec)}
}
\description{
zScores <x>
}
\keyword{vec.zScore}
