% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.zScore.underlying}
\alias{vec.zScore.underlying}
\title{vec.zScore.underlying}
\usage{
vec.zScore.underlying(x)
}
\arguments{
\item{x}{= a two-column data frame without NAs. The first column is numeric and the second logical.}
}
\description{
zScores the first column of <x> weighting only the second column
}
\seealso{
Other vec: \code{\link{vec.cat}}, \code{\link{vec.count}},
  \code{\link{vec.max}}, \code{\link{vec.min}},
  \code{\link{vec.named}}, \code{\link{vec.read}},
  \code{\link{vec.same}}, \code{\link{vec.swap}},
  \code{\link{vec.to.lags}}, \code{\link{vec.to.list}},
  \code{\link{vec.unique}}, \code{\link{vec.zScore}}
}
\keyword{vec.zScore.underlying}
