% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{weekday.to.name}
\alias{weekday.to.name}
\title{weekday.to.name}
\usage{
weekday.to.name(x)
}
\arguments{
\item{x}{= a vector of numbers between 0 and 6}
}
\description{
Converts to 0 = Sun, 1 = Mon, ..., 6 = Sat
}
\keyword{weekday.to.name}
