% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyy.ex.period}
\alias{yyyy.ex.period}
\title{yyyy.ex.period}
\usage{
yyyy.ex.period(x, y)
}
\arguments{
\item{x}{= vector of trade dates}

\item{y}{= return window in days or months depending on whether <x> is YYYYMMDD or YYYYMM}
}
\description{
the year in which the return window ends
}
\keyword{yyyy.ex.period}
