% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyy.ex.yy}
\alias{yyyy.ex.yy}
\title{yyyy.ex.yy}
\usage{
yyyy.ex.yy(x)
}
\arguments{
\item{x}{= an integer vector (non-negative!)}
}
\description{
returns a numeric vector of YYYY
}
\seealso{
Other yyyy: \code{\link{yyyy.ex.period}},
  \code{\link{yyyy.periods.count}}
}
\keyword{yyyy.ex.yy}
