% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyy.periods.count}
\alias{yyyy.periods.count}
\title{yyyy.periods.count}
\usage{
yyyy.periods.count(x)
}
\arguments{
\item{x}{= a string vector}
}
\description{
the number of periods that typically fall in a year
}
\seealso{
Other yyyy: \code{\link{yyyy.ex.period}},
  \code{\link{yyyy.ex.yy}}
}
\keyword{yyyy.periods.count}
