% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymm.ex.int}
\alias{yyyymm.ex.int}
\title{yyyymm.ex.int}
\usage{
yyyymm.ex.int(x)
}
\arguments{
\item{x}{= a vector of integers}
}
\description{
returns a vector of <yyyymm> months
}
\keyword{yyyymm.ex.int}
