% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymm.ex.qtr}
\alias{yyyymm.ex.qtr}
\title{yyyymm.ex.qtr}
\usage{
yyyymm.ex.qtr(x, y = 3)
}
\arguments{
\item{x}{= a vector of quarters}

\item{y}{= month, in the quarter, to return (defaults to the third)}
}
\description{
returns a specific yyyymm within the quarter
}
\keyword{yyyymm.ex.qtr}
