% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymm.exists}
\alias{yyyymm.exists}
\title{yyyymm.exists}
\usage{
yyyymm.exists(x)
}
\arguments{
\item{x}{= a string vector}
}
\description{
T if <x> is a month expressed in YYYYMM format
}
\keyword{yyyymm.exists}
