% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymm.seq}
\alias{yyyymm.seq}
\title{yyyymm.seq}
\usage{
yyyymm.seq(x, y, n = 1)
}
\arguments{
\item{x}{= a YYYYMM or YYYYMMDD}

\item{y}{= an isotypic element}

\item{n}{= quantum size in YYYYMM or YYYYMMDD}
}
\description{
returns a sequence of YYYYMM or YYYYMMDD between (and including) x and y
}
\seealso{
Other yyyymm: \code{\link{yyyymm.diff}},
  \code{\link{yyyymm.ex.int}}, \code{\link{yyyymm.ex.qtr}},
  \code{\link{yyyymm.lag}}, \code{\link{yyyymm.to.day}},
  \code{\link{yyyymm.to.int}}, \code{\link{yyyymm.to.qtr}},
  \code{\link{yyyymm.to.yyyy}}
}
\keyword{yyyymm.seq}
