% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymm.to.int}
\alias{yyyymm.to.int}
\title{yyyymm.to.int}
\usage{
yyyymm.to.int(x)
}
\arguments{
\item{x}{= a vector of <yyyymm> months}
}
\description{
returns a vector of integers
}
\keyword{yyyymm.to.int}
