% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymm.to.qtr}
\alias{yyyymm.to.qtr}
\title{yyyymm.to.qtr}
\usage{
yyyymm.to.qtr(x)
}
\arguments{
\item{x}{= a vector of yyyymm}
}
\description{
returns associated quarters
}
\seealso{
Other yyyymm: \code{\link{yyyymm.diff}},
  \code{\link{yyyymm.ex.int}}, \code{\link{yyyymm.ex.qtr}},
  \code{\link{yyyymm.exists}}, \code{\link{yyyymm.lag}},
  \code{\link{yyyymm.seq}}, \code{\link{yyyymm.to.day}},
  \code{\link{yyyymm.to.int}}, \code{\link{yyyymm.to.yyyy}}
}
\keyword{yyyymm.to.qtr}
