% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymm.to.yyyy}
\alias{yyyymm.to.yyyy}
\title{yyyymm.to.yyyy}
\usage{
yyyymm.to.yyyy(x)
}
\arguments{
\item{x}{= a vector of dates in yyyymm format}
}
\description{
Converts to yyyy years
}
\seealso{
Other yyyymm: \code{\link{yyyymm.diff}},
  \code{\link{yyyymm.ex.int}}, \code{\link{yyyymm.ex.qtr}},
  \code{\link{yyyymm.exists}}, \code{\link{yyyymm.lag}},
  \code{\link{yyyymm.seq}}, \code{\link{yyyymm.to.day}},
  \code{\link{yyyymm.to.int}}, \code{\link{yyyymm.to.qtr}}
}
\keyword{yyyymm.to.yyyy}
