% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.bulk}
\alias{yyyymmdd.bulk}
\title{yyyymmdd.bulk}
\usage{
yyyymmdd.bulk(x)
}
\arguments{
\item{x}{= a matrix/df indexed by YYYYMMDD}
}
\description{
Eliminates YYYYMMDD gaps
}
\keyword{yyyymmdd.bulk}
