% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.diff}
\alias{yyyymmdd.diff}
\title{yyyymmdd.diff}
\usage{
yyyymmdd.diff(x, y)
}
\arguments{
\item{x}{= a vector of weekdays}

\item{y}{= an isomekic vector of weekdays}
}
\description{
returns <x - y> in terms of weekdays
}
\keyword{yyyymmdd.diff}
