% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.ex.AllocMo}
\alias{yyyymmdd.ex.AllocMo}
\title{yyyymmdd.ex.AllocMo}
\usage{
yyyymmdd.ex.AllocMo(x)
}
\arguments{
\item{x}{= an object indexed by allocation months}
}
\description{
Returns an object indexed by flow dates
}
\keyword{yyyymmdd.ex.AllocMo}
