% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.ex.day}
\alias{yyyymmdd.ex.day}
\title{yyyymmdd.ex.day}
\usage{
yyyymmdd.ex.day(x)
}
\arguments{
\item{x}{= a vector of calendar dates}
}
\description{
Falls back to the closest weekday
}
\keyword{yyyymmdd.ex.day}
