% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.ex.txt}
\alias{yyyymmdd.ex.txt}
\title{yyyymmdd.ex.txt}
\usage{
yyyymmdd.ex.txt(x, y = "/", n = "MDY")
}
\arguments{
\item{x}{= a vector of dates in some format}

\item{y}{= separators used within <x>}

\item{n}{= order in which month, day and year are represented}
}
\description{
a vector of calendar dates in YYYYMMDD format
}
\seealso{
Other yyyymmdd: \code{\link{yyyymmdd.bulk}},
  \code{\link{yyyymmdd.diff}},
  \code{\link{yyyymmdd.ex.AllocMo}},
  \code{\link{yyyymmdd.ex.day}},
  \code{\link{yyyymmdd.ex.int}},
  \code{\link{yyyymmdd.ex.yyyymm}},
  \code{\link{yyyymmdd.exists}},
  \code{\link{yyyymmdd.lag}}, \code{\link{yyyymmdd.seq}},
  \code{\link{yyyymmdd.to.AllocMo.unique}},
  \code{\link{yyyymmdd.to.AllocMo}},
  \code{\link{yyyymmdd.to.CalYrDyOfWk}},
  \code{\link{yyyymmdd.to.int}},
  \code{\link{yyyymmdd.to.txt}},
  \code{\link{yyyymmdd.to.unity}},
  \code{\link{yyyymmdd.to.weekofmonth}},
  \code{\link{yyyymmdd.to.yyyymm}}
}
\keyword{yyyymmdd.ex.txt}
