% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.ex.yyyymm}
\alias{yyyymmdd.ex.yyyymm}
\title{yyyymmdd.ex.yyyymm}
\usage{
yyyymmdd.ex.yyyymm(x, y = T)
}
\arguments{
\item{x}{= a YYYYMM vector}

\item{y}{= a boolean (last/all weekdays)}
}
\description{
last/all weekdays in <x>
}
\seealso{
Other yyyymmdd: \code{\link{yyyymmdd.bulk}},
  \code{\link{yyyymmdd.diff}},
  \code{\link{yyyymmdd.ex.day}},
  \code{\link{yyyymmdd.ex.int}},
  \code{\link{yyyymmdd.ex.txt}},
  \code{\link{yyyymmdd.exists}},
  \code{\link{yyyymmdd.lag}}, \code{\link{yyyymmdd.seq}},
  \code{\link{yyyymmdd.to.AllocMo.unique}},
  \code{\link{yyyymmdd.to.AllocMo}},
  \code{\link{yyyymmdd.to.int}},
  \code{\link{yyyymmdd.to.txt}},
  \code{\link{yyyymmdd.to.weekofmonth}},
  \code{\link{yyyymmdd.to.yyyymm}}
}
\keyword{yyyymmdd.ex.yyyymm}
