% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.exists}
\alias{yyyymmdd.exists}
\title{yyyymmdd.exists}
\usage{
yyyymmdd.exists(x)
}
\arguments{
\item{x}{= a vector of calendar dates}
}
\description{
returns T if <x> is a weekday
}
\keyword{yyyymmdd.exists}
