% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.lag}
\alias{yyyymmdd.lag}
\title{yyyymmdd.lag}
\usage{
yyyymmdd.lag(x, y)
}
\arguments{
\item{x}{= a YYYYMMDD vector}

\item{y}{= an integer}
}
\description{
lags <x> by <y> weekdays
}
\keyword{yyyymmdd.lag}
