% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.seq}
\alias{yyyymmdd.seq}
\title{yyyymmdd.seq}
\usage{
yyyymmdd.seq(x, y, n = 1)
}
\arguments{
\item{x}{= a single weekday}

\item{y}{= a single weekday}

\item{n}{= a positive integer}
}
\description{
a sequence of weekdays starting at <x> and, if possible, ending at <y>
}
\keyword{yyyymmdd.seq}
