% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.to.AllocMo}
\alias{yyyymmdd.to.AllocMo}
\title{yyyymmdd.to.AllocMo}
\usage{
yyyymmdd.to.AllocMo(x, y = 23)
}
\arguments{
\item{x}{= a YYYYMMDD}

\item{y}{= calendar day allocations are known the next month}
}
\description{
Returns the month for which you need to get allocations Flows as of the 23rd of each month are known by the 24th. By this time allocations from #		:	the previous month are known
}
\keyword{yyyymmdd.to.AllocMo}
