% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.to.CalYrDyOfWk}
\alias{yyyymmdd.to.CalYrDyOfWk}
\title{yyyymmdd.to.CalYrDyOfWk}
\usage{
yyyymmdd.to.CalYrDyOfWk(x)
}
\arguments{
\item{x}{= a vector of dates in yyyymmdd format}
}
\description{
Converts to 0 = Sun, 1 = Mon, ..., 6 = Sat
}
\keyword{yyyymmdd.to.CalYrDyOfWk}
