% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.to.int}
\alias{yyyymmdd.to.int}
\title{yyyymmdd.to.int}
\usage{
yyyymmdd.to.int(x)
}
\arguments{
\item{x}{= a vector of YYYYMMDD}
}
\description{
Number of week days since Monday, 12/30/46
}
\keyword{yyyymmdd.to.int}
