% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.to.unity}
\alias{yyyymmdd.to.unity}
\title{yyyymmdd.to.unity}
\usage{
yyyymmdd.to.unity(x)
}
\arguments{
\item{x}{= a vector of dates in yyyymmdd format}
}
\description{
returns a vector of 1's corresponding to the length of <x>
}
\keyword{yyyymmdd.to.unity}
