% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.to.weekofmonth}
\alias{yyyymmdd.to.weekofmonth}
\title{yyyymmdd.to.weekofmonth}
\usage{
yyyymmdd.to.weekofmonth(x)
}
\arguments{
\item{x}{= a vector of dates in yyyymmdd format}
}
\description{
returns 1 if the date fell in the first week of the month, 2 if it fell in the second, etc.
}
\seealso{
Other yyyymmdd: \code{\link{yyyymmdd.bulk}},
  \code{\link{yyyymmdd.diff}},
  \code{\link{yyyymmdd.ex.AllocMo}},
  \code{\link{yyyymmdd.ex.day}},
  \code{\link{yyyymmdd.ex.int}},
  \code{\link{yyyymmdd.ex.txt}},
  \code{\link{yyyymmdd.ex.yyyymm}},
  \code{\link{yyyymmdd.exists}},
  \code{\link{yyyymmdd.lag}}, \code{\link{yyyymmdd.seq}},
  \code{\link{yyyymmdd.to.AllocMo}},
  \code{\link{yyyymmdd.to.CalYrDyOfWk}},
  \code{\link{yyyymmdd.to.int}},
  \code{\link{yyyymmdd.to.unity}},
  \code{\link{yyyymmdd.to.yyyymm}}
}
\keyword{yyyymmdd.to.weekofmonth}
