% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.to.yyyymm}
\alias{yyyymmdd.to.yyyymm}
\title{yyyymmdd.to.yyyymm}
\usage{
yyyymmdd.to.yyyymm(x, y = F)
}
\arguments{
\item{x}{= a vector of dates in yyyymmdd format}

\item{y}{= if T then falls back one month}
}
\description{
Converts to yyyymm format
}
\keyword{yyyymmdd.to.yyyymm}
