% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{zScore}
\alias{zScore}
\title{zScore}
\usage{
zScore(x)
}
\arguments{
\item{x}{= a numeric vector/matrix/data frame}
}
\description{
Converts <x>, if a numeric vector, or the rows of <x> otherwise, to a zScore
}
\keyword{zScore}
