% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{array.bind}
\alias{array.bind}
\title{array.bind}
\usage{
array.bind(x, y)
}
\arguments{
\item{x}{= an array}

\item{y}{= an array}
}
\description{
binds together <x> and <y> along the dimension they differ on
}
\seealso{
Other array: \code{\link{array.ex.list}},
  \code{\link{array.unlist}}
}
\keyword{array.bind}
