% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{array.ex.list}
\alias{array.ex.list}
\title{array.ex.list}
\usage{
array.ex.list(x, y, n)
}
\arguments{
\item{x}{= a list of mat objects or named vectors}

\item{y}{= T/F depending on whether you want union of rows}

\item{n}{= T/F depending on whether you want union of columns}
}
\description{
array
}
\seealso{
Other array: \code{\link{array.bind}},
  \code{\link{array.unlist}}
}
\keyword{array.ex.list}
