% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{array.unlist}
\alias{array.unlist}
\title{array.unlist}
\usage{
array.unlist(x, y)
}
\arguments{
\item{x}{= any numerical array}

\item{y}{= a vector of names for the columns of the output corresponding to the dimensions of <x>}
}
\description{
unlists the contents of an array
}
\seealso{
Other array: \code{\link{array.bind}},
  \code{\link{array.ex.list}}
}
\keyword{array.unlist}
