% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{avg.model}
\alias{avg.model}
\title{avg.model}
\usage{
avg.model(x)
}
\arguments{
\item{x}{= a numeric vector (results)}
}
\description{
constant-only (zero-variable) regression model
}
\seealso{
Other avg: \code{\link{avg.winsorized}},
  \code{\link{avg.wtd}}
}
\keyword{avg.model}
