% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{base.ex.int}
\alias{base.ex.int}
\title{base.ex.int}
\usage{
base.ex.int(x, y)
}
\arguments{
\item{x}{= a non-negative integer}

\item{y}{= a positive integer}
}
\description{
Expresses <x> in base <y>
}
\seealso{
Other base: \code{\link{base.to.int}}
}
\keyword{base.ex.int}
