% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{base.to.int}
\alias{base.to.int}
\title{base.to.int}
\usage{
base.to.int(x, y = 26)
}
\arguments{
\item{x}{= a vector of positive integers}

\item{y}{= a positive integer}
}
\description{
Evaluates the base <y> number <x>
}
\seealso{
Other base: \code{\link{base.ex.int}}
}
\keyword{base.to.int}
