% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bbk.bin.rets.summ}
\alias{bbk.bin.rets.summ}
\title{bbk.bin.rets.summ}
\usage{
bbk.bin.rets.summ(x, y, n = F)
}
\arguments{
\item{x}{= a matrix/data frame (rows indexed by time and columns indexed by bins)}

\item{y}{= a positive integer (number of rows covering a full year)}

\item{n}{= a boolean (report/ignore period count)}
}
\description{
Summarizes bin excess returns arithmetically
}
\seealso{
Other bbk: \code{\link{bbk.bin.rets.prd.summ}},
  \code{\link{bbk.bin.xRet}}, \code{\link{bbk.data}},
  \code{\link{bbk.drawdown}}, \code{\link{bbk.fwdRet}},
  \code{\link{bbk.holidays}}, \code{\link{bbk.matrix}},
  \code{\link{bbk.summ}}, \code{\link{bbk.turnover}},
  \code{\link{bbk}}
}
\keyword{bbk.bin.rets.summ}
