% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bbk.bin.rets.summ}
\alias{bbk.bin.rets.summ}
\title{bbk.bin.rets.summ}
\usage{
bbk.bin.rets.summ(x, y)
}
\arguments{
\item{x}{= a matrix/df with rows indexed by time and columns indexed by bins}

\item{y}{= number of rows of <x> needed to cover an entire calendar year}
}
\description{
Summarizes bin excess returns arithmetically
}
\seealso{
Other bbk: \code{\link{bbk.bin.rets.geom.summ}},
  \code{\link{bbk.bin.rets.prd.summ}},
  \code{\link{bbk.bin.xRet}}, \code{\link{bbk.data}},
  \code{\link{bbk.doW.bulk}}, \code{\link{bbk.drawdown}},
  \code{\link{bbk.fanChart}}, \code{\link{bbk.fwdRet}},
  \code{\link{bbk.histogram}}, \code{\link{bbk.holidays}},
  \code{\link{bbk.summ}}, \code{\link{bbk.turnover}},
  \code{\link{bbk}}
}
\keyword{bbk.bin.rets.summ}
