% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bbk.bin.xRet}
\alias{bbk.bin.xRet}
\title{bbk.bin.xRet}
\usage{
bbk.bin.xRet(x, y, n = 5, w = F, h = F, u)
}
\arguments{
\item{x}{= a matrix/data frame (predictors)}

\item{y}{= a matrix/data frame (forward returns)}

\item{n}{= an integer (number of bins)}

\item{w}{= a boolean (report/ignore universe return)}

\item{h}{= a boolean (do/don't provide detail)}

\item{u}{= a string vector (binning groups, can be missing)}
}
\description{
Returns equal weight bin returns through time
}
\seealso{
Other bbk: \code{\link{bbk.bin.rets.prd.summ}},
  \code{\link{bbk.bin.rets.summ}}, \code{\link{bbk.data}},
  \code{\link{bbk.drawdown}}, \code{\link{bbk.fwdRet}},
  \code{\link{bbk.holidays}}, \code{\link{bbk.matrix}},
  \code{\link{bbk.summ}}, \code{\link{bbk.turnover}},
  \code{\link{bbk}}
}
\keyword{bbk.bin.xRet}
