% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bbk.data}
\alias{bbk.data}
\title{bbk.data}
\usage{
bbk.data(x, y, n, w, h, u, v, g, r, s)
}
\arguments{
\item{x}{= predictor indexed by yyyymmdd or yyyymm}

\item{y}{= total return index indexed by the same date format as <x>}

\item{n}{= number of <prd.size>'s over which the predictor should be compounded/summed}

\item{w}{= T/F depending on whether <x> should be summed or compounded}

\item{h}{= predictor lag in days or months depending on whether <x> is YYYYMMDD or YYYYMM}

\item{u}{= delay in knowing data in days or months depending on whether <x> is YYYYMMDD or YYYYMM}

\item{v}{= day of the week you will trade on (5 = Fri)}

\item{g}{= return window in days or months depending on whether <x> is YYYYMMDD or YYYYMM}

\item{r}{= the index within which you are trading}

\item{s}{= T/F depending on whether spread changes, rather than returns, are needed}
}
\description{
fetches data required to compute standard model output
}
\seealso{
Other bbk: \code{\link{bbk.bin.rets.prd.summ}},
  \code{\link{bbk.bin.rets.summ}},
  \code{\link{bbk.bin.xRet}}, \code{\link{bbk.drawdown}},
  \code{\link{bbk.fwdRet}}, \code{\link{bbk.holidays}},
  \code{\link{bbk.matrix}}, \code{\link{bbk.summ}},
  \code{\link{bbk.turnover}}, \code{\link{bbk}}
}
\keyword{bbk.data}
