% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bbk.data}
\alias{bbk.data}
\title{bbk.data}
\usage{
bbk.data(x, y, floW, sum.flows, lag, delay, doW, retW, idx, prd.size)
}
\arguments{
\item{x}{= predictor indexed by yyyymmdd or yyyymm}

\item{y}{= total return index indexed by the same date format as <x>}

\item{floW}{= number of <prd.size>'s over which the predictor should be compounded/summed}

\item{sum.flows}{= T/F depending on whether <x> should be summed or compounded}

\item{lag}{= predictor lag in days or months depending on whether <x> is YYYYMMDD or YYYYMM}

\item{delay}{= delay in knowing data in days or months depending on whether <x> is YYYYMMDD or YYYYMM}

\item{doW}{= day of the week you will trade on (5 = Fri)}

\item{retW}{= return window in days or months depending on whether <x> is YYYYMMDD or YYYYMM}

\item{idx}{= the index within which you are trading}

\item{prd.size}{= size of each period in days or months depending on whether <x> is YYYYMMDD or YYYYMM}
}
\description{
fetches data required to compute standard model output
}
\seealso{
Other bbk: \code{\link{bbk.bin.rets.prd.summ}},
  \code{\link{bbk.bin.rets.summ}},
  \code{\link{bbk.bin.xRet}}, \code{\link{bbk.doW.bulk}},
  \code{\link{bbk.drawdown}}, \code{\link{bbk.fanChart}},
  \code{\link{bbk.fwdRet}}, \code{\link{bbk.histogram}},
  \code{\link{bbk.holidays}}, \code{\link{bbk.summ}},
  \code{\link{bbk.turnover}}, \code{\link{bbk}}
}
\keyword{bbk.data}
