% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bbk.drawdown}
\alias{bbk.drawdown}
\title{bbk.drawdown}
\usage{
bbk.drawdown(x)
}
\arguments{
\item{x}{= a numeric vector}
}
\description{
contiguous periods corresponding to max drawdown
}
\seealso{
Other bbk: \code{\link{bbk.bin.rets.prd.summ}},
  \code{\link{bbk.bin.rets.summ}},
  \code{\link{bbk.bin.xRet}}, \code{\link{bbk.data}},
  \code{\link{bbk.fwdRet}}, \code{\link{bbk.holidays}},
  \code{\link{bbk.matrix}}, \code{\link{bbk.summ}},
  \code{\link{bbk.turnover}}, \code{\link{bbk}}
}
\keyword{bbk.drawdown}
