% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bbk.fwdRet}
\alias{bbk.fwdRet}
\title{bbk.fwdRet}
\usage{
bbk.fwdRet(x, y, n, w, h)
}
\arguments{
\item{x}{= a matrix/data frame of predictors}

\item{y}{= a matrix/data frame of total return indices}

\item{n}{= the number of days in the return window}

\item{w}{= the number of days the predictors are lagged}

\item{h}{= the number of days needed for the predictors to be known}
}
\description{
returns a matrix/data frame of the same dimensions as <x>
}
\keyword{bbk.fwdRet}
