% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bbk.holidays}
\alias{bbk.holidays}
\title{bbk.holidays}
\usage{
bbk.holidays(x, y)
}
\arguments{
\item{x}{= a matrix/df of predictors, the rows of which are indexed by time}

\item{y}{= a matrix/df of the same dimensions as <x> containing associated forward returns}
}
\description{
Sets <x> to NA whenever <y> is NA
}
\seealso{
Other bbk: \code{\link{bbk.bin.rets.prd.summ}},
  \code{\link{bbk.bin.rets.summ}},
  \code{\link{bbk.bin.xRet}}, \code{\link{bbk.data}},
  \code{\link{bbk.doW.bulk}}, \code{\link{bbk.drawdown}},
  \code{\link{bbk.fanChart}}, \code{\link{bbk.fwdRet}},
  \code{\link{bbk.histogram}}, \code{\link{bbk.summ}},
  \code{\link{bbk.turnover}}, \code{\link{bbk}}
}
\keyword{bbk.holidays}
