% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bbk.matrix}
\alias{bbk.matrix}
\title{bbk.matrix}
\usage{
bbk.matrix(x, y, n, w = 5, h = 5, u = 5, v = F, g = 0, r = 2,
  s = NULL, b = F, p = "AnnMn")
}
\arguments{
\item{x}{= a matrix/data frame (predictors)}

\item{y}{= a matrix/data frame (total return indices)}

\item{n}{= an integer (flow window in days/months)}

\item{w}{= an integer (return window in days/months)}

\item{h}{= an integer (number of bins)}

\item{u}{= an integer (0 = Sun, 1 = Mon, etc., the day you trade)}

\item{v}{= a boolean (sum/compound)}

\item{g}{= an integer (lag in days/months)}

\item{r}{= an integer (delay in days/months)}

\item{s}{= a string (index within which you trade)}

\item{b}{= a boolean (spread changes/returns)}

\item{p}{= a string (e.g. AnnMn, Sharpe)}
}
\description{
standard model output summary value of <item> for "TxB" for various argument combinations
}
\seealso{
Other bbk: \code{\link{bbk.bin.rets.prd.summ}},
  \code{\link{bbk.bin.rets.summ}},
  \code{\link{bbk.bin.xRet}}, \code{\link{bbk.data}},
  \code{\link{bbk.drawdown}}, \code{\link{bbk.fwdRet}},
  \code{\link{bbk.holidays}}, \code{\link{bbk.summ}},
  \code{\link{bbk.turnover}}, \code{\link{bbk}}
}
\keyword{bbk.matrix}
