% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bbk}
\alias{bbk}
\title{bbk}
\usage{
bbk(x, y, n = 1, w = 5, h = 5, u = NULL, v = F, g = 0, r = 2,
  s = NULL, b = F)
}
\arguments{
\item{x}{= a matrix/data frame (predictors)}

\item{y}{= a matrix/data frame (total return indices)}

\item{n}{= a positive integer (flow window in days/months)}

\item{w}{= a positive integer (return window in days/months)}

\item{h}{= a positive integer (number of bins)}

\item{u}{= a non-negative integer (0 = Sun, 1 = Mon, etc., the day you trade)}

\item{v}{= a boolean (sum/compound)}

\item{g}{= a non-negative integer (lag in days/months)}

\item{r}{= a non-negative integer (delay in days/months)}

\item{s}{= a string (index within which you trade)}

\item{b}{= a boolean (spread changes/returns)}
}
\description{
standard model output
}
\seealso{
Other bbk: \code{\link{bbk.bin.rets.prd.summ}},
  \code{\link{bbk.bin.rets.summ}},
  \code{\link{bbk.bin.xRet}}, \code{\link{bbk.data}},
  \code{\link{bbk.drawdown}}, \code{\link{bbk.fwdRet}},
  \code{\link{bbk.holidays}}, \code{\link{bbk.matrix}},
  \code{\link{bbk.summ}}, \code{\link{bbk.turnover}}
}
\keyword{bbk}
