% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bbk.summ}
\alias{bbk.summ}
\title{bbk.summ}
\usage{
bbk.summ(x, y, n, w)
}
\arguments{
\item{x}{= bin returns}

\item{y}{= bin memberships}

\item{n}{= return window in days or months depending on whether <x> is YYYYMMDD or YYYYMM}

\item{w}{= quantum size (<n> is made up of non-overlapping windows of size <w>)}
}
\description{
summarizes by year and overall
}
\seealso{
Other bbk: \code{\link{bbk.bin.rets.prd.summ}},
  \code{\link{bbk.bin.rets.summ}},
  \code{\link{bbk.bin.xRet}}, \code{\link{bbk.data}},
  \code{\link{bbk.drawdown}}, \code{\link{bbk.fanChart}},
  \code{\link{bbk.fwdRet}}, \code{\link{bbk.histogram}},
  \code{\link{bbk.holidays}}, \code{\link{bbk.matrix}},
  \code{\link{bbk.turnover}}, \code{\link{bbk}}
}
\keyword{bbk.summ}
