% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bbk.summ}
\alias{bbk.summ}
\title{bbk.summ}
\usage{
bbk.summ(x, y, n, w)
}
\arguments{
\item{x}{= a matrix/data frame (bin returns)}

\item{y}{= a matrix/data frame (bin memberships)}

\item{n}{= a positive integer (return window in days/months)}

\item{w}{= a positive integer}
}
\description{
summarizes by year and overall
}
\seealso{
Other bbk: \code{\link{bbk.bin.rets.prd.summ}},
  \code{\link{bbk.bin.rets.summ}},
  \code{\link{bbk.bin.xRet}}, \code{\link{bbk.data}},
  \code{\link{bbk.drawdown.underlying}},
  \code{\link{bbk.drawdown}}, \code{\link{bbk.fwdRet}},
  \code{\link{bbk.holidays}}, \code{\link{bbk.matrix}},
  \code{\link{bbk.turnover}}, \code{\link{bbk}}
}
\keyword{bbk.summ}
