% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bbk.turnover}
\alias{bbk.turnover}
\title{bbk.turnover}
\usage{
bbk.turnover(x)
}
\arguments{
\item{x}{= a matrix/df of positive integers}
}
\description{
returns average name turnover per bin
}
\seealso{
Other bbk: \code{\link{bbk.bin.rets.prd.summ}},
  \code{\link{bbk.bin.rets.summ}},
  \code{\link{bbk.bin.xRet}}, \code{\link{bbk.data}},
  \code{\link{bbk.drawdown}}, \code{\link{bbk.fanChart}},
  \code{\link{bbk.fwdRet}}, \code{\link{bbk.histogram}},
  \code{\link{bbk.holidays}}, \code{\link{bbk.matrix}},
  \code{\link{bbk.summ}}, \code{\link{bbk}}
}
\keyword{bbk.turnover}
