% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bond.curve.expand}
\alias{bond.curve.expand}
\title{bond.curve.expand}
\usage{
bond.curve.expand(x)
}
\arguments{
\item{x}{= a numeric vector (interest rates)}
}
\description{
full yield curve
}
\seealso{
Other bond: \code{\link{bond.price}}
}
\keyword{bond.curve.expand}
