% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bond.price}
\alias{bond.price}
\title{bond.price}
\usage{
bond.price(x, y, n)
}
\arguments{
\item{x}{= numeric vector representing annual coupon rates}

\item{y}{= integer vector representing years to maturity}

\item{n}{= named vector of interest rates}
}
\description{
bond prices
}
\keyword{bond.price}
