% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{bond.price}
\alias{bond.price}
\title{bond.price}
\usage{
bond.price(x, y, n)
}
\arguments{
\item{x}{= a numeric vector (annual coupon rates)}

\item{y}{= an integer vector (years to maturity)}

\item{n}{= a numeric vector (interest rates)}
}
\description{
bond prices
}
\seealso{
Other bond: \code{\link{bond.curve.expand}}
}
\keyword{bond.price}
