% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{britten.jones.data}
\alias{britten.jones.data}
\title{britten.jones.data}
\usage{
britten.jones.data(x, y, n, w = NULL)
}
\arguments{
\item{x}{= a data frame of predictors}

\item{y}{= total return index of the same size as <x>}

\item{n}{= number of periods of forward returns used}

\item{w}{= the index within which you are trading}
}
\description{
returns data needed for a Britten-Jones analysis
}
\seealso{
Other britten: \code{\link{britten.jones.data.stack}},
  \code{\link{britten.jones}}
}
\keyword{britten.jones.data}
