% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{britten.jones.data.stack}
\alias{britten.jones.data.stack}
\title{britten.jones.data.stack}
\usage{
britten.jones.data.stack(rslt, df, retHz, prd.ret, n.beg, entity)
}
\arguments{
\item{rslt}{=}

\item{df}{=}

\item{retHz}{=}

\item{prd.ret}{=}

\item{n.beg}{=}

\item{entity}{=}
}
\description{
applies the Britten-Jones transformation to a subset and then stacks
}
\keyword{britten.jones.data.stack}
