% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{char.ex.int}
\alias{char.ex.int}
\title{char.ex.int}
\usage{
char.ex.int(x)
}
\arguments{
\item{x}{= a string of integers}
}
\description{
the characters whose ascii values correspond to <x>
}
\keyword{char.ex.int}
