% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{char.to.num}
\alias{char.to.num}
\title{char.to.num}
\usage{
char.to.num(x)
}
\arguments{
\item{x}{= a vector of strings}
}
\description{
coerces to numeric much more brutally than does as.numeric
}
\keyword{char.to.num}
