% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{classification.threshold}
\alias{classification.threshold}
\title{classification.threshold}
\usage{
classification.threshold(x, y)
}
\arguments{
\item{x}{= a 1/0 vector}

\item{y}{= a vector of predictors of the same length as <x>}
}
\description{
threshold value that causes fewest classification errors
}
\keyword{classification.threshold}
