% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{col.ex.int}
\alias{col.ex.int}
\title{col.ex.int}
\usage{
col.ex.int(x)
}
\arguments{
\item{x}{= a positive integer}
}
\description{
Returns the relevant excel column (1 = "A", 2 = "B", etc.)
}
\keyword{col.ex.int}
