% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{col.ex.int}
\alias{col.ex.int}
\title{col.ex.int}
\usage{
col.ex.int(x)
}
\arguments{
\item{x}{= a positive integer vector}
}
\description{
Returns the relevant excel column (1 = "A", 2 = "B", etc.)
}
\seealso{
Other col: \code{\link{col.lag}}, \code{\link{col.to.int}}
}
\keyword{col.ex.int}
