% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{col.lag}
\alias{col.lag}
\title{col.lag}
\usage{
col.lag(x, y)
}
\arguments{
\item{x}{= string representation of an excel column}

\item{y}{= an integer representing the desired column lag}
}
\description{
Lags <x> by <y> columns
}
\keyword{col.lag}
