% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{col.smallest}
\alias{col.smallest}
\title{col.smallest}
\usage{
col.smallest(x)
}
\arguments{
\item{x}{= a numeric matrix}
}
\description{
returns the column name of the smallest non-NA column for each row
}
\keyword{col.smallest}
