% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{col.to.int}
\alias{col.to.int}
\title{col.to.int}
\usage{
col.to.int(x)
}
\arguments{
\item{x}{= a vector of string representations of excel columns}
}
\description{
Returns the relevant associated integer (1 = "A", 2 = "B", etc.)
}
\keyword{col.to.int}
