% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{combinations.ex.int}
\alias{combinations.ex.int}
\title{combinations.ex.int}
\usage{
combinations.ex.int(x, y, n)
}
\arguments{
\item{x}{= a positive integer}

\item{y}{= a positive integer}

\item{n}{= a positive integer}
}
\description{
inverse of combinations.to.int; returns a logical vector #		:	of length <n>, <y> of which elements are T
}
\seealso{
Other combinations: \code{\link{combinations.next}},
  \code{\link{combinations.to.int}},
  \code{\link{combinations}}
}
\keyword{combinations.ex.int}
