% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{combinations.next}
\alias{combinations.next}
\title{combinations.next}
\usage{
combinations.next(x)
}
\arguments{
\item{x}{= a logical vector}
}
\description{
returns the next combination
}
\keyword{combinations.next}
