% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{combinations}
\alias{combinations}
\title{combinations}
\usage{
combinations(x, y)
}
\arguments{
\item{x}{= a vector}

\item{y}{= an integer between 1 and <length(x)>}
}
\description{
returns all possible combinations of <y> values of <x>
}
\keyword{combinations}
