% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{compound.flows}
\alias{compound.flows}
\title{compound.flows}
\usage{
compound.flows(x, y, n = F)
}
\arguments{
\item{x}{= a matrix/data frame}

\item{y}{= a positive integer (number of rows to compound/sum)}

\item{n}{= a boolean (sum/compound)}
}
\description{
compounded flows over <n> trailing periods indexed by last day in the flow window
}
\seealso{
Other compound: \code{\link{compound.sf}},
  \code{\link{compound}}
}
\keyword{compound.flows}
