% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{compound.sf}
\alias{compound.sf}
\title{compound.sf}
\usage{
compound.sf(x, y)
}
\arguments{
\item{x}{= a matrix/data frame}

\item{y}{= a boolean (sum/compound)}
}
\description{
compounds flows
}
\seealso{
Other compound: \code{\link{compound.flows}},
  \code{\link{compound}}
}
\keyword{compound.sf}
