% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{correl}
\alias{correl}
\title{correl}
\usage{
correl(x, y, n = T)
}
\arguments{
\item{x}{= a numeric vector/matrix/data frame}

\item{y}{= a numeric vector (can be missing)}

\item{n}{= a boolean (rank/regular correlations)}
}
\description{
the estimated correlation between <x> and <y> or the columns of <x>
}
\keyword{correl}
