% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{cpt.RgnSecJP}
\alias{cpt.RgnSecJP}
\title{cpt.RgnSecJP}
\usage{
cpt.RgnSecJP(x, y)
}
\arguments{
\item{x}{= an integer vector (sectors)}

\item{y}{= a country code vector}
}
\description{
makes Region-Sector groupings
}
\seealso{
Other cpt: \code{\link{cpt.RgnSec}}
}
\keyword{cpt.RgnSecJP}
