% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{Ctry.info}
\alias{Ctry.info}
\title{Ctry.info}
\usage{
Ctry.info(x, y)
}
\arguments{
\item{x}{= a vector of country codes}

\item{y}{= a column in the classif-ctry file}
}
\description{
handles the addition and removal of countries from an index
}
\examples{
Ctry.info("PK", "CtryNm")
}
\keyword{Ctry.info}
