% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{Ctry.msci.index.changes}
\alias{Ctry.msci.index.changes}
\title{Ctry.msci.index.changes}
\usage{
Ctry.msci.index.changes(x, y)
}
\arguments{
\item{x}{= a matrix/df of total returns indexed by the beginning #        :    of the period (trade date in yyyymmdd format)}

\item{y}{= an MSCI index such as ACWI/EAFE/EM}
}
\description{
handles the addition and removal of countries from an index
}
\keyword{Ctry.msci.index.changes}
