% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{Ctry.msci}
\alias{Ctry.msci}
\title{Ctry.msci}
\usage{
Ctry.msci(x)
}
\arguments{
\item{x}{= an index name such as ACWI/EAFE/EM}
}
\description{
Countries added or removed from the index in ascending order
}
\keyword{Ctry.msci}
