% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{day.ex.int}
\alias{day.ex.int}
\title{day.ex.int}
\usage{
day.ex.int(x)
}
\arguments{
\item{x}{= an integer vector}
}
\description{
the <x>th day after Thursday, January 1, 1970
}
\seealso{
Other day: \code{\link{day.ex.date}},
  \code{\link{day.lag}}, \code{\link{day.seq}},
  \code{\link{day.to.date}}, \code{\link{day.to.int}},
  \code{\link{day.to.weekday}}, \code{\link{day.to.week}}
}
\keyword{day.ex.int}
