% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{day.seq}
\alias{day.seq}
\title{day.seq}
\usage{
day.seq(x, y, n = 1)
}
\arguments{
\item{x}{= a single calendar date}

\item{y}{= a single calendar date}

\item{n}{= quantum size in calendar date}
}
\description{
returns a sequence of calendar dates between (and including) x and y
}
\keyword{day.seq}
