% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{day.to.int}
\alias{day.to.int}
\title{day.to.int}
\usage{
day.to.int(x)
}
\arguments{
\item{x}{= a vector of calendar dates}
}
\description{
number of days after Thursday, January 1, 1970
}
\seealso{
Other day: \code{\link{day.ex.date}},
  \code{\link{day.ex.int}}, \code{\link{day.lag}},
  \code{\link{day.seq}}, \code{\link{day.to.date}},
  \code{\link{day.to.weekday}}, \code{\link{day.to.week}}
}
\keyword{day.to.int}
