% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{day.to.week}
\alias{day.to.week}
\title{day.to.week}
\usage{
day.to.week(x, y)
}
\arguments{
\item{x}{= a calendar date vector}

\item{y}{= an integer (0 = Sun, 1 = Mon, etc., the day each week ends)}
}
\description{
maps days to weeks
}
\seealso{
Other day: \code{\link{day.ex.date}},
  \code{\link{day.ex.int}}, \code{\link{day.lag}},
  \code{\link{day.seq}}, \code{\link{day.to.date}},
  \code{\link{day.to.int}}, \code{\link{day.to.weekday}}
}
\keyword{day.to.week}
