% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{day.to.week}
\alias{day.to.week}
\title{day.to.week}
\usage{
day.to.week(x, y)
}
\arguments{
\item{x}{= a vector of calendar dates}

\item{y}{= an integer representing the day the week ends on 0 is Sun, 1 is Mon, ..., 6 is Sat}
}
\description{
maps days to weeks
}
\seealso{
Other day: \code{\link{day.ex.date}},
  \code{\link{day.ex.int}}, \code{\link{day.lag}},
  \code{\link{day.seq}}, \code{\link{day.to.date}},
  \code{\link{day.to.int}}, \code{\link{day.to.weekday}}
}
\keyword{day.to.week}
