% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{dir.all.files}
\alias{dir.all.files}
\title{dir.all.files}
\usage{
dir.all.files(x, y)
}
\arguments{
\item{x}{= a path such as "C:\\\\temp"}

\item{y}{= a string such as "*.txt"}
}
\description{
Returns all files in the folder including sub-directories
}
\seealso{
Other dir: \code{\link{dir.ensure}},
  \code{\link{dir.kill}}, \code{\link{dir.make}},
  \code{\link{dir.parameters}}, \code{\link{dir.parent}},
  \code{\link{dir.publications}}, \code{\link{dir.size}}
}
\keyword{dir.all.files}
