% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{dir.clear}
\alias{dir.clear}
\title{dir.clear}
\usage{
dir.clear(x, y)
}
\arguments{
\item{x}{= a path such as "C:\\\\temp"}

\item{y}{= a string such as "\\\\.txt"}
}
\description{
rids <x> of files of type <y>
}
\keyword{dir.clear}
