% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{dir.ensure}
\alias{dir.ensure}
\title{dir.ensure}
\usage{
dir.ensure(x)
}
\arguments{
\item{x}{= a vector of full file paths}
}
\description{
Creates necessary folders so files can be copied to <x>
}
\seealso{
Other dir: \code{\link{dir.all.files}},
  \code{\link{dir.clear}}, \code{\link{dir.kill}},
  \code{\link{dir.make}}, \code{\link{dir.parameters}},
  \code{\link{dir.parent}}, \code{\link{dir.publications}},
  \code{\link{dir.size}}
}
\keyword{dir.ensure}
