% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{dir.make}
\alias{dir.make}
\title{dir.make}
\usage{
dir.make(x)
}
\arguments{
\item{x}{= a vector of full folder paths}
}
\description{
creates folders <x>
}
\seealso{
Other dir: \code{\link{dir.all.files}},
  \code{\link{dir.clear}}, \code{\link{dir.ensure}},
  \code{\link{dir.kill}}, \code{\link{dir.parameters}},
  \code{\link{dir.parent}}, \code{\link{dir.publications}},
  \code{\link{dir.size}}
}
\keyword{dir.make}
