% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{dir.parameters}
\alias{dir.parameters}
\title{dir.parameters}
\usage{
dir.parameters(x)
}
\arguments{
\item{x}{= desired sub-folder}
}
\description{
returns full path to relevant parameters sub-folder
}
\seealso{
Other dir: \code{\link{dir.all.files}},
  \code{\link{dir.clear}}, \code{\link{dir.ensure}},
  \code{\link{dir.kill}}, \code{\link{dir.make}},
  \code{\link{dir.parent}}, \code{\link{dir.publications}},
  \code{\link{dir.size}}
}
\keyword{dir.parameters}
