% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{dir.parent}
\alias{dir.parent}
\title{dir.parent}
\usage{
dir.parent(x)
}
\arguments{
\item{x}{= a file vector}
}
\description{
returns paths to the parent directory
}
\seealso{
Other dir: \code{\link{dir.all.files}},
  \code{\link{dir.clear}}, \code{\link{dir.ensure}},
  \code{\link{dir.kill}}, \code{\link{dir.make}},
  \code{\link{dir.parameters}},
  \code{\link{dir.publications}}, \code{\link{dir.size}}
}
\keyword{dir.parent}
