% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{dir.publications}
\alias{dir.publications}
\title{dir.publications}
\usage{
dir.publications(x)
}
\arguments{
\item{x}{= desired sub-folder}
}
\description{
desired output directory for relevant publication
}
\keyword{dir.publications}
