% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{dup.code}
\alias{dup.code}
\title{dup.code}
\usage{
dup.code(x, y)
}
\arguments{
\item{x}{= a string vector}

\item{y}{= a string vector}
}
\description{
T/F depending on whether code is duplicated
}
\keyword{dup.code}
