% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{EHD}
\alias{EHD}
\title{EHD}
\usage{
EHD(x, y, n, w, h, u = NULL)
}
\arguments{
\item{x}{= input to or output of sql.connect}

\item{y}{= item (Flow/AssetsStart/AssetsEnd)}

\item{n}{= frequency (T/F for daily/weekly or D/W/M)}

\item{w}{= begin date in YYYYMMDD}

\item{h}{= end date in YYYYMMDD}

\item{u}{= vector of filters}
}
\description{
named vector of item between <w> and <h> sorted ascending
}
\keyword{EHD}
