% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{email}
\alias{email}
\title{email}
\usage{
email(x, y, n, w = "")
}
\arguments{
\item{x}{= the email address of the recipient}

\item{y}{= subject of the email}

\item{n}{= text of the email}

\item{w}{= path to an attachement}
}
\description{
emails <x>
}
\keyword{email}
