% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{email}
\alias{email}
\title{email}
\usage{
email(x, y, n, w = "", h = F)
}
\arguments{
\item{x}{= the email address of the recipient}

\item{y}{= subject of the email}

\item{n}{= text of the email}

\item{w}{= a vector of paths to attachement}

\item{h}{= T/F depending on whether you want to use html}
}
\description{
emails <x>
}
\keyword{email}
