% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{email}
\alias{email}
\title{email}
\usage{
email(x, y, n, w = "", h = F, u, v)
}
\arguments{
\item{x}{= the email address(es) of the recipient(s)}

\item{y}{= subject of the email}

\item{n}{= text of the email}

\item{w}{= a file vector}

\item{h}{= a boolean (use html/text)}

\item{u}{= the email address(es) being CC'ed}

\item{v}{= the email address(es) being BCC'ed}
}
\description{
emails <x>
}
\seealso{
Other email: \code{\link{email.exists}},
  \code{\link{email.kill}}, \code{\link{email.list}},
  \code{\link{email.record}}
}
\keyword{email}
