% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{err.raise.txt}
\alias{err.raise.txt}
\title{err.raise.txt}
\usage{
err.raise.txt(x, y, n)
}
\arguments{
\item{x}{= a vector}

\item{y}{= T/F depending on whether output goes on many lines}

\item{n}{= main line of error message}
}
\description{
error message
}
\seealso{
Other err: \code{\link{err.raise}}
}
\keyword{err.raise.txt}
