% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{extract.AnnMn.sf}
\alias{extract.AnnMn.sf}
\title{extract.AnnMn.sf}
\usage{
extract.AnnMn.sf(x, y)
}
\arguments{
\item{x}{= array (3D items/bins/parameters)}

\item{y}{= a string which must be one of AnnMn/AnnSd/Sharp/HitRate}
}
\description{
Subsets to "AnnMn" and re-labels columns
}
\seealso{
Other extract: \code{\link{extract.AnnMn.sf.wrapper}}
}
\keyword{extract.AnnMn.sf}
