% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{extract.AnnMn.sf.wrapper}
\alias{extract.AnnMn.sf.wrapper}
\title{extract.AnnMn.sf.wrapper}
\usage{
extract.AnnMn.sf.wrapper(x, y = "AnnMn")
}
\arguments{
\item{x}{= a list object, each element of which is a 3D object The first dimension has AnnMn/AnnSd/Sharp/HitRate The second dimension has bins Q1/Q2/Qna/Q3/Q4/Q5 The third dimension is some kind of parameter}

\item{y}{= a string which must be one of AnnMn/AnnSd/Sharp/HitRate}
}
\description{
Subsets to "AnnMn" and re-labels columns
}
\keyword{extract.AnnMn.sf.wrapper}
