% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.direct.super}
\alias{fcn.direct.super}
\title{fcn.direct.super}
\usage{
fcn.direct.super(x)
}
\arguments{
\item{x}{= a SINGLE function name}
}
\description{
names of all functions that directly depend on <x>
}
\seealso{
Other fcn: \code{\link{fcn.all.canonical}},
  \code{\link{fcn.all.roxygenize}},
  \code{\link{fcn.all.sub}}, \code{\link{fcn.all.super}},
  \code{\link{fcn.args.actual}},
  \code{\link{fcn.canonical}}, \code{\link{fcn.clean}},
  \code{\link{fcn.comments.parse}},
  \code{\link{fcn.dates.parse}}, \code{\link{fcn.date}},
  \code{\link{fcn.direct.sub}}, \code{\link{fcn.dir}},
  \code{\link{fcn.expressions.count}},
  \code{\link{fcn.extract.args}},
  \code{\link{fcn.extract.out}},
  \code{\link{fcn.indent.decrease}},
  \code{\link{fcn.indent.ignore}},
  \code{\link{fcn.indent.increase}},
  \code{\link{fcn.indent.proper}},
  \code{\link{fcn.indirect}}, \code{\link{fcn.lines.code}},
  \code{\link{fcn.lines.count}}, \code{\link{fcn.list}},
  \code{\link{fcn.mat.col}}, \code{\link{fcn.mat.num}},
  \code{\link{fcn.mat.vec}}, \code{\link{fcn.matrix}},
  \code{\link{fcn.nonNA}}, \code{\link{fcn.num.nonNA}},
  \code{\link{fcn.order}}, \code{\link{fcn.path}},
  \code{\link{fcn.roxygenize}}, \code{\link{fcn.sho}},
  \code{\link{fcn.simple}}, \code{\link{fcn.to.comments}},
  \code{\link{fcn.to.txt}}, \code{\link{fcn.vec.num}}
}
\keyword{fcn.direct.super}
