% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.mat.col}
\alias{fcn.mat.col}
\title{fcn.mat.col}
\usage{
fcn.mat.col(fcn, x, y, n)
}
\arguments{
\item{fcn}{= function mapping two vectors to a single value}

\item{x}{= a vector/matrix/dataframe}

\item{y}{= either missing or a numeric isomekic vector}

\item{n}{= T/F depending on whether inputs should be ranked}
}
\description{
applies <fcn> to the columns of <x> pairwise
}
\seealso{
Other fcn: \code{\link{fcn.all.canonical}},
  \code{\link{fcn.all.roxygenize}},
  \code{\link{fcn.all.sub}}, \code{\link{fcn.all.super}},
  \code{\link{fcn.args.actual}},
  \code{\link{fcn.canonical}}, \code{\link{fcn.clean}},
  \code{\link{fcn.comments.parse}},
  \code{\link{fcn.dates.parse}}, \code{\link{fcn.date}},
  \code{\link{fcn.direct.sub}},
  \code{\link{fcn.direct.super}}, \code{\link{fcn.dir}},
  \code{\link{fcn.extract.args}},
  \code{\link{fcn.extract.out}},
  \code{\link{fcn.indirect}},
  \code{\link{fcn.lines.count}}, \code{\link{fcn.mat.num}},
  \code{\link{fcn.mat.vec}}, \code{\link{fcn.matrix}},
  \code{\link{fcn.nonNA}}, \code{\link{fcn.num.nonNA}},
  \code{\link{fcn.order}}, \code{\link{fcn.path}},
  \code{\link{fcn.roxygenize}}, \code{\link{fcn.sho}},
  \code{\link{fcn.to.comments}}, \code{\link{fcn.to.txt}},
  \code{\link{fcn.vec.num}}
}
\keyword{fcn.mat.col}
