% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.mat.vec}
\alias{fcn.mat.vec}
\title{fcn.mat.vec}
\usage{
fcn.mat.vec(fcn, x, y, n)
}
\arguments{
\item{fcn}{= function mapping vector(s) to an isomekic vector}

\item{x}{= a vector/matrix/dataframe}

\item{y}{= a number/vector or matrix/dataframe with the same dimensions as <x>}

\item{n}{= T/F depending on whether you want <fcn> applied to columns or rows}
}
\description{
applies <fcn> to <x> if a vector or the columns/rows of <x> otherwise
}
\seealso{
Other fcn: \code{\link{fcn.all.canonical}},
  \code{\link{fcn.all.roxygenize}},
  \code{\link{fcn.all.sub}}, \code{\link{fcn.all.super}},
  \code{\link{fcn.args.actual}},
  \code{\link{fcn.canonical}}, \code{\link{fcn.clean}},
  \code{\link{fcn.comments.parse}},
  \code{\link{fcn.dates.parse}}, \code{\link{fcn.date}},
  \code{\link{fcn.direct.sub}},
  \code{\link{fcn.direct.super}}, \code{\link{fcn.dir}},
  \code{\link{fcn.expressions.count}},
  \code{\link{fcn.extract.args}},
  \code{\link{fcn.extract.out}},
  \code{\link{fcn.indent.decrease}},
  \code{\link{fcn.indent.else}},
  \code{\link{fcn.indent.ignore}},
  \code{\link{fcn.indent.increase}},
  \code{\link{fcn.indent.proper}},
  \code{\link{fcn.indirect}}, \code{\link{fcn.lines.code}},
  \code{\link{fcn.lines.count}}, \code{\link{fcn.list}},
  \code{\link{fcn.mat.col}}, \code{\link{fcn.mat.num}},
  \code{\link{fcn.matrix}}, \code{\link{fcn.nonNA}},
  \code{\link{fcn.num.nonNA}}, \code{\link{fcn.order}},
  \code{\link{fcn.path}}, \code{\link{fcn.roxygenize}},
  \code{\link{fcn.sho}}, \code{\link{fcn.simple}},
  \code{\link{fcn.to.comments}}, \code{\link{fcn.to.txt}},
  \code{\link{fcn.vec.grp}}, \code{\link{fcn.vec.num}}
}
\keyword{fcn.mat.vec}
