% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.mat.vec}
\alias{fcn.mat.vec}
\title{fcn.mat.vec}
\usage{
fcn.mat.vec(fcn, x, y, n)
}
\arguments{
\item{fcn}{= a function (vector to vector)}

\item{x}{= a numeric vector/matrix/data frame}

\item{y}{= a string vector/matrix/data frame}

\item{n}{= a boolean (apply to columns/rows)}
}
\description{
applies <fcn> to <x> if a numeric vector or the columns/rows of <x> otherwise
}
\seealso{
Other fcn: \code{\link{fcn.all.canonical}},
  \code{\link{fcn.all.roxygenize}},
  \code{\link{fcn.all.sub}}, \code{\link{fcn.all.super}},
  \code{\link{fcn.args.actual}},
  \code{\link{fcn.args.comment}},
  \code{\link{fcn.canonical}},
  \code{\link{fcn.comments.parse}},
  \code{\link{fcn.dates.parse}}, \code{\link{fcn.date}},
  \code{\link{fcn.direct.map}},
  \code{\link{fcn.direct.sub}},
  \code{\link{fcn.direct.super}}, \code{\link{fcn.dir}},
  \code{\link{fcn.expressions.count}},
  \code{\link{fcn.expressions.to.txt}},
  \code{\link{fcn.extract.args}},
  \code{\link{fcn.extract.out}}, \code{\link{fcn.has}},
  \code{\link{fcn.indent.proper}},
  \code{\link{fcn.lines.code}},
  \code{\link{fcn.lines.count}}, \code{\link{fcn.list}},
  \code{\link{fcn.lite}}, \code{\link{fcn.mat.col}},
  \code{\link{fcn.mat.num}}, \code{\link{fcn.nonNA}},
  \code{\link{fcn.num.nonNA}}, \code{\link{fcn.order}},
  \code{\link{fcn.pair.comment.wrapper}},
  \code{\link{fcn.pair.comment}}, \code{\link{fcn.path}},
  \code{\link{fcn.roxygenize}}, \code{\link{fcn.sho}},
  \code{\link{fcn.simple}}, \code{\link{fcn.to.comments}},
  \code{\link{fcn.to.txt}}, \code{\link{fcn.vec.grp}},
  \code{\link{fcn.vec.num}}
}
\keyword{fcn.mat.vec}
