% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.to.txt}
\alias{fcn.to.txt}
\title{fcn.to.txt}
\usage{
fcn.to.txt(x, y = F, n = F)
}
\arguments{
\item{x}{= a SINGLE function name}

\item{y}{= T/F vbl controlling whether comments are returned}

\item{n}{= T/F vbl controlling whether output is a string vector}
}
\description{
represents <x> as a string or string vector
}
\keyword{fcn.to.txt}
