% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fcn.vec.num}
\alias{fcn.vec.num}
\title{fcn.vec.num}
\usage{
fcn.vec.num(fcn, x, y)
}
\arguments{
\item{fcn}{= function mapping vector(s) to a single value}

\item{x}{= an element or vector}

\item{y}{= an element or isomekic vector}
}
\description{
applies <fcn> to <x>
}
\seealso{
Other fcn: \code{\link{fcn.all.canonical}},
  \code{\link{fcn.all.roxygenize}},
  \code{\link{fcn.all.sub}}, \code{\link{fcn.all.super}},
  \code{\link{fcn.args.actual}},
  \code{\link{fcn.canonical}}, \code{\link{fcn.clean}},
  \code{\link{fcn.comments.parse}},
  \code{\link{fcn.dates.parse}}, \code{\link{fcn.date}},
  \code{\link{fcn.direct.sub}},
  \code{\link{fcn.direct.super}}, \code{\link{fcn.dir}},
  \code{\link{fcn.extract.args}},
  \code{\link{fcn.extract.out}},
  \code{\link{fcn.indirect}},
  \code{\link{fcn.lines.count}}, \code{\link{fcn.mat.col}},
  \code{\link{fcn.mat.num}}, \code{\link{fcn.mat.vec}},
  \code{\link{fcn.matrix}}, \code{\link{fcn.nonNA}},
  \code{\link{fcn.num.nonNA}}, \code{\link{fcn.order}},
  \code{\link{fcn.path}}, \code{\link{fcn.roxygenize}},
  \code{\link{fcn.sho}}, \code{\link{fcn.to.comments}},
  \code{\link{fcn.to.txt}}
}
\keyword{fcn.vec.num}
