% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fetch}
\alias{fetch}
\title{fetch}
\usage{
fetch(x, y, n, w, h)
}
\arguments{
\item{x}{= a variable vector}

\item{y}{= a YYYYMM/flowdate}

\item{n}{= an integer (lookback)}

\item{w}{= R-object folder}

\item{h}{= classif file}
}
\description{
fetches <x> for the trailing <n> periods ending at <y>
}
\keyword{fetch}
