% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{file.time}
\alias{file.time}
\title{file.time}
\usage{
file.time(x)
}
\arguments{
\item{x}{= a file vector}
}
\description{
Returns the last modified date in yyyymmdd format
}
\seealso{
Other file: \code{\link{file.bkp}},
  \code{\link{file.date}},
  \code{\link{file.mtime.to.time}},
  \code{\link{file.to.last}}
}
\keyword{file.time}
