% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{find.data}
\alias{find.data}
\title{find.data}
\usage{
find.data(x, y = T)
}
\arguments{
\item{x}{= a logical vector}

\item{y}{= T/F depending on whether the position of the first/last true value of x is desired}
}
\description{
returns the position of the first/last true value of x
}
\keyword{find.data}
