% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{flowdate.diff}
\alias{flowdate.diff}
\title{flowdate.diff}
\usage{
flowdate.diff(x, y)
}
\arguments{
\item{x}{= a flowdate vector}

\item{y}{= a flowdate vector}
}
\description{
returns <x - y> in terms of flowdates
}
\seealso{
Other flowdate: \code{\link{flowdate.ex.AllocMo}},
  \code{\link{flowdate.ex.int}},
  \code{\link{flowdate.ex.yyyymm}},
  \code{\link{flowdate.exists}},
  \code{\link{flowdate.lag}}, \code{\link{flowdate.seq}},
  \code{\link{flowdate.to.int}}
}
\keyword{flowdate.diff}
