% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{flowdate.ex.int}
\alias{flowdate.ex.int}
\title{flowdate.ex.int}
\usage{
flowdate.ex.int(x)
}
\arguments{
\item{x}{= an integer or vector of integers}
}
\description{
the <x>th daily flow-publication date after Thursday, January 1, 1970
}
\seealso{
Other flowdate: \code{\link{flowdate.diff}},
  \code{\link{flowdate.ex.AllocMo}},
  \code{\link{flowdate.ex.yyyymm}},
  \code{\link{flowdate.exists}},
  \code{\link{flowdate.lag}}, \code{\link{flowdate.seq}},
  \code{\link{flowdate.to.int}}
}
\keyword{flowdate.ex.int}
