% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{flowdate.exists}
\alias{flowdate.exists}
\title{flowdate.exists}
\usage{
flowdate.exists(x)
}
\arguments{
\item{x}{= a vector of calendar dates}
}
\description{
returns T if <x> is a daily flow-publication date
}
\seealso{
Other flowdate: \code{\link{flowdate.diff}},
  \code{\link{flowdate.ex.int}},
  \code{\link{flowdate.ex.yyyymm}},
  \code{\link{flowdate.lag}}, \code{\link{flowdate.seq}},
  \code{\link{flowdate.to.int}}
}
\keyword{flowdate.exists}
