% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{flowdate.lag}
\alias{flowdate.lag}
\title{flowdate.lag}
\usage{
flowdate.lag(x, y)
}
\arguments{
\item{x}{= a vector of daily flow-publication dates}

\item{y}{= an integer}
}
\description{
lags <x> by <y> daily flow-publication dates
}
\keyword{flowdate.lag}
