% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{flowdate.seq}
\alias{flowdate.seq}
\title{flowdate.seq}
\usage{
flowdate.seq(x, y, n = 1)
}
\arguments{
\item{x}{= a single daily flow-publication date}

\item{y}{= a single daily flow-publication date}

\item{n}{= a positive integer}
}
\description{
a sequence of dly flow-pub dates starting at <x> and, if possible, ending at <y>
}
\seealso{
Other flowdate: \code{\link{flowdate.diff}},
  \code{\link{flowdate.ex.AllocMo}},
  \code{\link{flowdate.ex.int}},
  \code{\link{flowdate.ex.yyyymm}},
  \code{\link{flowdate.exists}},
  \code{\link{flowdate.lag}}, \code{\link{flowdate.to.int}}
}
\keyword{flowdate.seq}
