% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{flowdate.to.int}
\alias{flowdate.to.int}
\title{flowdate.to.int}
\usage{
flowdate.to.int(x)
}
\arguments{
\item{x}{= a vector of flow dates in YYYYMMDD format}
}
\description{
number of daily flow-publication dates after Friday, December 29, 2017
}
\seealso{
Other flowdate: \code{\link{flowdate.diff}},
  \code{\link{flowdate.ex.AllocMo}},
  \code{\link{flowdate.ex.int}},
  \code{\link{flowdate.ex.yyyymm}},
  \code{\link{flowdate.exists}},
  \code{\link{flowdate.lag}}, \code{\link{flowdate.seq}}
}
\keyword{flowdate.to.int}
