% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fop.IC.summ}
\alias{fop.IC.summ}
\title{fop.IC.summ}
\usage{
fop.IC.summ(x, y)
}
\arguments{
\item{x}{= a vector of IC's}

\item{y}{= an argument which is not used}
}
\description{
Summarizes IC's
}
\keyword{fop.IC.summ}
