% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fop.rank.xRet}
\alias{fop.rank.xRet}
\title{fop.rank.xRet}
\usage{
fop.rank.xRet(x, y)
}
\arguments{
\item{x}{= a matrix/df of predictors, the rows of which are indexed by time}

\item{y}{= an isomekic isoplatic matrix/df containing associated forward returns}
}
\description{
Ranks <x> only when <y> is available
}
\seealso{
Other fop: \code{\link{fop.Bin}},
  \code{\link{fop.IC.summ}}, \code{\link{fop.IC}},
  \code{\link{fop.grp.map}}, \code{\link{fop.stats}},
  \code{\link{fop.subset}}, \code{\link{fop.underlying}},
  \code{\link{fop}}
}
\keyword{fop.rank.xRet}
