% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fop.stats}
\alias{fop.stats}
\title{fop.stats}
\usage{
fop.stats(x, y, n)
}
\arguments{
\item{x}{= output of <fop>}

\item{y}{= a column in <x>}

\item{n}{= another column in <x> containing values of interest}
}
\description{
puts all the entries corresponding to <y> on one row
}
\seealso{
Other fop: \code{\link{fop.Bin}},
  \code{\link{fop.IC.summ}}, \code{\link{fop.IC}},
  \code{\link{fop.grp.map}}, \code{\link{fop.rank.xRet}},
  \code{\link{fop.subset}}, \code{\link{fop.wrapper}},
  \code{\link{fop}}
}
\keyword{fop.stats}
