% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fop.subset}
\alias{fop.subset}
\title{fop.subset}
\usage{
fop.subset(x, y = 100)
}
\arguments{
\item{x}{= output of <fop>}

\item{y}{= number}
}
\description{
Subsets to variations that have ALL combinations of Q1/TxB AnnMn/Sharpe in the <y>
}
\seealso{
Other fop: \code{\link{fop.Bin}},
  \code{\link{fop.IC.summ}}, \code{\link{fop.IC}},
  \code{\link{fop.grp.map}}, \code{\link{fop.rank.xRet}},
  \code{\link{fop.stats}}, \code{\link{fop.wrapper}},
  \code{\link{fop}}
}
\keyword{fop.subset}
