% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.all.dir}
\alias{ftp.all.dir}
\title{ftp.all.dir}
\usage{
ftp.all.dir(x, y, n, w, h, u)
}
\arguments{
\item{x}{= a remote folder}

\item{y}{= ftp site (can be missing)}

\item{n}{= user id (can be missing)}

\item{w}{= password (can be missing)}

\item{h}{= a string (ftp/sftp)}

\item{u}{= a boolean (ftp.use.epsv argument of getCurlHandle)}
}
\description{
remote-site directory listing of all sub-folders
}
\seealso{
Other ftp: \code{\link{ftp.all.files.underlying}},
  \code{\link{ftp.all.files}},
  \code{\link{ftp.credential}}, \code{\link{ftp.del}},
  \code{\link{ftp.dir.parse.ftp}},
  \code{\link{ftp.dir.parse.sftp}}, \code{\link{ftp.dir}},
  \code{\link{ftp.download}}, \code{\link{ftp.exists}},
  \code{\link{ftp.get}}, \code{\link{ftp.info}},
  \code{\link{ftp.kill}}, \code{\link{ftp.list}},
  \code{\link{ftp.missing.underlying}},
  \code{\link{ftp.missing}}, \code{\link{ftp.parent}},
  \code{\link{ftp.put}}, \code{\link{ftp.record}},
  \code{\link{ftp.rmdir}}, \code{\link{ftp.sql.factor}},
  \code{\link{ftp.upload}}
}
\keyword{ftp.all.dir}
