% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.all.files}
\alias{ftp.all.files}
\title{ftp.all.files}
\usage{
ftp.all.files(x, y, n, w, h, u)
}
\arguments{
\item{x}{= remote folder on an ftp site (e.g. "/ftpdata/mystuff")}

\item{y}{= ftp site (defaults to standard)}

\item{n}{= user id (defaults to standard)}

\item{w}{= password (defaults to standard)}

\item{h}{= protocol (either "ftp" or "sftp")}

\item{u}{= T/F flag for ftp.use.epsv argument of getCurlHandle}
}
\description{
remote-site directory listing of files (incl. sub-folders)
}
\seealso{
Other ftp: \code{\link{ftp.all.dir}},
  \code{\link{ftp.all.files.underlying}},
  \code{\link{ftp.break}}, \code{\link{ftp.credential}},
  \code{\link{ftp.del}}, \code{\link{ftp.dir.parse.77}},
  \code{\link{ftp.dir.parse.new}}, \code{\link{ftp.dir}},
  \code{\link{ftp.download}}, \code{\link{ftp.exists}},
  \code{\link{ftp.file}}, \code{\link{ftp.get}},
  \code{\link{ftp.info}}, \code{\link{ftp.kill}},
  \code{\link{ftp.list}}, \code{\link{ftp.missing}},
  \code{\link{ftp.parent}}, \code{\link{ftp.put}},
  \code{\link{ftp.record}}, \code{\link{ftp.rmdir}},
  \code{\link{ftp.sql.factor}},
  \code{\link{ftp.sql.other}}, \code{\link{ftp.upload}}
}
\keyword{ftp.all.files}
