% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.all.files.underlying}
\alias{ftp.all.files.underlying}
\title{ftp.all.files.underlying}
\usage{
ftp.all.files.underlying(x, y, n, w, h)
}
\arguments{
\item{x}{= remote folder on an ftp site (e.g. "/ftpdata/mystuff")}

\item{y}{= ftp site}

\item{n}{= user id}

\item{w}{= password}

\item{h}{= T/F depending on whether you want files or folders}
}
\description{
remote-site directory listing of files or folders
}
\seealso{
Other ftp: \code{\link{ftp.action}},
  \code{\link{ftp.all.dir}}, \code{\link{ftp.all.files}},
  \code{\link{ftp.credential}},
  \code{\link{ftp.delete.script.underlying}},
  \code{\link{ftp.delete.script}}, \code{\link{ftp.del}},
  \code{\link{ftp.dir.excise.crap}},
  \code{\link{ftp.dir.ftp.code}}, \code{\link{ftp.dir}},
  \code{\link{ftp.download}}, \code{\link{ftp.file.size}},
  \code{\link{ftp.get}}, \code{\link{ftp.info}},
  \code{\link{ftp.mkdir}}, \code{\link{ftp.parent}},
  \code{\link{ftp.put}}, \code{\link{ftp.remove}},
  \code{\link{ftp.rmdir}}, \code{\link{ftp.sql.factor}},
  \code{\link{ftp.sql.other}}, \code{\link{ftp.txt}},
  \code{\link{ftp.upload}}
}
\keyword{ftp.all.files.underlying}
