% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.credential}
\alias{ftp.credential}
\title{ftp.credential}
\usage{
ftp.credential(x, y = "ftp")
}
\arguments{
\item{x}{= one of ftp/user/pwd}

\item{y}{= one of ftp/sftp}
}
\description{
relevant ftp credential
}
\seealso{
Other ftp: \code{\link{ftp.all.dir}},
  \code{\link{ftp.all.files.underlying}},
  \code{\link{ftp.all.files}}, \code{\link{ftp.break}},
  \code{\link{ftp.del}}, \code{\link{ftp.dir}},
  \code{\link{ftp.download}}, \code{\link{ftp.exists}},
  \code{\link{ftp.file.size}}, \code{\link{ftp.file}},
  \code{\link{ftp.get}}, \code{\link{ftp.info}},
  \code{\link{ftp.is.dir}}, \code{\link{ftp.kill}},
  \code{\link{ftp.list}}, \code{\link{ftp.parent}},
  \code{\link{ftp.put}}, \code{\link{ftp.record}},
  \code{\link{ftp.rmdir}}, \code{\link{ftp.sql.factor}},
  \code{\link{ftp.sql.other}}, \code{\link{ftp.upload}}
}
\keyword{ftp.credential}
