% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.del}
\alias{ftp.del}
\title{ftp.del}
\usage{
ftp.del(x, y, n, w, h, u = "ftp")
}
\arguments{
\item{x}{= remote folder/file (e.g. "/ftpdata" or "/ftpdata/foo.txt")}

\item{y}{= a SINGLE file (e.g. "foo.txt") or missing if <x> is a file}

\item{n}{= ftp site (defaults to standard)}

\item{w}{= user id (defaults to standard)}

\item{h}{= password (defaults to standard)}

\item{u}{= protocol (either "ftp" or "sftp")}
}
\description{
deletes file <x> or file <y> on remote folder <x>
}
\keyword{ftp.del}
