% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.del}
\alias{ftp.del}
\title{ftp.del}
\usage{
ftp.del(x, y, n, w, h, u = 60)
}
\arguments{
\item{x}{= remote folder on an ftp site (e.g. "/ftpdata/mystuff")}

\item{y}{= a vector of remote file(s) (e.g. "foo.txt")}

\item{n}{= ftp site (defaults to standard)}

\item{w}{= user id (defaults to standard)}

\item{h}{= password (defaults to standard)}

\item{u}{= timeout in seconds}
}
\description{
deletes file <y> on remote site <x>
}
\seealso{
Other ftp: \code{\link{ftp.action}},
  \code{\link{ftp.all.dir}},
  \code{\link{ftp.all.files.underlying}},
  \code{\link{ftp.all.files}},
  \code{\link{ftp.credential}},
  \code{\link{ftp.dir.excise.crap}},
  \code{\link{ftp.dir.ftp.code}}, \code{\link{ftp.dir}},
  \code{\link{ftp.download}}, \code{\link{ftp.exists}},
  \code{\link{ftp.file.size}}, \code{\link{ftp.file}},
  \code{\link{ftp.get}}, \code{\link{ftp.info}},
  \code{\link{ftp.kill}}, \code{\link{ftp.list}},
  \code{\link{ftp.mkdir}}, \code{\link{ftp.parent}},
  \code{\link{ftp.put}}, \code{\link{ftp.record}},
  \code{\link{ftp.remove}}, \code{\link{ftp.rmdir}},
  \code{\link{ftp.sql.factor}},
  \code{\link{ftp.sql.other}}, \code{\link{ftp.txt}},
  \code{\link{ftp.upload}}
}
\keyword{ftp.del}
