% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.del}
\alias{ftp.del}
\title{ftp.del}
\usage{
ftp.del(x, y, n, w, h, u = "ftp")
}
\arguments{
\item{x}{= remote folder on an ftp site (e.g. "/ftpdata/mystuff")}

\item{y}{= a SINGLE remote file (e.g. "foo.txt")}

\item{n}{= ftp site (defaults to standard)}

\item{w}{= user id (defaults to standard)}

\item{h}{= password (defaults to standard)}

\item{u}{= protocol (either "ftp" or "sftp")}
}
\description{
deletes file <y> on remote site <x>
}
\seealso{
Other ftp: \code{\link{ftp.all.dir}},
  \code{\link{ftp.all.files.underlying}},
  \code{\link{ftp.all.files}}, \code{\link{ftp.break}},
  \code{\link{ftp.credential}},
  \code{\link{ftp.dir.parse.77}},
  \code{\link{ftp.dir.parse.new}}, \code{\link{ftp.dir}},
  \code{\link{ftp.download}}, \code{\link{ftp.exists}},
  \code{\link{ftp.file}}, \code{\link{ftp.get}},
  \code{\link{ftp.info}}, \code{\link{ftp.kill}},
  \code{\link{ftp.list}}, \code{\link{ftp.parent}},
  \code{\link{ftp.put}}, \code{\link{ftp.record}},
  \code{\link{ftp.rmdir}}, \code{\link{ftp.sql.factor}},
  \code{\link{ftp.sql.other}}, \code{\link{ftp.upload}}
}
\keyword{ftp.del}
