% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.dir.excise.crap}
\alias{ftp.dir.excise.crap}
\title{ftp.dir.excise.crap}
\usage{
ftp.dir.excise.crap(x, y, n)
}
\arguments{
\item{x}{= output from ftp directory listing}

\item{y}{= string demarcating the beginning of useful output}

\item{n}{= string demarcating the end of useful output}
}
\description{
cleans up output
}
\seealso{
Other ftp: \code{\link{ftp.all.dir}},
  \code{\link{ftp.all.files.underlying}},
  \code{\link{ftp.all.files}},
  \code{\link{ftp.delete.script.underlying}},
  \code{\link{ftp.delete.script}},
  \code{\link{ftp.dir.ftp.code}}, \code{\link{ftp.dir}},
  \code{\link{ftp.download.script}},
  \code{\link{ftp.file.size}},
  \code{\link{ftp.is.file.underlying}},
  \code{\link{ftp.is.file}}, \code{\link{ftp.put}},
  \code{\link{ftp.txt}},
  \code{\link{ftp.upload.script.underlying}},
  \code{\link{ftp.upload.script}}
}
\keyword{ftp.dir.excise.crap}
