% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.dir.excise.crap}
\alias{ftp.dir.excise.crap}
\title{ftp.dir.excise.crap}
\usage{
ftp.dir.excise.crap(x)
}
\arguments{
\item{x}{= output from ftp directory listing}
}
\description{
cleans up output
}
\seealso{
Other ftp: \code{\link{ftp.all.dir}},
  \code{\link{ftp.all.files}},
  \code{\link{ftp.dir.ftp.code}}, \code{\link{ftp.dir}},
  \code{\link{ftp.download.script}},
  \code{\link{ftp.file.size}}, \code{\link{ftp.put}},
  \code{\link{ftp.txt}}, \code{\link{ftp.upload.script}}
}
\keyword{ftp.dir.excise.crap}
