% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.dir.ftp.code}
\alias{ftp.dir.ftp.code}
\title{ftp.dir.ftp.code}
\usage{
ftp.dir.ftp.code(x, y, n, w)
}
\arguments{
\item{x}{= a folder on our ftp site}

\item{y}{= ftp site}

\item{n}{= user id}

\item{w}{= password}
}
\description{
generates ftp code for remote site directory listing
}
\keyword{ftp.dir.ftp.code}
