% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.dir}
\alias{ftp.dir}
\title{ftp.dir}
\usage{
ftp.dir(x, y, n, w)
}
\arguments{
\item{x}{= a folder on an ftp site}

\item{y}{= ftp site}

\item{n}{= user id}

\item{w}{= password}
}
\description{
remote site directory listing
}
\keyword{ftp.dir}
