% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.dir}
\alias{ftp.dir}
\title{ftp.dir}
\usage{
ftp.dir(x, y, n, w)
}
\arguments{
\item{x}{= remote folder on an ftp site (e.g. "/ftpdata/mystuff")}

\item{y}{= ftp site}

\item{n}{= user id}

\item{w}{= password}
}
\description{
remote site directory listing
}
\seealso{
Other ftp: \code{\link{ftp.all.files.in.dir}},
  \code{\link{ftp.dir.excise.crap}},
  \code{\link{ftp.dir.ftp.code}},
  \code{\link{ftp.download.script}},
  \code{\link{ftp.file.size}}, \code{\link{ftp.put}},
  \code{\link{ftp.txt}}
}
\keyword{ftp.dir}
