% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.exists}
\alias{ftp.exists}
\title{ftp.exists}
\usage{
ftp.exists(x, y)
}
\arguments{
\item{x}{= report name}

\item{y}{= date for which you want to send the report}
}
\description{
T/F depending on whether upload already happened
}
\seealso{
Other ftp: \code{\link{ftp.all.dir}},
  \code{\link{ftp.all.files.underlying}},
  \code{\link{ftp.all.files}}, \code{\link{ftp.break}},
  \code{\link{ftp.credential}}, \code{\link{ftp.del}},
  \code{\link{ftp.dir.parse.77}},
  \code{\link{ftp.dir.parse.new}}, \code{\link{ftp.dir}},
  \code{\link{ftp.download}}, \code{\link{ftp.file}},
  \code{\link{ftp.get}}, \code{\link{ftp.info}},
  \code{\link{ftp.kill}}, \code{\link{ftp.list}},
  \code{\link{ftp.missing}}, \code{\link{ftp.parent}},
  \code{\link{ftp.put}}, \code{\link{ftp.record}},
  \code{\link{ftp.rmdir}}, \code{\link{ftp.sql.factor}},
  \code{\link{ftp.sql.other}}, \code{\link{ftp.upload}}
}
\keyword{ftp.exists}
