% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.file.size}
\alias{ftp.file.size}
\title{ftp.file.size}
\usage{
ftp.file.size(x, y, n, w)
}
\arguments{
\item{x}{= a file on ftp site}

\item{y}{= ftp site}

\item{n}{= user id}

\item{w}{= password}
}
\description{
returns file size in KB
}
\seealso{
Other ftp: \code{\link{ftp.action}},
  \code{\link{ftp.all.dir}},
  \code{\link{ftp.all.files.underlying}},
  \code{\link{ftp.all.files}},
  \code{\link{ftp.credential}},
  \code{\link{ftp.delete.script.underlying}},
  \code{\link{ftp.delete.script}}, \code{\link{ftp.del}},
  \code{\link{ftp.dir.excise.crap}},
  \code{\link{ftp.dir.ftp.code}}, \code{\link{ftp.dir}},
  \code{\link{ftp.download}}, \code{\link{ftp.file}},
  \code{\link{ftp.get}}, \code{\link{ftp.info}},
  \code{\link{ftp.mkdir}}, \code{\link{ftp.parent}},
  \code{\link{ftp.put}}, \code{\link{ftp.remove}},
  \code{\link{ftp.rmdir}}, \code{\link{ftp.sql.factor}},
  \code{\link{ftp.sql.other}}, \code{\link{ftp.txt}},
  \code{\link{ftp.upload}}
}
\keyword{ftp.file.size}
