% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.file.size}
\alias{ftp.file.size}
\title{ftp.file.size}
\usage{
ftp.file.size(x, y, n, w)
}
\arguments{
\item{x}{= a file on ftp site}

\item{y}{= ftp site}

\item{n}{= user id}

\item{w}{= password}
}
\description{
returns file size in KB
}
\keyword{ftp.file.size}
