% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.file.size}
\alias{ftp.file.size}
\title{ftp.file.size}
\usage{
ftp.file.size(x, y, n, w, h = F)
}
\arguments{
\item{x}{= a file on ftp site}

\item{y}{= ftp site}

\item{n}{= user id}

\item{w}{= password}

\item{h}{= T/F flag for ftp.use.epsv argument of getURL}
}
\description{
returns file size in KB
}
\seealso{
Other ftp: \code{\link{ftp.all.dir}},
  \code{\link{ftp.all.files.underlying}},
  \code{\link{ftp.all.files}}, \code{\link{ftp.break}},
  \code{\link{ftp.credential}}, \code{\link{ftp.del}},
  \code{\link{ftp.dir}}, \code{\link{ftp.download}},
  \code{\link{ftp.exists}}, \code{\link{ftp.file}},
  \code{\link{ftp.get}}, \code{\link{ftp.info}},
  \code{\link{ftp.is.dir}}, \code{\link{ftp.kill}},
  \code{\link{ftp.list}}, \code{\link{ftp.parent}},
  \code{\link{ftp.put}}, \code{\link{ftp.record}},
  \code{\link{ftp.rmdir}}, \code{\link{ftp.sql.factor}},
  \code{\link{ftp.sql.other}}, \code{\link{ftp.upload}}
}
\keyword{ftp.file.size}
