% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.is.file}
\alias{ftp.is.file}
\title{ftp.is.file}
\usage{
ftp.is.file(x)
}
\arguments{
\item{x}{= a vector of file names}
}
\description{
T/F depending on whether each element is a file or folder
}
\seealso{
Other ftp: \code{\link{ftp.all.dir}},
  \code{\link{ftp.all.files}},
  \code{\link{ftp.delete.script.underlying}},
  \code{\link{ftp.delete.script}},
  \code{\link{ftp.dir.excise.crap}},
  \code{\link{ftp.dir.ftp.code}}, \code{\link{ftp.dir}},
  \code{\link{ftp.download.script}},
  \code{\link{ftp.file.size}}, \code{\link{ftp.put}},
  \code{\link{ftp.txt}}, \code{\link{ftp.upload.script}}
}
\keyword{ftp.is.file}
