% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.missing}
\alias{ftp.missing}
\title{ftp.missing}
\usage{
ftp.missing(x, y)
}
\arguments{
\item{x}{= arguments from higher function}

\item{y}{= argument names}
}
\description{
supplies missing arguments
}
\seealso{
Other ftp: \code{\link{ftp.all.dir}},
  \code{\link{ftp.all.files.underlying}},
  \code{\link{ftp.all.files}},
  \code{\link{ftp.credential}}, \code{\link{ftp.del}},
  \code{\link{ftp.dir.parse.ftp}},
  \code{\link{ftp.dir.parse.sftp}}, \code{\link{ftp.dir}},
  \code{\link{ftp.download}}, \code{\link{ftp.exists}},
  \code{\link{ftp.file}}, \code{\link{ftp.get}},
  \code{\link{ftp.info}}, \code{\link{ftp.kill}},
  \code{\link{ftp.list}},
  \code{\link{ftp.missing.underlying}},
  \code{\link{ftp.parent}}, \code{\link{ftp.put}},
  \code{\link{ftp.record}}, \code{\link{ftp.rmdir}},
  \code{\link{ftp.sql.factor}},
  \code{\link{ftp.sql.other}}, \code{\link{ftp.upload}}
}
\keyword{ftp.missing}
