% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.missing.underlying}
\alias{ftp.missing.underlying}
\title{ftp.missing.underlying}
\usage{
ftp.missing.underlying(x, y, n, w, h)
}
\arguments{
\item{x}{= ftp site (defaults to standard)}

\item{y}{= user id (defaults to standard)}

\item{n}{= password (defaults to standard)}

\item{w}{= a string (ftp/sftp)}

\item{h}{= a boolean (ftp.use.epsv argument of getURL)}
}
\description{
logical or YYYYMMDD vector indexed by remote file names
}
\seealso{
Other ftp: \code{\link{ftp.all.dir}},
  \code{\link{ftp.all.files.underlying}},
  \code{\link{ftp.all.files}},
  \code{\link{ftp.credential}}, \code{\link{ftp.del}},
  \code{\link{ftp.dir.parse.ftp}},
  \code{\link{ftp.dir.parse.sftp}}, \code{\link{ftp.dir}},
  \code{\link{ftp.download}}, \code{\link{ftp.exists}},
  \code{\link{ftp.file}}, \code{\link{ftp.get}},
  \code{\link{ftp.info}}, \code{\link{ftp.kill}},
  \code{\link{ftp.list}}, \code{\link{ftp.missing}},
  \code{\link{ftp.parent}}, \code{\link{ftp.put}},
  \code{\link{ftp.record}}, \code{\link{ftp.rmdir}},
  \code{\link{ftp.sql.factor}}, \code{\link{ftp.upload}}
}
\keyword{ftp.missing.underlying}
