% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.put}
\alias{ftp.put}
\title{ftp.put}
\usage{
ftp.put(x, y, n)
}
\arguments{
\item{x}{= name of the strategy}

\item{y}{= "daily" or "weekly"}

\item{n}{= location of the folder on the ftp server}
}
\description{
Writes ftp script to put the relevant file to the right folder
}
\keyword{ftp.put}
