% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.put}
\alias{ftp.put}
\title{ftp.put}
\usage{
ftp.put(x, y, n)
}
\arguments{
\item{x}{= name of the strategy}

\item{y}{= "daily" or "weekly"}

\item{n}{= location of the folder on the ftp server}
}
\description{
Writes ftp script to put the relevant file to the right folder
}
\seealso{
Other ftp: \code{\link{ftp.all.dir}},
  \code{\link{ftp.all.files.underlying}},
  \code{\link{ftp.all.files}},
  \code{\link{ftp.delete.script.underlying}},
  \code{\link{ftp.delete.script}},
  \code{\link{ftp.dir.excise.crap}},
  \code{\link{ftp.dir.ftp.code}}, \code{\link{ftp.dir}},
  \code{\link{ftp.download.script}},
  \code{\link{ftp.file.size}}, \code{\link{ftp.get}},
  \code{\link{ftp.info}},
  \code{\link{ftp.is.file.underlying}},
  \code{\link{ftp.is.file}}, \code{\link{ftp.sql.factor}},
  \code{\link{ftp.sql.other}}, \code{\link{ftp.txt}},
  \code{\link{ftp.upload.script.underlying}},
  \code{\link{ftp.upload.script}}
}
\keyword{ftp.put}
