% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.put}
\alias{ftp.put}
\title{ftp.put}
\usage{
ftp.put(x, y, n, w, h, u = "ftp", v)
}
\arguments{
\item{x}{= a remote folder}

\item{y}{= a file vector}

\item{n}{= ftp site (can be missing)}

\item{w}{= user id (can be missing)}

\item{h}{= password (can be missing)}

\item{u}{= a string (ftp/sftp)}

\item{v}{= a boolean (ftp.use.epsv argument of getCurlHandle)}
}
\description{
puts file <y> to remote site <x>, creating folders as needed
}
\seealso{
Other ftp: \code{\link{ftp.all.dir}},
  \code{\link{ftp.all.files.underlying}},
  \code{\link{ftp.all.files}},
  \code{\link{ftp.credential}}, \code{\link{ftp.del}},
  \code{\link{ftp.dir.parse.ftp}},
  \code{\link{ftp.dir.parse.sftp}}, \code{\link{ftp.dir}},
  \code{\link{ftp.download}}, \code{\link{ftp.exists}},
  \code{\link{ftp.get}}, \code{\link{ftp.info}},
  \code{\link{ftp.kill}}, \code{\link{ftp.list}},
  \code{\link{ftp.missing.underlying}},
  \code{\link{ftp.missing}}, \code{\link{ftp.parent}},
  \code{\link{ftp.record}}, \code{\link{ftp.rmdir}},
  \code{\link{ftp.sql.factor}}, \code{\link{ftp.upload}}
}
\keyword{ftp.put}
