% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.sql.factor}
\alias{ftp.sql.factor}
\title{ftp.sql.factor}
\usage{
ftp.sql.factor(x, y, n, w, h)
}
\arguments{
\item{x}{= vector of M/W/D depending on whether flows are monthly/weekly/daily}

\item{y}{= flow date in YYYYMMDD format}

\item{n}{= fund filter (e.g. Aggregate/Active/Passive/ETF/Mutual)}

\item{w}{= stock filter (e.g. All/China/Japan)}

\item{h}{= breakdown filter (e.g. All/GeoId/DomicileId)}
}
\description{
SQL code to validate <x> flows at the <y> level
}
\seealso{
Other ftp: \code{\link{ftp.all.dir}},
  \code{\link{ftp.all.files.underlying}},
  \code{\link{ftp.all.files}}, \code{\link{ftp.break}},
  \code{\link{ftp.credential}}, \code{\link{ftp.del}},
  \code{\link{ftp.dir}}, \code{\link{ftp.download}},
  \code{\link{ftp.exists}}, \code{\link{ftp.file.size}},
  \code{\link{ftp.file}}, \code{\link{ftp.get}},
  \code{\link{ftp.info}}, \code{\link{ftp.kill}},
  \code{\link{ftp.list}}, \code{\link{ftp.parent}},
  \code{\link{ftp.put}}, \code{\link{ftp.record}},
  \code{\link{ftp.rmdir}}, \code{\link{ftp.sql.other}},
  \code{\link{ftp.upload}}
}
\keyword{ftp.sql.factor}
