% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.sql.factor}
\alias{ftp.sql.factor}
\title{ftp.sql.factor}
\usage{
ftp.sql.factor(x, y, n, w, h, u = F)
}
\arguments{
\item{x}{= a variable}

\item{y}{= a flowdate}

\item{n}{= a filter (e.g. Aggregate/Active/Passive/ETF/Mutual)}

\item{w}{= DB - any of StockFlows/China/Japan/CSI300/Energy}

\item{h}{= a breakdown filter (All/GeoId/DomicileId, can be missing)}

\item{u}{= a boolean (chuck/keep securities held by just one fund)}
}
\description{
SQL code to validate <x> flows at the <y> level
}
\seealso{
Other ftp: \code{\link{ftp.all.dir}},
  \code{\link{ftp.all.files.underlying}},
  \code{\link{ftp.all.files}},
  \code{\link{ftp.credential}}, \code{\link{ftp.del}},
  \code{\link{ftp.dir.parse.ftp}},
  \code{\link{ftp.dir.parse.sftp}}, \code{\link{ftp.dir}},
  \code{\link{ftp.download}}, \code{\link{ftp.exists}},
  \code{\link{ftp.file}}, \code{\link{ftp.get}},
  \code{\link{ftp.info}}, \code{\link{ftp.kill}},
  \code{\link{ftp.list}},
  \code{\link{ftp.missing.underlying}},
  \code{\link{ftp.missing}}, \code{\link{ftp.parent}},
  \code{\link{ftp.put}}, \code{\link{ftp.record}},
  \code{\link{ftp.rmdir}}, \code{\link{ftp.upload}}
}
\keyword{ftp.sql.factor}
