% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.sql.factor}
\alias{ftp.sql.factor}
\title{ftp.sql.factor}
\usage{
ftp.sql.factor(x, y, n, w)
}
\arguments{
\item{x}{= vector of M/W/D depending on whether flows are monthly/weekly/daily}

\item{y}{= flow date in YYYYMMDD format}

\item{n}{= fund filter (e.g. Aggregate/Active/Passive/ETF/Mutual)}

\item{w}{= stock filter (e.g. All/China/Japan)}
}
\description{
SQL code to validate <x> flows at the <y> level
}
\seealso{
Other ftp: \code{\link{ftp.action}},
  \code{\link{ftp.all.dir}},
  \code{\link{ftp.all.files.underlying}},
  \code{\link{ftp.all.files}},
  \code{\link{ftp.credential}},
  \code{\link{ftp.delete.script.underlying}},
  \code{\link{ftp.delete.script}}, \code{\link{ftp.del}},
  \code{\link{ftp.dir.excise.crap}},
  \code{\link{ftp.dir.ftp.code}}, \code{\link{ftp.dir}},
  \code{\link{ftp.download}}, \code{\link{ftp.file.size}},
  \code{\link{ftp.get}}, \code{\link{ftp.info}},
  \code{\link{ftp.mkdir}}, \code{\link{ftp.parent}},
  \code{\link{ftp.put}}, \code{\link{ftp.remove}},
  \code{\link{ftp.rmdir}}, \code{\link{ftp.sql.other}},
  \code{\link{ftp.txt}}, \code{\link{ftp.upload}}
}
\keyword{ftp.sql.factor}
