% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.txt}
\alias{ftp.txt}
\title{ftp.txt}
\usage{
ftp.txt(x, y, n)
}
\arguments{
\item{x}{= ftp site}

\item{y}{= user id}

\item{n}{= password}
}
\description{
credentials needed to access ftp
}
\keyword{ftp.txt}
