% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.upload}
\alias{ftp.upload}
\title{ftp.upload}
\usage{
ftp.upload(x, y, n, w, h, u = "ftp", v)
}
\arguments{
\item{x}{= empty remote folder on an ftp site (e.g. "/ftpdata/mystuff")}

\item{y}{= local folder containing the data (e.g. "C:\\\\temp\\\\mystuff")}

\item{n}{= ftp site (defaults to standard)}

\item{w}{= user id (defaults to standard)}

\item{h}{= password (defaults to standard)}

\item{u}{= protocol (either "ftp" or "sftp")}

\item{v}{= T/F flag for ftp.use.epsv argument of getCurlHandle}
}
\description{
Copies up files from the local machine
}
\seealso{
Other ftp: \code{\link{ftp.all.dir}},
  \code{\link{ftp.all.files.underlying}},
  \code{\link{ftp.all.files}},
  \code{\link{ftp.credential}}, \code{\link{ftp.del}},
  \code{\link{ftp.dir.parse.ftp}},
  \code{\link{ftp.dir.parse.sftp}}, \code{\link{ftp.dir}},
  \code{\link{ftp.download}}, \code{\link{ftp.exists}},
  \code{\link{ftp.file}}, \code{\link{ftp.get}},
  \code{\link{ftp.info}}, \code{\link{ftp.kill}},
  \code{\link{ftp.list}},
  \code{\link{ftp.missing.underlying}},
  \code{\link{ftp.missing}}, \code{\link{ftp.parent}},
  \code{\link{ftp.put}}, \code{\link{ftp.record}},
  \code{\link{ftp.rmdir}}, \code{\link{ftp.sql.factor}}
}
\keyword{ftp.upload}
