% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ftp.upload.script}
\alias{ftp.upload.script}
\title{ftp.upload.script}
\usage{
ftp.upload.script(x, y, n, w, h)
}
\arguments{
\item{x}{= empty remote folder on an ftp site (e.g. "/ftpdata/mystuff")}

\item{y}{= local folder containing the data (e.g. "C:\\\\temp\\\\mystuff")}

\item{n}{= ftp site}

\item{w}{= user id}

\item{h}{= password}
}
\description{
returns ftp script to copy up files from the local machine
}
\seealso{
Other ftp: \code{\link{ftp.all.dir}},
  \code{\link{ftp.all.files}},
  \code{\link{ftp.delete.script.underlying}},
  \code{\link{ftp.delete.script}},
  \code{\link{ftp.dir.excise.crap}},
  \code{\link{ftp.dir.ftp.code}}, \code{\link{ftp.dir}},
  \code{\link{ftp.download.script}},
  \code{\link{ftp.file.size}}, \code{\link{ftp.is.file}},
  \code{\link{ftp.put}}, \code{\link{ftp.txt}},
  \code{\link{ftp.upload.script.underlying}}
}
\keyword{ftp.upload.script}
