% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fwd.probs.wrapper}
\alias{fwd.probs.wrapper}
\title{fwd.probs.wrapper}
\usage{
fwd.probs.wrapper(x, y, n, w, h, u, v, g, r)
}
\arguments{
\item{x}{= a matrix/data frame (predictors)}

\item{y}{= a matrix/data frame (total return indices)}

\item{n}{= a positive integer (flow window in days/months)}

\item{w}{= a boolean (sum/compound)}

\item{h}{= a non-negative integer (lag in days/months)}

\item{u}{= a non-negative integer (delay in days/months)}

\item{v}{= a non-negative integer (0 = Sun, 1 = Mon, etc., the day you trade)}

\item{g}{= a positive integer (return window in days/months)}

\item{r}{= a string (index within which you trade)}
}
\description{
probability that forward return is positive given predictor is positive
}
\seealso{
Other fwd: \code{\link{fwd.probs}}
}
\keyword{fwd.probs.wrapper}
