% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{fwd.probs.wrapper}
\alias{fwd.probs.wrapper}
\title{fwd.probs.wrapper}
\usage{
fwd.probs.wrapper(x, y, n, w, h, u, v, g, r)
}
\arguments{
\item{x}{= predictor indexed by yyyymmdd or yyyymm}

\item{y}{= total return index indexed by yyyymmdd or yyyymm}

\item{n}{= flow window in days}

\item{w}{= T/F depending on whether the predictor is to be summed or compounded}

\item{h}{= number of periods to lag the predictor}

\item{u}{= delay in knowing data}

\item{v}{= day of the week you will trade on (5 = Fri, NULL for monthlies)}

\item{g}{= a vector of forward return windows}

\item{r}{= the index within which you trade}
}
\description{
probability that forward return is positive given predictor is positive
}
\keyword{fwd.probs.wrapper}
