% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{html.and}
\alias{html.and}
\title{html.and}
\usage{
html.and(x)
}
\arguments{
\item{x}{= a string vector}
}
\description{
<x> stated in a grammatical phrase
}
\seealso{
Other html: \code{\link{html.flow.breakdown}},
  \code{\link{html.flow.english}},
  \code{\link{html.flow.underlying}},
  \code{\link{html.tbl}}
}
\keyword{html.and}
