% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{html.email}
\alias{html.email}
\title{html.email}
\usage{
html.email(x)
}
\arguments{
\item{x}{= a SINGLE flow date in YYYYMMDD format}
}
\description{
writes outgoing email report for <x>
}
\seealso{
Other html: \code{\link{html.and}},
  \code{\link{html.flow.breakdown}},
  \code{\link{html.flow.english}},
  \code{\link{html.flow.underlying}},
  \code{\link{html.image}}, \code{\link{html.list}},
  \code{\link{html.positioning}},
  \code{\link{html.problem}}, \code{\link{html.signature}},
  \code{\link{html.tbl}}, \code{\link{html.tenure}}
}
\keyword{html.email}
