% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{html.flow.breakdown}
\alias{html.flow.breakdown}
\title{html.flow.breakdown}
\usage{
html.flow.breakdown(x, y, n = 0)
}
\arguments{
\item{x}{= a named numeric vector}

\item{y}{= a string}

\item{n}{= a number representing miscellaneous flows}
}
\description{
html breaking down flows into constituents
}
\keyword{html.flow.breakdown}
