% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{html.flow.breakdown}
\alias{html.flow.breakdown}
\title{html.flow.breakdown}
\usage{
html.flow.breakdown(x, y, n = 0)
}
\arguments{
\item{x}{= a named numeric vector}

\item{y}{= a string}

\item{n}{= a number representing miscellaneous flows}
}
\description{
html breaking down flows into constituents
}
\seealso{
Other html: \code{\link{html.and}},
  \code{\link{html.flow.english}},
  \code{\link{html.flow.underlying}},
  \code{\link{html.image}}, \code{\link{html.list}},
  \code{\link{html.signature}}, \code{\link{html.tbl}}
}
\keyword{html.flow.breakdown}
