% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{html.flow.english}
\alias{html.flow.english}
\title{html.flow.english}
\usage{
html.flow.english(x, y, n, w)
}
\arguments{
\item{x}{= a named vector of integers (numbers need to be rounded)}

\item{y}{= a named text vector}

\item{n}{= line number(s) at which to insert a statement}

\item{w}{= statement(s) to be inserted}
}
\description{
writes a flow report in English
}
\seealso{
Other html: \code{\link{html.and}},
  \code{\link{html.email}},
  \code{\link{html.flow.breakdown}},
  \code{\link{html.flow.underlying}},
  \code{\link{html.image}}, \code{\link{html.list}},
  \code{\link{html.positioning}},
  \code{\link{html.problem}}, \code{\link{html.signature}},
  \code{\link{html.tbl}}, \code{\link{html.tenure}}
}
\keyword{html.flow.english}
