% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{html.flow.underlying}
\alias{html.flow.underlying}
\title{html.flow.underlying}
\usage{
html.flow.underlying(x)
}
\arguments{
\item{x}{= a numeric vector indexed by YYYYMMDD}
}
\description{
list object containing the following items: a) text - dates and text information about flows b) numbers - numeric summary of the flows
}
\seealso{
Other html: \code{\link{html.and}},
  \code{\link{html.flow.breakdown}},
  \code{\link{html.flow.english}},
  \code{\link{html.image}}, \code{\link{html.list}},
  \code{\link{html.signature}}, \code{\link{html.tbl}}
}
\keyword{html.flow.underlying}
