% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{html.image}
\alias{html.image}
\title{html.image}
\usage{
html.image(x, y)
}
\arguments{
\item{x}{= path to the image}

\item{y}{= percentage magnification}
}
\description{
html to attach an image
}
\seealso{
Other html: \code{\link{html.and}},
  \code{\link{html.email}},
  \code{\link{html.flow.breakdown}},
  \code{\link{html.flow.english}},
  \code{\link{html.flow.underlying}},
  \code{\link{html.list}}, \code{\link{html.positioning}},
  \code{\link{html.problem.underlying}},
  \code{\link{html.problem}}, \code{\link{html.signature}},
  \code{\link{html.tbl}}, \code{\link{html.tenure}}
}
\keyword{html.image}
