% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{html.positioning}
\alias{html.positioning}
\title{html.positioning}
\usage{
html.positioning(x, y)
}
\arguments{
\item{x}{= a matrix (predictors)}

\item{y}{= security names (correspond to columns of <x>, can be missing)}
}
\description{
writes a positioning report
}
\seealso{
Other html: \code{\link{html.and}},
  \code{\link{html.email}}, \code{\link{html.ex.utf8}},
  \code{\link{html.flow.breakdown}},
  \code{\link{html.flow.english}},
  \code{\link{html.flow.underlying}},
  \code{\link{html.image}}, \code{\link{html.list}},
  \code{\link{html.problem.underlying}},
  \code{\link{html.problem}}, \code{\link{html.signature}},
  \code{\link{html.tbl}}, \code{\link{html.tenure}}
}
\keyword{html.positioning}
