% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{html.problem}
\alias{html.problem}
\title{html.problem}
\usage{
html.problem(x, y, n)
}
\arguments{
\item{x}{= report names}

\item{y}{= string vector}

\item{n}{= logical vector (T = error, F = not, NA = no check)}
}
\description{
problem report
}
\keyword{html.problem}
