% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{html.problem.underlying}
\alias{html.problem.underlying}
\title{html.problem.underlying}
\usage{
html.problem.underlying(x, y, n)
}
\arguments{
\item{x}{= a string vector (report names)}

\item{y}{= a string vector}

\item{n}{= a boolean vector (error/no error, NA = no check)}
}
\description{
problem report
}
\seealso{
Other html: \code{\link{html.and}},
  \code{\link{html.email}}, \code{\link{html.ex.utf8}},
  \code{\link{html.flow.CountInOut}},
  \code{\link{html.flow.breakdown}},
  \code{\link{html.flow.english}},
  \code{\link{html.flow.underlying}},
  \code{\link{html.image}}, \code{\link{html.list}},
  \code{\link{html.positioning}},
  \code{\link{html.problem}}, \code{\link{html.signature}},
  \code{\link{html.tbl}}, \code{\link{html.tenure}}
}
\keyword{html.problem.underlying}
