% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{html.tbl}
\alias{html.tbl}
\title{html.tbl}
\usage{
html.tbl(x, y)
}
\arguments{
\item{x}{= matrix/data-frame}

\item{y}{= T/F depending on whether integer format is to be applied}
}
\description{
renders <x> in html
}
\seealso{
Other html: \code{\link{html.flow.english}},
  \code{\link{html.flow.underlying}}
}
\keyword{html.tbl}
