% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{html.tenure}
\alias{html.tenure}
\title{html.tenure}
\usage{
html.tenure(x, y, n)
}
\arguments{
\item{x}{= a named vector of integers}

\item{y}{= vector of length two for positive descriptions}

\item{n}{= vector of length two for negative descriptions}
}
\description{
describes how long securities/factors have belonged to a group
}
\seealso{
Other html: \code{\link{html.and}},
  \code{\link{html.email}},
  \code{\link{html.flow.breakdown}},
  \code{\link{html.flow.english}},
  \code{\link{html.flow.underlying}},
  \code{\link{html.image}}, \code{\link{html.list}},
  \code{\link{html.positioning}},
  \code{\link{html.problem.underlying}},
  \code{\link{html.problem}}, \code{\link{html.signature}},
  \code{\link{html.tbl}}
}
\keyword{html.tenure}
