% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{knapsack.count}
\alias{knapsack.count}
\title{knapsack.count}
\usage{
knapsack.count(x, y)
}
\arguments{
\item{x}{= a non-negative integer}

\item{y}{= a positive integer}
}
\description{
number of ways to subdivide <x> things amongst <y> people
}
\keyword{knapsack.count}
