% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{knapsack.ex.int}
\alias{knapsack.ex.int}
\title{knapsack.ex.int}
\usage{
knapsack.ex.int(x, y, n)
}
\arguments{
\item{x}{= a positive integer}

\item{y}{= a positive integer}

\item{n}{= a positive integer}
}
\description{
inverse of knapsack.to.int; returns a vector of length <n>, the elements of which sum to <y>
}
\keyword{knapsack.ex.int}
