% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{knapsack.next}
\alias{knapsack.next}
\title{knapsack.next}
\usage{
knapsack.next(x)
}
\arguments{
\item{x}{= an integer vector (non-negative!)}
}
\description{
next way to subdivide <sum(x)> things amongst <length(x)> people
}
\seealso{
Other knapsack: \code{\link{knapsack.count}},
  \code{\link{knapsack.ex.int}},
  \code{\link{knapsack.prev}},
  \code{\link{knapsack.to.int}}
}
\keyword{knapsack.next}
