% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{knapsack.prev}
\alias{knapsack.prev}
\title{knapsack.prev}
\usage{
knapsack.prev(x)
}
\arguments{
\item{x}{= a vector of non-negative integers}
}
\description{
inverse of knapsack.next
}
\keyword{knapsack.prev}
