% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{knapsack.prev}
\alias{knapsack.prev}
\title{knapsack.prev}
\usage{
knapsack.prev(x)
}
\arguments{
\item{x}{= an integer vector (non-negative!)}
}
\description{
inverse of knapsack.next
}
\seealso{
Other knapsack: \code{\link{knapsack.count}},
  \code{\link{knapsack.ex.int}},
  \code{\link{knapsack.next}},
  \code{\link{knapsack.to.int}}
}
\keyword{knapsack.prev}
