% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{knapsack.to.int}
\alias{knapsack.to.int}
\title{knapsack.to.int}
\usage{
knapsack.to.int(x)
}
\arguments{
\item{x}{= an integer vector (non-negative!)}
}
\description{
maps each particular way to subdivide <sum(x)> things #		:	amongst <length(x)> people to the number line
}
\seealso{
Other knapsack: \code{\link{knapsack.count}},
  \code{\link{knapsack.ex.int}},
  \code{\link{knapsack.next}}, \code{\link{knapsack.prev}},
  \code{\link{knapsack.process}}
}
\keyword{knapsack.to.int}
