% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{latin.to.arabic}
\alias{latin.to.arabic}
\title{latin.to.arabic}
\usage{
latin.to.arabic(x)
}
\arguments{
\item{x}{= a character vector}
}
\description{
<x> expressed as an arabic integer
}
\seealso{
Other latin: \code{\link{latin.ex.arabic}}
}
\keyword{latin.to.arabic}
