% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{load.dy.vbl.1obj}
\alias{load.dy.vbl.1obj}
\title{load.dy.vbl.1obj}
\usage{
load.dy.vbl.1obj(fcn, x, y, n, w, h, u)
}
\arguments{
\item{fcn}{= a function}

\item{x}{= a single YYYYMMDD}

\item{y}{= a single YYYYMMDD}

\item{n}{= passed down to <mk.fcn>}

\item{w}{= name under which the variable is to be stored}

\item{h}{= the YYYYMM for which the object is to be made}

\item{u}{= stock-flows environment}
}
\description{
Loads a daily variable
}
\keyword{load.dy.vbl.1obj}
