% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{load.dy.vbl}
\alias{load.dy.vbl}
\title{load.dy.vbl}
\usage{
load.dy.vbl(fcn, x, y, n, w, h, u)
}
\arguments{
\item{fcn}{= a function}

\item{x}{= a YYYYMMDD}

\item{y}{= a YYYYMMDD}

\item{n}{= passed down to <fcn>}

\item{w}{= name under which the variable is to be stored}

\item{h}{= R-object folder}

\item{u}{= stock-flows environment}
}
\description{
Loads a daily variable
}
\keyword{load.dy.vbl}
