% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{load.mo.vbl.1obj}
\alias{load.mo.vbl.1obj}
\title{load.mo.vbl.1obj}
\usage{
load.mo.vbl.1obj(fcn, x, y, n, w, h, u)
}
\arguments{
\item{fcn}{= a function}

\item{x}{= a single YYYYMM}

\item{y}{= a single YYYYMM}

\item{n}{= passed down to <mk.fcn>}

\item{w}{= name under which the variable is to be stored}

\item{h}{= the period for which the object is to be made}

\item{u}{= stock-flows environment}
}
\description{
Loads a monthly variable
}
\keyword{load.mo.vbl.1obj}
