% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{load.mo.vbl}
\alias{load.mo.vbl}
\title{load.mo.vbl}
\usage{
load.mo.vbl(fcn, x, y, n, w, h, u)
}
\arguments{
\item{fcn}{= a function}

\item{x}{= a YYYYMM}

\item{y}{= a YYYYMM}

\item{n}{= passed down to <fcn>}

\item{w}{= name under which the variable is to be stored}

\item{h}{= R-object folder}

\item{u}{= stock-flows environment}
}
\description{
Loads a monthly variable
}
\keyword{load.mo.vbl}
