% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{load.vbl.underlying}
\alias{load.vbl.underlying}
\title{load.vbl.underlying}
\usage{
load.vbl.underlying(fcn, x, y, n, w, h, u, v)
}
\arguments{
\item{fcn}{= a function}

\item{x}{= a YYYYMM/YYYYMMDD}

\item{y}{= a YYYYMM/YYYYMMDD}

\item{n}{= a string vector (passed down to <fcn>)}

\item{w}{= a string (name to store variable under)}

\item{h}{= R-object folder}

\item{u}{= a StockFlows environment}

\item{v}{= a boolean (daily/monthly)}
}
\description{
Loads a variable
}
\seealso{
Other load: \code{\link{load.dy.vbl.1obj}},
  \code{\link{load.dy.vbl}},
  \code{\link{load.mo.vbl.1obj}}, \code{\link{load.mo.vbl}}
}
\keyword{load.vbl.underlying}
