% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{map.rname}
\alias{map.rname}
\title{map.rname}
\usage{
map.rname(x, y)
}
\arguments{
\item{x}{= a vector/matrix/data-frame}

\item{y}{= a vector (usually string)}
}
\description{
returns a matrix/df, the row names of which match up with <y>
}
\keyword{map.rname}
