% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.correl}
\alias{mat.correl}
\title{mat.correl}
\usage{
mat.correl(x, y)
}
\arguments{
\item{x}{= a numeric vector/matrix/data frame}

\item{y}{= a numeric vector/matrix/data frame}
}
\description{
Returns the correlation of <x> & <y> if <x> is a numeric vector or those between the rows of <x> and <y> otherwise
}
\seealso{
Other mat: \code{\link{mat.compound}},
  \code{\link{mat.count}},
  \code{\link{mat.daily.to.monthly}},
  \code{\link{mat.daily.to.weekly}},
  \code{\link{mat.diff}}, \code{\link{mat.ex.array}},
  \code{\link{mat.ex.list}}, \code{\link{mat.ex.matrix}},
  \code{\link{mat.ex.vec}}, \code{\link{mat.index}},
  \code{\link{mat.lag}}, \code{\link{mat.last.to.first}},
  \code{\link{mat.periodic}}, \code{\link{mat.rank}},
  \code{\link{mat.read}}, \code{\link{mat.reverse}},
  \code{\link{mat.rollsum}}, \code{\link{mat.same}},
  \code{\link{mat.sort}}, \code{\link{mat.subset}},
  \code{\link{mat.to.last.Idx}}, \code{\link{mat.to.obs}},
  \code{\link{mat.to.xlModel}},
  \code{\link{mat.weekly.to.daily}},
  \code{\link{mat.write}}, \code{\link{mat.zScore}}
}
\keyword{mat.correl}
