% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.daily.to.weekly}
\alias{mat.daily.to.weekly}
\title{mat.daily.to.weekly}
\usage{
mat.daily.to.weekly(x, y)
}
\arguments{
\item{x}{= a matrix/df of daily data}

\item{y}{= an integer representing the day the week ends on #        :    0 is Sun, 1 is Mon, .., 6 is Sat}
}
\description{
returns latest data in each week in ascending order
}
\keyword{mat.daily.to.weekly}
