% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.ex.array}
\alias{mat.ex.array}
\title{mat.ex.array}
\usage{
mat.ex.array(x)
}
\arguments{
\item{x}{= an array}
}
\description{
a data frame with the first dimension forming the column space
}
\keyword{mat.ex.array}
