% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.ex.array3d}
\alias{mat.ex.array3d}
\title{mat.ex.array3d}
\usage{
mat.ex.array3d(x, y = 1:3)
}
\arguments{
\item{x}{= a three-dimensional numerical array}

\item{y}{= a vector of length 3}
}
\description{
unlists the contents of an array to a data frame
}
\seealso{
Other mat: \code{\link{mat.combine}},
  \code{\link{mat.compound}}, \code{\link{mat.correl}},
  \code{\link{mat.count}},
  \code{\link{mat.daily.to.monthly}},
  \code{\link{mat.daily.to.weekly}},
  \code{\link{mat.ex.array}}, \code{\link{mat.ex.matrix}},
  \code{\link{mat.ex.vec}}, \code{\link{mat.fake}},
  \code{\link{mat.index}}, \code{\link{mat.lag}},
  \code{\link{mat.last.to.first}}, \code{\link{mat.rank}},
  \code{\link{mat.read}}, \code{\link{mat.reverse}},
  \code{\link{mat.same}}, \code{\link{mat.subset}},
  \code{\link{mat.to.first.data.row}},
  \code{\link{mat.to.lags}}, \code{\link{mat.to.last.Idx}},
  \code{\link{mat.to.matrix}}, \code{\link{mat.to.obs}},
  \code{\link{mat.to.xlModel}}, \code{\link{mat.write}},
  \code{\link{mat.zScore}}
}
\keyword{mat.ex.array3d}
