% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.ex.matrix}
\alias{mat.ex.matrix}
\title{mat.ex.matrix}
\usage{
mat.ex.matrix(x, y = NULL)
}
\arguments{
\item{x}{= a matrix/data frame}

\item{y}{= a string vector (defaults to NULL)}
}
\description{
converts into a data frame
}
\seealso{
Other mat: \code{\link{mat.compound}},
  \code{\link{mat.correl}}, \code{\link{mat.count}},
  \code{\link{mat.daily.to.monthly}},
  \code{\link{mat.daily.to.weekly}},
  \code{\link{mat.diff}}, \code{\link{mat.ex.array}},
  \code{\link{mat.ex.list}}, \code{\link{mat.ex.vec}},
  \code{\link{mat.index}}, \code{\link{mat.lag}},
  \code{\link{mat.last.to.first}},
  \code{\link{mat.periodic}}, \code{\link{mat.rank}},
  \code{\link{mat.read}}, \code{\link{mat.reverse}},
  \code{\link{mat.rollsum}}, \code{\link{mat.same}},
  \code{\link{mat.sort}}, \code{\link{mat.subset}},
  \code{\link{mat.to.last.Idx}}, \code{\link{mat.to.obs}},
  \code{\link{mat.to.xlModel}},
  \code{\link{mat.weekly.to.daily}},
  \code{\link{mat.write}}, \code{\link{mat.zScore}}
}
\keyword{mat.ex.matrix}
