% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.lag.cols}
\alias{mat.lag.cols}
\title{mat.lag.cols}
\usage{
mat.lag.cols(x, y)
}
\arguments{
\item{x}{= a matrix/data frame of predictors}

\item{y}{= a vector of lags corresponding to each column}
}
\description{
lags columns of <x> individually based on <y>
}
\seealso{
Other mat: \code{\link{mat.combine}},
  \code{\link{mat.compound}}, \code{\link{mat.correl}},
  \code{\link{mat.count}},
  \code{\link{mat.daily.to.monthly}},
  \code{\link{mat.ex.array3d}}, \code{\link{mat.ex.array}},
  \code{\link{mat.ex.matrix}}, \code{\link{mat.ex.qtl}},
  \code{\link{mat.fake}}, \code{\link{mat.index}},
  \code{\link{mat.lag}}, \code{\link{mat.last.to.first}},
  \code{\link{mat.rank}}, \code{\link{mat.read}},
  \code{\link{mat.same}}, \code{\link{mat.subset}},
  \code{\link{mat.to.first.data.row}},
  \code{\link{mat.to.lags}}, \code{\link{mat.to.last.Idx}},
  \code{\link{mat.to.matrix}}, \code{\link{mat.to.obs}},
  \code{\link{mat.to.xlModel}}, \code{\link{mat.write}}
}
\keyword{mat.lag.cols}
