% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.last.to.first}
\alias{mat.last.to.first}
\title{mat.last.to.first}
\usage{
mat.last.to.first(x)
}
\arguments{
\item{x}{= a matrix/df}
}
\description{
Re-orders so the last column comes first
}
\keyword{mat.last.to.first}
