% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.rank}
\alias{mat.rank}
\title{mat.rank}
\usage{
mat.rank(x)
}
\arguments{
\item{x}{= a vector/matrix/data-frame}
}
\description{
ranks <x> if <x> is a vector or the rows of <x> otherwise
}
\seealso{
Other mat: \code{\link{mat.combine}},
  \code{\link{mat.compound}}, \code{\link{mat.correl}},
  \code{\link{mat.count}},
  \code{\link{mat.daily.to.monthly}},
  \code{\link{mat.daily.to.weekly}},
  \code{\link{mat.ex.array3d}}, \code{\link{mat.ex.array}},
  \code{\link{mat.ex.matrix}}, \code{\link{mat.ex.vec}},
  \code{\link{mat.fake}}, \code{\link{mat.index}},
  \code{\link{mat.lag}}, \code{\link{mat.last.to.first}},
  \code{\link{mat.read}}, \code{\link{mat.reverse}},
  \code{\link{mat.same}}, \code{\link{mat.subset}},
  \code{\link{mat.to.first.data.row}},
  \code{\link{mat.to.lags}}, \code{\link{mat.to.last.Idx}},
  \code{\link{mat.to.matrix}}, \code{\link{mat.to.obs}},
  \code{\link{mat.to.xlModel}}, \code{\link{mat.write}},
  \code{\link{mat.zScore}}
}
\keyword{mat.rank}
