% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.read}
\alias{mat.read}
\title{mat.read}
\usage{
mat.read(x = "C:\\\\temp\\\\write.csv", y = ",", n = 1, w = T)
}
\arguments{
\item{x}{= a path to a text file}

\item{y}{= the separator}

\item{n}{= the column containing the row names (or NULL if none)}

\item{w}{= T/F variable depending on whether <x> has a header}
}
\description{
reads the file into data frame
}
\seealso{
Other mat: \code{\link{mat.compound}},
  \code{\link{mat.correl}}, \code{\link{mat.count}},
  \code{\link{mat.daily.to.monthly}},
  \code{\link{mat.daily.to.weekly}},
  \code{\link{mat.ex.array3d}}, \code{\link{mat.ex.array}},
  \code{\link{mat.ex.matrix}}, \code{\link{mat.ex.vec}},
  \code{\link{mat.fake}}, \code{\link{mat.index}},
  \code{\link{mat.lag}}, \code{\link{mat.last.to.first}},
  \code{\link{mat.rank}}, \code{\link{mat.reverse}},
  \code{\link{mat.same}}, \code{\link{mat.subset}},
  \code{\link{mat.to.first.data.row}},
  \code{\link{mat.to.lags}}, \code{\link{mat.to.last.Idx}},
  \code{\link{mat.to.matrix}}, \code{\link{mat.to.obs}},
  \code{\link{mat.to.xlModel}}, \code{\link{mat.write}},
  \code{\link{mat.zScore}}
}
\keyword{mat.read}
