% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.read}
\alias{mat.read}
\title{mat.read}
\usage{
mat.read(x = "C:\\\\temp\\\\write.csv", y = ",", n = 1, w = T)
}
\arguments{
\item{x}{= a path to a text file}

\item{y}{= the separator}

\item{n}{= the column containing the row names (or NULL if none)}

\item{w}{= T/F variable depending on whether <x> has a header}
}
\description{
reads the file into data frame
}
\keyword{mat.read}
