% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.to.lags}
\alias{mat.to.lags}
\title{mat.to.lags}
\usage{
mat.to.lags(x, y, n = T, w = 1)
}
\arguments{
\item{x}{= a matrix/df indexed by time running FORWARDS}

\item{y}{= number of lagged values desired plus one}

\item{n}{= if T simple positional lagging is used. If F, yyyymm.lag is invoked}

\item{w}{= size of each period in terms of YYYYMMDD or YYYYMM depending on the rows of <x>}
}
\description{
a 3D array of <x> together with itself lagged 1, ..., <y> - 1 times
}
\seealso{
Other mat: \code{\link{mat.compound}},
  \code{\link{mat.correl}}, \code{\link{mat.count}},
  \code{\link{mat.daily.to.monthly}},
  \code{\link{mat.daily.to.weekly}},
  \code{\link{mat.diff}}, \code{\link{mat.ex.array}},
  \code{\link{mat.ex.matrix}}, \code{\link{mat.ex.vec}},
  \code{\link{mat.fake}}, \code{\link{mat.index}},
  \code{\link{mat.lag}}, \code{\link{mat.last.to.first}},
  \code{\link{mat.rank}}, \code{\link{mat.read}},
  \code{\link{mat.reverse}}, \code{\link{mat.rollsum}},
  \code{\link{mat.same}}, \code{\link{mat.sort}},
  \code{\link{mat.subset}},
  \code{\link{mat.to.first.data.row}},
  \code{\link{mat.to.last.Idx}}, \code{\link{mat.to.obs}},
  \code{\link{mat.to.xlModel}},
  \code{\link{mat.weekly.to.daily}},
  \code{\link{mat.write}}, \code{\link{mat.zScore}}
}
\keyword{mat.to.lags}
