% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.to.last.Idx}
\alias{mat.to.last.Idx}
\title{mat.to.last.Idx}
\usage{
mat.to.last.Idx(x)
}
\arguments{
\item{x}{= a matrix/data frame}
}
\description{
the last row index for which we have data
}
\seealso{
Other mat: \code{\link{mat.compound}},
  \code{\link{mat.correl}}, \code{\link{mat.count}},
  \code{\link{mat.daily.to.monthly}},
  \code{\link{mat.daily.to.weekly}},
  \code{\link{mat.diff}}, \code{\link{mat.ex.array}},
  \code{\link{mat.ex.list}}, \code{\link{mat.ex.matrix}},
  \code{\link{mat.ex.vec}}, \code{\link{mat.index}},
  \code{\link{mat.lag}}, \code{\link{mat.last.to.first}},
  \code{\link{mat.periodic}}, \code{\link{mat.rank}},
  \code{\link{mat.read}}, \code{\link{mat.reverse}},
  \code{\link{mat.rollsum}}, \code{\link{mat.sort}},
  \code{\link{mat.subset}}, \code{\link{mat.to.obs}},
  \code{\link{mat.to.xlModel}},
  \code{\link{mat.weekly.to.daily}},
  \code{\link{mat.write}}, \code{\link{mat.zScore}}
}
\keyword{mat.to.last.Idx}
