% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.to.last.Idx}
\alias{mat.to.last.Idx}
\title{mat.to.last.Idx}
\usage{
mat.to.last.Idx(x)
}
\arguments{
\item{x}{= a matrix/df}
}
\description{
the last row index for which we have data
}
\seealso{
Other mat: \code{\link{mat.combine}},
  \code{\link{mat.compound}}, \code{\link{mat.correl}},
  \code{\link{mat.count}},
  \code{\link{mat.daily.to.monthly}},
  \code{\link{mat.ex.array3d}}, \code{\link{mat.ex.array}},
  \code{\link{mat.ex.matrix}}, \code{\link{mat.ex.vec}},
  \code{\link{mat.fake}}, \code{\link{mat.index}},
  \code{\link{mat.lag}}, \code{\link{mat.last.to.first}},
  \code{\link{mat.rank}}, \code{\link{mat.read}},
  \code{\link{mat.reverse}}, \code{\link{mat.same}},
  \code{\link{mat.subset}},
  \code{\link{mat.to.first.data.row}},
  \code{\link{mat.to.lags}}, \code{\link{mat.to.matrix}},
  \code{\link{mat.to.obs}}, \code{\link{mat.to.xlModel}},
  \code{\link{mat.write}}, \code{\link{mat.zScore}}
}
\keyword{mat.to.last.Idx}
