% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.to.matrix}
\alias{mat.to.matrix}
\title{mat.to.matrix}
\usage{
mat.to.matrix(x)
}
\arguments{
\item{x}{= a matrix/data-frame with 3 columns corresponding respectively with the rows, columns and entries of the resulting matrix}
}
\description{
converts <x> to a matrix
}
\seealso{
Other mat: \code{\link{mat.combine}},
  \code{\link{mat.compound}}, \code{\link{mat.correl}},
  \code{\link{mat.count}},
  \code{\link{mat.daily.to.monthly}},
  \code{\link{mat.ex.array3d}},
  \code{\link{mat.ex.matrix}}, \code{\link{mat.ex.vec}},
  \code{\link{mat.fake}}, \code{\link{mat.index}},
  \code{\link{mat.lag.cols}}, \code{\link{mat.lag}},
  \code{\link{mat.last.to.first}}, \code{\link{mat.rank}},
  \code{\link{mat.read}}, \code{\link{mat.reverse}},
  \code{\link{mat.same}}, \code{\link{mat.subset}},
  \code{\link{mat.to.first.data.row}},
  \code{\link{mat.to.lags}}, \code{\link{mat.to.last.Idx}},
  \code{\link{mat.to.obs}}, \code{\link{mat.to.xlModel}},
  \code{\link{mat.write}}
}
\keyword{mat.to.matrix}
