% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mat.write}
\alias{mat.write}
\title{mat.write}
\usage{
mat.write(x, y, n = ",", w = T)
}
\arguments{
\item{x}{= any matrix/df or vector}

\item{y}{= file intended to receive the output}

\item{n}{= the separator}

\item{w}{= T/F depending on whether to write row names}
}
\description{
Writes <x> as a <n>-separated file to <y>
}
\keyword{mat.write}
