% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.1dActWtTrend.Ctry}
\alias{mk.1dActWtTrend.Ctry}
\title{mk.1dActWtTrend.Ctry}
\usage{
mk.1dActWtTrend.Ctry(x, y, n, w, h = "E")
}
\arguments{
\item{x}{= flowdate}

\item{y}{= factor (one of ActWtTrend/ActWtDiff/ActWtDiff2)}

\item{n}{= country list (one of Ctry/FX)}

\item{w}{= input to or output of sql.connect}

\item{h}{= vector of filters}
}
\description{
SQL query for daily ActWtTrend
}
\keyword{mk.1dActWtTrend.Ctry}
