% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.1dActWtTrend.Sec}
\alias{mk.1dActWtTrend.Sec}
\title{mk.1dActWtTrend.Sec}
\usage{
mk.1dActWtTrend.Sec(x, y, n, w = "E")
}
\arguments{
\item{x}{= flowdate}

\item{y}{= factor (one of ActWtTrend/ActWtDiff/ActWtDiff2)}

\item{n}{= input to or output of sql.connect}

\item{w}{= vector of filters}
}
\description{
SQL query for daily ActWtTrend
}
\keyword{mk.1dActWtTrend.Sec}
