% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.1dFloMo.Ctry.data}
\alias{mk.1dFloMo.Ctry.data}
\title{mk.1dFloMo.Ctry.data}
\usage{
mk.1dFloMo.Ctry.data(x, y, n, w)
}
\arguments{
\item{x}{= SQL statement}

\item{y}{= item (one of Flow/AssetsStart/AssetsEnd)}

\item{n}{= select-statement items}

\item{w}{= input to or output of sql.connect}
}
\description{
formats flow momentum output
}
\keyword{mk.1dFloMo.Ctry.data}
