% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.1dFloMo.Ctry}
\alias{mk.1dFloMo.Ctry}
\title{mk.1dFloMo.Ctry}
\usage{
mk.1dFloMo.Ctry(x, y, n, w, h, u = "E", v = F, g = F)
}
\arguments{
\item{x}{= flowdate/YYYYMMDD depending on whether daily/weekly}

\item{y}{= item (Flow/AssetsStart/AssetsEnd/PortfolioChange)}

\item{n}{= country list (one of Ctry/FX/Sector)}

\item{w}{= input to or output of sql.connect}

\item{h}{= frequency (T/F for daily/weekly or D/W/M)}

\item{u}{= vector of filters}

\item{v}{= T/F to use foreign or all allocations}

\item{g}{= T/F to use institutional or all share classes}
}
\description{
SQL query for daily/weekly CBE flow momentum
}
\keyword{mk.1dFloMo.Ctry}
