% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.1dFloMo.Ctry.rslt}
\alias{mk.1dFloMo.Ctry.rslt}
\title{mk.1dFloMo.Ctry.rslt}
\usage{
mk.1dFloMo.Ctry.rslt(x, y, n)
}
\arguments{
\item{x}{= item (one of Flow/AssetsStart/AssetsEnd)}

\item{y}{= flow momentum output}

\item{n}{= a named vector of country codes indexed by CountryId}
}
\description{
formats flow momentum output
}
\keyword{mk.1dFloMo.Ctry.rslt}
