% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.1dFloMo.Indy}
\alias{mk.1dFloMo.Indy}
\title{mk.1dFloMo.Indy}
\usage{
mk.1dFloMo.Indy(x, y, n, w, h)
}
\arguments{
\item{x}{= a flowdate/YYYYMMDD depending on whether daily/weekly}

\item{y}{= item (one of Flow/AssetsStart/AssetsEnd)}

\item{n}{= input to or output of sql.connect}

\item{w}{= frequency (T/F for daily/weekly or D/W/M)}

\item{h}{= one of US/UK/JP/EM/Eurozone/All (full global)}
}
\description{
SQL query for daily/weekly CBE flow momentum
}
\keyword{mk.1dFloMo.Indy}
