% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.1dFloMo.Sec}
\alias{mk.1dFloMo.Sec}
\title{mk.1dFloMo.Sec}
\usage{
mk.1dFloMo.Sec(x, y, n, w, h, u = F, v = F)
}
\arguments{
\item{x}{= flowdate/YYYYMMDD depending on whether daily/weekly}

\item{y}{= item (Flow/AssetsStart/AssetsEnd/PortfolioChange)}

\item{n}{= input to or output of sql.connect}

\item{w}{= frequency (T/F for daily/weekly or D/W/M)}

\item{h}{= a list object with the following elements: Region - one of US/UK/JP/EM/Eurozone/All (full global) Filter - a vector of filters Group - allocation bulking group (e.g. GeographicFocus/BenchIndex)}

\item{u}{= T/F to use foreign or all allocations}

\item{v}{= T/F to use institutional or all share classes}
}
\description{
SQL query for daily/weekly CBE flow momentum
}
\keyword{mk.1dFloMo.Sec}
