% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.1dFloMo.Sec.rslt}
\alias{mk.1dFloMo.Sec.rslt}
\title{mk.1dFloMo.Sec.rslt}
\usage{
mk.1dFloMo.Sec.rslt(x, y, n, w, h)
}
\arguments{
\item{x}{= item (one of Flow/AssetsStart/AssetsEnd)}

\item{y}{= flow momentum output}

\item{n}{= named vector of sector codes indexed by SectorId}

\item{w}{= frequency (T/F for daily/weekly or D/W/M)}

\item{h}{= IndustryId/SectorId}
}
\description{
formats flow momentum output
}
\keyword{mk.1dFloMo.Sec.rslt}
