% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.1dFloTrend.Ctry}
\alias{mk.1dFloTrend.Ctry}
\title{mk.1dFloTrend.Ctry}
\usage{
mk.1dFloTrend.Ctry(x, y, n, w, h, u = "E")
}
\arguments{
\item{x}{= a flowdate/YYYYMMDD depending on whether daily/weekly}

\item{y}{= factor (one of FloTrend/FloDiff/FloDiff2)}

\item{n}{= country list (one of Ctry/FX/Sector)}

\item{w}{= input to or output of sql.connect}

\item{h}{= T/F depending on whether daily/weekly}

\item{u}{= vector of filters}
}
\description{
SQL query for daily/weekly FloTrend
}
\keyword{mk.1dFloTrend.Ctry}
