% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.1dFloTrend.Sec}
\alias{mk.1dFloTrend.Sec}
\title{mk.1dFloTrend.Sec}
\usage{
mk.1dFloTrend.Sec(x, y, n, w, h)
}
\arguments{
\item{x}{= a flowdate/YYYYMMDD depending on whether daily/weekly}

\item{y}{= factor (one of FloTrend/FloDiff/FloDiff2)}

\item{n}{= input to or output of sql.connect}

\item{w}{= T/F depending on whether daily/weekly}

\item{h}{= vector of filters}
}
\description{
SQL query for daily/weekly FloTrend
}
\keyword{mk.1dFloTrend.Sec}
