% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.1mActPas.Ctry}
\alias{mk.1mActPas.Ctry}
\title{mk.1mActPas.Ctry}
\usage{
mk.1mActPas.Ctry(x, y)
}
\arguments{
\item{x}{= a YYYYMM}

\item{y}{= input to or output of sql.connect}
}
\description{
Generates the SQL query to get monthly AIS for countries
}
\keyword{mk.1mActPas.Ctry}
