% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.1mAllocMo}
\alias{mk.1mAllocMo}
\title{mk.1mAllocMo}
\usage{
mk.1mAllocMo(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMM}

\item{y}{= a string vector of variables to build with the last elements specifying the type of funds to use}

\item{n}{= list object containing the following items: a) classif - classif file b) conn - a connection, the output of odbcDriverConnect c) DB - any of StockFlows/Japan/CSI300/Energy}
}
\description{
Returns a flow variable with the same row space as <n>
}
\seealso{
Other mk: \code{\link{mk.1dFloMo}},
  \code{\link{mk.1mPerfTrend}}, \code{\link{mk.ActWt}},
  \code{\link{mk.Alpha.daily}}, \code{\link{mk.Alpha}},
  \code{\link{mk.FloAlphaLt.Ctry}},
  \code{\link{mk.FundsMem}},
  \code{\link{mk.JensensAlpha.fund}}, \code{\link{mk.Mem}},
  \code{\link{mk.SatoMem}}, \code{\link{mk.Wt}},
  \code{\link{mk.avail}}, \code{\link{mk.beta}},
  \code{\link{mk.isin}}, \code{\link{mk.sqlDump}},
  \code{\link{mk.vbl.chg}}, \code{\link{mk.vbl.diff}},
  \code{\link{mk.vbl.lag}}, \code{\link{mk.vbl.max}},
  \code{\link{mk.vbl.ratio}},
  \code{\link{mk.vbl.trail.fetch}},
  \code{\link{mk.vbl.trail.sum}}, \code{\link{mk.vbl.vol}}
}
\keyword{mk.1mAllocMo}
