% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.1mBullish.Ctry}
\alias{mk.1mBullish.Ctry}
\title{mk.1mBullish.Ctry}
\usage{
mk.1mBullish.Ctry(x, y)
}
\arguments{
\item{x}{= YYYYMM month}

\item{y}{= input to or output of sql.connect}
}
\description{
SQL query for monthly Bullish country indicator
}
\keyword{mk.1mBullish.Ctry}
