% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.1mBullish.rslt}
\alias{mk.1mBullish.rslt}
\title{mk.1mBullish.rslt}
\usage{
mk.1mBullish.rslt(x, y, n)
}
\arguments{
\item{x}{= SQL query}

\item{y}{= input to or output of sql.connect}

\item{n}{= map of sector code to name}
}
\description{
final result for monthly Bullish indicator
}
\keyword{mk.1mBullish.rslt}
