% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.1mBullish.Sec}
\alias{mk.1mBullish.Sec}
\title{mk.1mBullish.Sec}
\usage{
mk.1mBullish.Sec(x, y, n)
}
\arguments{
\item{x}{= a YYYYMM}

\item{y}{= input to or output of sql.connect}

\item{n}{= one of US/UK/JP/EM/Eurozone}
}
\description{
SQL query for monthly Bullish sector indicator
}
\keyword{mk.1mBullish.Sec}
