% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.1wFloMo.CtryFlow.data}
\alias{mk.1wFloMo.CtryFlow.data}
\title{mk.1wFloMo.CtryFlow.data}
\usage{
mk.1wFloMo.CtryFlow.data(x, y, n, w, h, u, v)
}
\arguments{
\item{x}{= YYYYMMDD (flowdate or month end)}

\item{y}{= a vector of FundHistory filters (first element MUST BE FundType)}

\item{n}{= item(s) (any of Flow/AssetsStart/AssetsEnd)}

\item{w}{= result object with element MAP}

\item{h}{= input to or output of sql.connect}

\item{u}{= frequency (T/F for daily/weekly or D/W/M)}

\item{v}{= filter to define cross-border funds}
}
\description{
data for country-flow computation
}
\keyword{mk.1wFloMo.CtryFlow.data}
