% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.1wFloMo.CtryFlow.local}
\alias{mk.1wFloMo.CtryFlow.local}
\title{mk.1wFloMo.CtryFlow.local}
\usage{
mk.1wFloMo.CtryFlow.local(x, y, n, w, h, u = T)
}
\arguments{
\item{x}{= a YYYYMMDD}

\item{y}{= a vector of FundHistory filters (first element MUST BE FundType)}

\item{n}{= item(s) (any of Flow/AssetsStart/AssetsEnd)}

\item{w}{= country list (one of Ctry/LatAm)}

\item{h}{= input to or output of sql.connect}

\item{u}{= T/F depending on whether weekly or daily data needed}
}
\description{
Country flows using locally-domiciled funds only
}
\keyword{mk.1wFloMo.CtryFlow.local}
