% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.1wFloMo.CtryFlow}
\alias{mk.1wFloMo.CtryFlow}
\title{mk.1wFloMo.CtryFlow}
\usage{
mk.1wFloMo.CtryFlow(x, y, n, w, h, u = "W")
}
\arguments{
\item{x}{= YYYYMMDD (flowdate or month end)}

\item{y}{= a vector of FundHistory filters (first element MUST BE FundType)}

\item{n}{= item(s) (any of Flow/AssetsStart/AssetsEnd)}

\item{w}{= country list (one of Ctry/LatAm)}

\item{h}{= input to or output of sql.connect}

\item{u}{= frequency (T/F for daily/weekly or D/W/M)}
}
\description{
Country flows using all funds
}
\keyword{mk.1wFloMo.CtryFlow}
