% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.1wFloMo.CtryFlow}
\alias{mk.1wFloMo.CtryFlow}
\title{mk.1wFloMo.CtryFlow}
\usage{
mk.1wFloMo.CtryFlow(x, y, n, w, h, u = T)
}
\arguments{
\item{x}{= YYYYMMDD}

\item{y}{= FundType (one of E/B)}

\item{n}{= item(s) (any of Flow/AssetsStart/AssetsEnd)}

\item{w}{= country list (one of Ctry/LatAm)}

\item{h}{= input to or output of sql.connect}

\item{u}{= T/F depending on whether weekly or daily data needed}
}
\description{
SQL query for country-flow percentage for date <x>
}
\seealso{
Other mk: \code{\link{mk.1dFloMo.Ctry}},
  \code{\link{mk.1dFloMo.Indy}},
  \code{\link{mk.1dFloMo.Sec}}, \code{\link{mk.1dFloMo}},
  \code{\link{mk.1mActPas.Ctry}},
  \code{\link{mk.1mActPas.Sec}},
  \code{\link{mk.1mAllocMo}},
  \code{\link{mk.1mBullish.Ctry}},
  \code{\link{mk.1mBullish.Sec}},
  \code{\link{mk.1mPerfTrend}}, \code{\link{mk.ActWt}},
  \code{\link{mk.Alpha.daily}}, \code{\link{mk.Alpha}},
  \code{\link{mk.EigenCentrality}},
  \code{\link{mk.FloAlphaLt.Ctry}},
  \code{\link{mk.Fragility}}, \code{\link{mk.FundsMem}},
  \code{\link{mk.HerdingLSV}}, \code{\link{mk.HoldValTot}},
  \code{\link{mk.JensensAlpha.fund}}, \code{\link{mk.Mem}},
  \code{\link{mk.SRIMem}}, \code{\link{mk.SatoMem}},
  \code{\link{mk.Wt}}, \code{\link{mk.avail}},
  \code{\link{mk.beta}}, \code{\link{mk.isin}},
  \code{\link{mk.sqlDump}}, \code{\link{mk.vbl.chg}},
  \code{\link{mk.vbl.diff}}, \code{\link{mk.vbl.lag}},
  \code{\link{mk.vbl.max}}, \code{\link{mk.vbl.ratio}},
  \code{\link{mk.vbl.scale}}, \code{\link{mk.vbl.sum}},
  \code{\link{mk.vbl.trail.fetch}},
  \code{\link{mk.vbl.trail.sum}}, \code{\link{mk.vbl.vol}}
}
\keyword{mk.1wFloMo.CtryFlow}
