% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.1wFloMo.CtryFlow.rslt}
\alias{mk.1wFloMo.CtryFlow.rslt}
\title{mk.1wFloMo.CtryFlow.rslt}
\usage{
mk.1wFloMo.CtryFlow.rslt(x)
}
\arguments{
\item{x}{= result object with names MAP, CBF, SCF & CBA}
}
\description{
Country flows using all funds
}
\keyword{mk.1wFloMo.CtryFlow.rslt}
