% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.1wFloMo.IndyFlow}
\alias{mk.1wFloMo.IndyFlow}
\title{mk.1wFloMo.IndyFlow}
\usage{
mk.1wFloMo.IndyFlow(x, y, n, w, h = T)
}
\arguments{
\item{x}{= YYYYMMDD}

\item{y}{= item(s) (any of Flow/AssetsStart/AssetsEnd)}

\item{n}{= input to or output of sql.connect}

\item{w}{= frequency (T/F for daily/weekly or D/W/M)}

\item{h}{= T/F depending for Industry/Sector flows}
}
\description{
Industry/Sector flows
}
\keyword{mk.1wFloMo.IndyFlow}
