% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.ActWt}
\alias{mk.ActWt}
\title{mk.ActWt}
\usage{
mk.ActWt(x, y, n)
}
\arguments{
\item{x}{= a YYYYMM}

\item{y}{= a string vector of names of the portfolio and benchmark}

\item{n}{= list object containing the following items: #        :    a) classif - classif file #        :    b) fldr - stock-flows folder}
}
\description{
Active weight
}
\keyword{mk.ActWt}
