% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.Alpha.daily}
\alias{mk.Alpha.daily}
\title{mk.Alpha.daily}
\usage{
mk.Alpha.daily(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMMDD}

\item{y}{= a string vector, the first two elements of which are #        :    universe and group to zScore on and within. This is then followed by #        :    a list of variables which are, in turn, followed by weights to put on variables #        :    and a logical vector indicating whether the variables are daily.}

\item{n}{= list object containing the following items: #        :    a) classif - classif file #        :    b) fldr - stock-flows folder}
}
\description{
makes Alpha
}
\keyword{mk.Alpha.daily}
