% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.Alpha}
\alias{mk.Alpha}
\title{mk.Alpha}
\usage{
mk.Alpha(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMM}

\item{y}{= a string vector, the first two elements of which are universe and group to zScore on and within. This is then followed by a list of variables which are, in turn, followed by weights to put on variables}

\item{n}{= list object containing the following items: a) classif - classif file b) fldr - stock-flows folder}
}
\description{
makes Alpha
}
\keyword{mk.Alpha}
