% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.avail}
\alias{mk.avail}
\title{mk.avail}
\usage{
mk.avail(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMM or YYYYMMDD}

\item{y}{= a string vector, the elements of which are: 1) folder to fetch data from 2+) variables to fetch}

\item{n}{= list object containing the following items: a) classif - classif file b) fldr - stock-flows folder}
}
\description{
Returns leftmost non-NA variable
}
\seealso{
Other mk: \code{\link{mk.1dFloMo}},
  \code{\link{mk.1mAllocMo}}, \code{\link{mk.1mPerfTrend}},
  \code{\link{mk.1wFloMo.CtryFlow}},
  \code{\link{mk.ActWt}}, \code{\link{mk.Alpha.daily}},
  \code{\link{mk.Alpha}}, \code{\link{mk.EigenCentrality}},
  \code{\link{mk.FloAlphaLt.Ctry}},
  \code{\link{mk.Fragility}}, \code{\link{mk.FundsMem}},
  \code{\link{mk.HerdingLSV}}, \code{\link{mk.HoldValTot}},
  \code{\link{mk.JensensAlpha.fund}}, \code{\link{mk.Mem}},
  \code{\link{mk.SRIMem}}, \code{\link{mk.SatoMem}},
  \code{\link{mk.Wt}}, \code{\link{mk.beta}},
  \code{\link{mk.isin}}, \code{\link{mk.sqlDump}},
  \code{\link{mk.vbl.chg}}, \code{\link{mk.vbl.diff}},
  \code{\link{mk.vbl.lag}}, \code{\link{mk.vbl.max}},
  \code{\link{mk.vbl.ratio}}, \code{\link{mk.vbl.scale}},
  \code{\link{mk.vbl.sum}},
  \code{\link{mk.vbl.trail.fetch}},
  \code{\link{mk.vbl.trail.sum}}, \code{\link{mk.vbl.vol}}
}
\keyword{mk.avail}
