% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.beta}
\alias{mk.beta}
\title{mk.beta}
\usage{
mk.beta(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMM}

\item{y}{= a string vector, the elements of which are: 1) benchmark (e.g. "Eafe") 2) number of trailing months of returns (e.g. 12)}

\item{n}{= list object containing the following items: a) classif - classif file b) fldr - stock-flows folder}
}
\description{
Computes monthly beta versus relevant benchmark
}
\seealso{
Other mk: \code{\link{mk.1dFloMo}},
  \code{\link{mk.1mAllocMo}}, \code{\link{mk.1mPerfTrend}},
  \code{\link{mk.ActWt}}, \code{\link{mk.Alpha.daily}},
  \code{\link{mk.Alpha}}, \code{\link{mk.FloAlphaLt.Ctry}},
  \code{\link{mk.FundsMem}},
  \code{\link{mk.JensensAlpha.fund}}, \code{\link{mk.Mem}},
  \code{\link{mk.PrcMo}}, \code{\link{mk.Wt}},
  \code{\link{mk.avail}}, \code{\link{mk.isin}},
  \code{\link{mk.sqlDump}},
  \code{\link{mk.trail.sum.dly.vbl}},
  \code{\link{mk.trail.sum.mo.vbl}},
  \code{\link{mk.vbl.chg}}, \code{\link{mk.vbl.diff}},
  \code{\link{mk.vbl.lag}}, \code{\link{mk.vbl.max}},
  \code{\link{mk.vbl.ratio}}
}
\keyword{mk.beta}
