% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.EigenCentrality}
\alias{mk.EigenCentrality}
\title{mk.EigenCentrality}
\usage{
mk.EigenCentrality(x, y, n)
}
\arguments{
\item{x}{= a YYYYMM}

\item{y}{= a string vector of variables to build with the last elements #        :    specifying the type of funds to use}

\item{n}{= list object containing the following items: #        :    a) classif - classif file #        :    b) conn - a connection, the output of odbcDriverConnect #        :    c) DB - any of StockFlows/China/Japan/CSI300/Energy}
}
\description{
Returns EigenCentrality with the same row space as <n>
}
\keyword{mk.EigenCentrality}
