% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.FloAlphaLt.Ctry}
\alias{mk.FloAlphaLt.Ctry}
\title{mk.FloAlphaLt.Ctry}
\usage{
mk.FloAlphaLt.Ctry(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMM}

\item{y}{= an object name (preceded by #) or the path to a ".csv" file}

\item{n}{= list object containing the following items: a) classif - classif file}
}
\description{
Monthly Country Flow Alpha
}
\seealso{
Other mk: \code{\link{mk.1dFloMo.Ctry}},
  \code{\link{mk.1dFloMo.Indy}},
  \code{\link{mk.1dFloMo.Sec}}, \code{\link{mk.1dFloMo}},
  \code{\link{mk.1mActPas.Ctry}},
  \code{\link{mk.1mActPas.Sec}},
  \code{\link{mk.1mAllocMo}},
  \code{\link{mk.1mBullish.Ctry}},
  \code{\link{mk.1mBullish.Sec}},
  \code{\link{mk.1mFloMo.Ctry}},
  \code{\link{mk.1mPerfTrend}},
  \code{\link{mk.1wFloMo.CtryFlow}},
  \code{\link{mk.1wFloMo.IndyFlow}},
  \code{\link{mk.ActWt}}, \code{\link{mk.Alpha.daily}},
  \code{\link{mk.Alpha}}, \code{\link{mk.EigenCentrality}},
  \code{\link{mk.Fragility}}, \code{\link{mk.FundsMem}},
  \code{\link{mk.HerdingLSV}},
  \code{\link{mk.JensensAlpha.fund}}, \code{\link{mk.Mem}},
  \code{\link{mk.SRIMem}}, \code{\link{mk.SatoMem}},
  \code{\link{mk.Wt}}, \code{\link{mk.avail}},
  \code{\link{mk.beta}}, \code{\link{mk.isin}},
  \code{\link{mk.sf.daily}}, \code{\link{mk.sqlDump}},
  \code{\link{mk.vbl.chg}}, \code{\link{mk.vbl.diff}},
  \code{\link{mk.vbl.lag}}, \code{\link{mk.vbl.max}},
  \code{\link{mk.vbl.ratio}}, \code{\link{mk.vbl.scale}},
  \code{\link{mk.vbl.sum}},
  \code{\link{mk.vbl.trail.fetch}},
  \code{\link{mk.vbl.trail.sum}}, \code{\link{mk.vbl.vol}}
}
\keyword{mk.FloAlphaLt.Ctry}
