% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.FloBeta}
\alias{mk.FloBeta}
\title{mk.FloBeta}
\usage{
mk.FloBeta(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMM}

\item{y}{= a string vector, the elements of which are: 1) connection string (e.g. "NEWUI") 2) number of trailing months of returns (e.g. 36)}

\item{n}{= list object containing the following items: a) classif - classif file b) fldr - stock-flows folder}
}
\description{
Computes monthly beta versus common fund flow shock
}
\keyword{mk.FloBeta}
