% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.isin}
\alias{mk.isin}
\title{mk.isin}
\usage{
mk.isin(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMM or YYYYMMDD}

\item{y}{= a string vector, the elements of which are: #        :    1) an object name (preceded by #) or the path to a ".csv" file #        :    2) defaults to "isin"}

\item{n}{= list object containing the following items: #        :    a) classif - classif file}
}
\description{
Looks up date from external file and maps on isin
}
\keyword{mk.isin}
