% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.isin}
\alias{mk.isin}
\title{mk.isin}
\usage{
mk.isin(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMM or YYYYMMDD}

\item{y}{= a string vector, the elements of which are: 1) an object name (preceded by #) or the path to a ".csv" file 2) defaults to "isin"}

\item{n}{= list object containing the following items: a) classif - classif file}
}
\description{
Looks up date from external file and maps on isin
}
\seealso{
Other mk: \code{\link{mk.1dFloMo}},
  \code{\link{mk.1mAllocMo}}, \code{\link{mk.1mPerfTrend}},
  \code{\link{mk.ActWt}}, \code{\link{mk.Alpha.daily}},
  \code{\link{mk.Alpha}}, \code{\link{mk.FloAlphaLt.Ctry}},
  \code{\link{mk.FundsMem}},
  \code{\link{mk.JensensAlpha.fund}}, \code{\link{mk.Mem}},
  \code{\link{mk.SatoMem}}, \code{\link{mk.Wt}},
  \code{\link{mk.avail}}, \code{\link{mk.beta}},
  \code{\link{mk.sqlDump}}, \code{\link{mk.vbl.chg}},
  \code{\link{mk.vbl.diff}}, \code{\link{mk.vbl.lag}},
  \code{\link{mk.vbl.max}}, \code{\link{mk.vbl.ratio}},
  \code{\link{mk.vbl.trail.fetch}},
  \code{\link{mk.vbl.trail.sum}}, \code{\link{mk.vbl.vol}}
}
\keyword{mk.isin}
