% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.Mem}
\alias{mk.Mem}
\title{mk.Mem}
\usage{
mk.Mem(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMM}

\item{y}{= a single FundId}

\item{n}{= list object containing the following items: a) classif - classif file b) conn - a connection, the output of odbcDriverConnect}
}
\description{
Returns a 1/0 membership vector
}
\keyword{mk.Mem}
