% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.SatoMem}
\alias{mk.SatoMem}
\title{mk.SatoMem}
\usage{
mk.SatoMem(x, y, n)
}
\arguments{
\item{x}{= an argument which is never used}

\item{y}{= path to a file containing isin's}

\item{n}{= list object containing the following items: #        :    a) classif - classif file}
}
\description{
Returns a 1/0 membership vector
}
\keyword{mk.SatoMem}
