% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.sf.daily}
\alias{mk.sf.daily}
\title{mk.sf.daily}
\usage{
mk.sf.daily(fcn, x, y, n, w)
}
\arguments{
\item{fcn}{= a fetch function}

\item{x}{= a list of flowdate vectors}

\item{y}{= input to sql.connect}

\item{n}{= maximum number of queries using the same connection}

\item{w}{= argument passed down to <fcn>}
}
\description{
gets data using query generated by <fcn>
}
\keyword{mk.sf.daily}
