% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.sf.daily}
\alias{mk.sf.daily}
\title{mk.sf.daily}
\usage{
mk.sf.daily(fcn, x, y, n, w)
}
\arguments{
\item{fcn}{= fetch function}

\item{x}{= list of vectors of flowdates}

\item{y}{= name of an SQL connection (e.g. "StockFlows", "NEWUI" etc.)}

\item{n}{= maximum number of queries using the same connection}

\item{w}{= argument passed down to <fcn>}
}
\description{
gets data using query generated by <fcn>
}
\keyword{mk.sf.daily}
