% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.sf.daily}
\alias{mk.sf.daily}
\title{mk.sf.daily}
\usage{
mk.sf.daily(fcn, x, y, n, w)
}
\arguments{
\item{fcn}{= fetch function}

\item{x}{= list of vectors of flowdates}

\item{y}{= name of an SQL connection (e.g. "StockFlows", "NEWUI" etc.)}

\item{n}{= maximum number of queries using the same connection}

\item{w}{= argument passed down to <fcn>}
}
\description{
gets data using query generated by <fcn>
}
\seealso{
Other mk: \code{\link{mk.1dFloMo.Ctry}},
  \code{\link{mk.1dFloMo.Indy}},
  \code{\link{mk.1dFloMo.Sec}}, \code{\link{mk.1dFloMo}},
  \code{\link{mk.1mActPas.Ctry}},
  \code{\link{mk.1mActPas.Sec}},
  \code{\link{mk.1mAllocMo}},
  \code{\link{mk.1mBullish.Ctry}},
  \code{\link{mk.1mBullish.Sec}},
  \code{\link{mk.1mFloMo.Ctry}},
  \code{\link{mk.1mPerfTrend}},
  \code{\link{mk.1wFloMo.CtryFlow}},
  \code{\link{mk.1wFloMo.IndyFlow}},
  \code{\link{mk.ActWt}}, \code{\link{mk.Alpha.daily}},
  \code{\link{mk.Alpha}}, \code{\link{mk.EigenCentrality}},
  \code{\link{mk.FloAlphaLt.Ctry}},
  \code{\link{mk.Fragility}}, \code{\link{mk.FundsMem}},
  \code{\link{mk.HerdingLSV}},
  \code{\link{mk.JensensAlpha.fund}}, \code{\link{mk.Mem}},
  \code{\link{mk.SRIMem}}, \code{\link{mk.SatoMem}},
  \code{\link{mk.Wt}}, \code{\link{mk.avail}},
  \code{\link{mk.beta}}, \code{\link{mk.isin}},
  \code{\link{mk.sqlDump}}, \code{\link{mk.vbl.chg}},
  \code{\link{mk.vbl.diff}}, \code{\link{mk.vbl.lag}},
  \code{\link{mk.vbl.max}}, \code{\link{mk.vbl.ratio}},
  \code{\link{mk.vbl.scale}}, \code{\link{mk.vbl.sum}},
  \code{\link{mk.vbl.trail.fetch}},
  \code{\link{mk.vbl.trail.sum}}, \code{\link{mk.vbl.vol}}
}
\keyword{mk.sf.daily}
