% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.sqlDump}
\alias{mk.sqlDump}
\title{mk.sqlDump}
\usage{
mk.sqlDump(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMM}

\item{y}{= a string vector, the elements of which are: 1) file to read from 2) variable to read 3) lag (defaults to zero)}

\item{n}{= list object containing the following items: a) fldr - stock-flows folder}
}
\description{
Returns variable with the same row space as <n>
}
\seealso{
Other mk: \code{\link{mk.1dFloMo}},
  \code{\link{mk.1mAllocMo}}, \code{\link{mk.1mPerfTrend}},
  \code{\link{mk.ActWt}}, \code{\link{mk.Alpha.daily}},
  \code{\link{mk.Alpha}}, \code{\link{mk.FloAlphaLt.Ctry}},
  \code{\link{mk.FundsMem}},
  \code{\link{mk.JensensAlpha.fund}}, \code{\link{mk.Mem}},
  \code{\link{mk.Wt}}, \code{\link{mk.avail}},
  \code{\link{mk.beta}}, \code{\link{mk.isin}},
  \code{\link{mk.vbl.chg}}, \code{\link{mk.vbl.diff}},
  \code{\link{mk.vbl.lag}}, \code{\link{mk.vbl.max}},
  \code{\link{mk.vbl.ratio}},
  \code{\link{mk.vbl.trail.fetch}},
  \code{\link{mk.vbl.trail.sum.dly}},
  \code{\link{mk.vbl.trail.sum.mo}},
  \code{\link{mk.vbl.vol}}
}
\keyword{mk.sqlDump}
