% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.sqlDump}
\alias{mk.sqlDump}
\title{mk.sqlDump}
\usage{
mk.sqlDump(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMM}

\item{y}{= a string vector, the elements of which are: #        :    1) file to read from #        :    2) variable to read #        :    3) lag (defaults to zero)}

\item{n}{= list object containing the following items: #        :    a) fldr - stock-flows folder}
}
\description{
Returns variable with the same row space as <n>
}
\keyword{mk.sqlDump}
