% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.SRIMem}
\alias{mk.SRIMem}
\title{mk.SRIMem}
\usage{
mk.SRIMem(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMM}

\item{y}{= a positive integer}

\item{n}{= list object containing the following items: a) classif - classif file b) conn - a connection, the output of odbcDriverConnect c) DB - any of StockFlows/China/Japan/CSI300/Energy}
}
\description{
1/0 depending on whether <y> or more SRI funds own the stock
}
\keyword{mk.SRIMem}
