% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.SRIMem}
\alias{mk.SRIMem}
\title{mk.SRIMem}
\usage{
mk.SRIMem(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMM}

\item{y}{= a positive integer}

\item{n}{= list object containing the following items: a) classif - classif file b) conn - a connection, the output of odbcDriverConnect c) DB - any of StockFlows/China/Japan/CSI300/Energy}
}
\description{
1/0 depending on whether <y> or more SRI funds own the stock
}
\seealso{
Other mk: \code{\link{mk.1dFloMo.Ctry}},
  \code{\link{mk.1dFloMo.Sec}}, \code{\link{mk.1dFloMo}},
  \code{\link{mk.1mActPas.Ctry}},
  \code{\link{mk.1mAllocMo}}, \code{\link{mk.1mPerfTrend}},
  \code{\link{mk.1wFloMo.CtryFlow}},
  \code{\link{mk.ActWt}}, \code{\link{mk.Alpha.daily}},
  \code{\link{mk.Alpha}}, \code{\link{mk.EigenCentrality}},
  \code{\link{mk.FloAlphaLt.Ctry}},
  \code{\link{mk.Fragility}}, \code{\link{mk.FundsMem}},
  \code{\link{mk.HerdingLSV}}, \code{\link{mk.HoldValTot}},
  \code{\link{mk.JensensAlpha.fund}}, \code{\link{mk.Mem}},
  \code{\link{mk.SatoMem}}, \code{\link{mk.Wt}},
  \code{\link{mk.avail}}, \code{\link{mk.beta}},
  \code{\link{mk.isin}}, \code{\link{mk.sqlDump}},
  \code{\link{mk.vbl.chg}}, \code{\link{mk.vbl.diff}},
  \code{\link{mk.vbl.lag}}, \code{\link{mk.vbl.max}},
  \code{\link{mk.vbl.ratio}}, \code{\link{mk.vbl.scale}},
  \code{\link{mk.vbl.sum}},
  \code{\link{mk.vbl.trail.fetch}},
  \code{\link{mk.vbl.trail.sum}}, \code{\link{mk.vbl.vol}}
}
\keyword{mk.SRIMem}
