% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.trail.sum.mo.vbl}
\alias{mk.trail.sum.mo.vbl}
\title{mk.trail.sum.mo.vbl}
\usage{
mk.trail.sum.mo.vbl(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMM}

\item{y}{= a string vector, the elements of which are: 1) 1m variable to fetch (e.g. "AllocMo"/"AllocDiff"/"AllocTrend") 2) T to sum or F to compound (e.g. "T") 3) number of months to sum over (e.g. "11") 4) number of months to lag (defaults to "0") 5) sub-folder to fetch basic variable from (defaults to "derived")}

\item{n}{= list object containing the following items: a) classif - classif file b) fldr - stock-flows folder}
}
\description{
makes 11 month AllocMo
}
\keyword{mk.trail.sum.mo.vbl}
