% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.vbl.chg}
\alias{mk.vbl.chg}
\title{mk.vbl.chg}
\usage{
mk.vbl.chg(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMM}

\item{y}{= variable name}

\item{n}{= list object containing the following items: a) classif - classif file b) fldr - stock-flows folder}
}
\description{
Makes the MoM change in the variable
}
\keyword{mk.vbl.chg}
