% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.vbl.max}
\alias{mk.vbl.max}
\title{mk.vbl.max}
\usage{
mk.vbl.max(x, y, n)
}
\arguments{
\item{x}{= a YYYYMM}

\item{y}{= a string vector of names of two variables}

\item{n}{= list object containing the following items: #        :    a) classif - classif file #        :    b) fldr - stock-flows folder}
}
\description{
Computes the maximum of the two variables
}
\keyword{mk.vbl.max}
