% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.vbl.ratio}
\alias{mk.vbl.ratio}
\title{mk.vbl.ratio}
\usage{
mk.vbl.ratio(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMM}

\item{y}{= a string vector, the elements of which are the numerator and denominator respectively.}

\item{n}{= list object containing the following items: a) classif - classif file b) fldr - stock-flows folder}
}
\description{
Computes the ratio of the two variables
}
\keyword{mk.vbl.ratio}
