% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.vbl.sum}
\alias{mk.vbl.sum}
\title{mk.vbl.sum}
\usage{
mk.vbl.sum(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMM}

\item{y}{= a string vector, the elements of which are the variables to be added.}

\item{n}{= list object containing the following items: a) classif - classif file b) fldr - stock-flows folder}
}
\description{
Computes the sum of the two variables
}
\keyword{mk.vbl.sum}
