% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.vbl.trail.sum}
\alias{mk.vbl.trail.sum}
\title{mk.vbl.trail.sum}
\usage{
mk.vbl.trail.sum(x, y, n)
}
\arguments{
\item{x}{= a single YYYYMM or YYYYMMDD}

\item{y}{= a string vector, the elements of which are: 1) variable to fetch (e.g. "AllocMo"/"AllocDiff"/"AllocTrend"/"Ret") 2) T to sum or F to compound (e.g. "T") 3) number of trailing periods to use (e.g. "11") 4) number of periods to lag (defaults to "0") 5) sub-folder to fetch basic variable from (defaults to "derived") 6) T/F depending on whether the compounded variable is daily (defaults to F, matters only if <x> is monthly)}

\item{n}{= list object containing the following items: a) classif - classif file b) fldr - stock-flows folder}
}
\description{
compounded variable over some trailing window
}
\seealso{
Other mk: \code{\link{mk.1dFloMo}},
  \code{\link{mk.1mAllocMo}}, \code{\link{mk.1mPerfTrend}},
  \code{\link{mk.ActWt}}, \code{\link{mk.Alpha.daily}},
  \code{\link{mk.Alpha}}, \code{\link{mk.FloAlphaLt.Ctry}},
  \code{\link{mk.FundsMem}},
  \code{\link{mk.JensensAlpha.fund}}, \code{\link{mk.Mem}},
  \code{\link{mk.Wt}}, \code{\link{mk.avail}},
  \code{\link{mk.beta}}, \code{\link{mk.isin}},
  \code{\link{mk.sqlDump}}, \code{\link{mk.vbl.chg}},
  \code{\link{mk.vbl.diff}}, \code{\link{mk.vbl.lag}},
  \code{\link{mk.vbl.max}}, \code{\link{mk.vbl.ratio}},
  \code{\link{mk.vbl.trail.fetch}},
  \code{\link{mk.vbl.vol}}
}
\keyword{mk.vbl.trail.sum}
