% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{mk.Wt}
\alias{mk.Wt}
\title{mk.Wt}
\usage{
mk.Wt(x, y, n)
}
\arguments{
\item{x}{= a YYYYMM}

\item{y}{= FundId of the fund of interest}

\item{n}{= list object containing the following items: #        :    a) classif - classif file #        :    b) conn - a connection, the output of odbcDriverConnect}
}
\description{
Generates the SQL query to get monthly index weight for individual stocks
}
\keyword{mk.Wt}
