% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{nameTo}
\alias{nameTo}
\title{nameTo}
\usage{
nameTo(x, y)
}
\arguments{
\item{x}{= a numeric vector/matrix/data frame (no NA's)}

\item{y}{= a numeric vector/matrix/data frame (no NA's)}
}
\description{
pct name turnover between <x> and <y> if <x> is a numeric vector or their rows otherwise
}
\keyword{nameTo}
