% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{nameTo.underlying}
\alias{nameTo.underlying}
\title{nameTo.underlying}
\usage{
nameTo.underlying(x, y, n)
}
\arguments{
\item{x}{= a vector of counts over the current period}

\item{y}{= a vector of counts over the previous period}

\item{n}{= a vector of numbers of names common between current and previous periods}
}
\description{
percent name turnover
}
\keyword{nameTo.underlying}
