% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{obj.lag}
\alias{obj.lag}
\title{obj.lag}
\usage{
obj.lag(x, y, n, w)
}
\arguments{
\item{x}{= an object vector}

\item{y}{= an integer vector}

\item{n}{= a function (object to value)}

\item{w}{= a function (bijective inverse of <n>)}
}
\description{
lags <x> by <y>
}
\seealso{
Other obj: \code{\link{obj.diff}}, \code{\link{obj.seq}}
}
\keyword{obj.lag}
