% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{obj.lag}
\alias{obj.lag}
\title{obj.lag}
\usage{
obj.lag(x, y, n, w)
}
\arguments{
\item{x}{= a vector of objects}

\item{y}{= an integer or vector of integers (if <x> and <y> are vectors then <y> isomekic)}

\item{n}{= a function mapping these objects to the number line}

\item{w}{= the bijective inverse of <n>}
}
\description{
lags <x> by <y>
}
\seealso{
Other obj: \code{\link{obj.diff}}, \code{\link{obj.list}},
  \code{\link{obj.seq}}
}
\keyword{obj.lag}
