% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{obj.seq}
\alias{obj.seq}
\title{obj.seq}
\usage{
obj.seq(x, y, n, w, h)
}
\arguments{
\item{x}{= an object}

\item{y}{= an object of the same type as <x>}

\item{n}{= a function mapping these objects to the number line}

\item{w}{= the bijective inverse of <n>}

\item{h}{= a positive integer}
}
\description{
returns a sequence of objects between (and including) <x> and <y>
}
\keyword{obj.seq}
