% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{obj.seq}
\alias{obj.seq}
\title{obj.seq}
\usage{
obj.seq(x, y, n, w, h)
}
\arguments{
\item{x}{= an object}

\item{y}{= an object}

\item{n}{= a function (vector to integer vector)}

\item{w}{= a function (bijective inverse of <n>)}

\item{h}{= a positive integer}
}
\description{
returns a sequence of objects between (and including) <x> and <y>
}
\seealso{
Other obj: \code{\link{obj.diff}}, \code{\link{obj.lag}}
}
\keyword{obj.seq}
