% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{obj.seq}
\alias{obj.seq}
\title{obj.seq}
\usage{
obj.seq(x, y, n, w, h)
}
\arguments{
\item{x}{= a SINGLE object}

\item{y}{= a SINGLE object of the same type as <x>}

\item{n}{= a function mapping these objects to the number line}

\item{w}{= the bijective inverse of <n>}

\item{h}{= a positive integer representing quantum size}
}
\description{
returns a sequence of objects between (and including) <x> and <y>
}
\seealso{
Other obj: \code{\link{obj.diff}}, \code{\link{obj.lag}},
  \code{\link{obj.list}}
}
\keyword{obj.seq}
