% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{permutations.next}
\alias{permutations.next}
\title{permutations.next}
\usage{
permutations.next(x)
}
\arguments{
\item{x}{= a vector of integers 1:length(<x>) in some order}
}
\description{
returns the next permutation in dictionary order
}
\seealso{
Other permutations: \code{\link{permutations.buckets.many}},
  \code{\link{permutations}}
}
\keyword{permutations.next}
