% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{permutations}
\alias{permutations}
\title{permutations}
\usage{
permutations(x)
}
\arguments{
\item{x}{= a string vector without NA's}
}
\description{
all possible permutations of <x>
}
\keyword{permutations}
