% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{portfolio.beta.wrapper}
\alias{portfolio.beta.wrapper}
\title{portfolio.beta.wrapper}
\usage{
portfolio.beta.wrapper(x, y, n)
}
\arguments{
\item{x}{= a file of total return indices indexed so that time runs forward}

\item{y}{= the name of the benchmark w.r.t. which beta is to be computed (e.g. "ACWorld")}

\item{n}{= the window in days over which beta is to be computed}
}
\description{
<n> day beta of columns of <x> with respect to benchmark <y>
}
\seealso{
Other portfolio: \code{\link{portfolio.beta}},
  \code{\link{portfolio.residual}}
}
\keyword{portfolio.beta.wrapper}
