% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{portfolio.residual}
\alias{portfolio.residual}
\title{portfolio.residual}
\usage{
portfolio.residual(x, y)
}
\arguments{
\item{x}{= a matrix/df}

\item{y}{= a matrix/df of the same dimensions as <x>}
}
\description{
residual of <x> after factoring out <y> for each row
}
\seealso{
Other portfolio: \code{\link{portfolio.beta.wrapper}}
}
\keyword{portfolio.residual}
