% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{position.ActWtDiff2}
\alias{position.ActWtDiff2}
\title{position.ActWtDiff2}
\usage{
position.ActWtDiff2(x)
}
\arguments{
\item{x}{= One of "StockFlows", "Quant" or "Regular"}
}
\description{
Current and week-over-week change of ActWtDiff2 on R1 Materials
}
\seealso{
Other position: \code{\link{position.floPct}}
}
\keyword{position.ActWtDiff2}
