% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{position.floPct}
\alias{position.floPct}
\title{position.floPct}
\usage{
position.floPct(x, y, n, w = 20, h = 5, u = F)
}
\arguments{
\item{x}{= a file (strategy name or path)}

\item{y}{= a string vector (to subset to, can be missing)}

\item{n}{= last publication date (can be missing)}

\item{w}{= an integer (size of lookback window)}

\item{h}{= an integer (prior period)}

\item{u}{= a boolean (sum/compound)}
}
\description{
Latest four-week flow percentage
}
\keyword{position.floPct}
