% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{principal.components.covar}
\alias{principal.components.covar}
\title{principal.components.covar}
\usage{
principal.components.covar(x, y, n)
}
\arguments{
\item{x}{= a matrix/df}

\item{y}{= number of principal components considered important}

\item{n}{= missing or a numeric vector of length <dim(x)[2]>}
}
\description{
covariance using first <y> components as factors. <n>, when present, becomes one of the factors. Residuals are considered independent.
}
\seealso{
Other principal: \code{\link{principal.components}}
}
\keyword{principal.components.covar}
