% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{principal.components.covar}
\alias{principal.components.covar}
\title{principal.components.covar}
\usage{
principal.components.covar(x, y)
}
\arguments{
\item{x}{= a matrix/df}

\item{y}{= number of principal components considered important}
}
\description{
covariance using first <y> components as factors
}
\seealso{
Other principal: \code{\link{principal.components.underlying}},
  \code{\link{principal.components}}
}
\keyword{principal.components.covar}
