% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{principal.components}
\alias{principal.components}
\title{principal.components}
\usage{
principal.components(x, y = 2)
}
\arguments{
\item{x}{= a matrix/data frame}

\item{y}{= an integer (num. princ. comp. wanted)}
}
\description{
first <y> principal components
}
\seealso{
Other principal: \code{\link{principal.components.covar}},
  \code{\link{principal.components.underlying}}
}
\keyword{principal.components}
