% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{proc.kill}
\alias{proc.kill}
\title{proc.kill}
\usage{
proc.kill(x, y)
}
\arguments{
\item{x}{= process name (e.g. "ftp.exe")}

\item{y}{= number of instances}
}
\description{
kills of all processes <x> if more than <y> of them are running
}
\seealso{
Other proc: \code{\link{proc.count}}
}
\keyword{proc.kill}
