% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{proc.kill}
\alias{proc.kill}
\title{proc.kill}
\usage{
proc.kill(x)
}
\arguments{
\item{x}{= process name (e.g. "ftp.exe")}
}
\description{
kills off all processes <x>
}
\seealso{
Other proc: \code{\link{proc.count}}
}
\keyword{proc.kill}
