% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{publications.data}
\alias{publications.data}
\title{publications.data}
\usage{
publications.data(x, y, n, w)
}
\arguments{
\item{x}{= a vector of desired dates}

\item{y}{= SQL query OR a function taking a date as argument}

\item{n}{= folder where the data live}

\item{w}{= one of StockFlows/Regular/Quant}
}
\description{
additional data is got and stale data removed
}
\keyword{publications.data}
