% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{publish.monthly.last}
\alias{publish.monthly.last}
\title{publish.monthly.last}
\usage{
publish.monthly.last(x, y = 23)
}
\arguments{
\item{x}{= a YYYYMMDD date}

\item{y}{= calendar day in the next month when allocations are known (usually the 23rd)}
}
\description{
date of last monthly publication
}
\seealso{
Other publish: \code{\link{publish.daily.last}},
  \code{\link{publish.date}},
  \code{\link{publish.weekly.last}}
}
\keyword{publish.monthly.last}
