% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{publish.monthly.last}
\alias{publish.monthly.last}
\title{publish.monthly.last}
\usage{
publish.monthly.last(x, y = 23, n = 0)
}
\arguments{
\item{x}{= a YYYYMMDD (can be missing)}

\item{y}{= calendar day allocations are known the next month}

\item{n}{= an integer}
}
\description{
date of last monthly publication
}
\seealso{
Other publish: \code{\link{publish.daily.last}},
  \code{\link{publish.weekly.last}}
}
\keyword{publish.monthly.last}
