% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{qa.flow}
\alias{qa.flow}
\title{qa.flow}
\usage{
qa.flow(x, y, n, w, h, u)
}
\arguments{
\item{x}{= a YYYYMM month}

\item{y}{= M/W/D depending on whether flows are monthly/weekly/daily}

\item{n}{= T for fund or F for share-class level}

\item{w}{= fund filter (e.g. Aggregate/Active/Passive/ETF/Mutual)}

\item{h}{= a connection, the output of odbcDriverConnect}

\item{u}{= stock filter (e.g. All/CN)}
}
\description{
Compares flow file to data from Quant server
}
\seealso{
Other qa: \code{\link{qa.columns}},
  \code{\link{qa.filter.map}}, \code{\link{qa.index}},
  \code{\link{qa.mat.read}}, \code{\link{qa.secMenu}}
}
\keyword{qa.flow}
