% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{qa.mat.read}
\alias{qa.mat.read}
\title{qa.mat.read}
\usage{
qa.mat.read(x, y, n, w, h, u = "ftp", v = F)
}
\arguments{
\item{x}{= remote file on an ftp site (e.g. "/ftpdata/mystuff/foo.txt")}

\item{y}{= local folder (e.g. "C:\\\\temp")}

\item{n}{= ftp site (defaults to standard)}

\item{w}{= user id (defaults to standard)}

\item{h}{= password (defaults to standard)}

\item{u}{= protocol (either "ftp" or "sftp")}

\item{v}{= T/F flag for ftp.use.epsv argument of getCurlHandle}
}
\description{
contents of <x> as a data frame
}
\seealso{
Other qa: \code{\link{qa.columns}},
  \code{\link{qa.filter.map}}, \code{\link{qa.flow}},
  \code{\link{qa.index}}, \code{\link{qa.secMenu}}
}
\keyword{qa.mat.read}
