% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{qtl.eq}
\alias{qtl.eq}
\title{qtl.eq}
\usage{
qtl.eq(x, y = 5)
}
\arguments{
\item{x}{= a vector/matrix/data-frame}

\item{y}{= number of desired bins}
}
\description{
performs an equal-weight binning on <x> if <x> is a vector or the rows of <x> otherwise
}
\seealso{
Other qtl: \code{\link{qtl.fast}},
  \code{\link{qtl.single.grp}}, \code{\link{qtl}}
}
\keyword{qtl.eq}
