% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{qtl.fast}
\alias{qtl.fast}
\title{qtl.fast}
\usage{
qtl.fast(x, y = 5)
}
\arguments{
\item{x}{= a vector}

\item{y}{= number of desired bins}
}
\description{
performs a FAST equal-weight binning on <x>. Can't handle NAs.
}
\seealso{
Other qtl: \code{\link{qtl.eq}},
  \code{\link{qtl.single.grp}}, \code{\link{qtl}}
}
\keyword{qtl.fast}
