% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{qtl}
\alias{qtl}
\title{qtl}
\usage{
qtl(x, y, n, w)
}
\arguments{
\item{x}{= a numeric vector}

\item{y}{= an integer (number of bins)}

\item{n}{= a numeric vector (weights)}

\item{w}{= a string vector (groups)}
}
\description{
<x> bucketed into <y> bins within groups <w> having equal weights <n>
}
\seealso{
Other qtl: \code{\link{qtl.eq}}, \code{\link{qtl.fast}},
  \code{\link{qtl.single.grp}}
}
\keyword{qtl}
