% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{qtl}
\alias{qtl}
\title{qtl}
\usage{
qtl(x, y, n, w)
}
\arguments{
\item{x}{= a vector}

\item{y}{= number of desired bins}

\item{n}{= a weight vector}

\item{w}{= a vector of groups (e.g. GSec)}
}
\description{
performs an equal-weight binning on <x> so that the members of <mem> are divided into <n> equal bins within each group <w>
}
\keyword{qtl}
