% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{qtl.single.grp}
\alias{qtl.single.grp}
\title{qtl.single.grp}
\usage{
qtl.single.grp(x, y, n)
}
\arguments{
\item{x}{= a vector}

\item{y}{= number of desired bins}

\item{n}{= a 1/0 membership vector}
}
\description{
performs an equal-weight binning on <x> so that the members of <n> are divided into <y> equal bins
}
\keyword{qtl.single.grp}
