% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{qtl.weighted}
\alias{qtl.weighted}
\title{qtl.weighted}
\usage{
qtl.weighted(x, y)
}
\arguments{
\item{x}{= an isomekic vector, lacking NA's or zeroes, that sums to unity}

\item{y}{= a number between zero and one}
}
\description{
returns a number <z> so that the sum of x[1:z] is as close as possible to <y>.
}
\seealso{
Other qtl: \code{\link{qtl.eq}},
  \code{\link{qtl.single.grp}},
  \code{\link{qtl.underlying}},
  \code{\link{qtl.zero.weight}}, \code{\link{qtl}}
}
\keyword{qtl.weighted}
