% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{qtl.zero.weight}
\alias{qtl.zero.weight}
\title{qtl.zero.weight}
\usage{
qtl.zero.weight(x, y, n, w)
}
\arguments{
\item{x}{= a vector of variables}

\item{y}{= a corresponding vector of bin assignments}

\item{n}{= a vector of variables that are to be assigned to bins}

\item{w}{= number of bins to divide <x> into}
}
\description{
assigns the members of <x> to bins
}
\seealso{
Other qtl: \code{\link{qtl.eq}}, \code{\link{qtl.fast}},
  \code{\link{qtl.single.grp}},
  \code{\link{qtl.underlying}}, \code{\link{qtl.weighted}},
  \code{\link{qtl}}
}
\keyword{qtl.zero.weight}
