% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{qtr.ex.int}
\alias{qtr.ex.int}
\title{qtr.ex.int}
\usage{
qtr.ex.int(x)
}
\arguments{
\item{x}{= an integer vector}
}
\description{
returns = a YYYYMM vector
}
\seealso{
Other qtr: \code{\link{qtr.lag}}, \code{\link{qtr.seq}},
  \code{\link{qtr.to.int}}
}
\keyword{qtr.ex.int}
