% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{qtr.to.int}
\alias{qtr.to.int}
\title{qtr.to.int}
\usage{
qtr.to.int(x)
}
\arguments{
\item{x}{= a qtr vector}
}
\description{
returns integers
}
\seealso{
Other qtr: \code{\link{qtr.ex.int}}, \code{\link{qtr.lag}},
  \code{\link{qtr.seq}}
}
\keyword{qtr.to.int}
