% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{recipient.exists}
\alias{recipient.exists}
\title{recipient.exists}
\usage{
recipient.exists(x)
}
\arguments{
\item{x}{= a string (report name)}
}
\description{
T/F depending on whether recipient list exists
}
\seealso{
Other recipient: \code{\link{recipient.read}}
}
\keyword{recipient.exists}
