% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{record.exists}
\alias{record.exists}
\title{record.exists}
\usage{
record.exists(x, y, n)
}
\arguments{
\item{x}{= report name}

\item{y}{= date for which you want to send the report}

\item{n}{= file name}
}
\description{
T/F depending on whether action already taken
}
\keyword{record.exists}
