% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{record.kill}
\alias{record.kill}
\title{record.kill}
\usage{
record.kill(x, y)
}
\arguments{
\item{x}{= report name}

\item{y}{= a file}
}
\description{
deletes entry <x> in the record <y>. Returns nothing.
}
\keyword{record.kill}
