% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{record.kill}
\alias{record.kill}
\title{record.kill}
\usage{
record.kill(x, y)
}
\arguments{
\item{x}{= a string (report name)}

\item{y}{= a file}
}
\description{
deletes entry <x> in the record <y>. Returns nothing.
}
\seealso{
Other record: \code{\link{record.exists}},
  \code{\link{record.read}},
  \code{\link{record.track.target}},
  \code{\link{record.track}}, \code{\link{record.write}}
}
\keyword{record.kill}
