% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{record.track}
\alias{record.track}
\title{record.track}
\usage{
record.track(x, y, n)
}
\arguments{
\item{x}{= a YYYYMMDD}

\item{y}{= a file}

\item{n}{= T/F depending on whether this is for standard or Asia process}
}
\description{
writes report for date <x> and type <y>
}
\keyword{record.track}
