% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{record.track}
\alias{record.track}
\title{record.track}
\usage{
record.track(x, y, n)
}
\arguments{
\item{x}{= a SINGLE YYYYMMDD date}

\item{y}{= file name}

\item{n}{= T/F depending on whether this is for standard or Asia process}
}
\description{
writes report for date <x> and type <y>
}
\seealso{
Other record: \code{\link{record.exists}},
  \code{\link{record.kill}}, \code{\link{record.read}},
  \code{\link{record.write}}
}
\keyword{record.track}
