% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{record.write}
\alias{record.write}
\title{record.write}
\usage{
record.write(x, y, n)
}
\arguments{
\item{x}{= report name}

\item{y}{= date for which you sent the report}

\item{n}{= file name}
}
\description{
updates the record. Returns nothing.
}
\keyword{record.write}
