% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{record.write}
\alias{record.write}
\title{record.write}
\usage{
record.write(x, y, n)
}
\arguments{
\item{x}{= a string (report name)}

\item{y}{= date for which you sent the report}

\item{n}{= a file}
}
\description{
updates the record. Returns nothing.
}
\seealso{
Other record: \code{\link{record.exists}},
  \code{\link{record.kill}}, \code{\link{record.read}},
  \code{\link{record.track.target}},
  \code{\link{record.track}}
}
\keyword{record.write}
