% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{refresh.predictors.daily}
\alias{refresh.predictors.daily}
\title{refresh.predictors.daily}
\usage{
refresh.predictors.daily(x, y, n, w, h = F)
}
\arguments{
\item{x}{= csv file containing the predictors}

\item{y}{= query needed to get full history}

\item{n}{= last part of the query that goes after the date restriction}

\item{w}{= one of StockFlows/Regular/Quant}

\item{h}{= T/F depending on whether you want changes in data to be ignored}
}
\description{
refreshes the text file contains flows data from SQL
}
\keyword{refresh.predictors.daily}
