% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{refresh.predictors.monthly}
\alias{refresh.predictors.monthly}
\title{refresh.predictors.monthly}
\usage{
refresh.predictors.monthly(x, y, n, w, h)
}
\arguments{
\item{x}{= a file (predictors)}

\item{y}{= query needed to get full history}

\item{n}{= a string (last part of query after date restriction)}

\item{w}{= a connection string}

\item{h}{= a boolean (ignore/report the fact last row has changed)}
}
\description{
refreshes the text file contains flows data from SQL
}
\seealso{
Other refresh: \code{\link{refresh.predictors.append}},
  \code{\link{refresh.predictors.daily}},
  \code{\link{refresh.predictors.script}},
  \code{\link{refresh.predictors.troika}},
  \code{\link{refresh.predictors.weekly}},
  \code{\link{refresh.predictors}}
}
\keyword{refresh.predictors.monthly}
