% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{refresh.predictors}
\alias{refresh.predictors}
\title{refresh.predictors}
\usage{
refresh.predictors(fcn, x, y, n, w, h, u)
}
\arguments{
\item{fcn}{= a function that returns the last complete publication period}

\item{x}{= csv file containing the predictors}

\item{y}{= query needed to get full history}

\item{n}{= last part of the query that goes after the date restriction}

\item{w}{= input to sql.connect}

\item{h}{= T/F depending on whether you want changes in data to be ignored}

\item{u}{= column corresponding to date in relevant sql table}
}
\description{
refreshes the text file contains flows data from SQL
}
\keyword{refresh.predictors}
