% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{refresh.predictors}
\alias{refresh.predictors}
\title{refresh.predictors}
\usage{
refresh.predictors(path, sql.query, sql.end.stub, connection.type,
  ignore.data.changes, date.field, publish.fcn)
}
\arguments{
\item{path}{= csv file containing the predictors}

\item{sql.query}{= query needed to get full history}

\item{sql.end.stub}{= last part of the query that goes after the date restriction}

\item{connection.type}{= one of StockFlows/Regular/Quant}

\item{ignore.data.changes}{= T/F depending on whether you want changes in data to be ignored}

\item{date.field}{= column corresponding to date in relevant sql table}

\item{publish.fcn}{= a function that returns the last complete publication period}
}
\description{
refreshes the text file contains flows data from SQL
}
\seealso{
Other refresh: \code{\link{refresh.predictors.append}},
  \code{\link{refresh.predictors.daily}},
  \code{\link{refresh.predictors.monthly}},
  \code{\link{refresh.predictors.script}},
  \code{\link{refresh.predictors.weekly}}
}
\keyword{refresh.predictors}
