% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{refresh.predictors}
\alias{refresh.predictors}
\title{refresh.predictors}
\usage{
refresh.predictors(fcn, x, y, n, w, h, u)
}
\arguments{
\item{fcn}{= a function (last complete publication period)}

\item{x}{= a file (predictors)}

\item{y}{= query needed to get full history}

\item{n}{= a string (last part of query after date restriction)}

\item{w}{= a connection string}

\item{h}{= a boolean (ignore/note data changes)}

\item{u}{= a column (corresponds to date in relevant SQL table)}
}
\description{
refreshes the text file contains flows data from SQL
}
\seealso{
Other refresh: \code{\link{refresh.predictors.append}},
  \code{\link{refresh.predictors.daily}},
  \code{\link{refresh.predictors.monthly}},
  \code{\link{refresh.predictors.script}},
  \code{\link{refresh.predictors.weekly}}
}
\keyword{refresh.predictors}
