% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{refresh.predictors.script}
\alias{refresh.predictors.script}
\title{refresh.predictors.script}
\usage{
refresh.predictors.script(x, y, n, w)
}
\arguments{
\item{x}{= query needed to get full history}

\item{y}{= a string (last part of query after date restriction)}

\item{n}{= a column (corresponds to date in relevant SQL table)}

\item{w}{= last date for which you already have data}
}
\description{
generates the SQL script to refresh predictors
}
\seealso{
Other refresh: \code{\link{refresh.predictors.append}},
  \code{\link{refresh.predictors.daily}},
  \code{\link{refresh.predictors.troika}},
  \code{\link{refresh.predictors.weekly}},
  \code{\link{refresh.predictors}}
}
\keyword{refresh.predictors.script}
