% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{reshape.long}
\alias{reshape.long}
\title{reshape.long}
\usage{
reshape.long(x, y, n = "id", w = "val")
}
\arguments{
\item{x}{= a data-frame (can't be a matrix)}

\item{y}{= column names to be compressed}

\item{n}{= name of the new column created for identifiers}

\item{w}{= name of the new column created for values}
}
\description{
<x> in long format
}
\keyword{reshape.long}
