% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{reshape.wide}
\alias{reshape.wide}
\title{reshape.wide}
\usage{
reshape.wide(x)
}
\arguments{
\item{x}{= a matrix/data-frame with 3 columns corresponding respectively with the rows, columns and entries of the resulting matrix}
}
\description{
converts <x> to a matrix
}
\keyword{reshape.wide}
