% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{reshape.wide}
\alias{reshape.wide}
\title{reshape.wide}
\usage{
reshape.wide(x)
}
\arguments{
\item{x}{= a matrix/data-frame with last columns corresponding to the entries of the resulting array}
}
\description{
converts <x> to an array
}
\keyword{reshape.wide}
