% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ret.ex.idx}
\alias{ret.ex.idx}
\title{ret.ex.idx}
\usage{
ret.ex.idx(x, y, n, w)
}
\arguments{
\item{x}{= a matrix/data frame (total return indices)}

\item{y}{= an integer (return window)}

\item{n}{= a boolean (label returns by beginning/end of the period)}

\item{w}{= a boolean (returns/spread changes)}
}
\description{
computes return
}
\seealso{
Other ret: \code{\link{ret.idx.gaps.fix}},
  \code{\link{ret.outliers}}, \code{\link{ret.to.idx}},
  \code{\link{ret.to.log}}
}
\keyword{ret.ex.idx}
