% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ret.ex.idx}
\alias{ret.ex.idx}
\title{ret.ex.idx}
\usage{
ret.ex.idx(x, y, n, w)
}
\arguments{
\item{x}{= a file of total return indices indexed so that time runs forward}

\item{y}{= number of periods over which the return is computed}

\item{n}{= if T simple positional lagging is used. If F, yyyymm.lag is invoked.}

\item{w}{= if T the result is labelled by the beginning of the period, else by the end.}
}
\description{
computes return
}
\seealso{
Other ret: \code{\link{ret.idx.gaps.fix}},
  \code{\link{ret.outliers}}, \code{\link{ret.to.idx}}
}
\keyword{ret.ex.idx}
