% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ret.outliers}
\alias{ret.outliers}
\title{ret.outliers}
\usage{
ret.outliers(x, y = 1.5)
}
\arguments{
\item{x}{= a vector of returns}

\item{y}{= outlier threshold}
}
\description{
Sets big ones to NA (a way to control for splits)
}
\keyword{ret.outliers}
