% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{ret.to.idx}
\alias{ret.to.idx}
\title{ret.to.idx}
\usage{
ret.to.idx(x)
}
\arguments{
\item{x}{== a file (total return indices)}
}
\description{
computes a total-return index
}
\seealso{
Other ret: \code{\link{ret.ex.idx}},
  \code{\link{ret.idx.gaps.fix}},
  \code{\link{ret.outliers}}, \code{\link{ret.to.log}}
}
\keyword{ret.to.idx}
