% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{rpt.email}
\alias{rpt.email}
\title{rpt.email}
\usage{
rpt.email(x, y, n, w, h, u, v)
}
\arguments{
\item{x}{= a folder}

\item{y}{= output type (".csv", ".pdf", etc.)}

\item{n}{= a boolean (do/don't check if latest)}

\item{w}{= a boolean (live/test)}

\item{h}{= a string (recipient email, can be missing)}

\item{u}{= a file (log, can be missing)}

\item{v}{= a string vector (report names, can be missing)}
}
\description{
emails report
}
\seealso{
Other rpt: \code{\link{rpt.email.send}}
}
\keyword{rpt.email}
