% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{rpt.email.send}
\alias{rpt.email.send}
\title{rpt.email.send}
\usage{
rpt.email.send(x, y, n, w, h, u)
}
\arguments{
\item{x}{= a string (report name)}

\item{y}{= the email address(es) of the recipient(s)}

\item{n}{= a flowdate}

\item{w}{= a boolean (live/test)}

\item{h}{= a file vector}

\item{u}{= a boolean (SendGrid/regular)}
}
\description{
emails report
}
\seealso{
Other rpt: \code{\link{rpt.email}}
}
\keyword{rpt.email.send}
