% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{rpt.email.send}
\alias{rpt.email.send}
\title{rpt.email.send}
\usage{
rpt.email.send(x, y, n, w, h)
}
\arguments{
\item{x}{= report name}

\item{y}{= the email address(es) of the recipient(s)}

\item{n}{= a flowdate}

\item{w}{= T/F depending on whether you're sending the email or testing}

\item{h}{= a file vector}
}
\description{
emails report
}
\keyword{rpt.email.send}
