% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{rrw}
\alias{rrw}
\title{rrw}
\usage{
rrw(prdBeg, prdEnd, vbls, univ, grp.nm, ret.nm, fldr, orth.factor = NULL,
  classif)
}
\arguments{
\item{prdBeg}{= a first-return date in yyyymm format representing the first month of the backtest}

\item{prdEnd}{= a first-return date in yyyymm format representing the last month of the backtest}

\item{vbls}{= vector of variables against which return is to be regressed}

\item{univ}{= universe (e.g. "R1Mem")}

\item{grp.nm}{= neutrality group (e.g. "GSec")}

\item{ret.nm}{= return variable (e.g. "Ret")}

\item{fldr}{= stock-flows folder}

\item{orth.factor}{= factor to orthogonalize all variables to (e.g. "PrcMo")}

\item{classif}{= classif file}
}
\description{
Runs regressions
}
\keyword{rrw}
