% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{rrw}
\alias{rrw}
\title{rrw}
\usage{
rrw(x, y, n, w, h, u, v, g = NULL, r)
}
\arguments{
\item{x}{= a first-return date in yyyymm format representing the first month of the backtest}

\item{y}{= a first-return date in yyyymm format representing the last month of the backtest}

\item{n}{= vector of variables against which return is to be regressed}

\item{w}{= universe (e.g. "R1Mem", or c("EafeMem", 1, "CountryCode", "JP"))}

\item{h}{= neutrality group (e.g. "GSec")}

\item{u}{= return variable (e.g. "Ret")}

\item{v}{= stock-flows folder}

\item{g}{= factor to orthogonalize all variables to (e.g. "PrcMo")}

\item{r}{= classif file}
}
\description{
regression results
}
\seealso{
Other rrw: \code{\link{rrw.factors}},
  \code{\link{rrw.underlying}}
}
\keyword{rrw}
