% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{rrw.underlying}
\alias{rrw.underlying}
\title{rrw.underlying}
\usage{
rrw.underlying(prd, vbls, univ, grp.nm, ret.nm, fldr, orth.factor, classif)
}
\arguments{
\item{prd}{= a first-return date in yyyymm format representing the return period of interest}

\item{vbls}{= vector of variables against which return is to be regressed}

\item{univ}{= universe (e.g. "R1Mem")}

\item{grp.nm}{= neutrality group (e.g. "GSec")}

\item{ret.nm}{= return variable (e.g. "Ret")}

\item{fldr}{= parent directory containing derived/data}

\item{orth.factor}{= factor to orthogonalize all variables to (e.g. "PrcMo")}

\item{classif}{= classif file}
}
\description{
Runs regressions
}
\keyword{rrw.underlying}
