% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{rrw.underlying}
\alias{rrw.underlying}
\title{rrw.underlying}
\usage{
rrw.underlying(x, y, n, w, h, u, v, g)
}
\arguments{
\item{x}{= a YYYYMM, first-return date representing the return period of interest}

\item{y}{= vector of variables against which return is to be regressed}

\item{n}{= universe (e.g. "R1Mem", or c("EafeMem", 1, "CountryCode", "JP"))}

\item{w}{= neutrality group (e.g. "GSec")}

\item{h}{= return variable (e.g. "Ret")}

\item{u}{= parent directory containing derived/data}

\item{v}{= factor to orthogonalize all variables to (e.g. "PrcMo")}

\item{g}{= classif file}
}
\description{
Runs regressions
}
\keyword{rrw.underlying}
