% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{separating.hyperplane}
\alias{separating.hyperplane}
\title{separating.hyperplane}
\usage{
separating.hyperplane(x, y)
}
\arguments{
\item{x}{= a membership vector (length dim(<x>)[2] - 1)}

\item{y}{= a matrix (first column is a membership vector)}
}
\description{
number of errors and distance from origin for best separating hyperlane
}
\keyword{separating.hyperplane}
