% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf.bin.nms}
\alias{sf.bin.nms}
\title{sf.bin.nms}
\usage{
sf.bin.nms(x, y)
}
\arguments{
\item{x}{= number of bins}

\item{y}{= T/F depending on whether you want universe returns returned}
}
\description{
returns bin names
}
\seealso{
Other sf: \code{\link{sf.detail}},
  \code{\link{sf.single.bsim}}, \code{\link{sf.subset}},
  \code{\link{sf.underlying.data.bin}},
  \code{\link{sf.underlying.data}},
  \code{\link{sf.underlying.summ}},
  \code{\link{sf.vec.to.array}}, \code{\link{sf}}
}
\keyword{sf.bin.nms}
