% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf.daily}
\alias{sf.daily}
\title{sf.daily}
\usage{
sf.daily(x, y, n, w, h, u, v, g = 5, r)
}
\arguments{
\item{x}{= first-return date in YYYYMMDD}

\item{y}{= first-return date in YYYYMMDD after <prdBeg>}

\item{n}{= membership (e.g. "R2Mem")}

\item{w}{= binning group (e.g. "GSec")}

\item{h}{= return variable}

\item{u}{= variable & lookback (e.g. c("1dFloMo", 3))}

\item{v}{= data folder}

\item{g}{= number of bins}

\item{r}{= classif file}
}
\description{
runs a daily stock-flows simulation FAST
}
\seealso{
Other sf: \code{\link{sf.bin.nms}},
  \code{\link{sf.detail}}, \code{\link{sf.single.bsim}},
  \code{\link{sf.subset}},
  \code{\link{sf.underlying.data.bin}},
  \code{\link{sf.underlying.data}},
  \code{\link{sf.underlying.summ}}, \code{\link{sf}}
}
\keyword{sf.daily}
