% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf.daily}
\alias{sf.daily}
\title{sf.daily}
\usage{
sf.daily(prdBeg, prdEnd, vbl.nm, univ, grp.nm, ret.nm, trail, sum.flows, fldr,
  vbl.lag, dly.vbl, retHz, classif)
}
\arguments{
\item{prdBeg}{= first-return date in YYYYMMDD}

\item{prdEnd}{= first-return date in YYYYMMDD (must postdate <prdBeg>)}

\item{vbl.nm}{= variable}

\item{univ}{= membership (e.g. "EafeMem" or c("GemMem", 1))}

\item{grp.nm}{= group within which binning is to be performed}

\item{ret.nm}{= return variable}

\item{trail}{= number of trailing periods to compound/sum over}

\item{sum.flows}{= T/F depending on whether you want flows summed or compounded.}

\item{fldr}{= data folder}

\item{vbl.lag}{= lags by <vbl.lag> weekdays or months depending on whether <dly.vbl> is true.}

\item{dly.vbl}{= whether the predictor is daily or monthly}

\item{retHz}{= forward return horizon in days}

\item{classif}{= classif file}
}
\description{
runs stock-flows simulation
}
\seealso{
Other sf: \code{\link{sf.bin.nms}},
  \code{\link{sf.detail}}, \code{\link{sf.single.bsim}},
  \code{\link{sf.subset}}, \code{\link{sf.summ}},
  \code{\link{sf.underlying.data}},
  \code{\link{sf.underlying.summ}},
  \code{\link{sf.underlying}}, \code{\link{sf}}
}
\keyword{sf.daily}
