% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf.daily}
\alias{sf.daily}
\title{sf.daily}
\usage{
sf.daily(x, y, n, w, h, u, v, g = 5, r, s = NULL, b)
}
\arguments{
\item{x}{= a flowdate (first-return date, backtest start)}

\item{y}{= a flowdate (first-return date, backtest end)}

\item{n}{= a universe (e.g. "R1Mem", or c("EafeMem", 1, "CountryCode", "JP"))}

\item{w}{= a variable (neutrality group)}

\item{h}{= a variable (return)}

\item{u}{= variable & lookback (e.g. c("1dFloMo", 3))}

\item{v}{= a folder (data)}

\item{g}{= an integer (number of bins)}

\item{r}{= classif file}

\item{s}{= an integer (NULL for daily or the day you trade, 0 = Sun, 1 = Mon, etc.)}

\item{b}{= an integer (return window in days, can be missing)}
}
\description{
runs a daily stock-flows simulation FAST
}
\seealso{
Other sf: \code{\link{sf.bin.nms}},
  \code{\link{sf.detail}}, \code{\link{sf.single.bsim}},
  \code{\link{sf.subset}},
  \code{\link{sf.underlying.data.bin}},
  \code{\link{sf.underlying.data}},
  \code{\link{sf.underlying.summ}}, \code{\link{sf}}
}
\keyword{sf.daily}
