% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf.detail}
\alias{sf.detail}
\title{sf.detail}
\usage{
sf.detail(fcn, prdBeg, prdEnd, univ, grp.nm, ret.nm, trail, fldr, nBins = 5,
  classif, weighting.factor = NULL)
}
\arguments{
\item{fcn}{= function that fetches data for the appropriate period and parameter}

\item{prdBeg}{= first-return date in YYYYMM}

\item{prdEnd}{= first-return date in YYYYMM after <prdBeg>}

\item{univ}{= membership (e.g. "EafeMem" or c("GemMem", 1))}

\item{grp.nm}{= group within which binning is to be performed}

\item{ret.nm}{= return variable}

\item{trail}{= variable parameter}

\item{fldr}{= data folder}

\item{nBins}{= number of bins}

\item{classif}{= classif file}

\item{weighting.factor}{= factor you want to use for Cap-weighted back-tests (defaults to NULL)}
}
\description{
runs a stock-flows simulation
}
\seealso{
Other sf: \code{\link{sf.bin.nms}}, \code{\link{sf.daily}},
  \code{\link{sf.single.bsim}}, \code{\link{sf.subset}},
  \code{\link{sf.underlying.data}},
  \code{\link{sf.underlying.summ}}, \code{\link{sf}}
}
\keyword{sf.detail}
