% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf.detail}
\alias{sf.detail}
\title{sf.detail}
\usage{
sf.detail(fcn, x, y, n, w, h, u, v, g = 5, r, s = NULL)
}
\arguments{
\item{fcn}{= function that fetches data for the appropriate period and parameter}

\item{x}{= first-return date in YYYYMM}

\item{y}{= first-return date in YYYYMM after <prdBeg>}

\item{n}{= membership (e.g. "EafeMem" or c("GemMem", 1))}

\item{w}{= group within which binning is to be performed}

\item{h}{= return variable}

\item{u}{= variable parameter}

\item{v}{= data folder}

\item{g}{= number of bins or numeric vector with last element T/F for dependent/independent}

\item{r}{= classif file}

\item{s}{= factor you want to use for Cap-weighted back-tests (defaults to NULL)}
}
\description{
runs a stock-flows simulation
}
\seealso{
Other sf: \code{\link{sf.bin.nms}}, \code{\link{sf.daily}},
  \code{\link{sf.single.bsim}}, \code{\link{sf.subset}},
  \code{\link{sf.underlying.data.bin}},
  \code{\link{sf.underlying.data}},
  \code{\link{sf.underlying.summ}}, \code{\link{sf}}
}
\keyword{sf.detail}
