% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf}
\alias{sf}
\title{sf}
\usage{
sf(prdBeg, prdEnd, vbl.nm, univ, grp.nm, ret.nm, trails, sum.flows, fldr,
  dly.vbl = T, vbl.lag = 0, nBins = 5, reverse.vbl = F, geom.comp = F,
  retHz = 1, classif)
}
\arguments{
\item{prdBeg}{= first-return date in YYYYMM}

\item{prdEnd}{= first-return date in YYYYMM after <prdBeg>}

\item{vbl.nm}{= variable}

\item{univ}{= membership (e.g. "EafeMem" or c("GemMem", 1))}

\item{grp.nm}{= group within which binning is to be performed}

\item{ret.nm}{= return variable}

\item{trails}{= number of trailing periods to compound/sum over}

\item{sum.flows}{= T/F depending on whether you want flows summed or compounded.}

\item{fldr}{= data folder}

\item{dly.vbl}{= if T then a daily predictor is assumed else a monthly one}

\item{vbl.lag}{= lags by <vbl.lag> weekdays or months depending on whether <dly.vbl> is true.}

\item{nBins}{= number of bins}

\item{reverse.vbl}{= T/F depending on whether you want the variable reversed}

\item{geom.comp}{= T/F depending on whether you want bin excess returns summarized geometrically or arithmetically}

\item{retHz}{= forward return horizon in months}

\item{classif}{= classif file}
}
\description{
runs a stock-flows simulation
}
\seealso{
Other sf: \code{\link{sf.bin.nms}},
  \code{\link{sf.daily.summ}}, \code{\link{sf.daily}},
  \code{\link{sf.single.bsim}}, \code{\link{sf.subset}},
  \code{\link{sf.summ}}, \code{\link{sf.underlying.data}},
  \code{\link{sf.underlying.summ}},
  \code{\link{sf.underlying}}
}
\keyword{sf}
