% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf}
\alias{sf}
\title{sf}
\usage{
sf(fcn, x, y, n, w, h, u, v, g = 5, r, s = 1, b = F)
}
\arguments{
\item{fcn}{= a fetch function}

\item{x}{= a YYYYMM (first-return date, backtest start)}

\item{y}{= a YYYYMM (first-return date, backtest end)}

\item{n}{= a universe (e.g. "R1Mem", or c("EafeMem", 1, "CountryCode", "JP"))}

\item{w}{= a variable (neutrality group)}

\item{h}{= a variable (return)}

\item{u}{= a string (parameter)}

\item{v}{= a folder (must have sub-folders data/derived)}

\item{g}{= an integer (number of bins)}

\item{r}{= classif file}

\item{s}{= an integer (forward return horizon)}

\item{b}{= a boolean (geometric/artihmetic returns)}
}
\description{
runs a stock-flows simulation
}
\seealso{
Other sf: \code{\link{sf.bin.nms}}, \code{\link{sf.daily}},
  \code{\link{sf.detail}}, \code{\link{sf.single.bsim}},
  \code{\link{sf.subset}},
  \code{\link{sf.underlying.data.bin}},
  \code{\link{sf.underlying.data}},
  \code{\link{sf.underlying.summ}}
}
\keyword{sf}
