% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf}
\alias{sf}
\title{sf}
\usage{
sf(fcn, x, y, n, w, h, u, v, g = 5, r, s = 1, b = F)
}
\arguments{
\item{fcn}{= function that fetches data for the appropriate period and parameter}

\item{x}{= first-return date in YYYYMM}

\item{y}{= first-return date in YYYYMM after <prdBeg>}

\item{n}{= membership (e.g. "EafeMem" or c("GemMem", 1))}

\item{w}{= group within which binning is to be performed}

\item{h}{= return variable}

\item{u}{= variable parameter}

\item{v}{= data folder}

\item{g}{= number of bins}

\item{r}{= classif file}

\item{s}{= forward return horizon in months}

\item{b}{= T/F depending on whether excess return summary is geometric or arithmetic}
}
\description{
runs a stock-flows simulation
}
\seealso{
Other sf: \code{\link{sf.bin.nms}},
  \code{\link{sf.detail}}, \code{\link{sf.single.bsim}},
  \code{\link{sf.subset}},
  \code{\link{sf.underlying.data.bin}},
  \code{\link{sf.underlying.data}},
  \code{\link{sf.underlying.summ}}
}
\keyword{sf}
