% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf.single.bsim}
\alias{sf.single.bsim}
\title{sf.single.bsim}
\usage{
sf.single.bsim(prdBeg, prdEnd, vbl.nm, univ, grp.nm, ret.nm, fldr,
  dly.vbl = F, trail = 1, sum.flows = T, vbl.lag = 0, uRet = F,
  nBins = 5, reverse.vbl = F, retHz = 1, classif)
}
\arguments{
\item{prdBeg}{= first-return date in YYYYMM}

\item{prdEnd}{= first-return date in YYYYMM after <prdBeg>}

\item{vbl.nm}{= variable}

\item{univ}{= membership (e.g. "EafeMem" or c("GemMem", 1))}

\item{grp.nm}{= group within which binning is to be performed}

\item{ret.nm}{= return variable}

\item{fldr}{= data folder}

\item{dly.vbl}{= T/F depending on whether the variable used is daily or monthly}

\item{trail}{= number of trailing periods to compound/sum over}

\item{sum.flows}{= if T, flows get summed. Otherwise they get compounded.}

\item{vbl.lag}{= lags by <vbl.lag> weekdays or months depending on whether <dly.vbl> is true.}

\item{uRet}{= T/F depending on whether the equal-weight universe return is desired}

\item{nBins}{= number of bins}

\item{reverse.vbl}{= T/F depending on whether you want the variable reversed}

\item{retHz}{= forward return horizon in months}

\item{classif}{= classif file}
}
\description{
runs a single quintile simulation
}
\seealso{
Other sf: \code{\link{sf.bin.nms}}, \code{\link{sf.daily}},
  \code{\link{sf.detail}}, \code{\link{sf.subset}},
  \code{\link{sf.underlying.data}},
  \code{\link{sf.underlying.summ}},
  \code{\link{sf.underlying}}, \code{\link{sf}}
}
\keyword{sf.single.bsim}
