% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf.single.bsim}
\alias{sf.single.bsim}
\title{sf.single.bsim}
\usage{
sf.single.bsim(fcn, x, y, n, w, h, u, v, g = 5, r, s = 1, b = T,
  p = NULL)
}
\arguments{
\item{fcn}{= a fetch function}

\item{x}{= a YYYYMM (first-return date, backtest start)}

\item{y}{= a YYYYMM (first-return date, backtest end) after <prdBeg>}

\item{n}{= a universe (e.g. "R1Mem", or c("EafeMem", 1, "CountryCode", "JP"))}

\item{w}{= a variable (neutrality group)}

\item{h}{= a variable (return)}

\item{u}{= a string (parameter)}

\item{v}{= a folder (must have sub-folders data/derived)}

\item{g}{= an integer vector (if vector, last element T/F for dep/indep binning))}

\item{r}{= classif file}

\item{s}{= an integer (forward return horizon)}

\item{b}{= a boolean (equal/cap-weight universe return)}

\item{p}{= a variable (factor for cap weighting, defaults to NULL)}
}
\description{
runs a single quintile simulation
}
\seealso{
Other sf: \code{\link{sf.bin.nms}}, \code{\link{sf.daily}},
  \code{\link{sf.detail}}, \code{\link{sf.subset}},
  \code{\link{sf.underlying.data.bin}},
  \code{\link{sf.underlying.data}},
  \code{\link{sf.underlying.summ}}, \code{\link{sf}}
}
\keyword{sf.single.bsim}
