% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf.subset}
\alias{sf.subset}
\title{sf.subset}
\usage{
sf.subset(x, y, n, w)
}
\arguments{
\item{x}{= membership (e.g. "EafeMem" or c("GemMem", 1))}

\item{y}{= a YYYYMM or YYYYMMDD}

\item{n}{= folder in which to find the data}

\item{w}{= classif file}
}
\description{
Returns a 1/0 mem vector
}
\keyword{sf.subset}
