% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf.subset}
\alias{sf.subset}
\title{sf.subset}
\usage{
sf.subset(x, y, n, w)
}
\arguments{
\item{x}{= membership (e.g. "EafeMem" or c("GemMem", 1))}

\item{y}{= a YYYYMM or YYYYMMDD}

\item{n}{= folder in which to find the data}

\item{w}{= classif file}
}
\description{
Returns a 1/0 mem vector
}
\seealso{
Other sf: \code{\link{sf.bin.nms}},
  \code{\link{sf.daily.summ}}, \code{\link{sf.daily}},
  \code{\link{sf.single.bsim}}, \code{\link{sf.summ}},
  \code{\link{sf.underlying.data}},
  \code{\link{sf.underlying.summ}},
  \code{\link{sf.underlying}}, \code{\link{sf}}
}
\keyword{sf.subset}
