% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf.summ}
\alias{sf.summ}
\title{sf.summ}
\usage{
sf.summ(fcn, x, y)
}
\arguments{
\item{fcn}{= a function that summarizes the data}

\item{x}{= a df of bin returns indexed by time}

\item{y}{= forward return horizon in months}
}
\description{
summarizes the back test
}
\seealso{
Other sf: \code{\link{sf.bin.nms}}, \code{\link{sf.daily}},
  \code{\link{sf.detail}}, \code{\link{sf.single.bsim}},
  \code{\link{sf.subset}},
  \code{\link{sf.underlying.data}},
  \code{\link{sf.underlying.summ}},
  \code{\link{sf.underlying}}, \code{\link{sf}}
}
\keyword{sf.summ}
