% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf.underlying.data.bin}
\alias{sf.underlying.data.bin}
\title{sf.underlying.data.bin}
\usage{
sf.underlying.data.bin(x, y, n, w)
}
\arguments{
\item{x}{= a numeric vector (or vector list)}

\item{y}{= an integer vector (if vector, last element T/F for dep/indep binning)}

\item{n}{= a numeric vector (weights)}

\item{w}{= a string vector (binning group)}
}
\description{
character vector of bin memberships
}
\seealso{
Other sf: \code{\link{sf.bin.nms}}, \code{\link{sf.daily}},
  \code{\link{sf.detail}}, \code{\link{sf.single.bsim}},
  \code{\link{sf.subset}},
  \code{\link{sf.underlying.data}},
  \code{\link{sf.underlying.summ}}, \code{\link{sf}}
}
\keyword{sf.underlying.data.bin}
