% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf.underlying.data.bin}
\alias{sf.underlying.data.bin}
\title{sf.underlying.data.bin}
\usage{
sf.underlying.data.bin(x, y, n, w)
}
\arguments{
\item{x}{= either vector or list of vectors}

\item{y}{= number of bins or numeric vector with last element T/F for dependent/independent}

\item{n}{= numeric vector of weighting factors}

\item{w}{= vector of binning groups}
}
\description{
character vector of bin memberships
}
\seealso{
Other sf: \code{\link{sf.bin.nms}},
  \code{\link{sf.detail}}, \code{\link{sf.single.bsim}},
  \code{\link{sf.subset}},
  \code{\link{sf.underlying.data}},
  \code{\link{sf.underlying.summ}}, \code{\link{sf}}
}
\keyword{sf.underlying.data.bin}
