% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf.underlying.data}
\alias{sf.underlying.data}
\title{sf.underlying.data}
\usage{
sf.underlying.data(fcn, x, y, n, w, h, u, v, g, r, s)
}
\arguments{
\item{fcn}{= function that fetches data for the appropriate period and parameter}

\item{x}{= membership (e.g. "EafeMem" or c("GemMem", 1))}

\item{y}{= return variable}

\item{n}{= the period for which you want returns}

\item{w}{= variable parameter}

\item{h}{= group within which binning is to be performed}

\item{u}{= number of bins or numeric vector with last element T/F for dependent/independent}

\item{v}{= data folder}

\item{g}{= forward return horizon in months}

\item{r}{= classif file}

\item{s}{= factor you want to use for Cap-weighted back-tests}
}
\description{
Gets data needed to back-test a single period
}
\seealso{
Other sf: \code{\link{sf.bin.nms}},
  \code{\link{sf.detail}}, \code{\link{sf.single.bsim}},
  \code{\link{sf.subset}},
  \code{\link{sf.underlying.data.bin}},
  \code{\link{sf.underlying.summ}},
  \code{\link{sf.vec.to.array}}, \code{\link{sf}}
}
\keyword{sf.underlying.data}
