% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf.underlying.data}
\alias{sf.underlying.data}
\title{sf.underlying.data}
\usage{
sf.underlying.data(fcn, univ, ret.nm, ret.prd, trail, grp, nBins, fldr, retHz,
  classif, weighting.factor)
}
\arguments{
\item{fcn}{= function that fetches data for the appropriate period and parameter}

\item{univ}{= membership (e.g. "EafeMem" or c("GemMem", 1))}

\item{ret.nm}{= return variable}

\item{ret.prd}{= the period for which you want returns}

\item{trail}{= variable parameter}

\item{grp}{= group within which binning is to be performed}

\item{nBins}{= number of bins}

\item{fldr}{= data folder}

\item{retHz}{= forward return horizon in months}

\item{classif}{= classif file}

\item{weighting.factor}{= factor you want to use for Cap-weighted back-tests}
}
\description{
Gets data needed to back-test a single period
}
\seealso{
Other sf: \code{\link{sf.bin.nms}}, \code{\link{sf.daily}},
  \code{\link{sf.detail}}, \code{\link{sf.single.bsim}},
  \code{\link{sf.subset}},
  \code{\link{sf.underlying.summ}}, \code{\link{sf}}
}
\keyword{sf.underlying.data}
