% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf.underlying.data}
\alias{sf.underlying.data}
\title{sf.underlying.data}
\usage{
sf.underlying.data(vbl.nm, univ, ret.nm, ret.prd, trail, sum.flows, grp,
  dly.vbl, nBins, fldr, vbl.lag, reverse.vbl, retHz, classif)
}
\arguments{
\item{vbl.nm}{= variable}

\item{univ}{= membership (e.g. "EafeMem" or c("GemMem", 1))}

\item{ret.nm}{= return variable}

\item{ret.prd}{= the period for which you want returns}

\item{trail}{= number of trailing periods to compound/sum over}

\item{sum.flows}{= if T, flows get summed. Otherwise they get compounded.}

\item{grp}{= group within which binning is to be performed}

\item{dly.vbl}{= if T then a daily predictor is assumed else a monthly one}

\item{nBins}{= number of bins}

\item{fldr}{= data folder}

\item{vbl.lag}{= lags by <vbl.lag> weekdays or months depending on whether <dly.vbl> is true.}

\item{reverse.vbl}{= T/F depending on whether you want the variable reversed}

\item{retHz}{= forward return horizon in months}

\item{classif}{= classif file}
}
\description{
Gets data needed to back-test a single period
}
\seealso{
Other sf: \code{\link{sf.bin.nms}}, \code{\link{sf.daily}},
  \code{\link{sf.single.bsim}}, \code{\link{sf.subset}},
  \code{\link{sf.summ}}, \code{\link{sf.underlying.summ}},
  \code{\link{sf.underlying}}, \code{\link{sf}}
}
\keyword{sf.underlying.data}
