% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf.underlying.data}
\alias{sf.underlying.data}
\title{sf.underlying.data}
\usage{
sf.underlying.data(fcn, x, y, n, w, h, u, v, g, r, s)
}
\arguments{
\item{fcn}{= function that fetches data for the appropriate period and parameter}

\item{x}{= the period for which you want returns}

\item{y}{= factor you want to use for Cap-weighted back-tests}

\item{n}{= membership (e.g. "EafeMem" or c("GemMem", 1))}

\item{w}{= group within which binning is to be performed}

\item{h}{= return variable}

\item{u}{= variable parameter}

\item{v}{= data folder}

\item{g}{= number of bins or numeric vector with last element T/F for dependent/independent}

\item{r}{= classif file}

\item{s}{= forward return horizon in months}
}
\description{
Gets data needed to back-test a single period
}
\keyword{sf.underlying.data}
