% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf.underlying.summ}
\alias{sf.underlying.summ}
\title{sf.underlying.summ}
\usage{
sf.underlying.summ(x, y)
}
\arguments{
\item{x}{= a matrix/data frame (columns bin/ret/mem/wgt)}

\item{y}{= a boolean (report/ignore universe return)}
}
\description{
Returns a named vector of bin returns
}
\seealso{
Other sf: \code{\link{sf.bin.nms}}, \code{\link{sf.daily}},
  \code{\link{sf.detail}}, \code{\link{sf.single.bsim}},
  \code{\link{sf.subset}},
  \code{\link{sf.underlying.data.bin}},
  \code{\link{sf.underlying.data}}, \code{\link{sf}}
}
\keyword{sf.underlying.summ}
