% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sf.underlying.summ}
\alias{sf.underlying.summ}
\title{sf.underlying.summ}
\usage{
sf.underlying.summ(x, y, n, w, h)
}
\arguments{
\item{x}{= vector of bins}

\item{y}{= corresponding numeric vector of forward returns}

\item{n}{= corresponding 1/0 universe membership vector}

\item{w}{= number of bins}

\item{h}{= T/F variable controlling whether universe return is returned}
}
\description{
Returns a named vector of bin returns
}
\seealso{
Other sf: \code{\link{sf.bin.nms}}, \code{\link{sf.daily}},
  \code{\link{sf.detail}}, \code{\link{sf.single.bsim}},
  \code{\link{sf.subset}},
  \code{\link{sf.underlying.data}},
  \code{\link{sf.underlying}}, \code{\link{sf}}
}
\keyword{sf.underlying.summ}
