% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sfpd.ActWtTrend}
\alias{sfpd.ActWtTrend}
\title{sfpd.ActWtTrend}
\usage{
sfpd.ActWtTrend(x, y, n, w, h)
}
\arguments{
\item{x}{= a matrix/data frame (flows)}

\item{y}{= a matrix/data frame (holdings)}

\item{n}{= a matrix/data frame (FundHistory)}

\item{w}{= a flowdate}

\item{h}{= a variable (ActWtTrend/ActWtDiff/ActWtDiff2)}
}
\description{
<h> factor
}
\seealso{
Other sfpd: \code{\link{sfpd.FloDollar}},
  \code{\link{sfpd.FloMo.underlying}},
  \code{\link{sfpd.FloMo}},
  \code{\link{sfpd.FloTrend.underlying}},
  \code{\link{sfpd.FloTrend}}, \code{\link{sfpd.Flow}},
  \code{\link{sfpd.Holdings}}, \code{\link{sfpd.ION}},
  \code{\link{sfpd.Wt}}, \code{\link{sfpd.filter}}
}
\keyword{sfpd.ActWtTrend}
