% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sfpd.filter}
\alias{sfpd.filter}
\title{sfpd.filter}
\usage{
sfpd.filter(x, y)
}
\arguments{
\item{x}{= a matrix/data frame (FundHistory)}

\item{y}{= a filter (one of All/Act/Pas/Etf/Mutual)}
}
\description{
T/F depending on whether records of <x> pass filter <y>
}
\seealso{
Other sfpd: \code{\link{sfpd.ActWtTrend}},
  \code{\link{sfpd.AllocMo}}, \code{\link{sfpd.FloDollar}},
  \code{\link{sfpd.FloMo.underlying}},
  \code{\link{sfpd.FloMo}},
  \code{\link{sfpd.FloTrend.holdings}},
  \code{\link{sfpd.FloTrend.parameterize}},
  \code{\link{sfpd.FloTrend.underlying}},
  \code{\link{sfpd.FloTrend}}, \code{\link{sfpd.Flow}},
  \code{\link{sfpd.Holdings}}, \code{\link{sfpd.ION}},
  \code{\link{sfpd.Wt}}
}
\keyword{sfpd.filter}
