% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sfpd.FloMo.underlying}
\alias{sfpd.FloMo.underlying}
\title{sfpd.FloMo.underlying}
\usage{
sfpd.FloMo.underlying(x, y, n, w, h = NULL)
}
\arguments{
\item{x}{= a matrix/data frame (flows)}

\item{y}{= a matrix/data frame (holdings)}

\item{n}{= a matrix/data frame (FundHistory)}

\item{w}{= a string vector (Flow/AssetsStart/AssetsEnd/PortfolioChange)}

\item{h}{= a column vector (in FundHistory, can be NULL)}
}
\description{
FloMo factor
}
\seealso{
Other sfpd: \code{\link{sfpd.ActWtTrend}},
  \code{\link{sfpd.AllocMo}}, \code{\link{sfpd.AllocSkew}},
  \code{\link{sfpd.FloDollar}}, \code{\link{sfpd.FloMo}},
  \code{\link{sfpd.FloTrend.holdings}},
  \code{\link{sfpd.FloTrend.parameterize}},
  \code{\link{sfpd.FloTrend.underlying}},
  \code{\link{sfpd.FloTrend}}, \code{\link{sfpd.Flow}},
  \code{\link{sfpd.Holdings}}, \code{\link{sfpd.ION}},
  \code{\link{sfpd.Wt}}, \code{\link{sfpd.filter}}
}
\keyword{sfpd.FloMo.underlying}
