% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sfpd.FloTrend}
\alias{sfpd.FloTrend}
\title{sfpd.FloTrend}
\usage{
sfpd.FloTrend(x, y, n, w, h, u, v)
}
\arguments{
\item{x}{= a matrix/data frame (flows)}

\item{y}{= a matrix/data frame (holdings)}

\item{n}{= a matrix/data frame (FundHistory)}

\item{w}{= a flowdate}

\item{h}{= a variable vector (FloTrend/FloDiff/FloDiff2)}

\item{u}{= a matrix/data frame (prior holdings)}

\item{v}{= a matrix/data frame (security history)}
}
\description{
<h> factor
}
\seealso{
Other sfpd: \code{\link{sfpd.ActWtTrend}},
  \code{\link{sfpd.AllocMo}}, \code{\link{sfpd.AllocSkew}},
  \code{\link{sfpd.FloDollar}},
  \code{\link{sfpd.FloMo.underlying}},
  \code{\link{sfpd.FloMo}},
  \code{\link{sfpd.FloTrend.holdings}},
  \code{\link{sfpd.FloTrend.parameterize.wrapper}},
  \code{\link{sfpd.FloTrend.parameterize}},
  \code{\link{sfpd.FloTrend.underlying}},
  \code{\link{sfpd.Flow}}, \code{\link{sfpd.Holdings}},
  \code{\link{sfpd.ION}}, \code{\link{sfpd.Wt}},
  \code{\link{sfpd.filter}}
}
\keyword{sfpd.FloTrend}
