% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sfpd.FloTrend}
\alias{sfpd.FloTrend}
\title{sfpd.FloTrend}
\usage{
sfpd.FloTrend(x, y, n, w, h, u, v)
}
\arguments{
\item{x}{= flows file}

\item{y}{= holdings file}

\item{n}{= fund history file}

\item{w}{= desired flow date to be reported}

\item{h}{= one of FloTrend/FloDiff/FloDiff2}

\item{u}{= holdings file from prior month}

\item{v}{= security history file}
}
\description{
<h> factor
}
\seealso{
Other sfpd: \code{\link{sfpd.ActWtTrend}},
  \code{\link{sfpd.FloDollar}},
  \code{\link{sfpd.FloMo.underlying}},
  \code{\link{sfpd.FloMo}}, \code{\link{sfpd.Holdings}},
  \code{\link{sfpd.ION}}, \code{\link{sfpd.Wt}},
  \code{\link{sfpd.filter}}
}
\keyword{sfpd.FloTrend}
