% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sfpd.FloTrend.underlying}
\alias{sfpd.FloTrend.underlying}
\title{sfpd.FloTrend.underlying}
\usage{
sfpd.FloTrend.underlying(x, y, n)
}
\arguments{
\item{x}{= a file}

\item{y}{= a flowdate}

\item{n}{= factor name}
}
\description{
<n> factor
}
\keyword{sfpd.FloTrend.underlying}
