% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sfpd.FloTrend.underlying}
\alias{sfpd.FloTrend.underlying}
\title{sfpd.FloTrend.underlying}
\usage{
sfpd.FloTrend.underlying(x, y, n)
}
\arguments{
\item{x}{= a matrix/data frame}

\item{y}{= a flowdate}

\item{n}{= a variable}
}
\description{
<n> factor
}
\seealso{
Other sfpd: \code{\link{sfpd.ActWtTrend}},
  \code{\link{sfpd.AllocMo}}, \code{\link{sfpd.AllocSkew}},
  \code{\link{sfpd.FloDollar}},
  \code{\link{sfpd.FloMo.underlying}},
  \code{\link{sfpd.FloMo}},
  \code{\link{sfpd.FloTrend.holdings}},
  \code{\link{sfpd.FloTrend.parameterize}},
  \code{\link{sfpd.FloTrend}}, \code{\link{sfpd.Flow}},
  \code{\link{sfpd.Holdings}}, \code{\link{sfpd.ION}},
  \code{\link{sfpd.Wt}}, \code{\link{sfpd.filter}}
}
\keyword{sfpd.FloTrend.underlying}
