% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sfpd.Holdings}
\alias{sfpd.Holdings}
\title{sfpd.Holdings}
\usage{
sfpd.Holdings(x, y)
}
\arguments{
\item{x}{= a single YYYYMM}

\item{y}{= input to or output of sql.connect (e.g. "SF2022")}
}
\description{
Generates the SQL query to get weights for individual stocks
}
\seealso{
Other sfpd: \code{\link{sfpd.ActWtTrend}},
  \code{\link{sfpd.FloDollar}},
  \code{\link{sfpd.FloMo.underlying}},
  \code{\link{sfpd.FloMo}},
  \code{\link{sfpd.FloTrend.underlying}},
  \code{\link{sfpd.FloTrend}}, \code{\link{sfpd.Flow}},
  \code{\link{sfpd.ION}}, \code{\link{sfpd.Wt}},
  \code{\link{sfpd.filter}}
}
\keyword{sfpd.Holdings}
