% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sfpd.ION}
\alias{sfpd.ION}
\title{sfpd.ION}
\usage{
sfpd.ION(x, y, n, w)
}
\arguments{
\item{x}{= a matrix/data frame (flows)}

\item{y}{= a matrix/data frame (holdings)}

\item{n}{= a matrix/data frame (FundHistory)}

\item{w}{= a flowdate}
}
\description{
Inflow/Outflow factors
}
\seealso{
Other sfpd: \code{\link{sfpd.ActWtTrend}},
  \code{\link{sfpd.FloDollar}},
  \code{\link{sfpd.FloMo.underlying}},
  \code{\link{sfpd.FloMo}},
  \code{\link{sfpd.FloTrend.underlying}},
  \code{\link{sfpd.FloTrend}}, \code{\link{sfpd.Flow}},
  \code{\link{sfpd.Holdings}}, \code{\link{sfpd.Wt}},
  \code{\link{sfpd.filter}}
}
\keyword{sfpd.ION}
