% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sim.direction.buy}
\alias{sim.direction.buy}
\title{sim.direction.buy}
\usage{
sim.direction.buy(x, y)
}
\arguments{
\item{x}{= a data frame, the output of <sim.fetch>}

\item{y}{= vector of group limits (names correspond to columns in <x>)}
}
\description{
percentage buy needed to get worst group under control
}
\keyword{sim.direction.buy}
