% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sim.direction}
\alias{sim.direction}
\title{sim.direction}
\usage{
sim.direction(x, y)
}
\arguments{
\item{x}{= a data frame}

\item{y}{= an integer vector (group limits, names correspond to columns in <x>)}
}
\description{
percentage needed to get worst group under control
}
\seealso{
Other sim: \code{\link{sim.direction.buy}},
  \code{\link{sim.direction.sell}},
  \code{\link{sim.fetch}}, \code{\link{sim.limits}},
  \code{\link{sim.optimal}}, \code{\link{sim.overall}},
  \code{\link{sim.seed}}, \code{\link{sim.summ}},
  \code{\link{sim.trade.grp}}, \code{\link{sim.trade.stk}}
}
\keyword{sim.direction}
