% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sim.direction.sell}
\alias{sim.direction.sell}
\title{sim.direction.sell}
\usage{
sim.direction.sell(x, y)
}
\arguments{
\item{x}{= a data frame, the output of <sim.fetch>}

\item{y}{= vector of group limits (names correspond to columns in <x>)}
}
\description{
percentage sell needed to get worst group under control
}
\seealso{
Other sim: \code{\link{sim.direction.buy}},
  \code{\link{sim.direction}}, \code{\link{sim.fetch}},
  \code{\link{sim.limits}}, \code{\link{sim.optimal}},
  \code{\link{sim.overall}}, \code{\link{sim.seed}},
  \code{\link{sim.summ}}, \code{\link{sim.trade.grp}},
  \code{\link{sim.trade.stk}}
}
\keyword{sim.direction.sell}
