% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sim.fetch}
\alias{sim.fetch}
\title{sim.fetch}
\usage{
sim.fetch(x, y, n, w, h = NULL)
}
\arguments{
\item{x}{= YYYYMM representing period of interest}

\item{y}{= string representing variable name}

\item{n}{= string representing universe name}

\item{w}{= stock-flow environment}

\item{h}{= risk factors}
}
\description{
data needed to run simulation
}
\keyword{sim.fetch}
