% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sim.fetch}
\alias{sim.fetch}
\title{sim.fetch}
\usage{
sim.fetch(x, y, n, w)
}
\arguments{
\item{x}{= YYYYMM representing period of interest}

\item{y}{= string representing variable name}

\item{n}{= string representing universe name}

\item{w}{= stock-flow environment}
}
\description{
data needed to run simulation
}
\seealso{
Other sim: \code{\link{sim.direction}},
  \code{\link{sim.limits}}, \code{\link{sim.optimal}},
  \code{\link{sim.overall}}, \code{\link{sim.summ}},
  \code{\link{sim.trade.grp}}, \code{\link{sim.trade.stk}}
}
\keyword{sim.fetch}
