% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sim.fetch}
\alias{sim.fetch}
\title{sim.fetch}
\usage{
sim.fetch(x, y, n, w, h = NULL)
}
\arguments{
\item{x}{= a YYYYMM}

\item{y}{= a variable}

\item{n}{= a string representing universe name}

\item{w}{= a StockFlows environment}

\item{h}{= a variable vector (risk factors)}
}
\description{
data needed to run simulation
}
\seealso{
Other sim: \code{\link{sim.direction.buy}},
  \code{\link{sim.direction.sell}},
  \code{\link{sim.direction}}, \code{\link{sim.limits}},
  \code{\link{sim.optimal}}, \code{\link{sim.overall}},
  \code{\link{sim.seed}}, \code{\link{sim.summ}},
  \code{\link{sim.trade.grp}}, \code{\link{sim.trade.stk}}
}
\keyword{sim.fetch}
