% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sim.optimal}
\alias{sim.optimal}
\title{sim.optimal}
\usage{
sim.optimal(x, y, n, w, h, u)
}
\arguments{
\item{x}{= a data frame with columns Alp/Bmk/Sec/Ctry}

\item{y}{= initial portfolio weight}

\item{n}{= percentage single-stock active-weight name limit}

\item{w}{= vector of group limits (names correspond to columns in <x>)}

\item{h}{= quintile to sell (stocks in bin <h> and higher are flushed)}

\item{u}{= integer between 0 and 100}
}
\description{
returns named vector of optimal portfolio weights
}
\keyword{sim.optimal}
