% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sim.overall}
\alias{sim.overall}
\title{sim.overall}
\usage{
sim.overall(x, y, n)
}
\arguments{
\item{x}{= list object, elements are the output of <sim.optimal>}

\item{y}{= named vector of turnover}

\item{n}{= vector of group limits (names correspond to columns in elements of <x>)}
}
\description{
summarizes simulation
}
\seealso{
Other sim: \code{\link{sim.direction.buy}},
  \code{\link{sim.direction.sell}},
  \code{\link{sim.direction}}, \code{\link{sim.fetch}},
  \code{\link{sim.limits}}, \code{\link{sim.optimal}},
  \code{\link{sim.summ}}, \code{\link{sim.trade.grp}},
  \code{\link{sim.trade.stk}}
}
\keyword{sim.overall}
