% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sim.overall}
\alias{sim.overall}
\title{sim.overall}
\usage{
sim.overall(x, y, n)
}
\arguments{
\item{x}{= a list (output of <sim.optimal>)}

\item{y}{= a numeric vector (turnover)}

\item{n}{= an integer vector (names correspond to columns in elements of <x>)}
}
\description{
summarizes simulation
}
\seealso{
Other sim: \code{\link{sim.direction.buy}},
  \code{\link{sim.direction.sell}},
  \code{\link{sim.direction}}, \code{\link{sim.fetch}},
  \code{\link{sim.limits}}, \code{\link{sim.optimal}},
  \code{\link{sim.seed}}, \code{\link{sim.summ}},
  \code{\link{sim.trade.grp}}, \code{\link{sim.trade.stk}}
}
\keyword{sim.overall}
