% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sim.summ}
\alias{sim.summ}
\title{sim.summ}
\usage{
sim.summ(x, y)
}
\arguments{
\item{x}{= data frame, the output of <sim.optimal>}

\item{y}{= vector of group limits (names correspond to columns of <x>)}
}
\description{
summarizes simulation
}
\seealso{
Other sim: \code{\link{sim.direction.buy}},
  \code{\link{sim.direction.sell}},
  \code{\link{sim.direction}}, \code{\link{sim.fetch}},
  \code{\link{sim.limits}}, \code{\link{sim.optimal}},
  \code{\link{sim.overall}}, \code{\link{sim.seed}},
  \code{\link{sim.trade.grp}}, \code{\link{sim.trade.stk}}
}
\keyword{sim.summ}
