% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sim.trade.grp}
\alias{sim.trade.grp}
\title{sim.trade.grp}
\usage{
sim.trade.grp(x, y, n)
}
\arguments{
\item{x}{= a data frame}

\item{y}{= a boolean (buy/sell)}

\item{n}{= an integer vector (group limits, names correspond to columns in <x>)}
}
\description{
max you can trade without breaching group limits
}
\seealso{
Other sim: \code{\link{sim.direction.buy}},
  \code{\link{sim.direction.sell}},
  \code{\link{sim.direction}}, \code{\link{sim.fetch}},
  \code{\link{sim.limits}}, \code{\link{sim.optimal}},
  \code{\link{sim.overall}}, \code{\link{sim.seed}},
  \code{\link{sim.summ}}, \code{\link{sim.trade.stk}}
}
\keyword{sim.trade.grp}
