% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sim.trade.stk}
\alias{sim.trade.stk}
\title{sim.trade.stk}
\usage{
sim.trade.stk(x, y, n, w)
}
\arguments{
\item{x}{= a data frame, the output of <sim.fetch>}

\item{y}{= T/F depending on whether you buy/sell}

\item{n}{= percentage single-stock active-weight name limit}

\item{w}{= T/F depending on whether you're fully investing the portfolio}
}
\description{
max you can trade without breaching name limits
}
\keyword{sim.trade.stk}
