% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloMo}
\alias{sql.1dFloMo}
\title{sql.1dFloMo}
\usage{
sql.1dFloMo(x, y, n, w, h, u = "All")
}
\arguments{
\item{x}{= vector of flow dates in YYYYMMDD (known two days later)}

\item{y}{= a string vector of factors to be computed, #        :    the last elements of which are the type of fund used}

\item{n}{= any of StockFlows/China/Japan/CSI300/Energy}

\item{w}{= T/F depending on whether you are checking ftp}

\item{h}{= one or more breakdown filters (e.g. All/GeoId/DomicileId)}

\item{u}{= share-class filter (one of All/Inst/Retail)}
}
\description{
Generates the SQL query to get the data for 1dFloMo for individual stocks
}
\keyword{sql.1dFloMo}
