% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloTrend.Alloc.from}
\alias{sql.1dFloTrend.Alloc.from}
\title{sql.1dFloTrend.Alloc.from}
\usage{
sql.1dFloTrend.Alloc.from(x, y, n, w, h)
}
\arguments{
\item{x}{= flowdate/YYYYMMDD depending on whether daily/weekly}

\item{y}{= temp table name (e.g. "#CTRY")}

\item{n}{= identifier column (SectorId/CountryId)}

\item{w}{= frequency (T/F for daily/weekly or D/W/M)}

\item{h}{= vector of filters}
}
\description{
SQL query for daily/weekly FloTrend
}
\keyword{sql.1dFloTrend.Alloc.from}
