% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1dFloTrend.Ctry}
\alias{sql.1dFloTrend.Ctry}
\title{sql.1dFloTrend.Ctry}
\usage{
sql.1dFloTrend.Ctry(x, y, n)
}
\arguments{
\item{x}{= one of Ctry/FX/Sector}

\item{y}{= one of Flo/ActWt}

\item{n}{= one of Diff/Diff2/Trend/AllocMo/AllocDiff/AllocTrend}
}
\description{
For Ctry/FX generates the SQL query to get daily 1d a) FloDiff		= sql.1dFloTrend.Ctry("?", "Flo", "Diff") b) FloTrend		= sql.1dFloTrend.Ctry("?", "Flo", "Trend") c) ActWtDiff		= sql.1dFloTrend.Ctry("?", "ActWt", "Diff") d) ActWtTrend		= sql.1dFloTrend.Ctry("?", "ActWt", "Trend") e) FloDiff2		= sql.1dFloTrend.Ctry("?", "Flo", "Diff2") f) ActWtDiff2		= sql.1dFloTrend.Ctry("?", "ActWt", "Diff2") g) AllocMo		= sql.1dFloTrend.Ctry("?", "Flo", "AllocMo") h) AllocDiff		= sql.1dFloTrend.Ctry("?", "Flo", "AllocDiff") i) AllocTrend		= sql.1dFloTrend.Ctry("?", "Flo", "AllocTrend") j) AllocSkew		= sql.1dFloTrend.Ctry("?", "ActWt", "AllocSkew")
}
\seealso{
Other sql: \code{\link{sql.1dActWtTrend.Ctry.underlying}},
  \code{\link{sql.1dActWtTrend.select}},
  \code{\link{sql.1dActWtTrend.topline.from}},
  \code{\link{sql.1dActWtTrend.topline}},
  \code{\link{sql.1dActWtTrend.underlying}},
  \code{\link{sql.1dActWtTrend}},
  \code{\link{sql.1dFloMo.Ctry.Allocations.GF.avg}},
  \code{\link{sql.1dFloMo.Ctry.Allocations.term}},
  \code{\link{sql.1dFloMo.Ctry.Allocations}},
  \code{\link{sql.1dFloMo.Ctry.List}},
  \code{\link{sql.1dFloMo.CtryFlow}},
  \code{\link{sql.1dFloMo.Ctry}},
  \code{\link{sql.1dFloMo.FI.underlying}},
  \code{\link{sql.1dFloMo.FI}},
  \code{\link{sql.1dFloMo.Rgn}},
  \code{\link{sql.1dFloMo.filter}},
  \code{\link{sql.1dFloMo.grp}},
  \code{\link{sql.1dFloMo.select.wrapper}},
  \code{\link{sql.1dFloMo.select}},
  \code{\link{sql.1dFloMo.underlying}},
  \code{\link{sql.1dFloMoAggr}}, \code{\link{sql.1dFloMo}},
  \code{\link{sql.1dFloTrend.Ctry.topline}},
  \code{\link{sql.1dFloTrend.Ctry.underlying}},
  \code{\link{sql.1dFloTrend.select}},
  \code{\link{sql.1dFloTrend.underlying}},
  \code{\link{sql.1dFloTrend}}, \code{\link{sql.1dFundCt}},
  \code{\link{sql.1dFundRet}}, \code{\link{sql.1dION}},
  \code{\link{sql.1mActPas.Ctry}},
  \code{\link{sql.1mActWt.underlying}},
  \code{\link{sql.1mActWt}},
  \code{\link{sql.1mAllocD.select}},
  \code{\link{sql.1mAllocD}},
  \code{\link{sql.1mAllocMo.select}},
  \code{\link{sql.1mAllocMo.underlying.from}},
  \code{\link{sql.1mAllocMo.underlying.pre}},
  \code{\link{sql.1mAllocMo}},
  \code{\link{sql.1mAllocSkew.topline.from}},
  \code{\link{sql.1mAllocSkew}},
  \code{\link{sql.1mChActWt}}, \code{\link{sql.1mFloMo}},
  \code{\link{sql.1mFundCt}},
  \code{\link{sql.1wFlow.Corp}},
  \code{\link{sql.ActWtDiff2}},
  \code{\link{sql.AggrAllocations}},
  \code{\link{sql.AllocTbl}}, \code{\link{sql.DailyFlo}},
  \code{\link{sql.Diff}}, \code{\link{sql.Dispersion}},
  \code{\link{sql.FloMo.Funds}},
  \code{\link{sql.FundHistory.macro}},
  \code{\link{sql.FundHistory.sf}},
  \code{\link{sql.FundHistory}}, \code{\link{sql.HSIdmap}},
  \code{\link{sql.HerdingLSV}},
  \code{\link{sql.Herfindahl}},
  \code{\link{sql.Holdings.bulk}},
  \code{\link{sql.Holdings}}, \code{\link{sql.ION}},
  \code{\link{sql.MonthlyAlloc}},
  \code{\link{sql.MonthlyAssetsEnd}}, \code{\link{sql.Mo}},
  \code{\link{sql.RDSuniv}}, \code{\link{sql.SRI}},
  \code{\link{sql.TopDownAllocs.items}},
  \code{\link{sql.TopDownAllocs}}, \code{\link{sql.Trend}},
  \code{\link{sql.and}}, \code{\link{sql.arguments}},
  \code{\link{sql.bcp}}, \code{\link{sql.connect}},
  \code{\link{sql.cross.border}},
  \code{\link{sql.datediff}}, \code{\link{sql.declare}},
  \code{\link{sql.drop}}, \code{\link{sql.exists}},
  \code{\link{sql.floTbl.to.Col}}, \code{\link{sql.into}},
  \code{\link{sql.in}}, \code{\link{sql.isin.old.to.new}},
  \code{\link{sql.label}}, \code{\link{sql.map.classif}},
  \code{\link{sql.nonneg}},
  \code{\link{sql.query.underlying}},
  \code{\link{sql.query}}, \code{\link{sql.sf.wtd.avg}},
  \code{\link{sql.tbl}}, \code{\link{sql.ui}},
  \code{\link{sql.unbracket}}
}
\keyword{sql.1dFloTrend.Ctry}
