% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.1mChActWt}
\alias{sql.1mChActWt}
\title{sql.1mChActWt}
\usage{
sql.1mChActWt(x, y)
}
\arguments{
\item{x}{= the YYYYMM for which you want data (known 24 days later)}

\item{y}{= a string vector, the elements of which are: 1) FundId for the fund used as the benchmark 2) BenchIndexId of the benchmark}
}
\description{
Generates the SQL query to get the following active weights: a) EqlChAct = equal weight average change in active weight b) BegChAct = beginning-of-period-asset weighted change in active weight c) EndChAct = end-of-period-asset weighted change in active weight d) BegPosChAct = beginning-of-period-asset weighted change in active weight (positive flows only) e) EndPosChAct = end-of-period-asset weighted change in active weight (positive flows only) f) BegNegChAct = beginning-of-period-asset weighted change in active weight (negative flows only) g) EndNegChAct = end-of-period-asset weighted change in active weight (negative flows only)
}
\seealso{
Other sql: \code{\link{sql.1dActWtTrend.Ctry.underlying}},
  \code{\link{sql.1dActWtTrend.topline}},
  \code{\link{sql.1dActWtTrend.underlying}},
  \code{\link{sql.1dActWtTrend}},
  \code{\link{sql.1dFloMo.Ctry.Allocations.GF.avg}},
  \code{\link{sql.1dFloMo.Ctry.Allocations.term}},
  \code{\link{sql.1dFloMo.Ctry.Allocations}},
  \code{\link{sql.1dFloMo.Ctry.List}},
  \code{\link{sql.1dFloMo.Ctry}},
  \code{\link{sql.1dFloMo.FI.underlying}},
  \code{\link{sql.1dFloMo.FI}},
  \code{\link{sql.1dFloMo.Rgn}},
  \code{\link{sql.1dFloMoAggr}}, \code{\link{sql.1dFloMo}},
  \code{\link{sql.1dFloTrend.Ctry.topline}},
  \code{\link{sql.1dFloTrend.Ctry.underlying}},
  \code{\link{sql.1dFloTrend.Ctry}},
  \code{\link{sql.1dFloTrend.underlying}},
  \code{\link{sql.1dFloTrend}},
  \code{\link{sql.1dFundRet}}, \code{\link{sql.1dION}},
  \code{\link{sql.1mActPas.Ctry}},
  \code{\link{sql.1mActWt.underlying}},
  \code{\link{sql.1mActWt}},
  \code{\link{sql.1mAllocMo.underlying}},
  \code{\link{sql.1mAllocMo}},
  \code{\link{sql.1mAllocSkew}},
  \code{\link{sql.AggrAllocations}},
  \code{\link{sql.AllocTbl}}, \code{\link{sql.DailyFlo}},
  \code{\link{sql.Diff}}, \code{\link{sql.FloMo.Funds}},
  \code{\link{sql.FundHistory}},
  \code{\link{sql.FundWtdExcl0}},
  \code{\link{sql.Holdings.bulk}},
  \code{\link{sql.Holdings}}, \code{\link{sql.ION}},
  \code{\link{sql.MonthlyAlloc}},
  \code{\link{sql.MonthlyAssetsEnd}}, \code{\link{sql.Mo}},
  \code{\link{sql.RDSuniv}},
  \code{\link{sql.TopDownAllocs.items}},
  \code{\link{sql.TopDownAllocs}}, \code{\link{sql.Trend}},
  \code{\link{sql.and}}, \code{\link{sql.bcp}},
  \code{\link{sql.connect}},
  \code{\link{sql.cross.border}},
  \code{\link{sql.datediff}}, \code{\link{sql.declare}},
  \code{\link{sql.drop}}, \code{\link{sql.exists}},
  \code{\link{sql.floTbl.to.Col}}, \code{\link{sql.into}},
  \code{\link{sql.in}}, \code{\link{sql.isin.old.to.new}},
  \code{\link{sql.label}}, \code{\link{sql.map.classif}},
  \code{\link{sql.nonneg}}, \code{\link{sql.query}},
  \code{\link{sql.stock.flows.wtd.avg}},
  \code{\link{sql.tbl}}, \code{\link{sql.ui}},
  \code{\link{sql.unbracket}}
}
\keyword{sql.1mChActWt}
