% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.DailyFlo}
\alias{sql.DailyFlo}
\title{sql.DailyFlo}
\usage{
sql.DailyFlo(x, y = T, n = T, w = "All", h = F)
}
\arguments{
\item{x}{= a vector of dates for which you want flows (known one day later)}

\item{y}{= T/F depending on whether to group by HFundId}

\item{n}{= T/F depending on whether StockFlows data are being used}

\item{w}{= share-class filter (one of All/Inst/Retail)}

\item{h}{= T/F depending on whether AssetsEnd is wanted}
}
\description{
Generates the SQL query to get the data for daily Flow
}
\keyword{sql.DailyFlo}
