% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.FundHistory}
\alias{sql.FundHistory}
\title{sql.FundHistory}
\usage{
sql.FundHistory(x, y, n, w)
}
\arguments{
\item{x}{= characters to place before each line of the SQL query part}

\item{y}{= either "All" or "Act" or "CBE"}

\item{n}{= T/F depending on whether StockFlows data are being used}

\item{w}{= columns needed in addition to HFundId}
}
\description{
SQL query to restrict to Global and Regional equity funds
}
\keyword{sql.FundHistory}
