% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.map.classif}
\alias{sql.map.classif}
\title{sql.map.classif}
\usage{
sql.map.classif(x, y, n)
}
\arguments{
\item{x}{= SQL queries to be submitted}

\item{y}{= a connection, the output of odbcDriverConnect}

\item{n}{= classif file}
}
\description{
Returns flow variables with the same row space as <w>
}
\keyword{sql.map.classif}
