% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.MonthlyAlloc}
\alias{sql.MonthlyAlloc}
\title{sql.MonthlyAlloc}
\usage{
sql.MonthlyAlloc(x, y = "")
}
\arguments{
\item{x}{= YYYYMM for which you want allocations (known 24 days after month end)}

\item{y}{= characters that get pasted in front of every line (usually tabs for indentation)}
}
\description{
Generates the SQL query to get the data for monthly allocations for StockFlows
}
\keyword{sql.MonthlyAlloc}
