% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.MonthlyAssetsEnd}
\alias{sql.MonthlyAssetsEnd}
\title{sql.MonthlyAssetsEnd}
\usage{
sql.MonthlyAssetsEnd(x, y = NULL, n = F, w = "All", h = "AssetsEnd")
}
\arguments{
\item{x}{= a single YYYYMMDD or list (where clause)}

\item{y}{= columns in addition to AssetsEnd}

\item{n}{= T/F depending on whether data are indexed by FundId}

\item{w}{= share-class filter (one of All/Inst/Retail)}

\item{h}{= name of AssetEnd column (e.g. "PortVal")}
}
\description{
Generates the SQL query to get the data for monthly Assets End
}
\keyword{sql.MonthlyAssetsEnd}
