% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.query}
\alias{sql.query}
\title{sql.query}
\usage{
sql.query(x, y, n = T)
}
\arguments{
\item{x}{= query needed for the update}

\item{y}{= input to or output of sql.connect}

\item{n}{= T/F depending on whether you wish to output number of rows of data got}
}
\description{
opens a connection, executes sql query, then closes the connection
}
\keyword{sql.query}
