% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.tbl}
\alias{sql.tbl}
\title{sql.tbl}
\usage{
sql.tbl(x, y, n, w, h)
}
\arguments{
\item{x}{= needed columns}

\item{y}{= table}

\item{n}{= where segment}

\item{w}{= group by segment}

\item{h}{= having}
}
\description{
Full SQL statement
}
\keyword{sql.tbl}
