% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{sql.TopDownAllocs}
\alias{sql.TopDownAllocs}
\title{sql.TopDownAllocs}
\usage{
sql.TopDownAllocs(x, y, n, w, h)
}
\arguments{
\item{x}{= the YYYYMM for which you want data (known 26 days later)}

\item{y}{= a string vector of top-down allocations wanted, #        :    the last element of which is the type of fund to be used.}

\item{n}{= any of StockFlows/Japan/CSI300/Energy}

\item{w}{= T/F depending on whether you are checking ftp}

\item{h}{= breakdown filter (e.g. All/GeoId/DomicileId)}
}
\description{
Generates the SQL query to get Active/Passive Top-Down Allocations
}
\keyword{sql.TopDownAllocs}
