% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{strat.dir}
\alias{strat.dir}
\title{strat.dir}
\usage{
strat.dir(x)
}
\arguments{
\item{x}{= frequency (e.g. "daily", "weekly" or "monthly")}
}
\description{
the folder where <x> factors live
}
\keyword{strat.dir}
