% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{strat.path}
\alias{strat.path}
\title{strat.path}
\usage{
strat.path(x, y)
}
\arguments{
\item{x}{= name of the strategy (e.g. "FX" or "FloPctCtry")}

\item{y}{= frequency (e.g. "daily", "weekly" or "monthly")}
}
\description{
Returns the full path to the factor file
}
\keyword{strat.path}
