% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{strat.path}
\alias{strat.path}
\title{strat.path}
\usage{
strat.path(x, y)
}
\arguments{
\item{x}{= name of the strategy (e.g. "FX" or "FloPctCtry")}

\item{y}{= frequency (e.g. "daily", "weekly" or "monthly")}
}
\description{
Returns the full path to the factor file
}
\seealso{
Other strat: \code{\link{strat.dir}},
  \code{\link{strat.email}}, \code{\link{strat.file}},
  \code{\link{strat.list}}
}
\keyword{strat.path}
