% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{strategy.file}
\alias{strategy.file}
\title{strategy.file}
\usage{
strategy.file(x, y)
}
\arguments{
\item{x}{= name of the strategy (e.g. "FX" or "PremSec-JP")}

\item{y}{= "daily" or "weekly"}
}
\description{
Returns the file in which the factor lives
}
\keyword{strategy.file}
