% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{stratrets.bbk}
\alias{stratrets.bbk}
\title{stratrets.bbk}
\usage{
stratrets.bbk(x, y)
}
\arguments{
\item{x}{= strategy}

\item{y}{= variable being back-tested (e.g. FloPct/AIS)}
}
\description{
named vector of TxB return spreads indexed by BoP
}
\keyword{stratrets.bbk}
