% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{stratrets.beta}
\alias{stratrets.beta}
\title{stratrets.beta}
\usage{
stratrets.beta(x, y, n, w)
}
\arguments{
\item{x}{= indicators}

\item{y}{= total return indices}

\item{n}{= benchmark (e.g. "ACWorld")}

\item{w}{= lookback over which beta is computed}
}
\description{
beta-adjusted indicator
}
\keyword{stratrets.beta}
