% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{stratrets.beta}
\alias{stratrets.beta}
\title{stratrets.beta}
\usage{
stratrets.beta(x, y, n, w)
}
\arguments{
\item{x}{= a matrix/data frame (predictors)}

\item{y}{= a matrix/data frame (total return indices)}

\item{n}{= a string (benchmark like "ACWorld")}

\item{w}{= an integer (lookback in days)}
}
\description{
beta-adjusted indicator
}
\seealso{
Other stratrets: \code{\link{stratrets.bbk}},
  \code{\link{stratrets.data}},
  \code{\link{stratrets.indicator}},
  \code{\link{stratrets.path}},
  \code{\link{stratrets.returns}},
  \code{\link{stratrets.subset.Ctry}},
  \code{\link{stratrets.subset}}, \code{\link{stratrets}}
}
\keyword{stratrets.beta}
