% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{stratrets.data}
\alias{stratrets.data}
\title{stratrets.data}
\usage{
stratrets.data(x, y)
}
\arguments{
\item{x}{= strategy}

\item{y}{= a variable}
}
\description{
list object containing arguments needed for function <bbk>
}
\keyword{stratrets.data}
