% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{stratrets.indicator}
\alias{stratrets.indicator}
\title{stratrets.indicator}
\usage{
stratrets.indicator(x, y, n, w, h)
}
\arguments{
\item{x}{= a file vector (strategies)}

\item{y}{= an integer (lookback)}

\item{n}{= a boolean (sum/compound)}

\item{w}{= if T, sector-adjustment is performed}

\item{h}{= an integer (delay, only used when <w>)}
}
\description{
data frame compounded across <y>
}
\seealso{
Other stratrets: \code{\link{stratrets.bbk}},
  \code{\link{stratrets.beta}},
  \code{\link{stratrets.data}},
  \code{\link{stratrets.path}},
  \code{\link{stratrets.returns}},
  \code{\link{stratrets.subset.Ctry}},
  \code{\link{stratrets.subset}}, \code{\link{stratrets}}
}
\keyword{stratrets.indicator}
