% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{stratrets.path}
\alias{stratrets.path}
\title{stratrets.path}
\usage{
stratrets.path(x, y, n, w, h)
}
\arguments{
\item{x}{= indicator type (e.g. Ctry/FX/SectorUK/Rgn/FI)}

\item{y}{= FundType (e.g. E/B)}

\item{n}{= filter (e.g. Aggregate, Act, SRI. etc.)}

\item{w}{= item (e.g. Flow/AssetsStart/Result)}

\item{h}{= variant (e.g. CB/SG/CBSG)}
}
\description{
path to strategy indicators
}
\seealso{
Other stratrets: \code{\link{stratrets.bbk}},
  \code{\link{stratrets.beta}},
  \code{\link{stratrets.data}},
  \code{\link{stratrets.indicator}},
  \code{\link{stratrets.returns}},
  \code{\link{stratrets.subset.Ctry}},
  \code{\link{stratrets.subset}}, \code{\link{stratrets}}
}
\keyword{stratrets.path}
