% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{stratrets}
\alias{stratrets}
\title{stratrets}
\usage{
stratrets(x)
}
\arguments{
\item{x}{= variable being back-tested (e.g. FloPct/AIS)}
}
\description{
data frame of TxB return spreads
}
\seealso{
Other stratrets: \code{\link{stratrets.bbk}},
  \code{\link{stratrets.beta}},
  \code{\link{stratrets.data}},
  \code{\link{stratrets.indicator}},
  \code{\link{stratrets.path}},
  \code{\link{stratrets.returns}},
  \code{\link{stratrets.subset.Ctry}},
  \code{\link{stratrets.subset}}
}
\keyword{stratrets}
