% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{stratrets.returns}
\alias{stratrets.returns}
\title{stratrets.returns}
\usage{
stratrets.returns(x)
}
\arguments{
\item{x}{= return type (e.g. Ctry/FX/SectorUK/Multi)}
}
\description{
data frame of daily returns
}
\seealso{
Other stratrets: \code{\link{stratrets.bbk}},
  \code{\link{stratrets.beta}},
  \code{\link{stratrets.data}},
  \code{\link{stratrets.indicator}},
  \code{\link{stratrets.path}},
  \code{\link{stratrets.subset.Ctry}},
  \code{\link{stratrets.subset}}, \code{\link{stratrets}}
}
\keyword{stratrets.returns}
