% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{stratrets.subset}
\alias{stratrets.subset}
\title{stratrets.subset}
\usage{
stratrets.subset(x, y)
}
\arguments{
\item{x}{= indicators}

\item{y}{= index to subset to}
}
\description{
subsets to columns used in the back-test
}
\seealso{
Other stratrets: \code{\link{stratrets.bbk}},
  \code{\link{stratrets.beta}},
  \code{\link{stratrets.data}},
  \code{\link{stratrets.indicator}},
  \code{\link{stratrets.returns}},
  \code{\link{stratrets.subset.Ctry}},
  \code{\link{stratrets}}
}
\keyword{stratrets.subset}
