% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{summ.multi}
\alias{summ.multi}
\title{summ.multi}
\usage{
summ.multi(fcn, x, y)
}
\arguments{
\item{fcn}{= a function (summary)}

\item{x}{= a data frame (bin returns)}

\item{y}{= an integer (return window in days/months)}
}
\description{
summarizes the multi-period back test
}
\keyword{summ.multi}
