% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.excise}
\alias{txt.excise}
\title{txt.excise}
\usage{
txt.excise(x, y)
}
\arguments{
\item{x}{= a vector of string}

\item{y}{= a vector of string}
}
\description{
cuts first instance of each element of <y> from <x>
}
\keyword{txt.excise}
