% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.expand}
\alias{txt.expand}
\title{txt.expand}
\usage{
txt.expand(x, y, n = "-", w = F)
}
\arguments{
\item{x}{= a vector of strings}

\item{y}{= a vector of strings}

\item{n}{= paste separator}

\item{w}{= T/F variable controlling paste order}
}
\description{
Returns all combinations OF <x> and <y> pasted together
}
\seealso{
Other txt: \code{\link{txt.anagram.underlying}},
  \code{\link{txt.anagram}}, \code{\link{txt.count}},
  \code{\link{txt.ex.file}}, \code{\link{txt.excise}},
  \code{\link{txt.has}}, \code{\link{txt.hdr}},
  \code{\link{txt.itrim}}, \code{\link{txt.left}},
  \code{\link{txt.name.format}}, \code{\link{txt.na}},
  \code{\link{txt.palindrome.entire}},
  \code{\link{txt.palindrome.partial}},
  \code{\link{txt.palindrome.tail}},
  \code{\link{txt.palindrome.underlying}},
  \code{\link{txt.palindrome}}, \code{\link{txt.parse}},
  \code{\link{txt.prepend}}, \code{\link{txt.regr}},
  \code{\link{txt.replace}}, \code{\link{txt.reverse}},
  \code{\link{txt.right}}, \code{\link{txt.space}},
  \code{\link{txt.to.char}}, \code{\link{txt.trim.end}},
  \code{\link{txt.trim.left}},
  \code{\link{txt.trim.right}}, \code{\link{txt.trim}}
}
\keyword{txt.expand}
