% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.gunning}
\alias{txt.gunning}
\title{txt.gunning}
\usage{
txt.gunning(x, y, n)
}
\arguments{
\item{x}{= a string representing a text passage}

\item{y}{= a file of potentially-usable capitalized words}

\item{n}{= a file of potentially-usable capitalized words considered "simple"}
}
\description{
the Gunning fog index measuring the number of years of  schooling beyond kindergarten needed to comprehend <x>
}
\seealso{
Other txt: \code{\link{txt.anagram.underlying}},
  \code{\link{txt.anagram}}, \code{\link{txt.core}},
  \code{\link{txt.count}}, \code{\link{txt.ex.file}},
  \code{\link{txt.ex.int}}, \code{\link{txt.excise}},
  \code{\link{txt.expand}}, \code{\link{txt.has}},
  \code{\link{txt.hdr}}, \code{\link{txt.itrim}},
  \code{\link{txt.left}}, \code{\link{txt.levenshtein}},
  \code{\link{txt.name.format}}, \code{\link{txt.na}},
  \code{\link{txt.parse}}, \code{\link{txt.prepend}},
  \code{\link{txt.regr}}, \code{\link{txt.replace}},
  \code{\link{txt.reverse}}, \code{\link{txt.right}},
  \code{\link{txt.space}}, \code{\link{txt.to.char}},
  \code{\link{txt.trim.end}}, \code{\link{txt.trim.left}},
  \code{\link{txt.trim.right}}, \code{\link{txt.trim}},
  \code{\link{txt.words}}
}
\keyword{txt.gunning}
