% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.has}
\alias{txt.has}
\title{txt.has}
\usage{
txt.has(x, y, n = F)
}
\arguments{
\item{x}{= a vector of strings}

\item{y}{= a single string}

\item{n}{= T/F depending on whether a logical vector is desired}
}
\description{
the elements of <x> that contain <y> if <n> is F or a logical vector otherwise
}
\seealso{
Other txt: \code{\link{txt.anagram.underlying}},
  \code{\link{txt.anagram}}, \code{\link{txt.core}},
  \code{\link{txt.count}}, \code{\link{txt.ex.file}},
  \code{\link{txt.ex.int}}, \code{\link{txt.excise}},
  \code{\link{txt.expand}}, \code{\link{txt.gunning}},
  \code{\link{txt.hdr}}, \code{\link{txt.itrim}},
  \code{\link{txt.left}}, \code{\link{txt.levenshtein}},
  \code{\link{txt.name.format}}, \code{\link{txt.na}},
  \code{\link{txt.parse}}, \code{\link{txt.prepend}},
  \code{\link{txt.regr}}, \code{\link{txt.replace}},
  \code{\link{txt.reverse}}, \code{\link{txt.right}},
  \code{\link{txt.space}}, \code{\link{txt.to.char}},
  \code{\link{txt.trim.end}}, \code{\link{txt.trim.left}},
  \code{\link{txt.trim.right}}, \code{\link{txt.trim}},
  \code{\link{txt.words}}
}
\keyword{txt.has}
