% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.name.format}
\alias{txt.name.format}
\title{txt.name.format}
\usage{
txt.name.format(x)
}
\arguments{
\item{x}{= a string vector}
}
\description{
capitalizes first letter of each word, rendering remaining letters in lower case
}
\seealso{
Other txt: \code{\link{txt.count}},
  \code{\link{txt.ex.file}}, \code{\link{txt.expand}},
  \code{\link{txt.has}}, \code{\link{txt.hdr}},
  \code{\link{txt.itrim}}, \code{\link{txt.left}},
  \code{\link{txt.na}}, \code{\link{txt.parse}},
  \code{\link{txt.prepend}}, \code{\link{txt.regr}},
  \code{\link{txt.replace}}, \code{\link{txt.reverse}},
  \code{\link{txt.right}}, \code{\link{txt.space}},
  \code{\link{txt.to.char}}, \code{\link{txt.trim.end}},
  \code{\link{txt.trim.left}},
  \code{\link{txt.trim.right}}, \code{\link{txt.trim}}
}
\keyword{txt.name.format}
