% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.prepend}
\alias{txt.prepend}
\title{txt.prepend}
\usage{
txt.prepend(x, y, n)
}
\arguments{
\item{x}{= a vector of strings}

\item{y}{= number of times to add <n>}

\item{n}{= string to add at the beginning}
}
\description{
adds <n> <y> times to the beginning of <x>
}
\keyword{txt.prepend}
