% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.prepend}
\alias{txt.prepend}
\title{txt.prepend}
\usage{
txt.prepend(x, y, n)
}
\arguments{
\item{x}{= a vector of strings}

\item{y}{= number of characters to add}

\item{n}{= the characters to add at the beginning}
}
\description{
bulks up each string to have at least <y> characters by adding <n> to the beginning of each string
}
\keyword{txt.prepend}
