% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.regr}
\alias{txt.regr}
\title{txt.regr}
\usage{
txt.regr(x, y = T)
}
\arguments{
\item{x}{= a vector of column names}

\item{y}{= T/F depending on whether regression has an intercept}
}
\description{
returns the string you need to regress the first column on the others
}
\seealso{
Other txt: \code{\link{txt.count}},
  \code{\link{txt.ex.file}}, \code{\link{txt.expand}},
  \code{\link{txt.has}}, \code{\link{txt.hdr}},
  \code{\link{txt.itrim}}, \code{\link{txt.left}},
  \code{\link{txt.name.format}}, \code{\link{txt.na}},
  \code{\link{txt.parse}}, \code{\link{txt.prepend}},
  \code{\link{txt.replace}}, \code{\link{txt.reverse}},
  \code{\link{txt.right}}, \code{\link{txt.space}},
  \code{\link{txt.to.char}}, \code{\link{txt.trim.end}},
  \code{\link{txt.trim.left}},
  \code{\link{txt.trim.right}}, \code{\link{txt.trim}}
}
\keyword{txt.regr}
