% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.replace}
\alias{txt.replace}
\title{txt.replace}
\usage{
txt.replace(x, y, n)
}
\arguments{
\item{x}{= a vector of strings}

\item{y}{= a string to be swapped out}

\item{n}{= a string to replace <y> with}
}
\description{
replaces all instances of <y> with <n>
}
\keyword{txt.replace}
