% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.replace}
\alias{txt.replace}
\title{txt.replace}
\usage{
txt.replace(x, y, n)
}
\arguments{
\item{x}{= a vector of strings}

\item{y}{= a string to be swapped out}

\item{n}{= a string to replace <txt.out> with}
}
\description{
replaces all instances of <txt.out> by <txt.by>
}
\seealso{
Other txt: \code{\link{txt.count}},
  \code{\link{txt.ex.file}}, \code{\link{txt.expand}},
  \code{\link{txt.has}}, \code{\link{txt.hdr}},
  \code{\link{txt.itrim}}, \code{\link{txt.left}},
  \code{\link{txt.name.format}}, \code{\link{txt.na}},
  \code{\link{txt.parse}}, \code{\link{txt.prepend}},
  \code{\link{txt.regr}}, \code{\link{txt.reverse}},
  \code{\link{txt.right}}, \code{\link{txt.space}},
  \code{\link{txt.to.char}}, \code{\link{txt.trim.end}},
  \code{\link{txt.trim.left}},
  \code{\link{txt.trim.right}}, \code{\link{txt.trim}}
}
\keyword{txt.replace}
