% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.space}
\alias{txt.space}
\title{txt.space}
\usage{
txt.space(x, y = " ")
}
\arguments{
\item{x}{= any integer}

\item{y}{= a single character}
}
\description{
returns <x> iterations of <y> pasted together
}
\seealso{
Other txt: \code{\link{txt.count}},
  \code{\link{txt.ex.file}}, \code{\link{txt.expand}},
  \code{\link{txt.hdr}}, \code{\link{txt.itrim}},
  \code{\link{txt.left}}, \code{\link{txt.name.format}},
  \code{\link{txt.na}}, \code{\link{txt.parse}},
  \code{\link{txt.prepend}}, \code{\link{txt.regr}},
  \code{\link{txt.replace}}, \code{\link{txt.reverse}},
  \code{\link{txt.right}}, \code{\link{txt.to.char}},
  \code{\link{txt.trim.end}}, \code{\link{txt.trim.left}},
  \code{\link{txt.trim.right}}, \code{\link{txt.trim}}
}
\keyword{txt.space}
