% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.trim.end}
\alias{txt.trim.end}
\title{txt.trim.end}
\usage{
txt.trim.end(fcn, x, y, n)
}
\arguments{
\item{fcn}{= a function that returns characters from the bad end}

\item{x}{= a vector of string}

\item{y}{= a vector of verboten strings, each of the same length}

\item{n}{= a functon that returns characters from the opposite end}
}
\description{
trims off leading or trailing elements of <y>
}
\keyword{txt.trim.end}
