% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.trim.left}
\alias{txt.trim.left}
\title{txt.trim.left}
\usage{
txt.trim.left(x, y)
}
\arguments{
\item{x}{= a vector of string}

\item{y}{= a vector of verboten strings, each of the same length}
}
\description{
trims off leading elements of <y>
}
\keyword{txt.trim.left}
