% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.trim.left}
\alias{txt.trim.left}
\title{txt.trim.left}
\usage{
txt.trim.left(x, y)
}
\arguments{
\item{x}{= a vector of string}

\item{y}{= a vector of verboten strings, each of the same length}
}
\description{
trims off leading elements of <y>
}
\seealso{
Other txt: \code{\link{txt.count}},
  \code{\link{txt.ex.file}}, \code{\link{txt.expand}},
  \code{\link{txt.hdr}}, \code{\link{txt.itrim}},
  \code{\link{txt.left}}, \code{\link{txt.name.format}},
  \code{\link{txt.na}}, \code{\link{txt.parse}},
  \code{\link{txt.prepend}}, \code{\link{txt.regr}},
  \code{\link{txt.replace}}, \code{\link{txt.reverse}},
  \code{\link{txt.right}}, \code{\link{txt.space}},
  \code{\link{txt.to.char}}, \code{\link{txt.trim.end}},
  \code{\link{txt.trim.right}}, \code{\link{txt.trim}}
}
\keyword{txt.trim.left}
