% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.trim}
\alias{txt.trim}
\title{txt.trim}
\usage{
txt.trim(x, y = " ")
}
\arguments{
\item{x}{= a vector of string}

\item{y}{= a vector of verboten strings, each of the same length}
}
\description{
trims leading/trailing spaces
}
\keyword{txt.trim}
