% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{txt.trim.right}
\alias{txt.trim.right}
\title{txt.trim.right}
\usage{
txt.trim.right(x, y)
}
\arguments{
\item{x}{= a vector of string}

\item{y}{= a single string or expression like "[\\t ]"}
}
\description{
trims off trailing occurences of <y>
}
\seealso{
Other txt: \code{\link{txt.anagram.underlying}},
  \code{\link{txt.anagram}}, \code{\link{txt.core}},
  \code{\link{txt.count}}, \code{\link{txt.ex.file}},
  \code{\link{txt.ex.int.cardinal.wrapper}},
  \code{\link{txt.ex.int.cardinal}},
  \code{\link{txt.ex.int.ordinal.wrapper}},
  \code{\link{txt.ex.int.ordinal}},
  \code{\link{txt.ex.int.underlying}},
  \code{\link{txt.ex.int}}, \code{\link{txt.excise}},
  \code{\link{txt.expand}}, \code{\link{txt.gunning}},
  \code{\link{txt.has}}, \code{\link{txt.hdr}},
  \code{\link{txt.itrim}}, \code{\link{txt.left}},
  \code{\link{txt.levenshtein}},
  \code{\link{txt.name.format}}, \code{\link{txt.na}},
  \code{\link{txt.parse}}, \code{\link{txt.prepend}},
  \code{\link{txt.regex}}, \code{\link{txt.regr}},
  \code{\link{txt.replace}}, \code{\link{txt.reverse}},
  \code{\link{txt.right}}, \code{\link{txt.space}},
  \code{\link{txt.to.char}}, \code{\link{txt.trim.left}},
  \code{\link{txt.trim}}, \code{\link{txt.words}}
}
\keyword{txt.trim.right}
