% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{urn.exact}
\alias{urn.exact}
\title{urn.exact}
\usage{
urn.exact(x, y)
}
\arguments{
\item{x}{= an integer vector}

\item{y}{= an integer vector (pointwise >= <x>)}
}
\description{
probability of drawing precisely <x> balls from an urn containing <y> balls
}
\keyword{urn.exact}
