% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.cat}
\alias{vec.cat}
\title{vec.cat}
\usage{
vec.cat(x)
}
\arguments{
\item{x}{= vector}
}
\description{
displays on screen
}
\seealso{
Other vec: \code{\link{vec.count}}, \code{\link{vec.cum}},
  \code{\link{vec.diff}}, \code{\link{vec.ex.filters}},
  \code{\link{vec.lag}}, \code{\link{vec.max}},
  \code{\link{vec.min}}, \code{\link{vec.named}},
  \code{\link{vec.read}}, \code{\link{vec.same}},
  \code{\link{vec.swap}}, \code{\link{vec.to.list}},
  \code{\link{vec.unique}}, \code{\link{vec.write}}
}
\keyword{vec.cat}
