% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.diff}
\alias{vec.diff}
\title{vec.diff}
\usage{
vec.diff(x, y)
}
\arguments{
\item{x}{= a numeric vector}

\item{y}{= an integer}
}
\description{
difference between <x> and itself lagged <y>
}
\seealso{
Other vec: \code{\link{vec.cat}}, \code{\link{vec.count}},
  \code{\link{vec.cum}}, \code{\link{vec.ex.filters}},
  \code{\link{vec.lag}}, \code{\link{vec.max}},
  \code{\link{vec.min}}, \code{\link{vec.named}},
  \code{\link{vec.read}}, \code{\link{vec.same}},
  \code{\link{vec.swap}}, \code{\link{vec.to.list}},
  \code{\link{vec.unique}}, \code{\link{vec.write}}
}
\keyword{vec.diff}
