% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.ex.filters}
\alias{vec.ex.filters}
\title{vec.ex.filters}
\usage{
vec.ex.filters(x)
}
\arguments{
\item{x}{= either "sf" or "macro"}
}
\description{
SQL query where clauses associated with filters
}
\seealso{
Other vec: \code{\link{vec.cat}}, \code{\link{vec.count}},
  \code{\link{vec.cum}}, \code{\link{vec.diff}},
  \code{\link{vec.first}}, \code{\link{vec.lag}},
  \code{\link{vec.last}}, \code{\link{vec.max}},
  \code{\link{vec.min}}, \code{\link{vec.named}},
  \code{\link{vec.read}}, \code{\link{vec.swap}},
  \code{\link{vec.to.list}}, \code{\link{vec.unique}}
}
\keyword{vec.ex.filters}
