% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.lag}
\alias{vec.lag}
\title{vec.lag}
\usage{
vec.lag(x, y)
}
\arguments{
\item{x}{= a vector indexed by time running FORWARDS}

\item{y}{= number of periods over which to lag}
}
\description{
<x> lagged <y> periods (simple positional lag)
}
\seealso{
Other vec: \code{\link{vec.cat}}, \code{\link{vec.count}},
  \code{\link{vec.cum}}, \code{\link{vec.diff}},
  \code{\link{vec.ex.filters}}, \code{\link{vec.max}},
  \code{\link{vec.min}}, \code{\link{vec.named}},
  \code{\link{vec.read}}, \code{\link{vec.same}},
  \code{\link{vec.swap}}, \code{\link{vec.to.list}},
  \code{\link{vec.unique}}
}
\keyword{vec.lag}
