% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.max}
\alias{vec.max}
\title{vec.max}
\usage{
vec.max(x, y)
}
\arguments{
\item{x}{= a numeric vector/matrix/data frame}

\item{y}{= a numeric vector/matrix/data frame}
}
\description{
Returns the piecewise maximum of <x> and <y>
}
\seealso{
Other vec: \code{\link{vec.cat}}, \code{\link{vec.count}},
  \code{\link{vec.cum}}, \code{\link{vec.diff}},
  \code{\link{vec.ex.filters}}, \code{\link{vec.first}},
  \code{\link{vec.lag}}, \code{\link{vec.last}},
  \code{\link{vec.min}}, \code{\link{vec.named}},
  \code{\link{vec.read}}, \code{\link{vec.same}},
  \code{\link{vec.swap}}, \code{\link{vec.to.list}},
  \code{\link{vec.unique}}
}
\keyword{vec.max}
