% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.min}
\alias{vec.min}
\title{vec.min}
\usage{
vec.min(x, y)
}
\arguments{
\item{x}{= a vector/matrix/dataframe}

\item{y}{= a number/vector or matrix/dataframe with the same dimensions as <x>}
}
\description{
Returns the piecewise minimum of <x> and <y>
}
\seealso{
Other vec: \code{\link{vec.cat}}, \code{\link{vec.count}},
  \code{\link{vec.max}}, \code{\link{vec.named}},
  \code{\link{vec.read}}, \code{\link{vec.same}},
  \code{\link{vec.swap}}, \code{\link{vec.to.lags}},
  \code{\link{vec.to.list}}, \code{\link{vec.unique}},
  \code{\link{vec.zScore}}
}
\keyword{vec.min}
