% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.named}
\alias{vec.named}
\title{vec.named}
\usage{
vec.named(x, y)
}
\arguments{
\item{x}{= a string vector (can be missing)}

\item{y}{= a string vector}
}
\description{
Returns a numeric vector with values <x> and names <y>
}
\seealso{
Other vec: \code{\link{vec.cat}}, \code{\link{vec.count}},
  \code{\link{vec.cum}}, \code{\link{vec.diff}},
  \code{\link{vec.ex.filters}}, \code{\link{vec.first}},
  \code{\link{vec.lag}}, \code{\link{vec.last}},
  \code{\link{vec.max}}, \code{\link{vec.min}},
  \code{\link{vec.read}}, \code{\link{vec.same}},
  \code{\link{vec.swap}}, \code{\link{vec.to.list}},
  \code{\link{vec.unique}}
}
\keyword{vec.named}
