% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.read}
\alias{vec.read}
\title{vec.read}
\usage{
vec.read(x, y = ",")
}
\arguments{
\item{x}{= path to a vector}

\item{y}{= the separator}
}
\description{
reads into <x> a named vector
}
\keyword{vec.read}
