% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.read}
\alias{vec.read}
\title{vec.read}
\usage{
vec.read(x, y, n = ",")
}
\arguments{
\item{x}{= path to a vector}

\item{y}{= T/F depending on whether the elements are named}

\item{n}{= separator}
}
\description{
reads into a vector
}
\seealso{
Other vec: \code{\link{vec.cat}}, \code{\link{vec.count}},
  \code{\link{vec.ex.filters}}, \code{\link{vec.max}},
  \code{\link{vec.min}}, \code{\link{vec.named}},
  \code{\link{vec.same}}, \code{\link{vec.swap}},
  \code{\link{vec.to.lags}}, \code{\link{vec.to.list}},
  \code{\link{vec.unique}}, \code{\link{vec.write}}
}
\keyword{vec.read}
