% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.read}
\alias{vec.read}
\title{vec.read}
\usage{
vec.read(x, y = ",")
}
\arguments{
\item{x}{= a file (vector)}

\item{y}{= the separator}
}
\description{
reads into <x> a named vector
}
\seealso{
Other vec: \code{\link{vec.cat}}, \code{\link{vec.count}},
  \code{\link{vec.cum}}, \code{\link{vec.diff}},
  \code{\link{vec.ex.filters}}, \code{\link{vec.first}},
  \code{\link{vec.lag}}, \code{\link{vec.last}},
  \code{\link{vec.max}}, \code{\link{vec.min}},
  \code{\link{vec.named}}, \code{\link{vec.swap}},
  \code{\link{vec.to.list}}, \code{\link{vec.unique}}
}
\keyword{vec.read}
