% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.same}
\alias{vec.same}
\title{vec.same}
\usage{
vec.same(x, y)
}
\arguments{
\item{x}{= a vector}

\item{y}{= an isomekic vector}
}
\description{
T/F depending on whether <x> and <y> are identical
}
\seealso{
Other vec: \code{\link{vec.cat}}, \code{\link{vec.count}},
  \code{\link{vec.max}}, \code{\link{vec.min}},
  \code{\link{vec.named}}, \code{\link{vec.read}},
  \code{\link{vec.swap}}, \code{\link{vec.to.lags}},
  \code{\link{vec.to.list}}, \code{\link{vec.unique}},
  \code{\link{vec.zScore.underlying}},
  \code{\link{vec.zScore}}
}
\keyword{vec.same}
