% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.to.lags}
\alias{vec.to.lags}
\title{vec.to.lags}
\usage{
vec.to.lags(x, y, n = T)
}
\arguments{
\item{x}{= a numeric vector (time flows forward)}

\item{y}{= number of lagged values desired plus one}

\item{n}{= T/F depending on whether time flows forwards}
}
\description{
a data frame of <x> together with itself lagged 1, ..., <y> - 1 times
}
\seealso{
Other vec: \code{\link{vec.cat}}, \code{\link{vec.count}},
  \code{\link{vec.ex.filters}}, \code{\link{vec.lag}},
  \code{\link{vec.max}}, \code{\link{vec.min}},
  \code{\link{vec.named}}, \code{\link{vec.read}},
  \code{\link{vec.same}}, \code{\link{vec.swap}},
  \code{\link{vec.to.list}}, \code{\link{vec.unique}},
  \code{\link{vec.write}}
}
\keyword{vec.to.lags}
