% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.to.list}
\alias{vec.to.list}
\title{vec.to.list}
\usage{
vec.to.list(x)
}
\arguments{
\item{x}{= string vector}
}
\description{
list object
}
\seealso{
Other vec: \code{\link{vec.cat}}, \code{\link{vec.count}},
  \code{\link{vec.max}}, \code{\link{vec.min}},
  \code{\link{vec.named}}, \code{\link{vec.read}},
  \code{\link{vec.same}}, \code{\link{vec.swap}},
  \code{\link{vec.to.lags}}, \code{\link{vec.unique}},
  \code{\link{vec.write}}
}
\keyword{vec.to.list}
