% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{vec.zScore}
\alias{vec.zScore}
\title{vec.zScore}
\usage{
vec.zScore(x, y, n)
}
\arguments{
\item{x}{= a vector}

\item{y}{= a 1/0 membership vector}

\item{n}{= a vector of groups (e.g. GSec)}
}
\description{
zScores <x>
}
\seealso{
Other vec: \code{\link{vec.cat}}, \code{\link{vec.count}},
  \code{\link{vec.max}}, \code{\link{vec.min}},
  \code{\link{vec.named}}, \code{\link{vec.read}},
  \code{\link{vec.same}}, \code{\link{vec.swap}},
  \code{\link{vec.to.lags}}, \code{\link{vec.to.list}},
  \code{\link{vec.unique}}
}
\keyword{vec.zScore}
