% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymm.diff}
\alias{yyyymm.diff}
\title{yyyymm.diff}
\usage{
yyyymm.diff(x, y)
}
\arguments{
\item{x}{= a vector of YYYYMM}

\item{y}{= an isomekic vector of YYYYMM}
}
\description{
returns <x - y> in terms of YYYYMM
}
\keyword{yyyymm.diff}
