% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymm.diff}
\alias{yyyymm.diff}
\title{yyyymm.diff}
\usage{
yyyymm.diff(x, y)
}
\arguments{
\item{x}{= a YYYYMM vector}

\item{y}{= a YYYYMM vector}
}
\description{
returns <x - y> in terms of YYYYMM
}
\seealso{
Other yyyymm: \code{\link{yyyymm.ex.int}},
  \code{\link{yyyymm.ex.qtr}}, \code{\link{yyyymm.exists}},
  \code{\link{yyyymm.lag}}, \code{\link{yyyymm.seq}},
  \code{\link{yyyymm.to.day}}, \code{\link{yyyymm.to.int}},
  \code{\link{yyyymm.to.qtr}}, \code{\link{yyyymm.to.yyyy}}
}
\keyword{yyyymm.diff}
