% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymm.ex.int}
\alias{yyyymm.ex.int}
\title{yyyymm.ex.int}
\usage{
yyyymm.ex.int(x)
}
\arguments{
\item{x}{= an integer vector}
}
\description{
returns = a YYYYMM vector
}
\seealso{
Other yyyymm: \code{\link{yyyymm.diff}},
  \code{\link{yyyymm.ex.qtr}}, \code{\link{yyyymm.exists}},
  \code{\link{yyyymm.lag}}, \code{\link{yyyymm.seq}},
  \code{\link{yyyymm.to.day}}, \code{\link{yyyymm.to.int}},
  \code{\link{yyyymm.to.qtr}}, \code{\link{yyyymm.to.yyyy}}
}
\keyword{yyyymm.ex.int}
