% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymm.ex.qtr}
\alias{yyyymm.ex.qtr}
\title{yyyymm.ex.qtr}
\usage{
yyyymm.ex.qtr(x)
}
\arguments{
\item{x}{= a vector of quarters}
}
\description{
returns quarter end in yyyymm
}
\seealso{
Other yyyymm: \code{\link{yyyymm.diff}},
  \code{\link{yyyymm.ex.int}}, \code{\link{yyyymm.lag}},
  \code{\link{yyyymm.seq}}, \code{\link{yyyymm.to.day}},
  \code{\link{yyyymm.to.int}}, \code{\link{yyyymm.to.qtr}},
  \code{\link{yyyymm.to.yyyy}}
}
\keyword{yyyymm.ex.qtr}
