% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymm.lag}
\alias{yyyymm.lag}
\title{yyyymm.lag}
\usage{
yyyymm.lag(x, y = 1)
}
\arguments{
\item{x}{= a vector of <yyyymm> months or <yyyymmdd> days}

\item{y}{= an integer or an isomekic vector of integers}
}
\description{
lags <x> by <y> months
}
\keyword{yyyymm.lag}
