% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymm.lag}
\alias{yyyymm.lag}
\title{yyyymm.lag}
\usage{
yyyymm.lag(x, y = 1, n = T)
}
\arguments{
\item{x}{= a qtr/YYYYMM/flowdate vector}

\item{y}{= an integer vector}

\item{n}{= T/F depending on whether you wish to lag by yyyymmdd or flowdate}
}
\description{
lags <x> by <y> periods
}
\keyword{yyyymm.lag}
