% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymm.seq}
\alias{yyyymm.seq}
\title{yyyymm.seq}
\usage{
yyyymm.seq(x, y, n = 1)
}
\arguments{
\item{x}{= a YYYYMM or YYYYMMDD or YYYY}

\item{y}{= an isotypic element}

\item{n}{= quantum size in YYYYMM or YYYYMMDD or YYYY}
}
\description{
returns a sequence between (and including) x and y
}
\keyword{yyyymm.seq}
