% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymm.seq}
\alias{yyyymm.seq}
\title{yyyymm.seq}
\usage{
yyyymm.seq(x, y, n = 1)
}
\arguments{
\item{x}{= a YYYY/YYYYMM/YYYYMMDD}

\item{y}{= a YYYY/YYYYMM/YYYYMMDD}

\item{n}{= a positive integer}
}
\description{
returns a sequence between (and including) x and y
}
\seealso{
Other yyyymm: \code{\link{yyyymm.diff}},
  \code{\link{yyyymm.ex.int}}, \code{\link{yyyymm.ex.qtr}},
  \code{\link{yyyymm.exists}}, \code{\link{yyyymm.lag}},
  \code{\link{yyyymm.to.day}}, \code{\link{yyyymm.to.int}},
  \code{\link{yyyymm.to.qtr}}, \code{\link{yyyymm.to.yyyy}}
}
\keyword{yyyymm.seq}
