% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymm.to.day}
\alias{yyyymm.to.day}
\title{yyyymm.to.day}
\usage{
yyyymm.to.day(x)
}
\arguments{
\item{x}{= a vector of months in yyyymm format}
}
\description{
Returns the last day in the month whether weekend or not.
}
\keyword{yyyymm.to.day}
