% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.bulk}
\alias{yyyymmdd.bulk}
\title{yyyymmdd.bulk}
\usage{
yyyymmdd.bulk(x)
}
\arguments{
\item{x}{= a matrix/data frame (indexed by YYYYMMDD)}
}
\description{
Eliminates YYYYMMDD gaps
}
\seealso{
Other yyyymmdd: \code{\link{yyyymmdd.diff}},
  \code{\link{yyyymmdd.ex.day}},
  \code{\link{yyyymmdd.ex.int}},
  \code{\link{yyyymmdd.ex.txt}},
  \code{\link{yyyymmdd.ex.yyyymm}},
  \code{\link{yyyymmdd.exists}},
  \code{\link{yyyymmdd.lag}}, \code{\link{yyyymmdd.seq}},
  \code{\link{yyyymmdd.to.AllocMo.unique}},
  \code{\link{yyyymmdd.to.AllocMo}},
  \code{\link{yyyymmdd.to.int}},
  \code{\link{yyyymmdd.to.txt}},
  \code{\link{yyyymmdd.to.weekofmonth}},
  \code{\link{yyyymmdd.to.yyyymm}}
}
\keyword{yyyymmdd.bulk}
