% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.ex.AllocMo}
\alias{yyyymmdd.ex.AllocMo}
\title{yyyymmdd.ex.AllocMo}
\usage{
yyyymmdd.ex.AllocMo(x)
}
\arguments{
\item{x}{= an object indexed by allocation months}
}
\description{
Returns an object indexed by flow dates
}
\seealso{
Other yyyymmdd: \code{\link{yyyymmdd.bulk}},
  \code{\link{yyyymmdd.diff}},
  \code{\link{yyyymmdd.ex.day}},
  \code{\link{yyyymmdd.ex.int}},
  \code{\link{yyyymmdd.ex.txt}},
  \code{\link{yyyymmdd.ex.yyyymm}},
  \code{\link{yyyymmdd.exists}},
  \code{\link{yyyymmdd.lag}}, \code{\link{yyyymmdd.seq}},
  \code{\link{yyyymmdd.to.AllocMo}},
  \code{\link{yyyymmdd.to.CalYrDyOfWk}},
  \code{\link{yyyymmdd.to.int}},
  \code{\link{yyyymmdd.to.unity}},
  \code{\link{yyyymmdd.to.weekofmonth}},
  \code{\link{yyyymmdd.to.yyyymm}}
}
\keyword{yyyymmdd.ex.AllocMo}
