% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.ex.txt}
\alias{yyyymmdd.ex.txt}
\title{yyyymmdd.ex.txt}
\usage{
yyyymmdd.ex.txt(x, y = "/", n = "MDY")
}
\arguments{
\item{x}{= a vector of dates in some format}

\item{y}{= separators used within <x>}

\item{n}{= order in which month, day and year are represented}
}
\description{
returns a vector of YYYYMMDD (formerly UIDate2yyyymmdd)
}
\keyword{yyyymmdd.ex.txt}
