% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.ex.yyyymm}
\alias{yyyymmdd.ex.yyyymm}
\title{yyyymmdd.ex.yyyymm}
\usage{
yyyymmdd.ex.yyyymm(x, y = T)
}
\arguments{
\item{x}{= a SINGLE month in yyyymm format}

\item{y}{= T/F variable depending on whether the last or all trading days in that month are desired}
}
\description{
Returns the last trading day or all trading days
}
\keyword{yyyymmdd.ex.yyyymm}
