% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.lag}
\alias{yyyymmdd.lag}
\title{yyyymmdd.lag}
\usage{
yyyymmdd.lag(x, y)
}
\arguments{
\item{x}{= a vector of yyyymmdd-dates that happen to fall on a weekday}

\item{y}{= a number}
}
\description{
lags <x> by <y> YYYYMMDD.
}
\keyword{yyyymmdd.lag}
