% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.lag}
\alias{yyyymmdd.lag}
\title{yyyymmdd.lag}
\usage{
yyyymmdd.lag(x, y)
}
\arguments{
\item{x}{= a vector of yyyymmdd-dates that happen to fall on a weekday}

\item{y}{= a number}
}
\description{
lags <x> by <y> YYYYMMDD.
}
\seealso{
Other yyyymmdd: \code{\link{yyyymmdd.bulk}},
  \code{\link{yyyymmdd.ex.AllocMo}},
  \code{\link{yyyymmdd.ex.day}},
  \code{\link{yyyymmdd.ex.int}},
  \code{\link{yyyymmdd.ex.txt}},
  \code{\link{yyyymmdd.ex.yyyymm}},
  \code{\link{yyyymmdd.exists}},
  \code{\link{yyyymmdd.seq}},
  \code{\link{yyyymmdd.to.AllocMo}},
  \code{\link{yyyymmdd.to.CalYrDyOfWk}},
  \code{\link{yyyymmdd.to.int}},
  \code{\link{yyyymmdd.to.unity}},
  \code{\link{yyyymmdd.to.weekofmonth}},
  \code{\link{yyyymmdd.to.yyyymm}}
}
\keyword{yyyymmdd.lag}
