% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.seq}
\alias{yyyymmdd.seq}
\title{yyyymmdd.seq}
\usage{
yyyymmdd.seq(x, y, n = 1)
}
\arguments{
\item{x}{= a single YYYYMMDD}

\item{y}{= a single YYYYMMDD}

\item{n}{= quantum size in YYYYMMDD}
}
\description{
returns a sequence of YYYYMMDD between (and including) x and y
}
\keyword{yyyymmdd.seq}
