% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.to.AllocMo}
\alias{yyyymmdd.to.AllocMo}
\title{yyyymmdd.to.AllocMo}
\usage{
yyyymmdd.to.AllocMo(x, y = 23)
}
\arguments{
\item{x}{= a YYYYMMDD}

\item{y}{= calendar day allocations are known the next month}
}
\description{
Returns the month for which you need to get allocations Flows as of the 23rd of each month are known by the 24th. By this time allocations from #		:	the previous month are known
}
\seealso{
Other yyyymmdd: \code{\link{yyyymmdd.bulk}},
  \code{\link{yyyymmdd.diff}},
  \code{\link{yyyymmdd.ex.day}},
  \code{\link{yyyymmdd.ex.int}},
  \code{\link{yyyymmdd.ex.txt}},
  \code{\link{yyyymmdd.ex.yyyymm}},
  \code{\link{yyyymmdd.exists}},
  \code{\link{yyyymmdd.lag}}, \code{\link{yyyymmdd.seq}},
  \code{\link{yyyymmdd.to.AllocMo.unique}},
  \code{\link{yyyymmdd.to.int}},
  \code{\link{yyyymmdd.to.txt}},
  \code{\link{yyyymmdd.to.weekofmonth}},
  \code{\link{yyyymmdd.to.yyyymm}}
}
\keyword{yyyymmdd.to.AllocMo}
