% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.to.AllocMo.unique}
\alias{yyyymmdd.to.AllocMo.unique}
\title{yyyymmdd.to.AllocMo.unique}
\usage{
yyyymmdd.to.AllocMo.unique(x, y, n)
}
\arguments{
\item{x}{= a YYYYMMDD vector}

\item{y}{= calendar day allocations are known the next month}

\item{n}{= T/F if month should be converted to day}
}
\description{
Checks each day in <x> has same allocation month. Error otherwise
}
\keyword{yyyymmdd.to.AllocMo.unique}
