% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.to.CalYrDyOfWk}
\alias{yyyymmdd.to.CalYrDyOfWk}
\title{yyyymmdd.to.CalYrDyOfWk}
\usage{
yyyymmdd.to.CalYrDyOfWk(x)
}
\arguments{
\item{x}{= a vector of dates in yyyymmdd format}
}
\description{
Converts to 0 = Sun, 1 = Mon, ..., 6 = Sat
}
\seealso{
Other yyyymmdd: \code{\link{yyyymmdd.bulk}},
  \code{\link{yyyymmdd.ex.AllocMo}},
  \code{\link{yyyymmdd.ex.day}},
  \code{\link{yyyymmdd.ex.int}},
  \code{\link{yyyymmdd.ex.txt}},
  \code{\link{yyyymmdd.ex.yyyymm}},
  \code{\link{yyyymmdd.exists}},
  \code{\link{yyyymmdd.lag}}, \code{\link{yyyymmdd.seq}},
  \code{\link{yyyymmdd.to.AllocMo}},
  \code{\link{yyyymmdd.to.int}},
  \code{\link{yyyymmdd.to.unity}},
  \code{\link{yyyymmdd.to.weekofmonth}},
  \code{\link{yyyymmdd.to.yyyymm}}
}
\keyword{yyyymmdd.to.CalYrDyOfWk}
