% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.to.int}
\alias{yyyymmdd.to.int}
\title{yyyymmdd.to.int}
\usage{
yyyymmdd.to.int(x)
}
\arguments{
\item{x}{= a YYYYMMDD vector}
}
\description{
number of weekdays after Thursday, January 1, 1970
}
\keyword{yyyymmdd.to.int}
