% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.to.weekofmonth}
\alias{yyyymmdd.to.weekofmonth}
\title{yyyymmdd.to.weekofmonth}
\usage{
yyyymmdd.to.weekofmonth(x)
}
\arguments{
\item{x}{= a vector of dates in yyyymmdd format}
}
\description{
returns 1 if the date fell in the first week of the month, 2 if it fell in the second, etc.
}
\keyword{yyyymmdd.to.weekofmonth}
