% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{yyyymmdd.to.yyyymm}
\alias{yyyymmdd.to.yyyymm}
\title{yyyymmdd.to.yyyymm}
\usage{
yyyymmdd.to.yyyymm(x, y = F)
}
\arguments{
\item{x}{= a vector of dates in yyyymmdd format}

\item{y}{= if T then falls back one month}
}
\description{
Converts to yyyymm format
}
\seealso{
Other yyyymmdd: \code{\link{yyyymmdd.bulk}},
  \code{\link{yyyymmdd.ex.AllocMo}},
  \code{\link{yyyymmdd.ex.day}},
  \code{\link{yyyymmdd.ex.int}},
  \code{\link{yyyymmdd.ex.txt}},
  \code{\link{yyyymmdd.ex.yyyymm}},
  \code{\link{yyyymmdd.exists}},
  \code{\link{yyyymmdd.lag}}, \code{\link{yyyymmdd.seq}},
  \code{\link{yyyymmdd.to.AllocMo}},
  \code{\link{yyyymmdd.to.CalYrDyOfWk}},
  \code{\link{yyyymmdd.to.int}},
  \code{\link{yyyymmdd.to.unity}},
  \code{\link{yyyymmdd.to.weekofmonth}}
}
\keyword{yyyymmdd.to.yyyymm}
