% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPFR.r
\name{zScore.underlying}
\alias{zScore.underlying}
\title{zScore.underlying}
\usage{
zScore.underlying(x)
}
\arguments{
\item{x}{= a matrix/data frame (last column is boolean without NA's)}
}
\description{
zScores the first columns of <x> using the last column as weight
}
\keyword{zScore.underlying}
